% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_adult.R
\docType{data}
\name{adult}
\alias{adult}
\title{Adult dataset}
\format{
A data frame with 32561 rows and 15 variables:
\describe{
  \item{salary}{factor, <=50K/>50K whether a person salary exceeds 50K a year or not}
  \item{age}{integer, age of person}
  \item{workclass}{factor, field of work}
  \item{fnlwgt}{numeric}
  \item{education}{factor, completed education degree}
  \item{education_num}{numeric, education number in converted from education
  factor, the bigger the better}
  \item{marital_status}{factor}
  \item{occupation}{factor, where this person works}
  \item{relationship}{factor, relationship information}
  \item{race}{factor, ethnicity of a person}
  \item{sex}{factor, gender of a person}
  \item{capital_gain}{numeric}
  \item{capital_loss}{numeric}
  \item{hours_per_week}{numeric, how many hours per week does this person work}
  \item{native_country}{factor, in which country was this person born}
}
}
\source{
Data from UCL \url{https://archive.ics.uci.edu/ml/datasets/adult}
}
\usage{
data(adult)
}
\description{
\code{adult} dataset consists of many columns containing various information
about relationship, hours worked per week, workclass etc... and about
salary, whether more than 50K a year or not. Lot's of possible protected
attributes such as sex, race age. Some columns contain level "unknown" and
these values are not removed and removing them depends on user as they might
contain some information.
}
