% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agestam_1985.R
\name{Agestam_1985_bark_addition_Pine}
\alias{Agestam_1985_bark_addition_Pine}
\title{Bark addition function for Pine from Agestam 1985}
\source{
Agestam, E. (1985) 'A growth simulator for mixed stands of pine, spruce
and birch in Sweden. Diss. Swedish University of Agricultural Sciences. Report no. 15. Dept. of Forest Yield Research. ISSN 0348-7636.
ISBN 91-576-2528-x. 150 pp. Garpenberg, Sweden; page. 46.
}
\usage{
Agestam_1985_bark_addition_Pine(basal_area_under_bark_Pine, SI_Pine, latitude)
}
\arguments{
\item{basal_area_under_bark_Pine}{Basal area Pine under bark, m2/ha.}

\item{SI_Pine}{Site Index H100 for Scots Pine from Hägglund (1974)}

\item{latitude}{Latitude, degrees.}
}
\value{
Bark area at breast height for Pine, m2/ha.
}
\description{
Bark addition function for Pine from Agestam 1985
}
\details{
Material:

Number of plots: 143

Multiple correlation coefficient: 0.968

Standard deviation about the function (sf) : 0.112

sf/standard deviation about the mean:  26\%
}
