% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agestam_1985.R
\name{Agestam_1985_basal_area_annual_increment_under_bark_Birch}
\alias{Agestam_1985_basal_area_annual_increment_under_bark_Birch}
\title{Basal area growth function for Birch from Agestam 1985}
\source{
Agestam, E. (1985) 'A growth simulator for mixed stands of pine, spruce
and birch in Sweden. Diss. Swedish University of Agricultural Sciences. Report no. 15. Dept. of Forest Yield Research. ISSN 0348-7636.
ISBN 91-576-2528-x. 150 pp. Garpenberg, Sweden; page. 29.
}
\usage{
Agestam_1985_basal_area_annual_increment_under_bark_Birch(
  basal_area_Birch_m2_ha_after_thinning,
  stem_count_Birch,
  age_at_breast_height_period_start,
  SI_H100_Birch,
  basal_area_other_species_m2_ha_after_thinning,
  increment_period_years
)
}
\arguments{
\item{basal_area_Birch_m2_ha_after_thinning}{Basal area of Birch after thinning, under bark. m2/ha.}

\item{stem_count_Birch}{Number of Birch stems per ha after thinning.}

\item{age_at_breast_height_period_start}{Age at breast height for Birch at the period start.}

\item{SI_H100_Birch}{Site index H100 for Birch according to height trajectory defined in Persson (1959)}

\item{basal_area_other_species_m2_ha_after_thinning}{Basal area under bark of other species after thinning, under bark. m2/ha.}

\item{increment_period_years}{Length of increment period, in years.}
}
\value{
Annual basal area growth under bark for Birch. m2/ha.
}
\description{
Basal area growth function for Birch from Agestam 1985
}
\details{
Material:

Number of growth periods: 126

Multiple correlation coefficient: 0.928

Standard deviation about the function (sf) : 0.396

sf/standard deviation about the mean:  38\%
}
