% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agestam_1985.R
\name{Agestam_1985_basal_area_annual_increment_under_bark_Pine}
\alias{Agestam_1985_basal_area_annual_increment_under_bark_Pine}
\title{Basal area growth function for Pine from Agestam 1985}
\source{
Agestam, E. (1985) 'A growth simulator for mixed stands of pine, spruce
and birch in Sweden. Diss. Swedish University of Agricultural Sciences. Report no. 15. Dept. of Forest Yield Research. ISSN 0348-7636.
ISBN 91-576-2528-x. 150 pp. Garpenberg, Sweden; page. 29.
}
\usage{
Agestam_1985_basal_area_annual_increment_under_bark_Pine(
  basal_area_Pine_m2_ha_after_thinning,
  stem_count_Pine,
  age_at_breast_height_period_start,
  SI_H100_Pine,
  basal_area_other_species_m2_ha_after_thinning,
  diameter_mean_basal_area_stem_Pine_cm,
  diameter_mean_basal_area_stem_all_species_cm,
  increment_period_years
)
}
\arguments{
\item{basal_area_Pine_m2_ha_after_thinning}{Basal area of Pine after thinning, under bark. m2/ha.}

\item{stem_count_Pine}{Number of Pine stems per ha after thinning.}

\item{age_at_breast_height_period_start}{Age at breast height for Pine at the period start.}

\item{SI_H100_Pine}{Site index H100 for Pine according to height trajectory defined in Hägglund, 1974.}

\item{basal_area_other_species_m2_ha_after_thinning}{Basal area under bark of other species after thinning, under bark. m2/ha.}

\item{diameter_mean_basal_area_stem_Pine_cm}{Diameter corresponding to mean basal area under bark, Pine. cm.}

\item{diameter_mean_basal_area_stem_all_species_cm}{Diameter corresponding to mean basal area under bark, all species on plot. cm.}

\item{increment_period_years}{Length of increment period, in years.}
}
\value{
Annual basal area growth under bark for Pine. m2/ha.
}
\description{
Basal area growth function for Pine from Agestam 1985
}
\details{
Material:

Number of growth periods: 505

Multiple correlation coefficient: 0.888

Standard deviation about the function (sf) : 0.245

sf/standard deviation about the mean:  46\%
}
\examples{
library(ggplot2)
ggplot()+geom_function(fun= function(x) Agestam_1985_basal_area_annual_increment_under_bark_Pine(basal_area_Pine_m2_ha_after_thinning = 15,stem_count_Pine = 800,age_at_breast_height_period_start = x,SI_H100_Pine = 25,basal_area_other_species_m2_ha_after_thinning = 15,diameter_mean_basal_area_stem_Pine_cm = 1,diameter_mean_basal_area_stem_all_species_cm = 1,increment_period_years = 10))+xlim(c(10,140))+ylim(c(0,0.5))
}
