% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agestam_1985.R
\name{Agestam_1985_mortality_above_bark_Birch}
\alias{Agestam_1985_mortality_above_bark_Birch}
\title{Mortality for Birch per annum according to Agestam 1985.}
\source{
Agestam, E. (1985) 'A growth simulator for mixed stands of pine, spruce
and birch in Sweden. Diss. Swedish University of Agricultural Sciences. Report no. 15. Dept. of Forest Yield Research. ISSN 0348-7636.
ISBN 91-576-2528-x. 150 pp. Garpenberg, Sweden; page. 50.
}
\usage{
Agestam_1985_mortality_above_bark_Birch(
  stems_ha_all_species,
  basal_area_m2_ha_Birch
)
}
\arguments{
\item{stems_ha_all_species}{Total number of stems per hectare.}

\item{basal_area_m2_ha_Birch}{Basal area m2/ha of Birch.}
}
\value{
m2/ha mortality above bark per annum.
}
\description{
Mortality for Birch per annum according to Agestam 1985.
}
\details{
Material:

Number of plots: 87

Multiple correlation coefficient: 0.484

Standard deviation about the function (sf) : 1.58

sf/standard deviation about the mean:  92\%
}
