% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agestam_1985.R
\name{Agestam_1985_mortality_above_bark_Pine}
\alias{Agestam_1985_mortality_above_bark_Pine}
\title{Mortality for Pine per annum according to Agestam 1985.}
\source{
Agestam, E. (1985) 'A growth simulator for mixed stands of pine, spruce
and birch in Sweden. Diss. Swedish University of Agricultural Sciences. Report no. 15. Dept. of Forest Yield Research. ISSN 0348-7636.
ISBN 91-576-2528-x. 150 pp. Garpenberg, Sweden; page. 50.
}
\usage{
Agestam_1985_mortality_above_bark_Pine(
  stems_Pine_ha,
  basal_area_above_bark_all_species,
  SI_Pine,
  age_at_breast_height
)
}
\arguments{
\item{stems_Pine_ha}{Number of Pine stems per hectare.}

\item{basal_area_above_bark_all_species}{Basal area above bark for all tree species. m2/ha.}

\item{SI_Pine}{Site Index H100 for Pine according to Hägglund (1974), meters.}

\item{age_at_breast_height}{Age at breast height of Pine stems.}
}
\value{
m2/ha mortality above bark per annum.
}
\description{
Mortality for Pine per annum according to Agestam 1985.
}
\details{
Material:

Number of plots: 134

Multiple correlation coefficient: 0.400

Standard deviation about the function (sf) : 1.99

sf/standard deviation about the mean:  93\%
}
