% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agestam_1985.R
\name{Agestam_1985_mortality_above_bark_Spruce}
\alias{Agestam_1985_mortality_above_bark_Spruce}
\title{Mortality for Spruce per annum according to Agestam 1985.}
\source{
Agestam, E. (1985) 'A growth simulator for mixed stands of pine, spruce
and birch in Sweden. Diss. Swedish University of Agricultural Sciences. Report no. 15. Dept. of Forest Yield Research. ISSN 0348-7636.
ISBN 91-576-2528-x. 150 pp. Garpenberg, Sweden; page. 50.
}
\usage{
Agestam_1985_mortality_above_bark_Spruce(
  stems_ha,
  basal_area_above_bark_all_species
)
}
\arguments{
\item{stems_ha}{Number of Spruce stems per hectare.}

\item{basal_area_above_bark_all_species}{Total basal area including bark, m2/ha.}
}
\value{
m2/ha mortality above bark per annum.
}
\description{
Mortality for Spruce per annum according to Agestam 1985.
}
\details{
Material:

Number of plots: 134

Multiple correlation coefficient: 0.447

Standard deviation about the function (sf) : 1.74

sf/standard deviation about the mean:  90\%
}
