% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agestam_1985.R
\name{Agestam_1985_mortality_diameter_cm}
\alias{Agestam_1985_mortality_diameter_cm}
\title{Basal area weighted mean diameter of dead trees from Agestam 1985}
\source{
Agestam, E. (1985) 'A growth simulator for mixed stands of pine, spruce
and birch in Sweden. Diss. Swedish University of Agricultural Sciences. Report no. 15. Dept. of Forest Yield Research. ISSN 0348-7636.
ISBN 91-576-2528-x. 150 pp. Garpenberg, Sweden; page. 53.
}
\usage{
Agestam_1985_mortality_diameter_cm(
  species,
  age_at_breast_height,
  basal_area_weighted_mean_diameter_cm_living
)
}
\arguments{
\item{species}{"Pinus sylvestris" or "Picea abies" or "Betula pendula" or "Betula pubescens"}

\item{age_at_breast_height}{Age of stems at breast height.}

\item{basal_area_weighted_mean_diameter_cm_living}{Basal area weighted mean diameter of living trees (in cm)}
}
\value{
Basal area weighted mean diameter of dead trees, in cm.
}
\description{
Basal area weighted mean diameter of dead trees from Agestam 1985
}
