% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agestam_1985.R
\name{Agestam_1985_volume_Birch_height}
\alias{Agestam_1985_volume_Birch_height}
\title{Volume function for Birch from Agestam (1985), including height.}
\source{
Agestam, E. (1985) 'A growth simulator for mixed stands of pine, spruce
and birch in Sweden. Diss. Swedish University of Agricultural Sciences. Report no. 15. Dept. of Forest Yield Research. ISSN 0348-7636.
ISBN 91-576-2528-x. 150 pp. Garpenberg, Sweden; page. 38.
}
\usage{
Agestam_1985_volume_Birch_height(
  basal_area_Birch_m2_ha,
  dominant_height_Birch_m,
  stems_ha_Birch,
  latitude
)
}
\arguments{
\item{basal_area_Birch_m2_ha}{Basal area over bark Pine, m2/ha.}

\item{dominant_height_Birch_m}{Dominant height of Birch stems, m2 /ha.}

\item{stems_ha_Birch}{Number of Birch stems per hectare.}

\item{latitude}{Latitude, degrees.}
}
\value{
Volume Birch m3sk/ha.
}
\description{
Volume function for Birch from Agestam (1985), including height.
}
\details{
Function: B1

Number of plots: 50

Multiple correlation coefficient: 0.998

Standard deviation about the function (sf): 0.069

sf/standard deviation about the mean: 6.4\%
}
