% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agestam_1985_volume_Pine_SI.R
\name{Agestam_1985_volume_Pine_SI}
\alias{Agestam_1985_volume_Pine_SI}
\title{Volume function for Pine from Agestam (1985), including SI.}
\source{
Agestam, E. (1985) 'A growth simulator for mixed stands of pine, spruce
and birch in Sweden. Diss. Swedish University of Agricultural Sciences. Report no. 15. Dept. of Forest Yield Research. ISSN 0348-7636.
ISBN 91-576-2528-x. 150 pp. Garpenberg, Sweden; page. 38.
}
\usage{
Agestam_1985_volume_Pine_SI(
  basal_area_Pine_m2_ha,
  SI_Pine_m,
  age_at_breast_height_Pine,
  stems_Pine_ha,
  latitude
)
}
\arguments{
\item{basal_area_Pine_m2_ha}{Basal area over bark Pine, m2/ha.}

\item{SI_Pine_m}{Site Index H100 of Pine, meters,  from Hägglund 1974.}

\item{age_at_breast_height_Pine}{Age at breast height of Pine stems.}

\item{stems_Pine_ha}{Pine stems per hectare.}

\item{latitude}{Latitude, degrees.}
}
\value{
Volume Pine m3sk/ha.
}
\description{
Volume function for Pine from Agestam (1985), including SI.
}
\details{
Function: T2

Number of plots: 141

Multiple correlation coefficient: 0.997

Standard deviation about the function (sf): 0.038

sf/standard deviation about the mean: 7.4\%
}
