% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agestam_1985_volume_Spruce_height.R
\name{Agestam_1985_volume_Spruce_height}
\alias{Agestam_1985_volume_Spruce_height}
\title{Volume function for Spruce from Agestam (1985), including height.}
\source{
Agestam, E. (1985) 'A growth simulator for mixed stands of pine, spruce
and birch in Sweden. Diss. Swedish University of Agricultural Sciences. Report no. 15. Dept. of Forest Yield Research. ISSN 0348-7636.
ISBN 91-576-2528-x. 150 pp. Garpenberg, Sweden; page. 38.
}
\usage{
Agestam_1985_volume_Spruce_height(
  basal_area_Spruce_m2_ha,
  dominant_height_Spruce_m,
  stems_Spruce_ha,
  age_at_breast_height_Spruce
)
}
\arguments{
\item{basal_area_Spruce_m2_ha}{Basal area over bark Spruce, m2/ha.}

\item{dominant_height_Spruce_m}{Dominant height of Spruce stems, m.}

\item{stems_Spruce_ha}{Spruce stems per hectare.}

\item{age_at_breast_height_Spruce}{Age at breast height of Spruce stems.}
}
\value{
Volume Spruce m3sk/ha.
}
\description{
Volume function for Spruce from Agestam (1985), including height.
}
\details{
Function: G1

Number of plots: 115

Multiple correlation coefficient: 0.995

Standard deviation about the function (sf): 0.058

sf/standard deviation about the mean: 10.2%
}
