% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agestam_1985.R
\name{Agestam_1985_volume_Spruce_SI}
\alias{Agestam_1985_volume_Spruce_SI}
\title{Volume function for Spruce from Agestam (1985), including SI.}
\source{
Agestam, E. (1985) 'A growth simulator for mixed stands of pine, spruce
and birch in Sweden. Diss. Swedish University of Agricultural Sciences. Report no. 15. Dept. of Forest Yield Research. ISSN 0348-7636.
ISBN 91-576-2528-x. 150 pp. Garpenberg, Sweden; page. 38.
}
\usage{
Agestam_1985_volume_Spruce_SI(
  basal_area_Spruce_m2_ha,
  SI_Spruce,
  age_at_breast_height_Spruce,
  stems_ha_Spruce,
  stems_ha_other
)
}
\arguments{
\item{basal_area_Spruce_m2_ha}{Basal area over bark Spruce, m2/ha.}

\item{SI_Spruce}{Site Index H100 of Spruce, meters,  from Hägglund 1972,1973.}

\item{age_at_breast_height_Spruce}{Age at breast height of Spruce stems.}

\item{stems_ha_Spruce}{Spruce stems per hectare.}

\item{stems_ha_other}{Stems per hectare of other tree species.}
}
\value{
Volume Spruce m3sk/ha.
}
\description{
Volume function for Spruce from Agestam (1985), including SI.
}
\details{
Function: G2

Number of plots: 143

Multiple correlation coefficient: 0.997

Standard deviation about the function (sf): 0.060

sf/standard deviation about the mean: 8.3%
}
