% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allen_2020_height_development_Norway_spruce.R
\name{Allen_2020_basal_area_after_thinning_Norway_Spruce_Norway}
\alias{Allen_2020_basal_area_after_thinning_Norway_Spruce_Norway}
\title{Get the basal area after thinning if thinning by stem numbers.}
\usage{
Allen_2020_basal_area_after_thinning_Norway_Spruce_Norway(
  basalAreaBeforeThinning,
  stemsBeforeThinning,
  stemsAfterThinning
)
}
\arguments{
\item{basalAreaBeforeThinning}{Basal Area per hectare before thinning (m2/ha)}

\item{stemsBeforeThinning}{Number of stems per hectare before thinning.}

\item{stemsAfterThinning}{Number of stems per hectare after thinning.}
}
\value{
Number of stems remaining after thinning per hectare.
}
\description{
Get the basal area after thinning if thinning by stem numbers.
}
\details{
Get the basal area after thinning if you are thinning by number of stems.
}
\seealso{
[forester::Allen_2020_stems_after_thinning_Norway_Spruce_Norway()]
}
