% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allen_2020_height_development_Norway_spruce.R
\name{Allen_2020_Basal_Area_Norway_Spruce_Norway}
\alias{Allen_2020_Basal_Area_Norway_Spruce_Norway}
\title{Basal area at the end of a growth period}
\source{
Allen, M.G. II, Antón-Fernández, C., Astrup, R. (2020) "A stand-level
growth and yield model for thinned and unthinned managed Norway spruce forests
in Norway". Scandinavian Journal of Forest Research. Vol. 35. pp. 238-251.
DOI: \url{https://doi.org/10.1080/02827581.2020.1773525}
}
\usage{
Allen_2020_Basal_Area_Norway_Spruce_Norway(
  basalAreaBeforeThinning,
  basalAreaAfterThinning,
  dominantHeight1,
  dominantHeightAtThinning,
  dominantHeight2,
  stemsBeforeThinning,
  stemsAfterThinning
)
}
\arguments{
\item{basalAreaBeforeThinning}{Basal area at period start before thinning}

\item{basalAreaAfterThinning}{Basal area after thinning.}

\item{dominantHeight1}{Dominant height at start of period}

\item{dominantHeightAtThinning}{Dominant height at thinning}

\item{dominantHeight2}{Dominant height at end of period.}

\item{stemsBeforeThinning}{Number of stems per hectare before thinning.}

\item{stemsAfterThinning}{Number of stems per hectare after thinning.}
}
\value{
Basal Area m2/ha at period end.
}
\description{
Basal area at the end of a growth period
}
