% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Allen_2020_stand_level_thinned_unthinned_P_abies.R
\name{Allen_2020_P_abies_five_year_increment}
\alias{Allen_2020_P_abies_five_year_increment}
\title{A stand-level growth and yield model for thinned and unthinned managed
Norway Spruce forests in Norway}
\source{
Allen, M.G. II, Antón-Fernández, C., Astrup, R. (2020) "A stand-level
growth and yield model for thinned and unthinned managed Norway spruce forests
in Norway". Scandinavian Journal of Forest Research. Vol. 35. pp. 238-251.
DOI: <https://doi.org/10.1080/02827581.2020.1773525>
}
\usage{
Allen_2020_P_abies_five_year_increment(
  stand_age,
  volume,
  SI_40,
  stems,
  stems_after_thinning,
  basal_area,
  basal_area_after_thinning,
  thinning_by = "basal_area_quotient",
  thinning_strength = 1,
  dominant_height_at_thinning
)
}
\arguments{
\item{stand_age}{numeric. Age of stand, in years.}

\item{SI_40}{numeric. Site index of stand at reference age 40.}

\item{stems}{numeric. Number of stems per hectare.}

\item{stems_after_thinning}{numeric. Number of stems per hectare.}

\item{basal_area}{numeric. Basal area at breast height (1.3 m) per hectare.}

\item{basal_area_after_thinning}{numeric. Basal area after thinning.}

\item{thinning_by}{default= "basal area quotient", "stems_percent"}

\item{thinning_strength}{If thinning_by=="basal area quotient, Basal Area after
thinning / Basal area before thinning. If thinning_by=="stems_percent": Stems
after thinning / Stems before thinning. For no thinning, 1.}

\item{dominant_height_at_thinning}{Dominant height of stand at thinning.
E.g. mean of tallest 10 \%.}
}
\value{
A data.frame. Stand description at stand age + 5 .
}
\description{
A stand-level growth and yield model for thinned and unthinned managed
Norway Spruce forests in Norway
}
