% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allen_2020_height_development_Norway_spruce.R
\name{Allen_2020_stems_after_thinning_Norway_Spruce_Norway}
\alias{Allen_2020_stems_after_thinning_Norway_Spruce_Norway}
\title{Get the number of stems after thinning if thinning by basal area.}
\usage{
Allen_2020_stems_after_thinning_Norway_Spruce_Norway(
  basalAreaBeforeThinning,
  basalAreaAfterThinning,
  stemsBeforeThinning
)
}
\arguments{
\item{basalAreaBeforeThinning}{Basal Area per hectare before thinning (m2/ha)}

\item{basalAreaAfterThinning}{Basal Area per hectare after thinning (m2/ha)}

\item{stemsBeforeThinning}{Number of stems per hectare before thinning.}
}
\value{
Number of stems remaining after thinning per hectare.
}
\description{
Get the number of stems after thinning if thinning by basal area.
}
\details{
Get the number of stems after thinning if you are thinning by basal area.
}
\seealso{
[forester::Allen_2020_basal_area_after_thinning_Norway_Spruce_Norway()]
}
