% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allen_2020_height_development_Norway_spruce.R
\name{Allen_2020_surviving_stems_Norway_Spruce_Norway}
\alias{Allen_2020_surviving_stems_Norway_Spruce_Norway}
\title{Calculate the number of surviving stems}
\source{
Allen, M.G. II, Antón-Fernández, C., Astrup, R. (2020) "A stand-level
growth and yield model for thinned and unthinned managed Norway spruce forests
in Norway". Scandinavian Journal of Forest Research. Vol. 35. pp. 238-251.
DOI: \url{https://doi.org/10.1080/02827581.2020.1773525}
}
\usage{
Allen_2020_surviving_stems_Norway_Spruce_Norway(
  stemsAfterThinning,
  basalAreaBeforeThinning,
  basalAreaAfterThinning,
  standAge,
  standAge2,
  SI40,
  IncrementPeriod = 5
)
}
\arguments{
\item{stemsAfterThinning}{Number of stems per hectare remaining after thinning}

\item{basalAreaBeforeThinning}{Basal Area per hectare before thinning (m2/ha)}

\item{basalAreaAfterThinning}{Basal area per hectare after thinning (m2/ha)}

\item{standAge}{Stand total age at period start.}

\item{standAge2}{Stand total age at end of period.}

\item{SI40}{SI40, e.g. [forester::Allen_2020_height_development_Norway_spruce()]}

\item{IncrementPeriod}{Always 5 (default).}
}
\value{
Number of surviving stems at the end of growth period.
}
\description{
Calculate the number of surviving stems
}
