% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allen_2020_height_development_Norway_spruce.R
\name{Allen_2020_Volume_Norway_Spruce_Norway}
\alias{Allen_2020_Volume_Norway_Spruce_Norway}
\title{Volume of Norway Spruce stand}
\source{
Allen, M.G. II, Antón-Fernández, C., Astrup, R. (2020) "A stand-level
growth and yield model for thinned and unthinned managed Norway spruce forests
in Norway". Scandinavian Journal of Forest Research. Vol. 35. pp. 238-251.
DOI: \url{https://doi.org/10.1080/02827581.2020.1773525}
}
\usage{
Allen_2020_Volume_Norway_Spruce_Norway(basalArea, dominantHeight, Age)
}
\arguments{
\item{basalArea}{Basal Area m2/ha}

\item{dominantHeight}{Dominant height of stand, m.}

\item{Age}{Stand Age (total).}
}
\value{
Volume, m3/ha.
}
\description{
Volume of Norway Spruce stand
}
