% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Andersson_1954_volume_small_trees_Birch_height_above_4_m.R
\name{Andersson_1954_volume_small_trees_Birch_height_above_4_m}
\alias{Andersson_1954_volume_small_trees_Birch_height_above_4_m}
\title{Andersson 1954 volume for small trees, Birch}
\source{
Andersson, S.-O. 1954. Funktioner och tabeller för kubering av småträd. Meddelanden från Statens Skogsforskningsinstitut 44:12, 29 s
Available: \url{https://www.skogskunskap.se/rakna-med-verktyg/mata-skogen/volymberakning/volymfunktioner/}
}
\usage{
Andersson_1954_volume_small_trees_Birch_height_above_4_m(diameter_cm, height_m)
}
\arguments{
\item{diameter_cm}{Diameter in cm.}

\item{height_m}{Height in m.}
}
\value{
Volume in dm3.
}
\description{
Suitable for trees taller than 4 m.
}
