% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Angstrom_1958_local_climate_Sweden.R
\name{Angstrom_1958_local_climate_Sweden}
\alias{Angstrom_1958_local_climate_Sweden}
\title{Sweden local climates}
\usage{
Angstrom_1958_local_climate_Sweden(latitude, longitude, epsg)
}
\arguments{
\item{latitude}{Latitude, decimal}

\item{longitude}{Longitude, decimal}

\item{epsg}{Reference system. Default 'WGS84': 4326. For SWEREF99TM: '3006'.}
}
\value{
Local Swedish Climate Code.
}
\description{
Finds the local climate according to Ångström (1958)
}
