% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Assmann_Franz_1965_height_trajectory_Germany_Bayern_Spruce.R
\name{Assmann_Franz_1965_height_trajectory_Germany_Bayern_Spruce}
\alias{Assmann_Franz_1965_height_trajectory_Germany_Bayern_Spruce}
\title{Dominant height trajectory for Norway Spruce in Bayern, Germany.}
\source{
Assmann, E. & Franz, F. 1965. Vorläufige Fichten-Ertragstafel für
Bayern. Forstwissenschaftliches Centralblatt 84(1/2):13-43. Available online (2022-04-06):
\url{https://www.waldwachstum.wzw.tum.de/fileadmin/publications/Assmann_1965_Vorlaeufige_Fichten-Ertragstafel.pdf}
}
\usage{
Assmann_Franz_1965_height_trajectory_Germany_Bayern_Spruce(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of the stand, m.}

\item{age}{Age of the stand.}

\item{age2}{Desired age of the stand.}

\item{output}{"Height" (default) or "Equation"}
}
\value{
Height of the stand in metres at the desired age.
}
\description{
Dominant height trajectory for Norway Spruce in Bayern, Germany.
}
