% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_mortality_Bengtsson_1978.R
\name{ba_mortality_thinned_Bengtsson_1978}
\alias{ba_mortality_thinned_Bengtsson_1978}
\title{Natural mortality, Bengtsson 1978}
\usage{
ba_mortality_thinned_Bengtsson_1978(standData)
}
\arguments{
\item{standData}{}
}
\value{
Proportion of the raw standing volume lost per annum.
}
\description{
Natural mortality. Functions based on National Forest Inventory
 surveys of dead trees during 1973-1975. Dead or wind-felled trees above 5 cm
 diameter at breast height. Avoided harsh storm damages during 1969. Largest
 insecurity pertains to which season the substrate in question died.

 A 'roof' was set at a maximum of 0.9% per annum and never let the loss be less than 0.1% per annum.

 The mortality level during the observed period was judged to have been
 higher than normal so at application the mortality given by the functions
 was reduced with 40 % in northern Sweden and 25 % in southern Sweden.

 \href{https://lagen.nu/sou/1978:7#sid293-img}{Available Online (Swedish)}
}
