% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_basal_area.R
\name{basal_area_ratio}
\alias{basal_area_ratio}
\title{Ratio of basal area}
\usage{
basal_area_ratio(treelist, species, basal_area)
}
\arguments{
\item{treelist}{A treelist object.}

\item{species}{Species of interest}

\item{basal_area}{Total basal area}
}
\value{
A ratio.
}
\description{
This function summarises the basal area of a given tree species as a percentage of the total in the stand for a tree list.
}
