% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{best_model_predict}
\alias{best_model_predict}
\title{Predict a single model}
\usage{
best_model_predict(best_models, test_data)
}
\arguments{
\item{best_models}{A list of models.}

\item{test_data}{A test dataset.}
}
\value{
A vector of predicitons between 0 and 1 values.
}
\description{
Predict a single model
}
