% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{boruta_selection}
\alias{boruta_selection}
\title{Perform Boruta algorithm for selecting most important features}
\usage{
boruta_selection(data, y)
}
\arguments{
\item{data}{A data source before preprocessing, that is one of the major R formats:
data.table, data.frame, matrix, and so on.}

\item{y}{A string that indicates a target column name.}
}
\value{
A dataset with removed columns that are not needed according to the Boruta algorithm.
}
\description{
Perform Boruta algorithm for selecting most important features
}
