% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norway_double_bark_functions.R
\name{Braastad_1966_double_bark_thickness_Norway_Betula}
\alias{Braastad_1966_double_bark_thickness_Norway_Betula}
\title{Double bark thickness for Birch in Norway}
\source{
Braastad, H. 1966. Volume tables for birch. Reports of the Norwegian
 Forest Research Institute, 21, 265-365.
}
\usage{
Braastad_1966_double_bark_thickness_Norway_Betula(diameter_cm)
}
\arguments{
\item{diameter_cm}{Diameter of tree at breast height, 1.3m.}
}
\value{
Double bark, in mm.
}
\description{
Double bark thickness for Birch in Norway
}
\author{
Modified from hansoleorka/skogR on the 2022-04-02
}
