% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Braastad_1966_volume_Norway_Betula.R
\name{Braastad_1966_volume_Norway_Betula}
\alias{Braastad_1966_volume_Norway_Betula}
\title{Simple Volume function for Birch in Norway}
\source{
Braastad, H. 1966. Volume tables for birch. Reports of the Norwegian
Forest Research Institute, 21, 265-365.
}
\usage{
Braastad_1966_volume_Norway_Betula(diameter_cm, height_m)
}
\arguments{
\item{diameter_cm}{Diameter of tree in cm at breast height, 1.3m}

\item{height_m}{Total height of tree, m.}
}
\value{
Volume of tree above bark, dm^3.
}
\description{
Simple Volume function for Birch in Norway
}
\author{
Modified from hansoleorka/skogR  2022-04-02.
}
