% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Braastad_1974_diameter_quotient_thinned_trees_to_unthinned_stand_Norway_Norway_Spruce.R
\name{Braastad_1974_diameter_quotient_thinned_trees_to_unthinned_stand_Norway_Norway_Spruce}
\alias{Braastad_1974_diameter_quotient_thinned_trees_to_unthinned_stand_Norway_Norway_Spruce}
\title{Diameter quotient between felled trees and unthinned stand for Norway Spruce.}
\source{
Braastad, H. 1974. Produksjonstabeller og tilvekstmodeller for gran. Yield tables and Growth models for Picea abies. Meddr. norsk inst. Skogforsk. 31:9. p. 371. function (VI.A-2)
}
\usage{
Braastad_1974_diameter_quotient_thinned_trees_to_unthinned_stand_Norway_Norway_Spruce(
  quadratic_mean_diameter_before,
  Basal_area_m2_ha_before,
  stems_remaining,
  SIH40,
  age
)
}
\arguments{
\item{quadratic_mean_diameter_before}{Quadratic mean diameter of stand before thinning. i.e. [forester::quadratic_mean_diameter()]}

\item{Basal_area_m2_ha_before}{Basal area m2 / ha before thinning.}

\item{stems_remaining}{Stems per hectare remaining after thinning.}

\item{SIH40}{Site Index H40, e.g. Tveite 1977 [forester::Tveite_1977_height_trajectory_Norway_Norway_Spruce()]. NB. Was included as unpublished.}

\item{age}{Age at breast height of stand.}
}
\value{
Diameter quotient.
}
\description{
Diameter quotient between felled trees and unthinned stand for Norway Spruce.
}
\details{
Multiple correlation coefficient: 0.463

Residual spread : 0.134

Spread about the mean: 16.8 percent.
}
