% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Braastad_1974.R
\name{Braastad_1974_QMD_tree_volume_Norway_Norway_Spruce}
\alias{Braastad_1974_QMD_tree_volume_Norway_Norway_Spruce}
\title{Volume of the BA weighted mean tree for Norway Spruce in Norway from Braastad 1974}
\source{
Braastad, H. 1974. Produksjonstabeller og tilvekstmodeller for gran. Yield tables and Growth models for Picea abies. Meddr. norsk inst. Skogforsk. 31:9. p. 397. Code section: 1160,1162,1163.
}
\usage{
Braastad_1974_QMD_tree_volume_Norway_Norway_Spruce(diameter_cm, height_m)
}
\arguments{
\item{diameter_cm}{Quadratic mean diameter, cm.}

\item{height_m}{Lorey's mean height, m. e.g. [forester::Tveite_1967_Loreys_mean_height_Norway_Norway_Spruce()]}
}
\value{
Volume of the BA weighted mean tree (m3).
}
\description{
Uses single tree volume functions from Vestjordet (1967).
Height is adjusted by the factor 0.967 such that the functions returned 'measured volume', cf. Eide 1923, as differed from volume estimates from means.
}
