% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Braastad_1977_BA_increment_Norway_Pine.R
\name{Braastad_1977_BA_increment_percent_Norway_Pine}
\alias{Braastad_1977_BA_increment_percent_Norway_Pine}
\title{Percent Basal Area Increment in stands of Scots Pine by Tveite 1976 for Braastad 1977}
\usage{
Braastad_1977_BA_increment_percent_Norway_Pine(
  SIH40,
  total_stand_age,
  BA_m2_ha_after_thinning,
  stems_after_thinning,
  diameter_mean_basal_area_stem_after_thinning,
  diameter_ratio_thinned_to_initial,
  thinned_BA_m2_ha_last_thinning,
  thinned_BA_m2_ha_second_last_thinning,
  thinned_BA_m2_ha_third_last_thinning
)
}
\arguments{
\item{SIH40}{Site Index H40 e.g. Site Index H40 , e.g. Tveite 1976 ([forester::Tveite_1976_height_trajectory_Pine_Norway()])}

\item{total_stand_age}{Total stand age, years.}

\item{BA_m2_ha_after_thinning}{Basal Area, m2/ha after thinning}

\item{stems_after_thinning}{Stems per hectare, after thinning.}

\item{diameter_mean_basal_area_stem_after_thinning}{Diameter corresponding to the mean basal area tree after thinning.}

\item{diameter_ratio_thinned_to_initial}{Ratio of the diameter of the thinned stems to that of the initial stand, e.g. [forester::Braastad_1977_diameter_quotient_thinned_trees_to_unthinned_stand_Norway_Pine()]}

\item{thinned_BA_m2_ha_last_thinning}{Basal Area m2/ha removed during the last thinning}

\item{thinned_BA_m2_ha_second_last_thinning}{Basal Area m2/ha removed during the last thinning}

\item{thinned_BA_m2_ha_third_last_thinning}{Basal Area m2/ha removed during the last thinning}
}
\value{
Basal Area Increment, Percent.
}
\description{
Percent Basal Area Increment in stands of Scots Pine by Tveite 1976 for Braastad 1977
}
