% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Braastad_1977_diameter_quotient_thinned_trees_to_unthinned_stand_Norway_Pine.R
\name{Braastad_1977_diameter_quotient_thinned_trees_to_unthinned_stand_Norway_Pine}
\alias{Braastad_1977_diameter_quotient_thinned_trees_to_unthinned_stand_Norway_Pine}
\title{Diameter Quotient of Thinned trees to remaining Standing trees for Scots Pine in Norway from Brastaad 1977.}
\source{
Braastad, Helge 1977. Tilvekstmodellprogram for furu. Growth model computer program for Pinus sylvestris. Meddr. Norsk inst. skogforsk 35:265-359. p 274 function (7)
}
\usage{
Braastad_1977_diameter_quotient_thinned_trees_to_unthinned_stand_Norway_Pine(
  dominant_height
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}
}
\value{
Ratio Diameter of thinned trees / Diameter of unthinned stand.
}
\description{
Diameter Quotient of Thinned trees to remaining Standing trees for Scots Pine in Norway from Brastaad 1977.
}
