% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Braastad_1977_Scots_Pine.R
\name{Braastad_1977_initial_stand_Norway_Pine}
\alias{Braastad_1977_initial_stand_Norway_Pine}
\title{Initial values for young Pine stands in Norway from Braastad 1977.}
\source{
Braastad, Helge 1977. Tilvekstmodellprogram for furu. Growth model computer program for Pinus sylvestris. Meddr. Norsk inst. skogforsk 35:265-359. p 274 function (4) function (5)
}
\usage{
Braastad_1977_initial_stand_Norway_Pine(
  dominant_height,
  stand_total_age,
  stems_per_ha,
  SIH40
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{stand_total_age}{Total age of stand.}

\item{stems_per_ha}{Number of stems per hectare.}

\item{SIH40}{Site Index H40 , e.g. Tveite 1976 ([forester::Tveite_1976_height_trajectory_Norway_Pine()])}
}
\value{
A list containing two elements: i) Lorey's Mean height. ii) Basal Area m2 per hectare.
}
\description{
Initial values for young Pine stands in Norway from Braastad 1977.
}
