% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Braastad_1977_Scots_Pine.R
\name{Braastad_1977_QMD_tree_volume_Norway_Pine}
\alias{Braastad_1977_QMD_tree_volume_Norway_Pine}
\title{Volume for Pine in Norway from Braastad 1977.}
\usage{
Braastad_1977_QMD_tree_volume_Norway_Pine(diameter_cm, height_m)
}
\arguments{
\item{diameter_cm}{Diameter corresponding the basal area mean tree.}

\item{height_m}{Lorey's mean height, m for the stand.}
}
\value{
Volume, m3sk.
}
\description{
Braastad uses two functions from Brantseg (1967). Lorey's height is weighted by 0.983 by basis of comparison with all experimental plots.
}
\examples{
Braastad_1977_QMD_tree_volume_Norway_Pine(23,19)
}
