% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Braastad_1977_stand_growth_model_Pine.R
\name{Braastad_1977_stand_growth_model_Pine}
\alias{Braastad_1977_stand_growth_model_Pine}
\title{Braastad 1977 growth model for Pine stands in Norway.}
\usage{
Braastad_1977_stand_growth_model_Pine(
  SIH40,
  Age,
  stems_per_ha,
  Basal_area_m2_ha = NULL,
  thinning_fix_stems = NULL,
  thinning_lower_relative_spacing = NULL,
  thinning_upper_relative_spacing = NULL,
  production_level = 1,
  height_reduction = 1,
  reduce_thinned_diameter_if_number_removed_stems_lower_than = 0,
  adjust_diameter_ratio_thinned = 1,
  BA_m2_ha_MAX = NULL
)
}
\arguments{
\item{SIH40}{Site Index H40, e.g. [forester::Tveite_1976_height_trajectory_Norway_Pine()]}

\item{Age}{Stand age at breast height.}

\item{stems_per_ha}{Number of stems per hectare.}

\item{Basal_area_m2_ha}{Optional. Default = NULL. Initial Basal area m2/ha.}

\item{thinning_fix_stems}{Optional. Should thinning always remove a fixed number of stems? Default is NULL.}

\item{thinning_lower_relative_spacing}{Set the lower density limit after thinning to a Hart - Becking Index., e.g. [forester::Hart_Becking_relative_spacing_index()]}

\item{thinning_upper_relative_spacing}{Set the upper density limit after thinning to a Hart - Becking Index., e.g. [forester::Hart_Becking_relative_spacing_index()]}

\item{production_level}{Default = 1. Set to lower to calculate production for stand less or more productive than average. e.g. 0.90 for 10\% lower production than average.}

\item{height_reduction}{Default=1. Increase the difference between dominant height and Lorey's mean height, e.g. for stands with extreme snow damage set the value 0.95 to reduce Lorey's mean height by 5\%}

\item{reduce_thinned_diameter_if_number_removed_stems_lower_than}{Default = 0. If the number of removed stems is less than this value, the stem diameters will be reduced by 10\%}

\item{adjust_diameter_ratio_thinned}{Default =1. Adjust ratio between the diameter of removed stems and the initial stand, calculated with [forester::Braastad_1977_diameter_quotient_thinned_trees_to_unthinned_stand_Norway_Pine()]}

\item{BA_m2_ha_MAX}{Optional. Default = NULL. Provide an upper limit to Basal Area after thinning given in m2/ha.}

\item{predict_ages}{A vector of ages to predict, starting with the Age.}
}
\description{
Braastad 1977 growth model for Pine stands in Norway.
}
