% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Braastad_1977_time_to_breast_height_Norway_Pine.R
\name{Braastad_1977_time_to_breast_height_Norway_Pine}
\alias{Braastad_1977_time_to_breast_height_Norway_Pine}
\title{Time to breast height for stands of Scots Pine in Norway from Braastad 1977.}
\source{
Braastad, Helge 1977. Tilvekstmodellprogram for furu. Growth model computer program for Pinus sylvestris. Meddr. Norsk inst. skogforsk 35:265-359. p 276 function (11)
}
\usage{
Braastad_1977_time_to_breast_height_Norway_Pine(SIH40)
}
\arguments{
\item{SIH40}{Site Index H40, e.g. [forester::Tveite_1976_height_trajectory_Norway_Pine()]}
}
\value{
Years until breast height 1.3 m.
}
\description{
Time to breast height for stands of Scots Pine in Norway from Braastad 1977.
}
\details{
Not applicable for SIHH40 lower than 8.

R = 0.377

Residual spread = ±3.8 yrs , ±35.3\%.
}
