% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brandel_1990.R
\name{Brandel_1990_volume_Pine}
\alias{Brandel_1990_volume_Pine}
\title{Brandel 1990 Volume for Scots Pine}
\source{
Brandel, G. 1990. Volymfunktioner för enskilda träd. Tall, gran och björk. SLU, inst. för skogsproduktion. Rapport nr 26.

Available: \url{https://www.heurekaslu.org/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf}
}
\usage{
Brandel_1990_volume_Pine(diameter_cm, height_m, latitude)
}
\arguments{
\item{diameter_cm}{Diameter in cm}

\item{height_m}{Height, metres.}

\item{latitude}{Latitude}
}
\value{
Volume, in dm3.
}
\description{
Volume functions for Scots Pine which are used in Heureka. PM Heureka 2004-01-20 Björn Elfving.
}
