% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brantseg_1967_volume_Norway_Pine.R
\name{Brantseg_1967_crown_base_height_Norway_Scots_Pine}
\alias{Brantseg_1967_crown_base_height_Norway_Scots_Pine}
\title{Height to crown base for Scots Pine in Norway}
\source{
Brantseg, A. 1967. Volume functions and tables for Scots pine.
 South Norway. Reports of the Norwegian Forest Research Institute, 12, 689-739.

 @author Modified from hansoleorka/skogR (2022-04-02)
}
\usage{
Brantseg_1967_crown_base_height_Norway_Scots_Pine(diameter_cm, height_m)
}
\arguments{
\item{diameter_cm}{Diameter of tree in cm at breast height, 1.3m.}

\item{height_m}{Total height of tree, m.}
}
\value{
Height to the tree crown base.
}
\description{
Height to crown base for Scots Pine in Norway
}
