% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norway_double_bark_functions.R
\name{Brantseg_1967_double_bark_thickness_Norway_Scots_Pine}
\alias{Brantseg_1967_double_bark_thickness_Norway_Scots_Pine}
\title{Double bark thickness for Norway Spruce in Norway}
\source{
Brantseg, A. 1967. Volume functions and tables for Scots pine.
South Norway. Reports of the Norwegian Forest Research Institute, 12, 689-739.
}
\usage{
Brantseg_1967_double_bark_thickness_Norway_Scots_Pine(diameter_cm, height_m)
}
\arguments{
\item{diameter_cm}{Diameter of tree at breast height, 1.3m.}

\item{height_m}{Total height of tree, m.}
}
\value{
Double bark, in mm.
}
\description{
Double bark thickness for Norway Spruce in Norway
}
\author{
Modified from hansoleorka/skogR on the 2022-04-02
}
