% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brantseg_1967_volume_Norway_Pine.R
\name{Brantseg_1967_volume_detailed_Norway_Scots_Pine}
\alias{Brantseg_1967_volume_detailed_Norway_Scots_Pine}
\title{Detailed volume function for Scots Pine in Norway}
\source{
Brantseg, A. 1967. Volume functions and tables for Scots pine.
 South Norway. Reports of the Norwegian Forest Research Institute, 12, 689-739.
}
\usage{
Brantseg_1967_volume_detailed_Norway_Scots_Pine(
  diameter_cm,
  height_m,
  crown_base_height.m,
  double_bark.mm,
  above_bark = TRUE
)
}
\arguments{
\item{diameter_cm}{Diameter at breast height, in cm.}

\item{height_m}{Total tree height, m.}

\item{crown_base_height.m}{Height to crown base, m.}

\item{double_bark.mm}{Thickness of double bark, mm.}

\item{above_bark}{TRUE (default) or FALSE.}
}
\value{
Volume, dm^3
}
\description{
Detailed volume function for Scots Pine in Norway
}
\seealso{
[forester::Brantseg_1967_volume_Norway_Scots_Pine()]
[forester::Brantseg_1967_crown_base_height_Norway_Scots_Pine()]
}
\author{
Modified from hansoleorka/skogR (2022-04-02)  NB inconsistency in comments.
}
