% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brantseg_1967_volume_Norway_Pine.R
\name{Brantseg_1967_volume_Norway_Scots_Pine}
\alias{Brantseg_1967_volume_Norway_Scots_Pine}
\title{Simple Volume function for Scots Pine in Norway}
\source{
Brantseg, A. 1967. Volume functions and tables for Scots pine.
 South Norway. Reports of the Norwegian Forest Research Institute, 12, 689-739.
}
\usage{
Brantseg_1967_volume_Norway_Scots_Pine(
  diameter_cm,
  height_m,
  above_bark = TRUE
)
}
\arguments{
\item{diameter_cm}{Diameter of tree at breast height, 1.3 m.}

\item{height_m}{Height of tree, m.}

\item{above_bark}{TRUE (default) or FALSE}
}
\value{
Volume of tree in DM^3
}
\description{
Simple Volume function for Scots Pine in Norway
}
\seealso{
[forester::Brantseg_1967_volume_detailed_Norway_Scots_Pine()]
}
\author{
Modified from hansoleorka/skogR (2022-04-02)
}
