% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Brantseg_1969_diameter_quotient_thinned_trees_Norway_Pine.R
\name{Brantseg_1969_diameter_quotient_thinned_trees_to_remaining_Norway_Pine}
\alias{Brantseg_1969_diameter_quotient_thinned_trees_to_remaining_Norway_Pine}
\title{Diameter Quotient of Thinned trees to remaining Standing trees for Scots Pine in Norway from Brantseg 1969.}
\source{
Brantseg, A. 1969. Furu sønnafjells. Produksjonstabeller. Meddr. norske SkogforsVes. 26: 1-291. p. 53 function 13. in
Braastad, Helge 1977. Tilvekstmodellprogram for furu. Growth model computer program for Pinus sylvestris. Meddr. Norsk inst. skogforsk 35:265-359. p 274 function (6)
}
\usage{
Brantseg_1969_diameter_quotient_thinned_trees_to_remaining_Norway_Pine(
  mean_height_Lorey
)
}
\arguments{
\item{mean_height_Lorey}{Lorey's mean height, e.g. [forester::Braastad_1977_initial_stand_Norway_Pine()] element 1.}
}
\value{
Diameter of Thinned trees / Diameter of remaining standing trees.
}
\description{
Diameter Quotient of Thinned trees to remaining Standing trees for Scots Pine in Norway from Brantseg 1969.
}
\examples{
Brantseg_1969_diameter_quotient_thinned_trees_to_remaining_Norway_Pine(9)
}
