% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralHelpers.R
\name{CajanusDg}
\alias{CajanusDg}
\alias{Oppermann_Cajanus_stand_QMD}
\alias{Oppermann_Cajanus_mean_diameter}
\alias{Oppermann_Cajanus_stand_BA}
\title{Calculate basal area weighted mean stem diameter from arithmetic mean stem diameter and
its' standard deviation e.g. Oppermann 1905, Cajanus 1912.}
\source{
Meyer, W. 1930. Diameter distribution Series in Evenaged Forest
Stands. Yale School of Forestry Bulletin 28. 105 pp. p.24. Available online:
 \url{https://elischolar.library.yale.edu/cgi/viewcontent.cgi?article=1027&context=yale_fes_bulletin}

Reformulation : Eriksson, H. 1976. Yield of Norway spruce in Sweden.
Research Notes. Dept. of. Forest Yield Research. Nr. 41. Royal College of
Forestry. Stockholm. p. 165.
}
\usage{
Oppermann_Cajanus_stand_QMD(diameter, diameter_sd)

Oppermann_Cajanus_mean_diameter(QMD, diameter_sd)

Oppermann_Cajanus_stand_BA(stems, diameter, diameter_sd)
}
\arguments{
\item{diameter}{Mean arithmetic diameter of the stand in cm.}

\item{diameter_sd}{Standard deviation of the stem distribution of the stand.}

\item{QMD}{The basal area weighted mean diameter (cm) of the stand (Dg).}

\item{stems}{Stems per hectare.}
}
\value{
Stand QMD.

The mean diameter of the stand, cm.

Stand total basal area
}
\description{
A relationship commonly attributed to either Oppermann (1905)
or Cajanus (1912).
}
