% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carbonnier_1954_volume_Larch.R
\name{Carbonnier_1954_volume_Larch}
\alias{Carbonnier_1954_volume_Larch}
\title{Volume of Larch trees, from Carbonnier 1954.}
\source{
Carbonnier, C. 1954. Funktioner för kubering av europeisk, sibirisk och japansk lärk. MS.

PM for Heureka 2004-01-20 Björn Elfving. Available: \url{https://www.heurekaslu.org/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf}
}
\usage{
Carbonnier_1954_volume_Larch(diameter_cm, height_m)
}
\arguments{
\item{diameter_cm}{Diameter at breast height, cm.}

\item{height_m}{Height of tree, metres.}
}
\value{
Volume of tree, dm3.
}
\description{
Volume of Larch trees, from Carbonnier 1954.
}
