% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carbonnier_1975.R
\name{Carbonnier_1975_form_factor_Oak}
\alias{Carbonnier_1975_form_factor_Oak}
\title{Calculate the form factor for Oak stands, from Carbonnier (1975)}
\source{
Carbonnier, C. (1975) Produktionen i kulturbestånd av ek i södra Sverige:
Yield of Oak plantations in southern Sweden. Studia Forestalia Suecica. Nr. 125.
Royal College of Forestry. Stockholm. ISBN 91-38-02278-8. p. 72.
}
\usage{
Carbonnier_1975_form_factor_Oak(Lorey_mean_height)
}
\arguments{
\item{Lorey_mean_height}{Mean height of stand according to Lorey's formula, e.g. [forester::Lorey_mean_height]}
}
\value{
Form factor for stand.
}
\description{
Calculate the form factor for Oak stands, from Carbonnier (1975)
}
