% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carbonnier_1975.R
\name{Carbonnier_1975_initial_stand_mean_diameter}
\alias{Carbonnier_1975_initial_stand_mean_diameter}
\title{Mean diameter for initial Oak stands from Carbonnier 1975}
\source{
Carbonnier, C. (1975) Produktionen i kulturbestånd av ek i södra Sverige:
Yield of Oak plantations in southern Sweden. Studia Forestalia Suecica. Nr. 125.
Royal College of Forestry. Stockholm. ISBN 91-38-02278-8. p. 18.
}
\usage{
Carbonnier_1975_initial_stand_mean_diameter(stems, dominant_height_m)
}
\arguments{
\item{stems}{Stems per hectare}

\item{dominant_height_m}{Dominant height, m.}
}
\value{
Mean stem diameter, cm.
}
\description{
Mean diameter for initial Oak stands from Carbonnier 1975
}
\details{
s= 0.872
}
