% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carbonnier_1975.R
\name{Carbonnier_1975_mean_height_before_thinning_Oak}
\alias{Carbonnier_1975_mean_height_before_thinning_Oak}
\title{Calculate the mean height in an Oak stand before a thinning.}
\source{
Carbonnier, C. (1975) Produktionen i kulturbestånd av ek i södra Sverige:
Yield of Oak plantations in southern Sweden. Studia Forestalia Suecica. Nr. 125.
Royal College of Forestry. Stockholm. ISBN 91-38-02278-8. p. 72.
}
\usage{
Carbonnier_1975_mean_height_before_thinning_Oak(dominant_height_m, stems)
}
\arguments{
\item{dominant_height_m}{Dominant height of stand, in metres.}

\item{stems}{Number of stems per hectare.}
}
\value{
Mean height of the Oak stand, in meters.
}
\description{
Calculate the mean height in an Oak stand before a thinning.
}
