% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Carbonnier_1975_Sweden_height_trajectories_Oak.R
\name{Carbonnier_1975_Sweden_height_trajectories_Oak}
\alias{Carbonnier_1975_Sweden_height_trajectories_Oak}
\title{Carbonnier 1975 Height growth of Oak plantations in Sweden.}
\source{
Carbonnier, C. (1975) Produktionen i kulturbestånd av ek i södra Sverige:
Yield of Oak plantations in southern Sweden. Studia Forestalia Suecica. Nr. 125.
Royal College of Forestry. Stockholm. ISBN 91-38-02278-8. p. 89.
}
\usage{
Carbonnier_1975_Sweden_height_trajectories_Oak(
  total_age,
  top_height_m,
  age_2,
  output = "SIH100"
)
}
\arguments{
\item{total_age}{total age of stand or tree.}

\item{top_height_m}{Top height of tree or stand in m.}

\item{age_2}{Necessary if output is "Height". The age for which height along
the same curve is to be computed.}

\item{output}{One of "SIH100", "Height", or "Equation".}
}
\value{
If output is "SIH100", the numeric value for the height (m) a stand
will reach at age 100.

 If output is "Height", the numeric value for the height (m) a stand will
 reach at age_2.

 If output is "Equation", a text response with the equation for that height
 curve.
}
\description{
Largely based on the Fortran IV script appended to the original source.
Checked against table VI.1 (ibid), p. 66.

This function was written by Björn Hägglund following the same method as outlined
in Hägglund 1972.
}
\details{
multiple correlation coefficient: 0.995

100*spread around the function/spread around the mean: 10.1%

spread around the mean of ln(height): 0.30096
}
\examples{
Carbonnier_1975_Sweden_height_trajectories_Oak(total_age = 50,top_height_m = 16.9,age_2 = 100,output = "Height")
}
