% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditions.R
\name{check_conditions}
\alias{check_conditions}
\title{Checking conditions for dataset}
\usage{
check_conditions(data, target = NULL, type = NULL)
}
\arguments{
\item{data}{dataset, used for training models. Class of data_train is one of those classes: data.frame, matrix, data.table or dgCMatrix. NOTE: data_train includes target column.}

\item{target}{character, indicating name of the target column in data_train.}

\item{type}{character, defining the task. Option is "regression" or "classification", namely binary classification.}
}
\value{
A verified dataframe, which is ready for creating ML models.
}
\description{
\code{check_conditions} verifies whether the input data with the defined target and defined type of problem
is ready for further steps in creating Machine Learning models.
}
\details{
Function transforms different data structures into dataframe, checks the presence of defined target column in data and the validity
of type.
}
\references{
forester library \url{https://modeloriented.github.io/forester/}
}
