% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_cor}
\alias{check_cor}
\title{Search for strongly correlated values (Spearman's rank for numerical, Crammer's V for
categorical)}
\usage{
check_cor(df, y = NULL, time = NULL, status = NULL, verbose = TRUE)
}
\arguments{
\item{df}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{y}{A string that indicates a target column name for regression or classification.
Either y, or pair: time, status can be used. By default NULL.}

\item{time}{A string that indicates a time column name for survival analysis task.
Either y, or pair: time, status can be used. By default NULL.}

\item{status}{A string that indicates a status column name for survival analysis task.
Either y, or pair: time, status can be used. By default NULL.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
the process, if FALSE gives none.}
}
\value{
A list with every line of the sub-report.
}
\description{
Search for strongly correlated values (Spearman's rank for numerical, Crammer's V for
categorical)
}
