% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Runs data check pipeline to seek for potential problems with the data}
\usage{
check_data(data, y)
}
\arguments{
\item{data}{A data source, that is one of major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string which indicates a target column name.}
}
\value{
A list with every line of the report.
}
\description{
Runs data check pipeline to seek for potential problems with the data
}
\examples{
check_data(iris[1:100,], 'Species')
check_data(lisbon, 'Price')
check_data(compas, 'Two_yr_Recidivism')
check_data(iris, 'Species')
check_data(lymph, 'class')
}
