% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_dim}
\alias{check_dim}
\title{Search for dimensionality problems in the dataset}
\usage{
check_dim(df, verbose = TRUE)
}
\arguments{
\item{df}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
the process, if FALSE gives none.}
}
\value{
A list with every line of the sub-report.
}
\description{
Search for dimensionality problems in the dataset
}
