% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_duplicate_col}
\alias{check_duplicate_col}
\title{Searches for duplicates between columns}
\usage{
check_duplicate_col(df)
}
\arguments{
\item{df}{A data source, that is one of major R formats: data.table, data.frame,
matrix and so on.}
}
\value{
A list with every line of the sub-report.
}
\description{
Searches for duplicates between columns
}
