% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metric.R
\name{check_metric}
\alias{check_metric}
\title{Checking conditions for metrics}
\usage{
check_metric(metric, type)
}
\arguments{
\item{metric}{character, defining metric which should be used in further tasks.}

\item{type}{character, defining the task. Option is "regression" or "classification", namely binary classification.}
}
\value{
A verified metric
}
\description{
\code{check_metric} verifies whether the given metric is chosen properly.
}
\references{
forester library \url{https://modeloriented.github.io/forester/}
}
