% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_outliers}
\alias{check_outliers}
\title{Searches for outliers via mean standard deviation, median abosulte deviation
and inter quantile range}
\usage{
check_outliers(df)
}
\arguments{
\item{df}{A data source, that is one of major R formats: data.table, data.frame,
matrix and so on.}
}
\value{
A list with every line of the sub-report.
}
\description{
Searches for outliers via mean standard deviation, median abosulte deviation
and inter quantile range
}
