% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Chrimes_2004_basal_area_increment_uneven_aged_Spruce.R
\name{Chrimes_2004_basal_area_increment_uneven_aged_Spruce}
\alias{Chrimes_2004_basal_area_increment_uneven_aged_Spruce}
\title{Basal area increment for Uneven-aged Spruce from Chrimes 2004}
\source{
Chrimes, D. & Lundqvist, L. 2004. Manuscript: Simulated volume increment of managed uneven-aged Picea abies stands in central Sweden. (PhD. dissertation, Swedish University of Agricultural Sciences). Chapter I p. 8.
}
\usage{
Chrimes_2004_basal_area_increment_uneven_aged_Spruce(
  diameter_cm,
  residual_stand_basal_area_m2_ha,
  overtopping_BA_m2_ha
)
}
\arguments{
\item{diameter_cm}{Diameter of tree at breast height 1.3 m.}

\item{residual_stand_basal_area_m2_ha}{Stand Basal Area at breast height, m2 / ha.}

\item{overtopping_BA_m2_ha}{The BA sum of all diameter class with a mean height of at least 2 m higher than the target diameter class.}
}
\value{
Over bark Basal area increment for a class. m2/ha/yr
}
\description{
Basal area increment for Uneven-aged Spruce from Chrimes 2004
}
\details{
Details from page 6.

r^2 = 0.69
df= 606
mse= 82.73
F= 150.6
se(\%)=25.88
p<0.0001
}
