% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chrimes_2004_form_quotient_uneven_aged_Spruce.R
\name{Chrimes_2004_form_quotient_uneven_aged_Spruce}
\alias{Chrimes_2004_form_quotient_uneven_aged_Spruce}
\title{Form quotient for Uneven aged Spruce from Chrimes 2004.}
\usage{
Chrimes_2004_form_quotient_uneven_aged_Spruce(
  diameter_cm,
  stand_basal_area_m2_ha
)
}
\arguments{
\item{diameter_cm}{Diameter of tree at breast height 1.3 m.}

\item{stand_basal_area_m2_ha}{Stand Basal Area at breast height, m2 / ha.}
}
\value{
Form quotient
}
\description{
Form quotient for Uneven aged Spruce from Chrimes 2004.
}
\details{
Chrimes, D. & Lundqvist, L. 2004. Manuscript: Simulated volume increment of managed uneven-aged Picea abies stands in central Sweden. (PhD. dissertation, Swedish University of Agricultural Sciences). Chapter I p. 7.

r^2 = 0.83
df = 3079
mse = 0.93
cv(\%)= 10.64
}
