% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chrimes_2004_mean_height_uneven_aged_Spruce.R
\name{Chrimes_2004_mean_height_uneven_aged_Spruce}
\alias{Chrimes_2004_mean_height_uneven_aged_Spruce}
\title{Height equation from Chrimes 2004}
\source{
Chrimes, D. & Lundqvist, L. 2004. Manuscript: Simulated volume increment of managed uneven-aged Picea abies stands in central Sweden. (PhD. dissertation, Swedish University of Agricultural Sciences). Chapter I p. 7.
}
\usage{
Chrimes_2004_mean_height_uneven_aged_Spruce(diameter_cm)
}
\arguments{
\item{diameter_cm}{Diameter of tree at breast height 1.3 m.}
}
\value{
Height of tree, metres.
}
\description{
Height equation from Chrimes 2004
}
\details{
r^2 = 0.95
df= 619
mse = 1.86
cv(\%)= 1.58
}
