% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_theme_forester.R
\name{colors_discrete_forester}
\alias{colors_discrete_forester}
\alias{colors_diverging_forester}
\title{Forester color palettes for ggplot2 objects}
\usage{
colors_discrete_forester(n = 2)

colors_diverging_forester()
}
\arguments{
\item{n}{number of colors for color palette}
}
\value{
color palette as vector of charactes
}
\description{
Forester color palettes for ggplot2 objects
}
