% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Function for comparing models}
\usage{
compare_models(models, metric)
}
\arguments{
\item{models}{list of models to compare. Models shoud be either \code{forester_model} or \code{explainer} object.}

\item{metric}{string containing name of mertic based on which the model will be selected. 
Metric shoud be written in small letters}
}
\value{
An object of the class \code{forester_model}.
}
\description{
Function \code{compare_models} enables user to compare various model with
the same type of task. Function compares them based on metric which can be chosen by the user 
and returns the best model
}
\examples{
\donttest{
library(DALEX)
data(apartments, package="DALEX")

lightgbm_explained <- make_lightgbm(apartments, "m2.price", "regression")
ranger_explained <- make_ranger(apartments, "m2.price", "regression")

models <- list(lightgbm_explained, ranger_explained)

best_model <- compare_models(models, "rmse")
}
}
\references{
Explanatory Model Analysis. Explore, Explain, and Examine Predictive Models. \url{https://ema.drwhy.ai/}
}
