% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_shape_file.R
\docType{data}
\name{county_shape_file}
\alias{county_shape_file}
\title{Swedish County Shape Files.}
\format{A large SpatialPolygonsDataFrame}
\source{
\url{https://www.heurekaslu.se/wiki/Definition:CountyCode}
}
\usage{
county_shape_file
}
\description{
Contains polygons corresponding to Swedish counties.
}
\keyword{datasets}
