% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_sweden.R
\name{county_sweden}
\alias{county_sweden}
\title{Coordinate to county Sweden}
\usage{
county_sweden(latitude, longitude, epsg = 4326)
}
\arguments{
\item{latitude}{Latitude, decimal}

\item{longitude}{Longitude, decimal}

\item{epsg}{Reference system. Default 'WGS84': 4326. For SWEREF99TM: '3006'.}
}
\value{
Text, Swedish County
}
\description{
Returns the Swedish county which the coordinate is placed in.
 If there is no match, county will be 'Other', and a warning will be shown.
}
