% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ranked_list.R
\name{create_ranked_list}
\alias{create_ranked_list}
\title{Create final ranked_list}
\usage{
create_ranked_list(
  basic,
  random,
  Bayes,
  type,
  score,
  sort_by = "auto",
  metric_function = NULL,
  metric_function_name = "metric_function",
  metric_function_decreasing = TRUE
)
}
\arguments{
\item{basic}{A data.frame with ranked list of basic models.}

\item{random}{A data.frame with ranked list of random search models.}

\item{Bayes}{A data.frame with ranked list of Bayes Optimization models.}

\item{type}{A string which determines if Machine Learning task is the
`binary_clf` or `regression`.}

\item{score}{A data frame with metrics values for given models, by `score_models`.}

\item{sort_by}{String with name of metric to sort by.
For `auto` models going to be sorted by `mse` for regression and `f1` for classification.}

\item{metric_function}{The self-created function.
It should look like name(predictions, observed) and return the numeric value.
In case of using `metrics` param with value other than `auto` or `all`, is needed to use value `metric_function`
in order to see given metric in report. If `sort_by` is equal to `auto` models are sorted by `metric_function`.}

\item{metric_function_name}{The name of the column with values of param `metric_function`.
By default `metric_function_name` is `metric_function`.}

\item{metric_function_decreasing}{A logical value indicating how metric_function should be sorted. `TRUE` by default.}
}
\value{
A data.frame ranked list sorted by default metric.
}
\description{
Create final ranked_list
}
