% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_preprocessing.R
\name{custom_preprocessing}
\alias{custom_preprocessing}
\title{Tailored custom preprocessing function which removes unnecessary data,
imputes the missing fields and selects most important features.}
\usage{
custom_preprocessing(
  data,
  y,
  na_indicators = c(""),
  removal_parameters = list(active_modules = c(duplicate_cols = TRUE, id_like_cols =
    TRUE, static_cols = TRUE, sparse_cols = TRUE, corrupt_rows = TRUE, correlated_cols =
    TRUE), id_names = c("id", "nr", "number", "idx", "identification", "index"),
    static_threshold = 0.99, sparse_columns_threshold = 0.7, sparse_rows_threshold = 0.7,
    high_correlation_threshold = 0.7),
  imputation_parameters = list(imputation_method = "median-other", k = 10, m = 5),
  feature_selection_parameters = list(feature_selection_method = "VI", max_features =
    "default", nperm = 1, cutoffPermutations = 20, threadsNumber = NULL, ),
  verbose = FALSE
)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string that indicates a target column name.}

\item{na_indicators}{A list containing the values that will be treated as NA
indicators. By default the list is c(''). WARNING Do not include NA or NaN,
as these are already checked in other criterion.}

\item{removal_parameters}{A list containing the parameters used in the removal
of unnecessary data. It needs to be provided as presented below with exactly
the same column names.

removal_parameters = list(
active_modules = c(duplicate_cols = TRUE, id_like_cols = TRUE,
                  static_cols = TRUE, sparse_cols = TRUE,
                  corrupt_rows = TRUE, correlated_cols = TRUE),
id_names = c('id', 'nr', 'number', 'idx', 'identification', 'index'),
static_threshold = 0.99,
sparse_columns_threshold = 0.7,
sparse_rows_threshold = 0.7,
high_correlation_threshold = 0.7
)
The parameters are described below:
\itemize{
\item \code{`active_modules`} A logical vector describing active removal modules. By default it
is set as `c(duplicate_cols = TRUE, id_like_cols = TRUE, static_cols = TRUE,
sparse_cols = TRUE, corrupt_rows = TRUE, correlated_cols = TRUE)`, which is
equal to c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE). Setting corrupt_rows to FALSE
still results in the removal of observations without target value.
\item \code{`id_names`} A vector of strings indicating which column names are perceived
as ID-like. By default the list is: ['id', 'nr', 'number', 'idx', 'identification', 'index'].
\item \code{`static_threshold`} A numeric value from [0,1] range, which indicates the maximum
threshold of dominating values for column If feature has more dominating
values it is going to be removed. By default set to 1, which indicates that
all values are equal.
\item \code{`sparse_columns_threshold`}A numeric value from [0,1] range, which indicates the maximum
threshold of missing values for columns If column has more missing fields
it is going to be removed. By default set to 0.7.
\item \code{`sparse_rows_threshold`}A numeric value from [0,1] range, which indicates the maximum
threshold of missing values for observation. If observation has more missing fields
it is going to be removed. By default set to 0.7.
\code{`na_indicators`}A list containing the values that will be treated as NA
indicators. By default the list is c(''). WARNING Do not include NA or NaN,
as these are already checked in other criterion.
\item \code{`high_correlation_threshold`}A numeric value from [0,1] range, which indicates when we consider
the correlation to be high. If feature surpasses this threshold it is going to
be removed. By default set to 0.7.
}}

\item{verbose}{}
}
\description{
Tailored custom preprocessing function which removes unnecessary data,
imputes the missing fields and selects most important features.
}
