% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{delete_correlated_values}
\alias{delete_correlated_values}
\title{Delete correlated values}
\usage{
delete_correlated_values(data, y, verbose = TRUE)
}
\arguments{
\item{data}{A data source before preprocessing, that is one of the major R formats:
data.table, data.frame, matrix, and so on.}

\item{y}{A string that indicates a target column name.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
the process, if FALSE gives none.}
}
\value{
A list with 2 items: data set with removed correlated columns and
names of removed columns.
}
\description{
The deleting process starts with the extraction of correlated pairs of columns.
Later, they are sorted decreasingly by correlation value and we count the number
of occurrences to sort column names later by the amount of them. In the end we
remove columns iteratively from the most occurring one, so we will remove the
bare minimum of columns in the end.
}
