% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuction_test.R
\name{desc}
\alias{desc}
\title{Descriptive statistics}
\usage{
desc(x)
}
\arguments{
\item{x}{Um vetor numérico.}
}
\value{
Uma lista com estatísticas descritivas.
}
\description{
Esta função realiza a estatística descritiva para um vetor numérico.
}
\details{
Um detalhe importante.
}
