% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_young_tree_age.R
\name{determine_young_tree_age}
\alias{determine_young_tree_age}
\title{Determine age at breast height for young trees}
\usage{
determine_young_tree_age(
  height_dm,
  species,
  mean_height,
  si,
  source_sprout = 0
)
}
\arguments{
\item{height_dm}{Height in decimetres.}

\item{species}{One of Pinus sylvestris, Picea abies, Betula pendula, Betula pubescens}

\item{mean_height}{Arithmetic mean height of future crop trees (1600 stems per ha)}

\item{si}{dm, SI H100 spruce for spruce, H100 Pine for pine.}

\item{source_sprout}{1 for sprout, else 0}
}
\value{
age at breast height of young trees
}
\description{
Determine age at breast height for young trees
}
