% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{draw_boxplot}
\alias{draw_boxplot}
\title{Draw boxplot of resuduals - for regression}
\usage{
draw_boxplot(observed, predictions, models)
}
\arguments{
\item{observed}{A vector of the true values of target.}

\item{predictions}{A list of predictions for each trained model.}

\item{models}{A list of models.}
}
\description{
Draw boxplot of resuduals - for regression
}
