% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{draw_confusion_matrix}
\alias{draw_confusion_matrix}
\title{Draw confusion matrix for the model}
\usage{
draw_confusion_matrix(best_models, test_data, observed)
}
\arguments{
\item{best_models}{A list of models.}

\item{test_data}{A test dataset.}

\item{observed}{A vector of observed target values.}
}
\value{
A confusion matrix ggplot object.
}
\description{
Draw confusion matrix for the model
}
