% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{draw_radar_plot}
\alias{draw_radar_plot}
\title{Plot radar chart of one metric}
\usage{
draw_radar_plot(score_frame, type)
}
\arguments{
\item{score_frame}{The result of `score_model` function.}

\item{type}{A string which determines if machine learning task is the
`classification` or `regression`.}
}
\description{
Plot radar chart of one metric
}
\examples{
train_out <- train(lisbon, 'Price', verbose = FALSE)
draw_radar_plot(train_out$ranked_list[1:10, ], train_out$type)
}
