% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{draw_rmse_plot}
\alias{draw_rmse_plot}
\title{Draws train vs test RMSE plot for models}
\usage{
draw_rmse_plot(
  train_observed,
  train_predictions,
  test_observed,
  test_predictions,
  models
)
}
\arguments{
\item{train_observed}{A vector of the true values for training dataset.}

\item{train_predictions}{A list of lists of predictions on the training dataset for models but just for c('ranger', 'xgboost', 'decision_tree')}

\item{test_observed}{A vector of the true values for test dataset.}

\item{test_predictions}{A list of lists of predictions on test dataset for models but just for c('ranger', 'xgboost', 'decision_tree')}

\item{models}{A list of models.}
}
\description{
Draws train vs test RMSE plot for models
}
