% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{draw_roc_plot}
\alias{draw_roc_plot}
\title{Draw AUC ROC curve for the best model}
\usage{
draw_roc_plot(best_models, test_data, observed)
}
\arguments{
\item{best_models}{A list of models.}

\item{test_data}{A test dataset.}

\item{observed}{A vector of observed target values.}
}
\value{
An AUC ROC ggplot object.
}
\description{
Draw AUC ROC curve for the best model
}
