% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Eko_PM_1985_volume_northern_central_Sweden_Birch.R
\name{Eko_PM_1985_volume_northern_central_Sweden_Birch}
\alias{Eko_PM_1985_volume_northern_central_Sweden_Birch}
\title{Volume function for Birch in northern and central Sweden.}
\source{
Ekö, Per Magnus (1985) "En produktionsmodell för skog i Sverige, baserad på bestånd från
riksskogstaxeringens provytor: A growth simulator for Swedish forests, based on
data from the national forest survey. Rapporter nr. 16. Swedish University of
Agricultural Sciences, dept. of Silviculture. Umeå. ISBN 91-576-2386-4. ISSN
 0348-8969.  p. 107

 @description Multiple correlation coefficient: 0.95

 Standard deviation about the function: 0.39

 Sf/Standard deviation about the mean: 0.32
}
\usage{
Eko_PM_1985_volume_northern_central_Sweden_Birch(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  latitude,
  altitude,
  fertilised,
  thinned,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)
}
\arguments{
\item{basal_area.m2}{Basal area over bark of Birch (m2/ha)}

\item{basal_area_other_species}{Basal area of other species on the plot. (m2/ha)}

\item{age_at_breast_height}{Age at breast height, calculated as the mean age of the two thickest trees (years).}

\item{stem_number_ha}{Number of stems per hectare.}

\item{SI}{Site index H100 Spruce , m.}

\item{latitude}{Degrees}

\item{altitude}{Meters above sea level.}

\item{fertilised}{1 if the stand has been fertilised, otherwise 0.}

\item{thinned}{1 if the stand has been thinned, otherwise 0.}

\item{basal_area_weighted_mean_diameter}{Diameter corresponding to mean basal area (m).}

\item{basal_area_weighted_mean_diameter_other_species}{Diameter corresponding to mean basal area (m) of other species on plot.}
}
\value{
m3sk.
}
\description{
Volume function for Birch in northern and central Sweden.
}
