% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eko_PM_1985_volume_northern_Sweden_Pine.R
\name{Eko_PM_1985_volume_northern_Sweden_Pine}
\alias{Eko_PM_1985_volume_northern_Sweden_Pine}
\title{Volume function for Scots Pine in northern Sweden.}
\source{
Ekö, Per Magnus (1985) "En produktionsmodell för skog i Sverige, baserad på bestånd från
riksskogstaxeringens provytor: A growth simulator for Swedish forests, based on
data from the national forest survey. Rapporter nr. 16. Swedish University of
Agricultural Sciences, dept. of Silviculture. Umeå. ISBN 91-576-2386-4. ISSN
 0348-8969.  p. 105

 @description Multiple correlation coefficient: 0.98

 Standard deviation about the function: 0.23

 Sf/Standard deviation about the mean: 0.17
}
\usage{
Eko_PM_1985_volume_northern_Sweden_Pine(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  thinned_previous_five_years,
  thinned_before_previous_five_years,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)
}
\arguments{
\item{basal_area.m2}{Basal area over bark of Scots Pine (m2/ha)}

\item{basal_area_other_species}{Basal area of other species on the plot. (m2/ha)}

\item{age_at_breast_height}{Age at breast height, calculated as the mean age of the two thickest trees (years).}

\item{stem_number_ha}{Number of stems per hectare.}

\item{SI}{Site index H100 Pine , m.}

\item{thinned_previous_five_years}{1 if the stand has been thinned during the last five years, otherwise 0.}

\item{thinned_before_previous_five_years}{1 if the stand has been thinned before the last five years, otherwise 0.}

\item{basal_area_weighted_mean_diameter}{Diameter corresponding to mean basal area (m).}

\item{basal_area_weighted_mean_diameter_other_species}{Diameter corresponding to mean basal area (m) of other species on plot.}
}
\value{
m3sk.
}
\description{
Volume function for Scots Pine in northern Sweden.
}
