% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eko_PM_1985_volume_southern_Sweden_Pine.R
\name{Eko_PM_1985_volume_southern_Sweden_Pine}
\alias{Eko_PM_1985_volume_southern_Sweden_Pine}
\title{Volume function for Pine in southern Sweden.}
\source{
Ekö, Per Magnus (1985) "En produktionsmodell för skog i Sverige, baserad på bestånd från
riksskogstaxeringens provytor: A growth simulator for Swedish forests, based on
data from the national forest survey. Rapporter nr. 16. Swedish University of
Agricultural Sciences, dept. of Silviculture. Umeå. ISBN 91-576-2386-4. ISSN
 0348-8969.  p. 105

 @description Multiple correlation coefficient: 0.98

 Standard deviation about the function: 0.23

 Sf/standard deviation about the mean: 0.18
}
\usage{
Eko_PM_1985_volume_southern_Sweden_Pine(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  thinned,
  vegetation,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)
}
\arguments{
\item{basal_area.m2}{Basal area over bark of Pine (m2/ha)}

\item{basal_area_other_species}{Basal area of other species on the plot. (m2/ha)}

\item{age_at_breast_height}{Age at breast height, calculated as the mean age of the two thickest trees (years).}

\item{stem_number_ha}{Number of stems per hectare.}

\item{SI}{Site index H100 Pine , m.}

\item{thinned}{1 if the stand has been thinned before, otherwise 0.}

\item{vegetation}{Variable indicating vegetation type scaled from -5 to +4 as follows:

\tabular{llr}{
Field Layer Code (NFI) \tab Description \tab Index \cr
1\tab Tall herbs w/o dwarf shrubs \tab 4 \cr
2\tab Tall herbs with bilberry \tab 2.5 \cr
3\tab Tall herbs with cowberry \tab 2 \cr
4\tab Low herbs w/o dwarf shrubs \tab 3 \cr
5\tab Low herbs with bilberry \tab 2.5 \cr
6\tab Low herbs with cowberry \tab 2 \cr
7\tab No field layer \tab 3 \cr
8\tab broad-leafed grasses \tab 2.5 \cr
9\tab narrow-leaved grasses \tab 1.5 \cr
10\tab Sedge, tall \tab -3 \cr
11\tab Sedge, low \tab -3 \cr
12\tab Horsetail \tab 1 \cr
13\tab Bilberry \tab 0 \cr
14\tab Cowberry \tab -0.5 \cr
15\tab Crowberry \tab -3 \cr
16\tab Poor shrub \tab -5 \cr
17\tab Lichen-rich \tab -0.5 \cr
18\tab Lichen-dominated \tab -1 \cr
}}

\item{basal_area_weighted_mean_diameter}{Diameter corresponding to mean basal area (m).}

\item{basal_area_weighted_mean_diameter_other_species}{Diameter corresponding to mean basal area (m) of other species on plot.}

\item{latitude}{Degrees}

\item{altitude}{Meters above sea level}
}
\value{
m3sk.
}
\description{
Volume function for Pine in southern Sweden.
}
