% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eko_PM_2008_MAI_MAX_estimate_smoothed.R
\name{Eko_PM_2008_MAI_MAX_estimate_smoothed}
\alias{Eko_PM_2008_MAI_MAX_estimate_smoothed}
\title{MAI max as estimated by smoothing function in Ekö PM et al 2008.}
\source{
Ekö, P-M, Johansson, U., Petersson, N. Bergqvist, J., Elfving, B.,
Frisk, J. (2008) Current growth differences of Norway spruce (Picea abies),
Scots Pine (Pinus sylvestris) and birch (Betula pendula and Betula pubescens)
in different regions in Sweden. Scandinavian Journal of Forest Research Vol.
23:4. p. 307-318. DOI: \url{https://doi.org/10.1080/02827580802249126}
}
\usage{
Eko_PM_2008_MAI_MAX_estimate_smoothed(
  species,
  SI,
  county,
  altitude,
  vegetation
)
}
\arguments{
\item{species}{One of : "Betula pendula", "Betula pubescens","Picea abies",
"Pinus sylvestris"}

\item{SI}{Site Index. \strong{OBSERVE: H100 for Picea abies,
H100 for Pinus sylvestris,
H50 for Betula spp.}

For Picea abies northern Sweden
[forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce()].
For Picea abies southern Sweden e.g.
[forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce()].
For Pinus sylvestris e.g.
[forester::Hagglund_1974_Sweden_height_trajectories_Pine()].
For Betula pendula / Betula pubescens e.g.
[forester::Eriksson_1997_height_trajectory_Sweden_Birch()].
If SIH not available, use Site factors: e.g. [forester::SIS_estimate],
for Birch: [forester::Eko_PM_2008_stand_factors_estimate_SI_Sweden_Birch()]}

\item{county}{Swedish county, e.g. [forester::county_sweden()]}

\item{altitude}{Meters above sea level.}

\item{vegetation}{Vegetation type according to code following Swedish
National forest inventory FALTSKIKT:
\tabular{ll}{
Code \tab Vegetation \cr
1 \tab  Rich-herb without shrubs \cr
2 \tab Rich-herb with shrubs/bilberry \cr
3 \tab Rich-herb with shrubs/lingonberry \cr
4 \tab Low-herb without shrubs \cr
5 \tab Low-herb with shrubs/bilberry \cr
6 \tab Low-herb with shrubs/lingonberry \cr
7 \tab No field layer \cr
8 \tab Broadleaved grass \cr
9 \tab Thinleaved grass \cr
10 \tab Sedge, high \cr
11 \tab Sedge, low \cr
12 \tab Horsetail, Equisetum ssp. \cr
13 \tab Bilberry \cr
14 \tab Lingonberry \cr
15 \tab Crowberry \cr
16 \tab Poor shrub \cr
17 \tab Lichen, frequent occurrence \cr
18 \tab Lichen, dominating \cr
}}
}
\value{
Numeric. MAI max.
}
\description{
MAI max as estimated by smoothing function in Ekö PM et al 2008.
}
\examples{
Eko_PM_2008_MAI_MAX_estimate_smoothed(species="Picea abies",SI=36county="Kronoberg",altitude=100,vegetation=4)
}
