% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Eko_PM_2008_stand_factors_estimate_SI_Sweden_Birch.R
\name{Eko_PM_2008_stand_factors_estimate_SI_Sweden_Birch}
\alias{Eko_PM_2008_stand_factors_estimate_SI_Sweden_Birch}
\title{Estimate SIH50 with stand factors for Birch from Ekö et al 2008.}
\source{
Ekö, P-M, Johansson, U., Petersson, N. Bergqvist, J., Elfving, B., Frisk, J. (2008) Current growth differences of Norway spruce (Picea abies), Scots Pine (Pinus sylvestris) and birch (Betula pendula and Betula pubescens) in different regions in Sweden. Scandinavian Journal of Forest Research Vol. 23:4. p. 307-318. DOI: \url{https://doi.org/10.1080/02827580802249126}
}
\usage{
Eko_PM_2008_stand_factors_estimate_SI_Sweden_Birch(
  altitude,
  latitude,
  vegetation,
  ground_layer,
  lateral_water,
  soil_moisture
)
}
\arguments{
\item{altitude}{Altitude in metres above sea level.}

\item{latitude}{Latitude, degrees.}

\item{vegetation}{Vegetation type according to follows Swedish National forest inventory FALTSKIKT:
\tabular{ll}{
Code \tab Vegetation \cr
1 \tab  Rich-herb without shrubs \cr
2 \tab Rich-herb with shrubs/bilberry \cr
3 \tab Rich-herb with shrubs/lingonberry \cr
4 \tab Low-herb without shrubs \cr
5 \tab Low-herb with shrubs/bilberry \cr
6 \tab Low-herb with shrubs/lingonberry \cr
7 \tab No field layer \cr
8 \tab Broadleaved grass \cr
9 \tab Thinleaved grass \cr
10 \tab Sedge, high \cr
11 \tab Sedge, low \cr
12 \tab Horsetail, Equisetum ssp. \cr
13 \tab Bilberry \cr
14 \tab Lingonberry \cr
15 \tab Crowberry \cr
16 \tab Poor shrub \cr
17 \tab Lichen, frequent occurrence \cr
18 \tab Lichen, dominating \cr
}}

\item{ground_layer}{According to Swedish NFI, roughly:

  \tabular{ll}{
Code \tab Ground layer \cr
1 \tab Lichen type (>50\% of existing ground layer)\cr
2 \tab Lichen-rich bogmoss type (>25\% lichen + >50\% Sphagnum of existing ground layer)\cr
3 \tab Lichen rich (>25\% lichen, & not >50\% Sphagnum of existing ground layer) \cr
4 \tab Bogmoss type (\emph{Sphagnum} > 50\% of existing ground layer) \cr
5 \tab Swamp moss type (\emph{Polytrichum commune, P. gracile, P. strictum, Sphagnum, Depranocladus, Scorpidium, Paludella, Calliergon, Tomentypnum, Campylium.}) \cr
6 \tab Fresh moss type (\emph{Hylocomium splendens, Ptilium crista-castrensis, Dicranum ssp.}) \cr
}}

\item{lateral_water}{Type 1="Missing",Type 2="Seldom",3="shorter periods",4="longer periods", 5="slope".}

\item{soil_moisture}{Type 1="Dry/torr",2="Mesic/frisk",3="Mesic-moist/frisk-fuktig",4="Moist/fuktig",5="Wet/Blöt"}
}
\value{
Site Index for Birch H50
}
\description{
Estimate SIH50 with stand factors for Birch from Ekö et al 2008.
}
\details{
Northern Sweden, mesic (SD=2.86,R^2=61\%)

Northern Sweden, moist (SD=2.94,R^2=48\%)

Southern Sweden, mesic (SD=2.95,R^2=18\%)

Southern Sweden, moist (SD=2.88,R^2=28\%)
}
