% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_1985_volume_Pinus_contorta.R
\name{Elfving_1985_volume_Pinus_contorta}
\alias{Elfving_1985_volume_Pinus_contorta}
\title{Volume of Lodgepole Pine trees, from Elfving 1985.}
\source{
Elfving, B. 1985. Nya data om contortatallens produktion. SLU, Inst. f. skogsskötsel. Arbetsrapporter, nr. 3.
}
\usage{
Elfving_1985_volume_Pinus_contorta(diameter_cm, height_m)
}
\arguments{
\item{diameter_cm}{Diameter at breast height, cm.}

\item{height_m}{Height of tree, metres.}
}
\value{
Volume of tree, dm3.
}
\description{
Volume of Lodgepole Pine trees, from Elfving 1985.
}
