% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_1992_stocking_cultivated_regeneration.R
\name{Elfving_1992_stocking_cultivated_regeneration}
\alias{Elfving_1992_stocking_cultivated_regeneration}
\title{Stocking (0-1) in natural regenerations from Elfving 1992.}
\source{
Elfving, B. (1992). Återväxtens etablering och utveckling till röjningstidspunkten. SLU, Institutionen för Skogsskötsel. Arbetsrapport nr. 67.
Available, p.5 : \url{https://www.heurekaslu.se/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf}
}
\usage{
Elfving_1992_stocking_cultivated_regeneration(
  latitude,
  altitude,
  demanded_seedlings_per_ha = 2500,
  spacing,
  growth_periods_since_cut = 12,
  county,
  sown,
  spruce,
  scarification,
  prescribed_burning,
  Jonson_site_class
)
}
\arguments{
\item{latitude}{Latitude north, decimal degrees.}

\item{altitude}{Meters above sea level.}

\item{demanded_seedlings_per_ha}{Number of seedlings demanded for full stocking. Default 2500 as per implementation.}

\item{spacing}{Average square spacing used at sowing or planting.}

\item{growth_periods_since_cut}{Number of growth periods since the cut. Default 12 as per implementation.}

\item{county}{County, e.g. [forester::county_sweden()]}

\item{sown}{TRUE, if the site has been sown. Otherwise FALSE.}

\item{spruce}{TRUE, if the site has been cultivated with Spruce. Otherwise FALSE.}

\item{scarification}{TRUE if site has been scarified, otherwise FALSE.}

\item{prescribed_burning}{TRUE if a prescribed burn has been done after the cut, otherwise FALSE.}

\item{Jonson_site_class}{Site class according to Jonson 1914. e.g. [forester::Hagglund_1981_si_to_bonitet()] and  [forester::Jonson_1914_MAI_to_class()]}
}
\value{
A list with elements (1) : Stocking, varying between 0-1, and (2): Young stand quality W.
}
\description{
OBSERVE! Plausible retransformation error?
Distance from equator calculated as 111 km * latitude.
Jämtland county has been interpreted as both Jämtland - Härjedalens landskap and Jämtland - Jämtlands landskap.

OBSERVE! This function includes the modifications to coefficients as defined on p. 6 in the source:
\emph{Increased efficiency in provenance selection and scarification has also been considered by assumed influence on the cofficients. For cultivations the negative effect by increasing height above see level has been reduced from -0.0514 to -0.0257 and the positive effect by scarification has been increased from 0.0757 to 0.2.}
}
\details{
n = 3528

R^2 = 0.492

s-res = 0.391

Mean of dependents (?): 2.3059
}
