% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_1992_stocking_natural_regeneration.R
\name{Elfving_1992_stocking_natural_regeneration}
\alias{Elfving_1992_stocking_natural_regeneration}
\title{Stocking (0-1) in natural regenerations from Elfving 1992.}
\source{
Elfving, B. (1992). Återväxtens etablering och utveckling till röjningstidspunkten. SLU, Institutionen för Skogsskötsel. Arbetsrapport nr. 67.
Available, p.5 : \url{https://www.heurekaslu.se/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf}
}
\usage{
Elfving_1992_stocking_natural_regeneration(
  latitude,
  altitude,
  proportion_cultivated,
  demanded_seedlings_per_ha = 2500,
  growth_periods_since_cut = 12,
  soil_moisture,
  county,
  scarification,
  prescribed_burning,
  untreated,
  uncleaned,
  seed_trees_per_ha,
  Jonson_site_class,
  area
)
}
\arguments{
\item{latitude}{Latitude north, decimal degrees.}

\item{altitude}{Meters above sea level.}

\item{proportion_cultivated}{Proportion of cultivated seedlings according to inventory.}

\item{demanded_seedlings_per_ha}{Number of seedlings demanded for full stocking. Default 2500 as per implementation.}

\item{growth_periods_since_cut}{Number of growth periods since the cut. Default 12 as per implementation.}

\item{soil_moisture}{Type 1="Dry/torr",2="Mesic/frisk",3="Mesic-moist/frisk-fuktig",4="Moist/fuktig",5="Wet/Blöt"}

\item{county}{County, e.g. [forester::county_sweden()]}

\item{scarification}{TRUE if site has been scarified, otherwise FALSE.}

\item{prescribed_burning}{TRUE if a prescribed burn has been done after the cut, otherwise FALSE.}

\item{untreated}{TRUE if there were no measures done to promote regeneration after the final felling, otherwise FALSE.}

\item{uncleaned}{TRUE if bushes and small trees remain after clear-felling, otherwise FALSE.}

\item{seed_trees_per_ha}{Number of seed trees per hectare.}

\item{Jonson_site_class}{Site class according to Jonson 1914. e.g. [forester::Hagglund_1981_si_to_bonitet()] and  [forester::Jonson_1914_MAI_to_class()]}

\item{area}{Area of site in hectares.}
}
\value{
A list with elements (1) : Stocking, varying between 0-1, and (2): Young stand quality W.
}
\description{
OBSERVE! Plausible retransformation error?
Distance from equator calculated as 111 km * latitude.
Jämtland county has been interpreted as both Jämtland - Härjedalens landskap and Jämtland - Jämtlands landskap.

OBSERVE! This function includes the modifications to coefficients as defined on p. 6 in the source:
\emph{For natural regenerations the effect by scarification has been modified to 0.30 on mesic sites and 0.15 on other sites.}
}
\details{
n = 2462

R^2 = 0.492

s-res = 0.475

Mean of dependents (?): 1.8585
}
