% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_2003_height_trajectory_Sweden_Spruce.R
\name{Elfving_2003_height_trajectory_Sweden_Spruce}
\alias{Elfving_2003_height_trajectory_Sweden_Spruce}
\title{Height trajectory for Norway Spruce in Sweden from Elfving 2003.}
\source{
Elfving B. (2003) Övre höjdens utveckling i granplanteringar. Arbetsrapporter 185:1-8. Inst. f. skogsskötsel. Swedish University of Agricultural Sciences.
}
\usage{
Elfving_2003_height_trajectory_Sweden_Spruce(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Total age.}

\item{age2}{Total age at output age.}

\item{output}{One of "SIH100","Equation" or "Height".}
}
\value{

}
\description{
81 Norway Spruce stands.

293 measurement periods.
}
