% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_2004_BA_increment_Beech.R
\name{Elfving_2004_BA_increment_Beech}
\alias{Elfving_2004_BA_increment_Beech}
\title{Individual tree basal area growth for Beech from Elfving 2004}
\source{
Elfving, B. (2010) Translated, re-formulated Pro-Memoria for HEUREKA based on Manuscript 2004-01-26. 'Individual-tree basal area growth functions for all Swedish forests'. Available: \url{https://www.heurekaslu.se/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf}
}
\usage{
Elfving_2004_BA_increment_Beech(
  diameter_cm,
  Basal_area_weighted_mean_diameter_cm,
  distance_to_coast_km,
  BA_sum_of_trees_with_larger_diameter,
  Basal_area_Beech,
  Basal_area_plot,
  Basal_area_stand,
  computed_tree_age,
  latitude,
  altitude,
  vegetation,
  SIS100,
  divided_plot = 0,
  edge_effect,
  uneven_aged,
  fertilised = 0,
  thinned,
  last_thinned
)
}
\arguments{
\item{diameter_cm}{Diameter of the tree at breast height, 1.3 m}

\item{Basal_area_weighted_mean_diameter_cm}{Basal area weighted mean diameter of the trees on the plot, cm. \eqn{(\sum{diameter_cm^3} / \sum{diameter_cm^2})}}

\item{distance_to_coast_km}{Distance in kilometres to the nearest Swedish coast, e.g. [forester::coast_distance()]}

\item{BA_sum_of_trees_with_larger_diameter}{Basal area sum of trees on the plot with larger diameters than that of the target tree.}

\item{Basal_area_Beech}{Basal area Beech on the plot. m2/ha.}

\item{Basal_area_plot}{Basal area on the plot, m2/ha.}

\item{Basal_area_stand}{Basal area in the surrounding stand, m2/ha.}

\item{computed_tree_age}{Age at breast height computed by [forester::Elfving_2003_single_tree_age_estimation()]}

\item{latitude}{Latitude, decimal degrees.}

\item{vegetation}{Vegetation type according to follows Swedish National forest inventory FALTSKIKT:
\tabular{ll}{
Code \tab Vegetation \cr
1 \tab  Rich-herb without shrubs \cr
2 \tab Rich-herb with shrubs/bilberry \cr
3 \tab Rich-herb with shrubs/lingonberry \cr
4 \tab Low-herb without shrubs \cr
5 \tab Low-herb with shrubs/bilberry \cr
6 \tab Low-herb with shrubs/lingonberry \cr
7 \tab No field layer \cr
8 \tab Broadleaved grass \cr
9 \tab Thinleaved grass \cr
10 \tab Sedge, high \cr
11 \tab Sedge, low \cr
12 \tab Horsetail, Equisetum ssp. \cr
13 \tab Bilberry \cr
14 \tab Lingonberry \cr
15 \tab Crowberry \cr
16 \tab Poor shrub \cr
17 \tab Lichen, frequent occurrence \cr
18 \tab Lichen, dominating \cr
}}

\item{SIS100}{Site Index as measured from [forester::SIS_estimate()]}

\item{divided_plot}{1 for plots described in different parts, where the other part is **not** open land.  0 for full plots (default). (divided_plot and edge_effect cannot be TRUE at the same time)}

\item{edge_effect}{1 for a partitioned plot where the other part **is** open land. 0 for full plots (default). (divided_plot and edge_effect cannot be TRUE at the same time)}

\item{uneven_aged}{1 if less than 80\% of the main stand volume is within a 20-year age-span.}

\item{fertilised}{1 if the plot has been fertilised, otherwise 0 (default).}

\item{thinned}{1 if the plot has been thinned, otherwise 0.}

\item{last_thinned}{Number of growth seasons since the stand was last thinned.}
}
\value{
The 5 - year increase of Basal Area (m^2)
}
\description{
Based on data from permanent plots at the National Forest Inventory.
}
\details{
R^2 = 0.722
}
