% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_2009_BA_stand.R
\name{Elfving_2009_BA_stand}
\alias{Elfving_2009_BA_stand}
\title{Stand Basal Area Development Elfving 2009}
\source{
\url{https://www.heurekaslu.se/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf} page 50 in text.
}
\usage{
Elfving_2009_BA_stand(
  vegetation,
  stand_age,
  Basal_area_Conifer_m2_ha,
  Basal_area_Pine_m2_ha,
  Basal_area_Birch_m2_ha,
  Basal_area_after_thinning,
  Basal_area_before_thinning,
  Basal_area_stand,
  stems_after_thinning,
  peatland,
  soil_moisture,
  SIS100,
  latitude,
  altitude,
  ditched,
  thinned,
  last_thinned
)
}
\arguments{
\item{vegetation}{Vegetation type according to follows Swedish National forest inventory FALTSKIKT:
\tabular{ll}{
Code \tab Vegetation \cr
1 \tab  Rich-herb without shrubs \cr
2 \tab Rich-herb with shrubs/bilberry \cr
3 \tab Rich-herb with shrubs/lingonberry \cr
4 \tab Low-herb without shrubs \cr
5 \tab Low-herb with shrubs/bilberry \cr
6 \tab Low-herb with shrubs/lingonberry \cr
7 \tab No field layer \cr
8 \tab Broadleaved grass \cr
9 \tab Thinleaved grass \cr
10 \tab Sedge, high \cr
11 \tab Sedge, low \cr
12 \tab Horsetail, Equisetum ssp. \cr
13 \tab Bilberry \cr
14 \tab Lingonberry \cr
15 \tab Crowberry \cr
16 \tab Poor shrub \cr
17 \tab Lichen, frequent occurrence \cr
18 \tab Lichen, dominating \cr
}}

\item{stand_age}{Age at start of period}

\item{Basal_area_Conifer_m2_ha}{Conifer basal area}

\item{Basal_area_Pine_m2_ha}{Pine basal area}

\item{Basal_area_Birch_m2_ha}{Birch basal area}

\item{Basal_area_after_thinning}{Basal area m2 per ha after thinning.}

\item{Basal_area_before_thinning}{Basal area m2 per ha before thinning.}

\item{Basal_area_stand}{Basal area in the surrounding stand, m2/ha.}

\item{stems_after_thinning}{Stems per hectare after thinning.}

\item{peatland}{1 if plot is Peatland, 0 for others (default).}

\item{soil_moisture}{Type 1="Dry/torr",2="Mesic/frisk",3="Mesic-moist/frisk-fuktig",4="Moist/fuktig",5="Wet/Blöt"}

\item{SIS100}{Site Index as measured from [forester::SIS_estimate()]}

\item{latitude}{degrees latitude}

\item{altitude}{metres above sea level}

\item{ditched}{True or False; if there is a ditch within 25 m of the plot centre.}

\item{thinned}{1 if the plot has been thinned, otherwise 0.}

\item{last_thinned}{Number of growth seasons since the stand was last thinned.}
}
\value{
Basal Area.
}
\description{
This relates to the function given in text in the source: the calibrated function BEY2 suggested for application in Heureka.
}
