% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUGIN_UNGSKOG.R
\name{Elfving_2010_NYSKOG_young_stand_height_development}
\alias{Elfving_2010_NYSKOG_young_stand_height_development}
\title{Height development in young stands NYSKOG}
\usage{
Elfving_2010_NYSKOG_young_stand_height_development(
  species,
  H100_Spruce,
  H100_Pine,
  total_age
)
}
\arguments{
\item{species}{One of : Pinus sylvestris, Picea abies, Betula pendula, Betula pubescens, Populus tremula, Fagus sylvatica, Quercus robur, Pinus contorta.}

\item{H100_Spruce}{Site index H100 [forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce()],[forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce()]}

\item{H100_Pine}{Site index H100 [forester::Hagglund_1974_Sweden_height_trajectories_Pine()]}

\item{total_age}{Total age of stand.}
}
\value{
Estimated height, m.
}
\description{
A group of functions for mean height development in young stands of scots pine, norway spruce, birch (B. pubescens), birch (B. pendula), aspen, beech, oak or pinus contorta.
}
