% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUGIN_UNGSKOG.R
\name{Elfving_2010_regeneration_stocking}
\alias{Elfving_2010_regeneration_stocking}
\title{Stocking level in cultivated or natural regenerations}
\source{
Elfving, B. (1992). Återväxtens etablering och utveckling till röjningstidspunkten. SLU, Institutionen för Skogsskötsel. Arbetsrapport nr. 67.
Available, p.5 : \url{https://www.heurekaslu.se/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf}

Elfving, B. 1982. Hugins ungskogstaxering 1976-1979. SLU, Projekt HUGIN. Rapport 27.

Nyström, K. & Söderberg, U. 1987. Tillväxtberäkningen för ungskog i Hugin-systemet. SLU. Inst. f. skogsskötsel. Arbetsrapport 18.
}
\usage{
Elfving_2010_regeneration_stocking(
  altitude,
  latitude,
  age,
  regeneration_method,
  proportion_cultivated,
  N_full,
  ground_water,
  county = "county",
  scarification,
  burnt,
  no_treatment,
  uncleaned,
  number_seed_trees,
  H100_Spruce,
  area,
  square_spacing,
  soil_moisture
)
}
\arguments{
\item{altitude}{Meters above sea level.}

\item{latitude}{Decimal latitude.}

\item{age}{age, number of vegetation periods between cut and inventory.}

\item{regeneration_method}{"natural" or "cultivated-sown" or "cultivated-planted"}

\item{proportion_cultivated}{percent cultivated, for partly planted sites.}

\item{N_full}{number seedlings demanded for full stocking. In thousands?}

\item{scarification}{1 if scarified, 0 if no treatment.}

\item{burnt}{1 if prescribed burning, 0 if no treatment.}

\item{no_treatment}{1 if no measures were taken to promote regeneration.}

\item{uncleaned}{1 if bushes and small trees remain after clear-cut.}

\item{number_seed_trees}{number of seed trees per hectare}

\item{H100_Spruce}{e.g. [forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce],
[forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce],
[forester::Hagglund_1974_Sweden_height_trajectories_Pine]}

\item{area}{size of regeneration area in hectares.}

\item{square_spacing}{Average square spacing used for sowing/planting}

\item{soil_moisture}{Soil moisture code [forester::Sweden_soil_types('moisture')]}

\item{main_species}{Main species "Picea abies" or "Pinus sylvestris".}
}
\value{
A vector of SLH: Stocking quality, between 0 and 1, and W, a measure of the young stand quality between 0 and 1.
}
\description{
OBSERVE! Plausible retransformation error?
Distance from equator calculated as 111 km * latitude.
Jämtland county has been interpreted as both Jämtland - Härjedalens landskap and Jämtland - Jämtlands landskap.

OBSERVE! This function includes the modifications to coefficients as defined on p. 6 in the source:
\emph{Increased efficiency in provenance selection and scarification has also been considered by assumed influence on the cofficients. For cultivations the negative effect by increasing height above see level has been reduced from -0.0514 to -0.0257 and the positive effect by scarification has been increased from 0.0757 to 0.2.}
}
\details{
\strong{Summary statistics for natural regenerations}
\itemize{
\item n = 2462.
\item R^2 = 0.492
\item s-res = 0.475
\item mean of the dependents: 1.8585
}

\strong{Summary statistics for cultivated or sown stands}
\itemize{
\item n = 3528
\item R^2 = 0.492
\item s-res = 0.391
\item mean of the dependents: 2.3059
}
}
