% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfving_ba_stand_level_2005.R
\name{elfving_ba_stand_level_2005}
\alias{elfving_ba_stand_level_2005}
\title{Stand Basal Area Development Elfving 2005}
\usage{
elfving_ba_stand_level_2005(
  vegetation,
  age,
  conifer_ba,
  pine_ba,
  birch_ba,
  ba_followed_trees,
  ba_total,
  stem_number_followed,
  peat_vegetation,
  soil_moisture,
  sis,
  latitude,
  altitude,
  ditch_within_25,
  edge_to_open_area_within_20,
  divided_plot,
  years_since_last_thinned,
  ba_surrounding_stand,
  establishment_year
)
}
\arguments{
\item{vegetation}{Variable indicating vegetation type scaled from -5 to +4 as follows:

\tabular{llr}{
Field Layer Code (NFI) \tab Description \tab Index \cr
1\tab Tall herbs w/o dwarf shrubs \tab 4 \cr
2\tab Tall herbs with bilberry \tab 2.5 \cr
3\tab Tall herbs with cowberry \tab 2 \cr
4\tab Low herbs w/o dwarf shrubs \tab 3 \cr
5\tab Low herbs with bilberry \tab 2.5 \cr
6\tab Low herbs with cowberry \tab 2 \cr
7\tab No field layer \tab 3 \cr
8\tab broad-leafed grasses \tab 2.5 \cr
9\tab narrow-leaved grasses \tab 1.5 \cr
10\tab Sedge, tall \tab -3 \cr
11\tab Sedge, low \tab -3 \cr
12\tab Horsetail \tab 1 \cr
13\tab Bilberry \tab 0 \cr
14\tab Cowberry \tab -0.5 \cr
15\tab Crowberry \tab -3 \cr
16\tab Poor shrub \tab -5 \cr
17\tab Lichen-rich \tab -0.5 \cr
18\tab Lichen-dominated \tab -1 \cr
}}

\item{age}{Age at start of period}

\item{conifer_ba}{Conifer basal area}

\item{pine_ba}{Pine basal area}

\item{birch_ba}{Birch basal area}

\item{ba_followed_trees}{basal area per ha for followed trees}

\item{ba_total}{total basal area per ha at start of period}

\item{stem_number_followed}{stem number per ha for followed trees}

\item{peat_vegetation}{True or False; if there is peat on the plot.}

\item{soil_moisture}{Soil moisture code (Dry/mesic/mesic-moist/moist,wet)}

\item{sis}{site index according to site factors for site-indicative species, mean between measurement 1 and 2.}

\item{latitude}{degrees latitude}

\item{altitude}{metres above sea level}

\item{ditch_within_25}{True or False; if there is a ditch within 25 m of the plot centre.}

\item{edge_to_open_area_within_20}{True or False; if there is a stand edge towards and open area within 20 m of the plot centre.}

\item{divided_plot}{True or False; if the plot is divided.}

\item{years_since_last_thinned}{Years since the plot was last thinned. If never, 0}

\item{ba_surrounding_stand}{Basal area of the surrounding stand}

\item{establishment_year}{Year established.}
}
\value{

}
\description{
Description available: \url{https://www.heurekaslu.se/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf} page 50.
}
