% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_Hagglund_1975.R
\name{Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Pine}
\alias{Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Pine}
\title{Basal Area per hectare for young Pine stands in northern Sweden from Elfving & Hagglund 1975.}
\source{
Elfving, B., Hägglund, B. (1975) Utgångslägen för produktionsprognoser: Tall och gran i Sverige.
/ Initial stands for yield forecasts: Scots pine and Norway spruce in Sweden. Research Notes #38. Dept. of
Forest Yield Research. Royal College of Forestry. Stockholm. p. 53. pp. 75.
}
\usage{
Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Pine(
  latitude,
  altitude,
  stand_density = 0.65,
  stems_per_ha = "function",
  SI,
  dominant_height_m,
  broadleaves_percent_of_basal_area,
  pre_commercial_thinning,
  even_or_somewhat_uneven_aged = TRUE
)
}
\arguments{
\item{altitude}{meters above sea level.}

\item{stand_density}{0-1}

\item{stems_per_ha}{number of stems per hectare.}

\item{SI}{Site index H100 m.}

\item{dominant_height_m}{Dominant height of stand.}

\item{broadleaves_percent_of_basal_area}{Percent of basal area composed of broadleaves.}
}
\value{
Basal area per hectare, m^2.
}
\description{
Basal Area per hectare for young Pine stands in northern Sweden from Elfving & Hagglund 1975.
}
\details{
F. 6.1

N = 426

Mean basal area per hectare  = 7.190

Standard deviation about the mean = 0.496

Standard deviation about the function = 0.334

R = 0.76
}
\examples{

ggplot()+
 geom_function(aes(linetype="66N 400masl 16m"), fun= function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Pine(latitude = 66,altitude = 400,
                                                                                                                                             stand_density = x, stems_per_ha = "function",SI = 16,dominant_height_m = 12.5,broadleaves_percent_of_basal_area = 0,pre_commercial_thinning = FALSE,even_or_somewhat_uneven_aged = TRUE)
 )+
 geom_function(aes(linetype="62N 200masl 24m"), fun= function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Pine(latitude = 62,altitude = 200,
                                                                                                                                             stand_density = x, stems_per_ha = "function",SI = 24,dominant_height_m = 12.5,broadleaves_percent_of_basal_area = 0,pre_commercial_thinning = FALSE,even_or_somewhat_uneven_aged = TRUE)
 )+
 xlim(c(0.4,1.0))
}
