% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_Hagglund_1975.R
\name{Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Spruce}
\alias{Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Spruce}
\title{Basal Area per hectare for young Spruce stands in northern Sweden from Elfving & Hagglund 1975.}
\source{
Elfving, B., Hägglund, B. (1975) Utgångslägen för produktionsprognoser: Tall och gran i Sverige.
/ Initial stands for yield forecasts: Scots pine and Norway spruce in Sweden. Research Notes #38. Dept. of
Forest Yield Research. Royal College of Forestry. Stockholm. p. 53. pp. 75.
}
\usage{
Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Spruce(
  altitude,
  stand_density = 0.65,
  stems_per_ha = "function",
  SI,
  broadleaves_percent_of_basal_area = 0,
  dominant_height_m,
  spatial_distribution,
  pre_commercial_thinning,
  uneven_aged = FALSE
)
}
\arguments{
\item{altitude}{meters above sea level.}

\item{stand_density}{0-1}

\item{stems_per_ha}{number of stems per hectare. or "function".}

\item{SI}{Site index H100 m}

\item{dominant_height_m}{Dominant height of stand.}

\item{spatial_distribution}{1 if even. 2 if somewhat uneven. 3 if grouped.}

\item{uneven_aged}{TRUE/FALSE if stand is uneven-aged.}
}
\value{
Basal area per hectare, m^2.
}
\description{
Basal Area per hectare for young Spruce stands in northern Sweden from Elfving & Hagglund 1975.
}
\details{
F. 6.3

N = 262

Mean basal area per hectare  = 7.411

Standard deviation about the mean = 0.452

Standard deviation about the function = 0.273

R = 0.80
}
\examples{
ggplot()+
geom_function(aes(linetype="16,500"), fun = function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Spruce(altitude = 500,stand_density = x,stems_per_ha = "function",SI = 16,broadleaves_percent_of_basal_area = 0,dominant_height_m = 12.5,spatial_distribution = 1,pre_commercial_thinning = FALSE,uneven_aged = FALSE))+
  geom_function(aes(linetype="16,200"), fun = function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Spruce(altitude = 200,stand_density = x,stems_per_ha = "function",SI = 16,broadleaves_percent_of_basal_area = 0,dominant_height_m = 12.5,spatial_distribution = 1,pre_commercial_thinning = FALSE,uneven_aged = FALSE))+
  geom_function(aes(linetype="28,200"), fun = function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Spruce(altitude = 200,stand_density = x,stems_per_ha = "function",SI = 28,broadleaves_percent_of_basal_area = 0,dominant_height_m = 12.5,spatial_distribution = 1,pre_commercial_thinning = FALSE,uneven_aged = FALSE))+
  xlim(c(0.4,1.0))+
  scale_y_continuous(limits = c(8,28),breaks=c(seq(8,28,2)))
}
