% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Pine.R
\name{Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Pine}
\alias{Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Pine}
\title{Basal Area per hectare for young Pine stands in southern Sweden from Elfving & Hagglund 1975.}
\source{
Elfving, B., Hägglund, B. (1975) Utgångslägen för produktionsprognoser: Tall och gran i Sverige.
/ Initial stands for yield forecasts: Scots pine and Norway spruce in Sweden. Research Notes #38. Dept. of
Forest Yield Research. Royal College of Forestry. Stockholm. p. 53. pp. 75.
}
\usage{
Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Pine(
  altitude,
  stand_density = 0.65,
  stems_per_ha = "function",
  SI,
  dominant_height_m,
  uneven_aged = FALSE,
  age_at_breast_height = "function",
  regeneration,
  pre_commercial_thinning
)
}
\arguments{
\item{altitude}{meters above sea level.}

\item{stand_density}{0-1}

\item{stems_per_ha}{number of stems per hectare.}

\item{SI}{Site index H100}

\item{uneven_aged}{TRUE/FALSE if stand is uneven-aged.}

\item{age_at_breast_height}{Age at breast height, years.}

\item{regeneration}{How was the stand established? One of "culture", "natural regeneration" or "unknown". Only required if age_at_breast_height=="function"}

\item{pre_commercial_thinning}{TRUE/FALSE, has pre-commercial thinning been done?}

\item{dominant_height}{Dominant height of stand.}
}
\value{
Basal area per hectare, m^2.
}
\description{
Basal Area per hectare for young Pine stands in southern Sweden from Elfving & Hagglund 1975.
}
\details{
F. 6.2

N = 235

Mean basal area per hectare  = 7.465

Standard deviation about the mean = 0.440

Standard deviation about the function = 0.280

R = 0.71
}
\examples{

ggplot()+
geom_function(aes(linetype="h100=16"), fun= function (x) Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Pine(altitude = 300,
                                                                                                                                     stand_density = x,
                                                                                                                                     stems_per_ha = "function" ,
                                                                                                                                     SI = 16 ,
                                                                                                                                     dominant_height_m = 12.5 ,
                                                                                                                                     uneven_aged = FALSE,
                                                                                                                                     age_at_breast_height = "function" ,
                                                                                                                                     regeneration = "culture",
                                                                                                                                     pre_commercial_thinning = FALSE))+
  geom_function(aes(linetype="h100=24"), fun= function (x) Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Pine(altitude = 300,
                                                                                                                                       stand_density = x,
                                                                                                                                       stems_per_ha = "function" ,
                                                                                                                                       SI = 24 ,
                                                                                                                                       dominant_height_m = 12.5 ,
                                                                                                                                       uneven_aged = FALSE,
                                                                                                                                       age_at_breast_height = "function" ,
                                                                                                                                       regeneration = "culture",
                                                                                                                                       pre_commercial_thinning = FALSE))+
  xlim(c(0.4,1))+
  scale_y_continuous(limits=c(8,24),breaks=seq(8,24,2))

}
