% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_Hagglund_1975.R
\name{Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Spruce}
\alias{Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Spruce}
\title{Basal Area per hectare for young Spruce stands in northern Sweden from Elfving & Hagglund 1975.}
\source{
Elfving, B., Hägglund, B. (1975) Utgångslägen för produktionsprognoser: Tall och gran i Sverige.
/ Initial stands for yield forecasts: Scots pine and Norway spruce in Sweden. Research Notes #38. Dept. of
Forest Yield Research. Royal College of Forestry. Stockholm. p. 53. pp. 75.
}
\usage{
Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Spruce(
  altitude,
  stems_per_ha = "function",
  dominant_height_m,
  broadleaves_percent_of_basal_area,
  spatial_distribution,
  pre_commercial_thinning,
  stand_density,
  SI,
  age_at_breast_height = "function",
  even_or_somewhat_uneven_aged,
  regeneration = "culture"
)
}
\arguments{
\item{altitude}{meters above sea level.}

\item{stems_per_ha}{number of stems per hectare.}

\item{dominant_height_m}{Dominant height of stand.}

\item{broadleaves_percent_of_basal_area}{Percent of basal area composed of broadleaves.}

\item{spatial_distribution}{1 if even. 2 if somewhat uneven. 3 if grouped.}

\item{pre_commercial_thinning}{TRUE if pre_commercial thinning has taken place.}

\item{stand_density}{0-1. Required only if stems_per_ha or age_at_breast_height is "function".}

\item{SI}{Site index H100 m. Required only if stems_per_ha or age_at_breast_height is "function".}

\item{age_at_breast_height}{Age of stand at breast height. Required only if stems_per_ha is "function".}

\item{even_or_somewhat_uneven_aged}{TRUE if even_aged or slightly uneven-aged.}

\item{regeneration}{One of "culture", "natural regeneration" or "unknown.". Required only if age_to_breast_height is "function".}
}
\value{
Basal area per hectare, m^2.
}
\description{
Basal Area per hectare for young Spruce stands in northern Sweden from Elfving & Hagglund 1975.
}
\details{
F. 6.4

N = 183

Mean basal area per hectare  = 7.531

Standard deviation about the mean = 0.470

Standard deviation about the function = 0.306

R = 0.74
}
\examples{

ggplot()+
geom_function(aes(linetype="G20"), fun = function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Spruce(altitude = 300,
                                                                                                                                   stems_per_ha = "function" ,
                                                                                                                                   dominant_height_m = 12.5 ,
                                                                                                                                   broadleaves_percent_of_basal_area = 0,
                                                                                                                                   spatial_distribution = 1 ,
                                                                                                                                   pre_commercial_thinning = FALSE,
                                                                                                                                   stand_density = x,
                                                                                                                                   SI = 20,
                                                                                                                                   age_at_breast_height = "function" ,
                                                                                                                                   even_or_somewhat_uneven_aged = TRUE ,
                                                                                                                                   regeneration = "culture"))+
  geom_function(aes(linetype="G36"), fun = function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Spruce(altitude = 300,
                                                                                                                                     stems_per_ha = "function" ,
                                                                                                                                     dominant_height_m = 12.5 ,
                                                                                                                                     broadleaves_percent_of_basal_area = 0,
                                                                                                                                     spatial_distribution = 1 ,
                                                                                                                                     pre_commercial_thinning = FALSE,
                                                                                                                                     stand_density = x,
                                                                                                                                     SI = 36,
                                                                                                                                     age_at_breast_height = "function" ,
                                                                                                                                     even_or_somewhat_uneven_aged = TRUE ,
                                                                                                                                     regeneration = "culture"))+
  xlim(c(0.4,1))+
  scale_y_continuous(limits=c(8,28), breaks = seq(8,28,2))
}
