% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_Hagglund_1975.R
\name{Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Pine}
\alias{Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Pine}
\title{Stems per hectare for young stands of Scots Pine in northern Sweden.}
\source{
Elfving, B., Hägglund, B. (1975) Utgångslägen för produktionsprognoser: Tall och gran i Sverige.
/ Initial stands for yield forecasts: Scots pine and Norway spruce in Sweden. Research Notes #38. Dept. of
Forest Yield Research. Royal College of Forestry. Stockholm. p. 42. pp. 75.
}
\usage{
Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Pine(
  latitude,
  altitude,
  stand_density = 0.65,
  dominant_height_m,
  pre_commercial_thinning,
  even_or_somewhat_uneven_aged,
  uneven_aged
)
}
\arguments{
\item{latitude}{Latitude, degrees N.}

\item{altitude}{Altitude, meters above sea level.}

\item{stand_density}{0-1.}

\item{dominant_height_m}{Dominant height of stand, in metres.}

\item{pre_commercial_thinning}{Has the stand been pre-commercially thinned?}

\item{even_or_somewhat_uneven_aged}{TRUE/FALSE}

\item{uneven_aged}{TRUE/FALSE}
}
\value{
Stems thicker than 2.5 cm per ha.
}
\description{
For stems thicker than 2.5 cm at breast height.
}
\details{
Function 5.1.


Number of sites = 426.

Mean stems per ha: 7.2151

Standard deviation mean stems per ha: 0.571


standard deviation about the function: 0.394

R = 0.73

Cp = 12. (Daniel & Wood 1971, pp. 86)
}
