% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_Hagglund_1975.R
\name{Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Spruce}
\alias{Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Spruce}
\title{Stems per hectare for young stands of norway Spruce in northern Sweden.}
\source{
Elfving, B., Hägglund, B. (1975) Utgångslägen för produktionsprognoser: Tall och gran i Sverige.
/ Initial stands for yield forecasts: Scots pine and Norway spruce in Sweden. Research Notes #38. Dept. of
Forest Yield Research. Royal College of Forestry. Stockholm. p. 42. pp. 75.
}
\usage{
Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Spruce(
  stand_density = 0.65,
  pre_commercial_thinning,
  SI,
  altitude,
  broadleaves_percent_of_basal_area,
  even_or_somewhat_uneven_aged,
  uneven_aged
)
}
\arguments{
\item{stand_density}{0-1.}

\item{pre_commercial_thinning}{Has the stand been pre-commercially thinned?}

\item{SI}{Site index H 100.}

\item{altitude}{meters above sea level.}

\item{broadleaves_percent_of_basal_area}{Percent of basal area composed of broadleaves.}

\item{even_or_somewhat_uneven_aged}{TRUE / FALSE}

\item{uneven_aged}{TRUE/FALSE}
}
\value{
Stems thicker than 2.5 cm per ha.
}
\description{
For stems thicker than 2.5 cm at breast height.
}
\details{
Function 5.3.


Number of sites = 262.

Mean stems per ha: 7.6012

Standard deviation mean stems per ha: 0.533


standard deviation about the function: 0.416

R = 0.64

Cp = 5 (Daniel & Wood 1971, pp. 86)
}
\examples{

ggplot()+
geom_function(aes(linetype=">22, 200masl"), fun= function(x)
 Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Spruce(stand_density = x,
                                                                          pre_commercial_thinning = FALSE,
                                                                         SI = 23,
                                                                         altitude = 200,
                                                                         broadleaves_percent_of_basal_area = 0,
                                                                         even_or_somewhat_uneven_aged = TRUE,
                                                                          uneven_aged = FALSE) )+
 geom_function(aes(linetype="<22, 500masl"), fun= function(x)
   Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Spruce(stand_density = x,
                                                                            pre_commercial_thinning = FALSE,
                                                                            SI = 21,
                                                                            altitude = 500,
                                                                            broadleaves_percent_of_basal_area = 0,
                                                                            even_or_somewhat_uneven_aged = TRUE,
                                                                            uneven_aged = FALSE) )+
 xlim(c(0.4,0.9))+
 geom_function(aes(linetype="h<22, 200masl"), fun= function (x)
   Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Spruce(stand_density = x,
                                                                            pre_commercial_thinning = FALSE,
                                                                           SI = 21,
                                                                           altitude = 200,
                                                                           broadleaves_percent_of_basal_area = 0,
                                                                            even_or_somewhat_uneven_aged = TRUE,
                                                                           uneven_aged = FALSE ))



}
