% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Elfving_Hagglund_1975_initial_stems_young_forests_southern_Sweden_Pine.R
\name{Elfving_Hagglund_1975_initial_stems_young_forests_southern_Sweden_Pine}
\alias{Elfving_Hagglund_1975_initial_stems_young_forests_southern_Sweden_Pine}
\title{Stems per hectare for young stands of Scots Pine in southern Sweden.}
\source{
Elfving, B., Hägglund, B. (1975) Utgångslägen för produktionsprognoser: Tall och gran i Sverige.
/ Initial stands for yield forecasts: Scots pine and Norway spruce in Sweden. Research Notes #38. Dept. of
Forest Yield Research. Royal College of Forestry. Stockholm. p. 42. pp. 75.
}
\usage{
Elfving_Hagglund_1975_initial_stems_young_forests_southern_Sweden_Pine(
  stand_density = 0.65,
  pre_commercial_thinning,
  age_at_breast_height = "function",
  latitude = 60,
  SI,
  dominant_height_m = 12.5,
  regeneration = "culture"
)
}
\arguments{
\item{stand_density}{0-1.}

\item{pre_commercial_thinning}{Has the stand been pre-commercially thinned?}

\item{age_at_breast_height}{Stand age at breast height (1.3 m). For 'function', see [forester::Hagglund_age_to_height()]}

\item{latitude}{Latitude in degrees. Only Required if "age_at_breast_height" is "function".}

\item{SI}{SI H100 m. Only Required if "age_at_breast_height" is "function".}

\item{dominant_height_m}{Dominant height of stand. Defaults to 12.5 m. Only Required if "age_at_breast_height" is "function".}

\item{regeneration}{Method of stand establishment. One of "culture","natural regeneration" or "unknown". Only Required if "age_at_breast_height" is "function".}
}
\value{
Stems thicker than 2.5 cm per ha.
}
\description{
For stems thicker than 2.5 cm at breast height.
}
\details{
Function 5.2.


Number of sites = 235.

Mean stems per ha: 7.3192

Standard deviation mean stems per ha: 0.523


standard deviation about the function: 0.371

R = 0.72

Cp = 7. (Daniel & Wood 1971, pp. 86)
}
