% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bengtsson_1978_mortality.R
\name{Elfving_HUGIN_annual_BA_mortality_percent}
\alias{Elfving_HUGIN_annual_BA_mortality_percent}
\title{Annual Basal Area Mortality}
\source{
Elfving, Björn. 2014-03-17. PM. Modellering av naturlig avgång i Heureka. (Swedish). URL:\url{https://www.heurekaslu.se/w/images/f/f4/HeurekaMortality-PM140317.pdf}

Bengtsson, G. 1981-02-26. Stencil. Beräkning av den naturliga avgången ur virkesförrådet i Hugin-systemet. 8 sid. Umeå: Sveriges lantbruksuniversitet, inst. f. skogstaxering.
}
\usage{
Elfving_HUGIN_annual_BA_mortality_percent(
  northern_Sweden = TRUE,
  species = "Pinus sylvestris",
  DAge = 20
)
}
\arguments{
\item{northern_Sweden}{Logical. Site located > 60 deg. N?}

\item{species}{One of 'Pinus sylvestris','Picea abies','Betula pendula','Betula pubescens', or other deciduous tree species.}

\item{DAge}{Basal Area Weighted Mean Age, not including tree standards.}
}
\value{
Annual Mortality in Percent of Basal Area.
}
\description{
Annual Mortality in terms of percent of the basal area at start of period.
}
\details{
N.B. At application it at first glance looks as if from ProdMod2 the
mean diameter of the outgoing trees should be set to the QMD of the stand at
the start of the period.
}
