% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Elfving_Kiviste_1997_height_trajectory_Sweden_Pine.R
\name{Elfving_Kiviste_1997_height_trajectory_Sweden_Pine}
\alias{Elfving_Kiviste_1997_height_trajectory_Sweden_Pine}
\title{Height trajectory for Scots Pine in Sweden from Elfving, Kiviste 1997.}
\source{
Elfving, B., Kiviste, A. (1997) Construction of site index equations for
Pinus sylvestris L. using permanent plot data in Sweden. For. Ecol. Manage. Vol. 98. Issue 2. pp. 125-134.
Available: \url{https://doi.org/10.1016/S0378-1127(97)00077-7}
}
\usage{
Elfving_Kiviste_1997_height_trajectory_Sweden_Pine(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Total age.}

\item{age2}{Total age at output age.}

\item{output}{One of "SIH100","Equation" or "Height".}
}
\value{

}
\description{
156 Scots pine stands, whereof most were long term experiments maintained by the
Dept. of Forest Yield Research.

Suitable for Scots Pine of cultivated origin between ages 10 - 80.

RMSE = 0.401.
}
