% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_Hagglund_1975.R
\name{ElfvingHagglund1975}
\alias{ElfvingHagglund1975}
\alias{Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Pine}
\alias{Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Spruce}
\alias{Elfving_Hagglund_1975_initial_stems_young_forests_southern_Sweden_Pine}
\alias{Elfving_Hagglund_1975_initial_stems_young_forests_southern_Sweden_Spruce}
\alias{Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Pine}
\alias{Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Spruce}
\alias{Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Pine}
\alias{Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Spruce}
\title{Stems per hectare for young stands of Scots Pine in northern Sweden.}
\source{
Elfving, B., Hägglund, B. (1975) Utgångslägen för produktionsprognoser: Tall och gran i Sverige.
/ Initial stands for yield forecasts: Scots pine and Norway spruce in Sweden. Research Notes #38. Dept. of
Forest Yield Research. Royal College of Forestry. Stockholm. p. 53. pp. 75.
}
\usage{
Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Pine(
  latitude,
  altitude,
  stand_density = 0.65,
  dominant_height,
  PCT,
  even_or_somewhat_uneven_aged,
  uneven_aged
)

Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Spruce(
  stand_density = 0.65,
  PCT,
  SI,
  altitude,
  broadleaves_percent_of_basal_area,
  even_or_somewhat_uneven_aged,
  uneven_aged
)

Elfving_Hagglund_1975_initial_stems_young_forests_southern_Sweden_Pine(
  stand_density = 0.65,
  PCT,
  age_at_breast_height = "function",
  latitude = 60,
  SI,
  dominant_height = 12.5,
  regeneration = "culture"
)

Elfving_Hagglund_1975_initial_stems_young_forests_southern_Sweden_Spruce(
  stand_density = 0.65,
  PCT,
  SI,
  age_at_breast_height,
  altitude,
  broadleaves_percent_of_basal_area,
  even_or_somewhat_uneven_aged
)

Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Pine(
  latitude,
  altitude,
  stand_density = 0.65,
  stems = "function",
  SI,
  dominant_height,
  broadleaves_percent_of_basal_area,
  PCT,
  even_or_somewhat_uneven_aged = TRUE
)

Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Spruce(
  altitude,
  stand_density = 0.65,
  stems = "function",
  SI,
  broadleaves_percent_of_basal_area = 0,
  dominant_height,
  spatial_distribution,
  PCT,
  uneven_aged = FALSE
)

Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Pine(
  altitude,
  stand_density = 0.65,
  stems = "function",
  SI,
  dominant_height,
  uneven_aged = FALSE,
  age_at_breast_height = "function",
  regeneration,
  PCT
)

Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Spruce(
  altitude,
  stems = "function",
  dominant_height,
  broadleaves_percent_of_basal_area,
  spatial_distribution,
  PCT,
  stand_density,
  SI,
  age_at_breast_height = "function",
  even_or_somewhat_uneven_aged,
  regeneration = "culture"
)
}
\arguments{
\item{latitude}{Latitude, degrees N.}

\item{altitude}{meters above sea level.}

\item{stand_density}{0-1}

\item{dominant_height}{Dominant height of stand, meters.}

\item{PCT}{1 if Pre-commercial thinning has occurred, otherwise 0.}

\item{even_or_somewhat_uneven_aged}{TRUE/FALSE if stand is even, or somewhat uneven-aged.}

\item{uneven_aged}{TRUE/FALSE if stand is uneven-aged.}

\item{SI}{Site index H100 m.}

\item{broadleaves_percent_of_basal_area}{Percent of basal area composed of broadleaves.}

\item{age_at_breast_height}{Age at breast height, years.}

\item{regeneration}{How was the stand established? One of "culture", "natural regeneration" or "unknown". Only required if age_at_breast_height=="function"}

\item{stems}{Number of Stems / ha.}

\item{spatial_distribution}{1 if even. 2 if somewhat uneven. 3 if grouped.}
}
\value{
Number of Stems thicker than 2.5 cm per ha.

Stems thicker than 2.5 cm per ha.

Stems thicker than 2.5 cm per ha.

Stems thicker than 2.5 cm per ha.

Basal area per hectare, m^2.

Basal area per hectare, m^2.

Basal area per hectare, m^2.

Basal area per hectare, m^2.
}
\description{
Initial stand conditions for Norway Spruce and Scots Pine in Sweden.
}
\details{
*OBSERVE*
**For stems thicker than 2.5 cm at breast height.**

Function 5.1.


Number of sites = 426.

Mean stems per ha: 7.2151

Standard deviation mean stems per ha: 0.571


standard deviation about the function: 0.394

R = 0.73

Cp = 12. (Daniel & Wood 1971, pp. 86)

*OBSERVE*
**For stems thicker than 2.5 cm at breast height.**

Function 5.3.


Number of sites = 262.

Mean stems per ha: 7.6012

Standard deviation mean stems per ha: 0.533


standard deviation about the function: 0.416

R = 0.64

Cp = 5 (Daniel & Wood 1971, pp. 86)

*OBSERVE*
**For stems thicker than 2.5 cm at breast height.**

Function 5.2.


Number of sites = 235.

Mean stems per ha: 7.3192

Standard deviation mean stems per ha: 0.523


standard deviation about the function: 0.371

R = 0.72

Cp = 7. (Daniel & Wood 1971, pp. 86)

*OBSERVE*
**For stems thicker than 2.5 cm at breast height.**

Function 5.4.


Number of sites = 183.

Mean stems per ha: 7.6675

Standard deviation mean stems per ha: 0.501


standard deviation about the function: 0.381

R = 0.67

Cp = 5 (Daniel & Wood 1971, pp. 86)

F. 6.1

N = 426

Mean basal area per hectare  = 7.190

Standard deviation about the mean = 0.496

Standard deviation about the function = 0.334

R = 0.76

F. 6.3

N = 262

Mean basal area per hectare  = 7.411

Standard deviation about the mean = 0.452

Standard deviation about the function = 0.273

R = 0.80

F. 6.2

N = 235

Mean basal area per hectare  = 7.465

Standard deviation about the mean = 0.440

Standard deviation about the function = 0.280

R = 0.71

F. 6.4

N = 183

Mean basal area per hectare  = 7.531

Standard deviation about the mean = 0.470

Standard deviation about the function = 0.306

R = 0.74
}
\examples{

ggplot()+
geom_function(aes(linetype=">22, 200masl"), fun= function(x)
 Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Spruce(stand_density = x,
                                                                          PCT = FALSE,
                                                                         SI = 23,
                                                                         altitude = 200,
                                                                         broadleaves_percent_of_basal_area = 0,
                                                                         even_or_somewhat_uneven_aged = TRUE,
                                                                          uneven_aged = FALSE) )+
 geom_function(aes(linetype="<22, 500masl"), fun= function(x)
   Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Spruce(stand_density = x,
                                                                            PCT = FALSE,
                                                                            SI = 21,
                                                                            altitude = 500,
                                                                            broadleaves_percent_of_basal_area = 0,
                                                                            even_or_somewhat_uneven_aged = TRUE,
                                                                            uneven_aged = FALSE) )+
 xlim(c(0.4,0.9))+
 geom_function(aes(linetype="h<22, 200masl"), fun= function (x)
   Elfving_Hagglund_1975_initial_stems_young_forests_northern_Sweden_Spruce(stand_density = x,
                                                                            PCT = FALSE,
                                                                           SI = 21,
                                                                           altitude = 200,
                                                                           broadleaves_percent_of_basal_area = 0,
                                                                            even_or_somewhat_uneven_aged = TRUE,
                                                                           uneven_aged = FALSE ))



ggplot()+
geom_function(aes(linetype="G20"), fun= function(x) Elfving_Hagglund_1975_initial_stems_young_forests_southern_Sweden_Spruce(
 stand_density = x,PCT = FALSE,SI = 20,age_at_breast_height = 44.6,altitude=100,broadleaves_percent_of_basal_area = 0,even_or_somewhat_uneven_aged = TRUE
) )+
 geom_function(aes(linetype="G32"), fun= function(x) Elfving_Hagglund_1975_initial_stems_young_forests_southern_Sweden_Spruce(
   stand_density = x,PCT = FALSE,SI = 32,age_at_breast_height = 21,altitude=100,broadleaves_percent_of_basal_area = 0,even_or_somewhat_uneven_aged = TRUE
 ) )+
xlim(c(0.4,1.0))


ggplot()+
 geom_function(aes(linetype="66N 400masl 16m"), fun= function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Pine(latitude = 66,altitude = 400,
                                                                                                                                             stand_density = x, stems = "function",SI = 16,dominant_height = 12.5,broadleaves_percent_of_basal_area = 0,PCT = FALSE,even_or_somewhat_uneven_aged = TRUE)
 )+
 geom_function(aes(linetype="62N 200masl 24m"), fun= function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Pine(latitude = 62,altitude = 200,
                                                                                                                                             stand_density = x, stems = "function",SI = 24,dominant_height = 12.5,broadleaves_percent_of_basal_area = 0,PCT = FALSE,even_or_somewhat_uneven_aged = TRUE)
 )+
 xlim(c(0.4,1.0))
ggplot()+
geom_function(aes(linetype="16,500"), fun = function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Spruce(altitude = 500,stand_density = x,stems = "function",SI = 16,broadleaves_percent_of_basal_area = 0,dominant_height = 12.5,spatial_distribution = 1,PCT = FALSE,uneven_aged = FALSE))+
  geom_function(aes(linetype="16,200"), fun = function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Spruce(altitude = 200,stand_density = x,stems = "function",SI = 16,broadleaves_percent_of_basal_area = 0,dominant_height = 12.5,spatial_distribution = 1,PCT = FALSE,uneven_aged = FALSE))+
  geom_function(aes(linetype="28,200"), fun = function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_northern_Sweden_Spruce(altitude = 200,stand_density = x,stems = "function",SI = 28,broadleaves_percent_of_basal_area = 0,dominant_height = 12.5,spatial_distribution = 1,PCT = FALSE,uneven_aged = FALSE))+
  xlim(c(0.4,1.0))+
  scale_y_continuous(limits = c(8,28),breaks=c(seq(8,28,2)))

ggplot()+
geom_function(aes(linetype="h100=16"), fun= function (x) Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Pine(altitude = 300,
                                                                                                                                     stand_density = x,
                                                                                                                                     stems = "function" ,
                                                                                                                                     SI = 16 ,
                                                                                                                                     dominant_height = 12.5 ,
                                                                                                                                     uneven_aged = FALSE,
                                                                                                                                     age_at_breast_height = "function" ,
                                                                                                                                     regeneration = "culture",
                                                                                                                                     PCT = FALSE))+
  geom_function(aes(linetype="h100=24"), fun= function (x) Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Pine(altitude = 300,
                                                                                                                                       stand_density = x,
                                                                                                                                       stems = "function" ,
                                                                                                                                       SI = 24 ,
                                                                                                                                       dominant_height = 12.5 ,
                                                                                                                                       uneven_aged = FALSE,
                                                                                                                                       age_at_breast_height = "function" ,
                                                                                                                                       regeneration = "culture",
                                                                                                                                       PCT = FALSE))+
  xlim(c(0.4,1))+
  scale_y_continuous(limits=c(8,24),breaks=seq(8,24,2))


ggplot()+
geom_function(aes(linetype="G20"), fun = function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Spruce(altitude = 300,
                                                                                                                                   stems = "function" ,
                                                                                                                                   dominant_height = 12.5 ,
                                                                                                                                   broadleaves_percent_of_basal_area = 0,
                                                                                                                                   spatial_distribution = 1 ,
                                                                                                                                   PCT = FALSE,
                                                                                                                                   stand_density = x,
                                                                                                                                   SI = 20,
                                                                                                                                   age_at_breast_height = "function" ,
                                                                                                                                   even_or_somewhat_uneven_aged = TRUE ,
                                                                                                                                   regeneration = "culture"))+
  geom_function(aes(linetype="G36"), fun = function(x) Elfving_Hagglund_1975_initial_basal_area_young_forests_southern_Sweden_Spruce(altitude = 300,
                                                                                                                                     stems = "function" ,
                                                                                                                                     dominant_height = 12.5 ,
                                                                                                                                     broadleaves_percent_of_basal_area = 0,
                                                                                                                                     spatial_distribution = 1 ,
                                                                                                                                     PCT = FALSE,
                                                                                                                                     stand_density = x,
                                                                                                                                     SI = 36,
                                                                                                                                     age_at_breast_height = "function" ,
                                                                                                                                     even_or_somewhat_uneven_aged = TRUE ,
                                                                                                                                     regeneration = "culture"))+
  xlim(c(0.4,1))+
  scale_y_continuous(limits=c(8,28), breaks = seq(8,28,2))
}
