% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_2004_Individual_Based_Tree_Model_NFI.R
\name{ElfvingIBM}
\alias{ElfvingIBM}
\alias{Elfving_2004_BA_increment_Aspen}
\alias{Elfving_2004_BA_increment_Beech}
\alias{Elfving_2004_BA_increment_Birch}
\alias{Elfving_2004_BA_increment_Noble}
\alias{Elfving_2004_BA_increment_Oak}
\alias{Elfving_2004_BA_increment_Pine}
\alias{Elfving_2004_BA_increment_Spruce}
\alias{Elfving_2004_BA_increment_Trivial}
\title{Basal Area Increment function for Aspen}
\source{
[Elfving, B. (2010) Translated, re-formulated Pro-Memoria for HEUREKA based on Manuscript 2004-01-26. 'Individual-tree basal area growth functions for all Swedish forests'.](https://www.heurekaslu.se/w/images/9/93/Heureka_prognossystem_\\%28Elfving_rapportutkast\\%29.pdf)
}
\usage{
Elfving_2004_BA_increment_Aspen(
  diameter_cm,
  Basal_area_weighted_mean_diameter_cm,
  distance_to_coast_km,
  BA_sum_of_trees_with_larger_diameter,
  BA_Aspen,
  Basal_area_plot,
  Basal_area_stand,
  computed_tree_age,
  latitude,
  altitude,
  vegetation,
  uneven_aged,
  fertilised = 0,
  thinned,
  last_thinned
)

Elfving_2004_BA_increment_Beech(
  diameter_cm,
  Basal_area_weighted_mean_diameter_cm,
  distance_to_coast_km,
  BA_sum_of_trees_with_larger_diameter,
  BA_Beech,
  Basal_area_plot,
  Basal_area_stand,
  computed_tree_age,
  latitude,
  altitude,
  vegetation,
  SIS,
  divided_plot = 0,
  edge_effect,
  uneven_aged,
  fertilised = 0,
  thinned,
  last_thinned
)

Elfving_2004_BA_increment_Birch(
  diameter_cm,
  Basal_area_weighted_mean_diameter_cm,
  distance_to_coast_km,
  BA_sum_of_trees_with_larger_diameter,
  BA_Birch,
  Basal_area_plot,
  Basal_area_stand,
  computed_tree_age,
  latitude,
  altitude,
  vegetation,
  edge_effect,
  uneven_aged,
  fertilised = 0,
  thinned,
  last_thinned
)

Elfving_2004_BA_increment_Noble(
  diameter_cm,
  BA_sum_of_trees_with_larger_diameter,
  Basal_area_plot,
  Basal_area_stand,
  altitude,
  gotland = FALSE,
  vegetation,
  thinned,
  last_thinned
)

Elfving_2004_BA_increment_Oak(
  diameter_cm,
  BA_sum_of_trees_with_larger_diameter,
  BA_Oak,
  Basal_area_plot,
  Basal_area_stand,
  altitude,
  gotland = FALSE,
  vegetation,
  thinned,
  last_thinned,
  edge_effect
)

Elfving_2004_BA_increment_Pine(
  diameter_cm,
  Basal_area_weighted_mean_diameter_cm,
  BA_sum_of_trees_with_larger_diameter,
  BA_Spruce,
  BA_Pine,
  Basal_area_plot,
  Basal_area_stand,
  computed_tree_age,
  latitude,
  gotland = FALSE,
  altitude,
  SIS,
  vegetation,
  divided_plot = 0,
  edge_effect,
  uneven_aged,
  fertilised = 0,
  thinned,
  last_thinned
)

Elfving_2004_BA_increment_Spruce(
  diameter_cm,
  Basal_area_weighted_mean_diameter_cm,
  SS_diam,
  stems,
  BA_sum_of_trees_with_larger_diameter,
  BA_Spruce,
  BA_Pine,
  Basal_area_plot,
  Basal_area_stand,
  computed_tree_age,
  latitude,
  gotland = FALSE,
  altitude,
  SIS,
  vegetation,
  divided_plot = 0,
  edge_effect,
  uneven_aged,
  fertilised = 0,
  thinned,
  last_thinned
)

Elfving_2004_BA_increment_Trivial(
  diameter_cm,
  BA_sum_of_trees_with_larger_diameter,
  Basal_area_plot,
  computed_tree_age,
  vegetation,
  SIS,
  uneven_aged,
  thinned,
  last_thinned
)
}
\arguments{
\item{diameter_cm}{Diameter of the tree at breast height, 1.3 m}

\item{Basal_area_weighted_mean_diameter_cm}{Basal area weighted mean diameter of the trees on the plot, cm. \eqn{(\sum{diameter_cm^3} / \sum{diameter_cm^2})}}

\item{distance_to_coast_km}{Distance in kilometres to the nearest Swedish coast, e.g. [forester::coast_distance()]}

\item{BA_sum_of_trees_with_larger_diameter}{Basal area sum of trees on the plot with larger diameters than that of the target tree.}

\item{BA_Aspen}{Basal area Aspen on the plot. m2/ha.}

\item{Basal_area_plot}{Basal area on the plot, m2/ha.}

\item{Basal_area_stand}{Basal area in the surrounding stand, m2/ha.}

\item{computed_tree_age}{Age at breast height computed by [forester::Elfving_2003_single_tree_age_estimation()]}

\item{latitude}{Latitude, decimal degrees.}

\item{altitude}{Altitude, meters above sea level.}

\item{vegetation}{Vegetation type according to follows Swedish National forest inventory FALTSKIKT:
\tabular{ll}{
Code \tab Vegetation \cr
1 \tab  Rich-herb without shrubs \cr
2 \tab Rich-herb with shrubs/bilberry \cr
3 \tab Rich-herb with shrubs/lingonberry \cr
4 \tab Low-herb without shrubs \cr
5 \tab Low-herb with shrubs/bilberry \cr
6 \tab Low-herb with shrubs/lingonberry \cr
7 \tab No field layer \cr
8 \tab Broadleaved grass \cr
9 \tab Thinleaved grass \cr
10 \tab Sedge, high \cr
11 \tab Sedge, low \cr
12 \tab Horsetail, Equisetum ssp. \cr
13 \tab Bilberry \cr
14 \tab Lingonberry \cr
15 \tab Crowberry \cr
16 \tab Poor shrub \cr
17 \tab Lichen, frequent occurrence \cr
18 \tab Lichen, dominating \cr
}}

\item{uneven_aged}{1 if less than 80\% of the main stand volume is within a 20-year age-span.}

\item{fertilised}{1 if the plot has been fertilised, otherwise 0 (default).}

\item{thinned}{1 if the plot has been thinned, otherwise 0.}

\item{last_thinned}{Number of growth seasons since the stand was last thinned.}

\item{BA_Beech}{Basal area Beech on the plot. m2/ha.}

\item{SIS}{Site Index predicted from Stand factors [forester::Hagglund_Lundmark_1979_Site_Index_Prediction()]}

\item{divided_plot}{1 if plot has been divided, otherwise 0. Always 0 for prediction.}

\item{edge_effect}{1 if plot has been divided towards open land, otherwise 0. Both divided_plot and edge_effect cannot be 1 at the same time. Always 0 for prediction.}

\item{BA_Birch}{Basal area Birch on the plot. m2/ha.}

\item{gotland}{1 if plot is on the island of Gotland, otherwise 0.}

\item{BA_Oak}{Basal area Oak on the plot. m2/ha.}

\item{BA_Spruce}{Basal area Norway Spruce on the plot. m2/ha.}

\item{BA_Pine}{Basal area Scots Pine on the plot. m2/ha.}

\item{SS_diam}{Sum of Squared Diameters, cm.}

\item{stems}{Number of stems per hectare.}
}
\value{
The 5 - year increase of Basal Area (m^2)
}
\description{
Based on data from permanent plots at the National Forest Inventory.
}
\details{
Elfving Basal Area Increment functions for Individual Trees from NFI data.

R^2 = 0.723

R^2 = 0.722

R^2 = 0.555

R^2 = 0.639. The noble tree species are Ash \emph{Fraxinus excelsior}, Elm \emph{Ulmus glabra}, Linden \emph{Tilia cordata} , Maple \emph{Acer platanoides}, Hornbeam \emph{Carpinus betulus} and Wild Cherry \emph{Prunus avium}.

R^2 = 0.729

R^2 = 0.758

R^2 = 0.753

R^2 = 0.569. The trivial tree species include mostly: Alder \emph{Alnus}, Rowan \emph{Sorbus aucuparia} and Willow \emph{Salix spp.}
}
