% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_2010_BA_Mortality.R
\name{ElfvingMortality2010}
\alias{ElfvingMortality2010}
\alias{Elfving_2010_BasalArea_Mortality_Percent_Spruce_Pine_Sweden}
\alias{Elfving_2010_GMax_Spruce_Pine_Sweden}
\title{Mortality in Swedish stands of Scots Pine and Norway Spruce.}
\source{
Elfving, B. Natural mortality in thinning and fertilisation experiments with pine and spruce in Sweden. Forest Ecology and Management. Vol. 260. Issue 3. Available online:
\url{https://doi.org/10.1016/j.foreco.2010.04.025}
}
\usage{
Elfving_2010_BasalArea_Mortality_Percent_Spruce_Pine_Sweden(
  dominant_height,
  H100,
  latitude,
  stems,
  ThinningProportionBAStart,
  ThinningForm = 1,
  fertilised = 0
)

Elfving_2010_GMax_Spruce_Pine_Sweden(dominant_height, H100, stems, latitude)
}
\arguments{
\item{dominant_height}{Dominant height of stand.}

\item{H100}{Site Index H100 [forester::Elfving_Kiviste_1997_height_trajectory_Sweden_Pine()], [forester::Elfving_2003_height_trajectory_Sweden_Spruce()]}

\item{latitude}{Degrees N.}

\item{stems}{Stems/ha at period start.}

\item{ThinningProportionBAStart}{Proportion of basal area thinned at start of period. 0 - 0.8}

\item{ThinningForm}{Boolean. 1 if thinned from above or unthinned, otherwise 0.}
}
\value{
Annual mortality in percent of basal area at start of period.
}
\description{
Mortality in Swedish stands of Scots Pine and Norway Spruce.
}
