% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eriksson_1976_Spruce.R
\name{Eriksson1976}
\alias{Eriksson1976}
\alias{Eriksson_1976_bark_procent}
\alias{Eriksson_1976_bark_subtraction_procent}
\alias{Eriksson_1976_basal_area_before_first_thinning}
\alias{Eriksson_1976_basal_increment_under_bark}
\alias{Eriksson_1976_dry_weight_increment_tons_per_hectare}
\alias{Eriksson_1976_form_height}
\alias{Eriksson_1976_self_thinning_diameter_quotient}
\alias{Eriksson_1976_self_thinning}
\alias{Eriksson_1976_number_self_thinned_stems}
\alias{Eriksson_1976_volume_increment_under_bark_Spruce}
\title{Bark Percent addition}
\source{
Eriksson, Harry. 1976. Granens Produktion i Sverige: Yield of Norway spruce in Sweden. Research Notes Nr. 41. Dept. of Forest Yield Research. Royal College of Forestry. Stockholm.
}
\usage{
Eriksson_1976_bark_procent(age, SI, QMD_UB, BA_UB)

Eriksson_1976_bark_subtraction_procent(QMD, SI, BA, age)

Eriksson_1976_basal_area_before_first_thinning(dominant_height, stems, planted)

Eriksson_1976_basal_increment_under_bark(
  BA1,
  SI,
  thinningPercent,
  QMDThinned_UB,
  QMD1,
  stemsThinned,
  stems1,
  incrementPeriod,
  dominant_height,
  age
)

Eriksson_1976_dry_weight_increment_tons_per_hectare(
  BA1,
  SI,
  thinningPercent,
  QMDThinned_UB,
  QMD1,
  stemsThinned,
  stems1,
  incrementPeriod,
  dominant_height,
  age,
  BA5CV
)

Eriksson_1976_form_height(dominant_height, SI, QMD)

Eriksson_1976_self_thinning_diameter_quotient(dominant_height)

Eriksson_1976_self_thinning(stems1, dominant_height, SI)

Eriksson_1976_number_self_thinned_stems(dominant_height, stems1, BA1, SI)

Eriksson_1976_volume_increment_under_bark_Spruce(
  BA1,
  SI,
  thinningPercent,
  QMDThinned_UB,
  QMD1,
  stemsThinned,
  stems1,
  incrementPeriod,
  dominant_height,
  age
)
}
\arguments{
\item{age}{Age at breast height of 100 largest trees per hectare in years.}

\item{SI}{SI H100 Spruce, meters. e.g. [forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce()][forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce()]}

\item{QMD_UB}{QMD under bark, cm.}

\item{BA_UB}{Basal area under bark per hectare, m2.}

\item{QMD}{QMD over bark, cm.}

\item{BA}{Basal area over bark per hectare, m2.}

\item{dominant_height}{Dominant height, m.}

\item{stems}{Number of trees per hectare.}

\item{planted}{If site was planted, TRUE. If the site was founded by pre-commercial thinning to a certain plant spacing, FALSE.}

\item{BA1}{Basal area (under bark) per hectare before thinning.}

\item{thinningPercent}{Thinning grade in percent of BA1.}

\item{QMDThinned_UB}{QMD under bark of thinned stems}

\item{QMD1}{QMD (over bark?) of stand before thinning, cm.}

\item{stemsThinned}{Number of thinned stems per hectare.}

\item{stems1}{Number of trees per hectare before thinning.}

\item{incrementPeriod}{Number of years to increment.}

\item{BA5CV}{Group structure index (Coefficient of variation for the basal area under bark on 5-meter circular plots inside the sample plot.) (\%).}
}
\value{
(\%)

\%

Basal Area, m2 / ha

Annual increment in volume under bark per hectare, m3.

Annual increment in volume under bark per hectare, m3.

Form Height: F = VOL / BA

Quotient of QMD of self-thinned stems to QMD at start of period.

Annual self thinning in basal area over bark per hectare, m2.

Number of self-thinned stems per ha.

Volume increment under bark in cubic meters per hectare.
}
\description{
Yield of Norway spruce in Sweden
}
\details{
Author does not recommend the use of this function if dominant height exceeds 16 meters, since self thinning increases rapidly if the number of stems then also is high.
In ranges between 1'100 and 10'000 stems per ha and a dominant height between 7 and 16 metres the function should provide acceptable results. Results may be less accurate at lower site indexes due to no underlying data.

Coefficient of variation for basal area on a 5 plot has been calculated as per p.112.

R = 0.405. Gives the quotient between the quadratic mean diameter for self-thinned trees (QMD4) and the quadratic mean diameter for all trees on the plot (QMD1)

Author recommends that the function not be used when stems per ha exceed 4500 and dominant height at the same time is more than 15 m. Author still sees this as the best available information when it was written.
}
