% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eriksson_1973_volume_Aspen.R
\name{Eriksson_1973_volume_Aspen}
\alias{Eriksson_1973_volume_Aspen}
\title{Eriksson 1973 Volume for Aspen}
\source{
Eriksson, H. 1973. Volymfunktioner för stående träd av ask, asp, klibbal och contortatall. Skogshögskolan, institutionen för skogsproduktion. Rapp. o. Upps. nr 26.

PM for Heureka 2004-01-20 Björn Elfving. Available: \url{https://www.heurekaslu.org/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf}
}
\usage{
Eriksson_1973_volume_Aspen(diameter_cm, height_m)
}
\arguments{
\item{diameter_cm}{Diameter at breast height, cm.}

\item{height_m}{Height, in metres.}
}
\value{
Volume, dm3
}
\description{
Eriksson 1973 Volume for Aspen
}
