% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eriksson_1976_Spruce.R
\name{Eriksson_1976_bark_procent}
\alias{Eriksson_1976_bark_procent}
\title{Bark procent of basal area at breast height}
\source{
From p. 82-83 in Eriksson, H. (1976) "Granens produktion i Sverige", translated: "Yield of Norway spruce in Sweden". Report no. 41. Dept. of Forest Yield Research. Royal College of Forestry. Stockholm.
}
\usage{
Eriksson_1976_bark_procent(
  age_bh_100_largest_trees_per_ha_years,
  SI,
  diameter_of_mean_basal_area_under_bark_cm,
  basal_area_under_bark_ha_m2
)
}
\arguments{
\item{age_bh_100_largest_trees_per_ha_years}{Age of 100 largest trees at breast height per hectare in years.}

\item{SI}{SI H100 Spruce, m.}

\item{diameter_of_mean_basal_area_under_bark_cm}{Diameter corresponding to the mean basal area under bark, cm.}

\item{basal_area_under_bark_ha_m2}{Basal area under bark per hectare, m2.}
}
\value{
\%
}
\description{
Bark procent of basal area at breast height
}
