% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eriksson_1976_Spruce.R
\name{Eriksson_1976_bark_subtraction_procent}
\alias{Eriksson_1976_bark_subtraction_procent}
\title{Bark subtraction procent}
\source{
From p. 82-83 in Eriksson, H. (1976) "Granens produktion i Sverige", translated: "Yield of Norway spruce in Sweden". Report no. 41. Dept. of Forest Yield Research. Royal College of Forestry. Stockholm.
}
\usage{
Eriksson_1976_bark_subtraction_procent(
  diameter_of_mean_basal_area_over_bark_cm,
  SI,
  basal_area_over_bark_ha_m2,
  age_bh_100_largest_trees_per_ha_years
)
}
\arguments{
\item{diameter_of_mean_basal_area_over_bark_cm}{Diameter corresponding to the mean basal area over bark, cm.}

\item{SI}{Site index H100 m.}

\item{basal_area_over_bark_ha_m2}{Basal area over bark per hectare, m2.}

\item{age_bh_100_largest_trees_per_ha_years}{Age of 100 largest trees at breast height per hectare in years.}
}
\value{
\%
}
\description{
Bark subtraction procent
}
