% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eriksson_1976_Spruce.R
\name{Eriksson_1976_basal_area_before_first_thinning}
\alias{Eriksson_1976_basal_area_before_first_thinning}
\title{Basal area m2 per hectare before first thinning}
\source{
From p. 99 in Eriksson, H. (1976) "Granens produktion i Sverige", translated: "Yield of Norway spruce in Sweden". Report no. 41. Dept. of Forest Yield Research. Royal College of Forestry. Stockholm.
}
\usage{
Eriksson_1976_basal_area_before_first_thinning(
  dominant_height_m,
  number_of_trees_per_ha,
  planted
)
}
\arguments{
\item{dominant_height_m}{Dominant height, m.}

\item{number_of_trees_per_ha}{Number of trees per hectare.}

\item{planted}{If site was planted, TRUE. If the site was founded by pre-commercial thinning to a certain plant spacing, FALSE.}
}
\description{
Author does not reccomend the use of this function if dominant height exceeds 16 meters, since self thinning increases rapidly if the number of stems then also is high.
In ranges between 1'100 and 10'000 stems per ha and a dominant height between 7 and 16 metres the function should provide acceptable results. Results may be less accurate at lower site indexes due to no underlying data.
}
