% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eriksson_1976_Spruce.R
\name{Eriksson_1976_form_height}
\alias{Eriksson_1976_form_height}
\title{Form height for spruce stands in Eriksson 1976}
\source{
From p. 78 in Eriksson, H. (1976) "Granens produktion i Sverige", translated: "Yield of Norway spruce in Sweden". Report no. 41. Dept. of Forest Yield Research. Royal College of Forestry. Stockholm.
}
\usage{
Eriksson_1976_form_height(
  dominant_height_dm,
  SI,
  diameter_of_mean_basal_area_over_bark_cm
)
}
\arguments{
\item{dominant_height_dm}{Dominant height, dm.}

\item{SI}{Site index, metres.}

\item{diameter_of_mean_basal_area_over_bark_cm}{Diameter corresponding to mean basal area over bark, cm.}
}
\value{
Form Height
}
\description{
Form height for spruce stands in Eriksson 1976
}
