% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eriksson_1976_Spruce.R
\name{Eriksson_1976_number_self_thinned_stems}
\alias{Eriksson_1976_number_self_thinned_stems}
\title{Calculate the number of self-thinned stems during a period as by Eriksson 1976.}
\source{
Eriksson, H. (1976) "Granens produktion i Sverige", translated: "Yield of Norway spruce in Sweden". Report no. 41. Dept. of Forest Yield Research. Royal College of Forestry. Stockholm. p. 107.
}
\usage{
Eriksson_1976_number_self_thinned_stems(
  dominant_height_m,
  number_trees_per_ha_period_start,
  basal_area_over_bark_m2_ha_before_thinning,
  SI
)
}
\arguments{
\item{dominant_height_m}{Dominant height of stand, meters.}

\item{number_trees_per_ha_period_start}{Number of trees per hectare at start of period.}

\item{basal_area_over_bark_m2_ha_before_thinning}{Basal area above bark m2/ha before thinning.}
}
\value{
numeric.
}
\description{
Calculate the number of self-thinned stems during a period as by Eriksson 1976.
}
