% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eriksson_1976_self_thinning.R
\name{Eriksson_1976_self_thinning}
\alias{Eriksson_1976_self_thinning}
\title{Eriksson 1976 self thinning}
\source{
From p. 86-88 in Eriksson, H. (1976) "Granens produktion i Sverige", translated: "Yield of Norway spruce in Sweden". Report no. 41. Dept. of Forest Yield Research. Royal College of Forestry. Stockholm.
}
\usage{
Eriksson_1976_self_thinning(
  number_trees_per_ha_period_start,
  dominant_height_m,
  SI
)
}
\arguments{
\item{number_trees_per_ha_period_start}{Number of trees per hectare at the start of the period.}

\item{dominant_height_m}{Dominant height, m.}

\item{SI100}{Site Index H100 Spruce from Hägglund 1972/1973, e.g. [forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce()], [forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce()]}
}
\value{
Annual self thinning in basal area over bark per hectare, m2.
}
\description{
Author recommends that the function not be used when stems per ha exceed 4500 and dominant height at the same time is more than 15 m. Author still sees this as the best available information when it was written.
}
