% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Eriksson_1976_volume_increment_under_bark_Spruce.R
\name{Eriksson_1976_volume_increment_under_bark_Spruce}
\alias{Eriksson_1976_volume_increment_under_bark_Spruce}
\title{Annual Volume Increment under bark per hectare, m3}
\source{
From p. 64;65. in Eriksson, H. (1976) "Granens produktion i Sverige", translated: "Yield of Norway spruce in Sweden". Report no. 41. Dept. of Forest Yield Research. Royal College of Forestry. Stockholm.
}
\usage{
Eriksson_1976_volume_increment_under_bark_Spruce(
  basal_area_ha_before_thinning.m2,
  site_index,
  thinning_percent_ba_under_bark,
  diameter_of_mean_basal_area_of_thinned_trees_cm,
  diameter_of_mean_basal_area_before_thinning_cm,
  number_thinned_trees_per_ha,
  number_trees_per_ha_before_thinning,
  increment_period_years,
  dominant_height_dm,
  age_bh_100_largest_trees_per_ha_years
)
}
\arguments{
\item{site_index}{Spruce}

\item{thinning_percent_ba_under_bark}{Thinning as percentage of basal area under bark before thinning.}

\item{diameter_of_mean_basal_area_of_thinned_trees_cm}{Diameter corresponding to mean basal area under bark of the thinned trees, cm.}

\item{diameter_of_mean_basal_area_before_thinning_cm}{Diameter corresponding to the mean basal area of the stand before thinning, cm.}

\item{number_thinned_trees_per_ha}{Number of thinned trees per hectare.}

\item{number_trees_per_ha_before_thinning}{Number of trees per hectare in the stand before thinning.}

\item{increment_period_years}{Length of increment period, years.}

\item{dominant_height_dm}{Dominant height, dm.}

\item{age_bh_100_largest_trees_per_ha_years}{Age at breast height of the 100 largest trees per hectare, years.}

\item{basal_area_ha_before_thinning_m2}{Basal area under bark per hectare before thinning, m2.}
}
\value{
Annual increment in basal area under bark per hectare, m2.
}
\description{
Annual Volume Increment under bark per hectare, m3
}
