% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eriksson_1986_humidity.R
\name{Eriksson_1986_humidity}
\alias{Eriksson_1986_humidity}
\title{Humidity according to polygons from the Swedish National Forest Inventory.}
\source{
The polygons delineating contour lines were provided from the Swedish National Forest Inventory.
These polygons origins are unknown, but are very close to fig. 34 'Humidity during the vegetation period (mm)' from Eriksson, B. (1986) Nederbörds- och humiditetsklimatet i Sverige under vegetationsperiod. Sveriges Meterologiska och Hydrologiska Institut (SMHI), rapporter i meteorologi och klimatologi (RMK) 46. Available:
\url{https://www.smhi.se/polopoly_fs/1.166663!/RMK_46\%20Nederbörds-\%20och\%20humiditetsklimat\%20i\%20Sverige\%20under\%20vegetationsperioden..pdf}
}
\usage{
Eriksson_1986_humidity(latitude, longitude, epsg = 4326)
}
\arguments{
\item{latitude}{Latitude degrees}

\item{longitude}{Longitude degrees}

\item{epsg}{Reference system, default 4326 (WGS84). For SWEREF99TM, 3006.}
}
\value{
Humidity during the vegetation period (mm).
}
\description{
Provides simple interface for an estimation of the humidity during the vegetation period for Swedish sites.

The humidity during the vegetation period is 'the difference between precipitation amount and 'real' evapotranspiration.'.
}
\examples{
Eriksson_1986_humidity(latitude = 64.4,longitude = 17.8,epsg=4326)
}
