% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eriksson_1976_Spruce.R
\name{Eriksson_frequency}
\alias{Eriksson_frequency}
\alias{Eriksson_1976_volume_frequency}
\alias{Eriksson_1976_stem_frequency}
\title{Volume and Stem Frequencies for even-aged Norway Spruce stands from the Great Yield
Investigation (Näslund 1971).}
\source{
Eriksson, H. 1976. Yield of Norway spruce in Sweden. Research Notes.
Dept. of. Forest Yield Research. Nr. 41. Royal College of Forestry.
Stockholm. p. 161-165; 279-284.
}
\usage{
Eriksson_1976_volume_frequency(
  age,
  dominant_height,
  QMD,
  volume,
  stems,
  class_width_cm = 2,
  class_middle = 1.5
)

Eriksson_1976_stem_frequency(
  age,
  dominant_height,
  QMD,
  stems,
  class_width_cm = 2,
  class_middle = 1.5
)
}
\arguments{
\item{age}{Stand age at breast height (1.3 m)}

\item{dominant_height}{Dominant height, in metres.}

\item{QMD}{Quadratic Mean Diameter (cm).}

\item{volume}{Volume of stand, cu.m.}

\item{stems}{Stems per hectare.}

\item{class_width_cm}{Default 2 cm. Range between upper and lower class limit.}

\item{class_middle}{Arithmetic mean of the sum of the upper and lower limit.}
}
\value{
Frequency, in percent.
}
\description{
This is an attempt to follow the instructions given by Eriksson.
From functions relating easily gathered information to a fitted stem and
volume distribution respectively, a skewed normal distribution is estimated.
From an expanded series, the frequency of each class is then calculated.
}
\details{
TESTED ONLY FOR 2 CM class width!
}
