% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Explain forester model}
\usage{
explain(models, test_data, y, verbose = FALSE)
}
\arguments{
\item{models}{A single model or the model list created with the `train()` function.}

\item{test_data}{A test dataset returned from `train()` function.}

\item{y}{A target variable. It can be either
(1) a vector of the same number of observations as `data` or
(2) a character name of variable in the `data` that contains
the target variable.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
the process, if FALSE gives none.}
}
\value{
A list of DALEX explainers for 5 models of different engines.
}
\description{
The `explain()` function is a wrapper for `DALEX` methods for model explanations.
If possible it will use methods for tree-based models.
}
\examples{
\dontrun{
data(lisbon)
train_output   <- train(lisbon, 'Price', verbose = FALSE, random_evals = 2, bayes_iter = 1)
best_explainer <- explain(train_output$best_models_on_valid[[1]][1],
                          train_output$valid_data, train_output$y)
best_explainers <- explain(train_output$best_models_on_valid[[1]],
                           train_output$valid_data, train_output$y)
}
}
