% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_engineering_function}
\alias{feature_engineering_function}
\title{Automatic Function for Feature Engineering}
\usage{
feature_engineering_function(
  data_train,
  data_test,
  target,
  type,
  threshold_na = 0.5,
  remove_outliers = FALSE,
  fill_na = FALSE,
  scaling = NULL,
  num_features = NULL
)
}
\arguments{
\item{data_train}{dataset, used for training models. Class of data_train is one of those classes: data.frame, matrix, data.table or dgCMatrix. NOTE: data_train includes target column.}

\item{data_test}{dataset having target column, used for evaluating models' performance, having same structure with data_train.}

\item{target}{character, indicating name of the target column in data_train.}

\item{type}{character, defining the task. Option is "regression" or "classification", namely binary classification.}

\item{threshold_na}{number, ranged from 0 to 1. Function removes columns having percentage of NAs higher than value of threshold_na.}

\item{remove_outliers}{logical, default is FALSE. If TRUE, all outliers in numeric columns will be removed, except for target column.}

\item{fill_na}{logical, default is FALSE. If TRUE, missing values in target column are removed, missing values in categorical columns are replaced by mode and
missing values in numeric columns are substituted by median of corresponding columns.}

\item{scaling}{character, default is NULL. Parameter is used for scaling features. Options are "standardize", "minmax" and NULL.}

\item{num_features}{numeric, default is NULL. Parameter indicates number of most important features, which are chosen from the train dataset. Automatically, those important
features will be kept in the train and test datasets.}
}
\value{
A list contains 2 objects: processed data_train and process data_test.
}
\description{
Function \code{feature_engineering} covers fundamental techniques in ML Feature Engineering process.
}
\details{
With adjustable arguments, users can independently decide how function deals with outliers, NAs value,
feature scaling and important features selection. Furthermore, options for sampling methods will pop up in
case of classification problem.
}
\references{
forester library \url{https://modeloriented.github.io/forester/}
}
