% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_removal.R
\name{find_duplicate_columns}
\alias{find_duplicate_columns}
\title{Find duplicate columns}
\usage{
find_duplicate_columns(data)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}
}
\value{
A numeric 0-1 matrix.
}
\description{
It is a helper function for the rm_duplicate_columns(). It creates a 2 dimensional
matrix which has 0s when the columns are not duplicates and 1s otherwise.
}
