% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finland_vegetation_types.R
\name{Finland_vegetation_types}
\alias{Finland_vegetation_types}
\title{AK Cajander Forest Vegetation Types.}
\source{
Cajander, A.K. 1949. Forest types and their significance. Acta Forestalia Fennica. 56:5. 71 pp.
Available (16/03/2022): \url{https://www.silvafennica.fi/pdf/article7396.pdf}
}
\usage{
Finland_vegetation_types(short = TRUE)
}
\arguments{
\item{short}{TRUE (default) or FALSE, for short or long list, respectively.}
}
\value{
Tibble.
}
\description{
A function which prints a tibble of AK Cajander's Forest Types.
}
