% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norway_double_bark_functions.R
\name{Fitje_1995_double_bark_thickness_Norway_Scots_Pine}
\alias{Fitje_1995_double_bark_thickness_Norway_Scots_Pine}
\title{Double bark thickness for Norway Spruce in Norway}
\source{
Fitje, A. 1995.
}
\usage{
Fitje_1995_double_bark_thickness_Norway_Scots_Pine(diameter_cm)
}
\arguments{
\item{diameter_cm}{Diameter of tree at breast height, 1.3m.}
}
\value{
Double bark, in mm.
}
\description{
Double bark thickness for Norway Spruce in Norway
}
\author{
Modified from hansoleorka/skogR on the 2022-04-02
}
