% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestable.R
\name{forestable}
\alias{forestable}
\title{Create a forest plot column in a table}
\usage{
forestable(
  left_side_data,
  estimate,
  ci_low,
  ci_high,
  right_side_data = NULL,
  theme = NULL,
  estimate_precision = 1,
  ggplot_width = 30,
  null_line_at = 0,
  file_path = here::here("forestable_plot.png"),
  dpi = 600,
  display = TRUE,
  blank_na = TRUE,
  font_family = "mono",
  estimate_col_name = "Estimate",
  stripe_colour = "#eff3f2",
  x_scale_linear = TRUE,
  xlim = NULL
)
}
\arguments{
\item{left_side_data}{A data frame with the information to be displayed to the left of the forest plot.}

\item{estimate}{A vector containing the point estimates to be displayed in the forest plot.}

\item{ci_low}{A vector containing the lower confidence bounds}

\item{ci_high}{A vector containing the upper confidence bounds}

\item{right_side_data}{A data frame (optional) containing the information to be displayed on the right side of the table. If not supplied, an Estimate column is generated automatically.}

\item{theme}{A theme object for the table (optional)}

\item{estimate_precision}{Integer. The number of decimal places on the estimate (default 1)}

\item{ggplot_width}{The width of forest plot in characters (default 26)}

\item{null_line_at}{numeric, default 0. Change to 1 if using relative measures such as OR, RR.}

\item{file_path}{Where to save the image, default "forestable_plot.png" in the current working directory.}

\item{dpi}{The image resolution in dpi, default 600}

\item{display}{Show the table in RStudio viewer? Default TRUE}

\item{blank_na}{Should missing values in the left side table be displayed as blank? Default TRUE, if FALSE, NA values will be shown}

\item{font_family}{The font to use for the ggplot and table}

\item{estimate_col_name}{The name for the generated estimate column. Default "Estimate"}

\item{stripe_colour}{Colour to use for the table stripes, default "#eff3f2"}

\item{x_scale_linear}{Default TRUE, change to FALSE for log scale}

\item{xlim}{Manually specify limits for the x axis as a vector length 2, i.e. c(low, high)}
}
\value{
image
}
\description{
Creates an png image with the forest plot column inserted in the supplied data frame.
}
