% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forester_palette.R
\name{forester_palette}
\alias{forester_palette}
\title{Return colors from palette}
\usage{
forester_palette(...)
}
\arguments{
\item{...}{List of names of colors.}
}
\value{
A list of colors' codes.
}
\description{
Return colors from palette
}
\examples{
forester_palette()
forester_palette('brown')
forester_palette('green', 'dark green')
}
