% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_models_details.R
\name{format_models_details}
\alias{format_models_details}
\title{Formats info about models}
\usage{
format_models_details(models)
}
\arguments{
\item{models}{The result of `choose_best_models()` function or just a list of models
from engine.}
}
\value{
Prints formatted text..
}
\description{
Formats info about models
}
\examples{
data(iris)
iris_bin <- iris[1:100, ]
type <- guess_type(iris_bin, 'Species')
preprocessed_data <- preprocessing(iris_bin, 'Species')
preprocessed_data <- preprocessed_data$data
split_data <-
  train_test_balance(preprocessed_data, 'Species', type = type, balance = FALSE)
train_data <-
  prepare_data(split_data$train,
               'Species',
               c('ranger', 'xgboost', 'decision_tree', 'lightgbm', 'catboost'))
test_data <- prepare_data(split_data$test,
            'Species',
             engine = c('ranger', 'xgboost', 'decision_tree', 'lightgbm', 'catboost'),
             predict = TRUE,
             train = split_data$train)
models <-  train_models(train_data,
       'Species',
       engine = c('ranger', 'xgboost', 'decision_tree', 'lightgbm', 'catboost'),
       type = type)
predictions <-
  predict_models(models,
                 test_data,
                 'Species',
                 engine = c('ranger', 'xgboost', 'decision_tree', 'lightgbm', 'catboost'),
                 type = type)
score <-
    score_models(models,
                 predictions,
                 observed = split_data$test$Species,
                 type = type)
best_models <- choose_best_models(models,
              score = score,
              number = 3)
format_models_details(best_models)
}
