% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulator.R
\name{formulator}
\alias{formulator}
\title{Formula maker}
\usage{
formulator(data, ID, lhs, constant = "constant", term = "terms")
}
\arguments{
\item{data}{A data.frame, more information on structure in details.}

\item{ID}{The ID of the column you are interested in.}

\item{lhs}{Your response variable, Default: "myVariable".}

\item{constant}{A string defining the 'constant term'. Default: 'constant'.}
}
\value{
A function in functional form.
}
\description{
Creates functions from data frame information. Data frame specifications is given in details.
}
\details{
The data frame must consist of:
* a column containing a row for each of the terms used in any of the functions (NA for not used).
* a column for each function with the respective value for each term.

By default, If any term is called *constant* this rows value will not be multiplied by the term.


\tabular{lrrr}{
Term \tab F1 \tab \tab F2 \cr
constant \tab 1 \tab \tab 1.5 \cr
b1 \tab -3.23 \tab \tab NA \cr
b2 \tab NA \tab \tab 32
}
}
\examples{
#Not run
formulator(data, "Beech", "`my Variable`")
myVariable ~ 0.2081 + 1.7491\\*lndp1 + -0.2167\\*d
}
