% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{guess_type}
\alias{guess_type}
\title{Guess task type by the target value from the dataset}
\usage{
guess_type(data, y, max_unique = 15)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{y}{A string that indicates a target column name.}

\item{max_unique}{An integer describing the maximal number of unique
values in `y`.}
}
\value{
A string describing the type of ml task: `binary_clf`, `multiclass`,
`regression`, or `survival`.
}
\description{
For now, supported types are binary classification and regression.
Multilabel classification is planned to be added later on.
}
