% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gustavsen_1980_height_trajectories.R
\name{gustavsen1980}
\alias{gustavsen1980}
\alias{Gustavsen_1980_height_trajectory_Finland_Scots_Pine}
\alias{Gustavsen_1980_height_trajectory_Finland_Norway_Spruce}
\title{Dominant height trajectories for Scots Pine and Norway Spruce in Finland
from Gustavsen, 1980.}
\source{
Gustavsen, H-G (1980). Talousmetsien kasvupaikkaluokittelu
valtapituuden avulla. (Site index curves for conifer stands in Finland.)
Folia Forestalia 454: 1-31. Available (12/03/2021):
\url{https://jukuri.luke.fi/handle/10024/522037}
}
\usage{
Gustavsen_1980_height_trajectory_Finland_Scots_Pine(dominant_height, age, age2)

Gustavsen_1980_height_trajectory_Finland_Norway_Spruce(
  dominant_height,
  age,
  age2
)
}
\arguments{
\item{dominant_height}{Dominant height, in metres.}

\item{age}{Total age}

\item{age2}{Total age at output.}
}
\description{
"Site index curves for natural Scots Pine.. and Norway spruce
stands in Finland, based on 1069 temporary sample plots of 0.1 ha in pine-dominated
and 617 plots in spruce-dominated stands from the 3rd national forest inventory
(1951-53)".
}
