% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Gustavsen_1984_height_trajectory_Finland_Birch.R
\name{Gustavsen_1984_height_trajectory_Finland_Birch}
\alias{Gustavsen_1984_height_trajectory_Finland_Birch}
\title{Dominant height trajectory for naturally regenerated Birch in Finland}
\source{
Gustavsen, H.G., Mielikäinen, K. (1984). Luontaisesti syntyneiden
koivikoiden kasvupaikka-luokittelu valtapituuden avulla. (Site Index Curves
for Natural Birch Stands in Finland). Folia For. 597: 1-20.
Available (26/03/2022): \url{https://jukuri.luke.fi/handle/10024/522180}
}
\usage{
Gustavsen_1984_height_trajectory_Finland_Birch(dominant_height, age, age2)
}
\arguments{
\item{dominant_height}{Dominant height, in metres.}

\item{age}{Total age}

\item{age2}{Total age at output}
}
\value{
Dominant height at age2.
}
\description{
'For natural Betula pendula Roth and Betula pubescens Ehrh.
stands on mineral soils in Finland...using 341 observation
periods on permanent plots and stemanalyses of dominant trees on temporary
plots... the site index classes H50=12-18 will mainly serve the classification
of Betula pubescens stands under practical conditions. The upper classes of
the system, H50 = 20-26, will mainly be site index classes for Betula pendula
stands...
\strong{The site index curves and yield values calculated for natural birch
stands show that the results cannot be used for the classifications of
plantations of Betula pendula.}'

\strong{NB Increment is for next year. Therefore to find height at a lower
age than at outset, function optimises to find height at output age which can
be incremented to the outset height, age pair.}
}
\details{
ln(increment Hdom): N = 341. R^2= 0.578. s_m= 0.647, s_f = 0.422.
}
