% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Gustavsen_1984_height_trajectory_Finland_Birch.R
\name{Gustavsen_1984_time_to_breast_height}
\alias{Gustavsen_1984_time_to_breast_height}
\title{Smoothing of time required to reach breast height according to Gustavsen 1984.}
\usage{
Gustavsen_1984_time_to_breast_height(SI50)
}
\arguments{
\item{SI50}{Site index 50, according to Gustavsen 1984.}
}
\value{
Number of years required to reach breast height 1.3m.
}
\description{
Smoothing of time required to reach breast height according to Gustavsen 1984.
}
\details{
log-linear function log(t1.3)~H50.
}
\seealso{
[forester::Gustavsen_1984_height_trajectory_Finland_Birch()]
}
\author{
Carl Vigren SLU
}
