% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Hagglund_1972_northern_Sweden_Height_trajectories_Spruce.R
\name{Hagglund_1972_northern_Sweden_Height_trajectories_Spruce}
\alias{Hagglund_1972_northern_Sweden_Height_trajectories_Spruce}
\title{Hägglund 1972 Height growth of Spruce in Northern Sweden.}
\source{
Hägglund, Björn (1972) Om övre höjdens utveckling för gran i norra
 Sverige: Site index curves for Norway Spruce in northern Sweden. Diss. Dept.
 of Forest Yield Research. Royal College of Forestry. Report 21. 298 pp. Stockholm.
}
\usage{
Hagglund_1972_northern_Sweden_Height_trajectories_Spruce(
  dominant_height,
  age,
  age2,
  latitude,
  culture = 1,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{height of tree or stand in m.}

\item{age}{Age of stand or tree at breast height 1.3 m.}

\item{age2}{Necessary if output is "Height". The age for which height along the same curve is to be computed.}

\item{latitude}{Latitude in degrees. For function 8.4, set to 0.}

\item{culture}{1 (Default) if stand is culture, otherwise 0.}

\item{output}{One of "SIH100", "Height", or "Equation".}
}
\value{
If output is "SIH100", the numeric value for the height (m) a stand
will reach at age 100.

 If output is "Height", the numeric value for the height (m) a stand will
 reach at age2.

 If output is "Equation", a named list with 2 elements: I) "Equation": text response with the equation for that height
 curve. II) "T13": The time taken to reach breast height 1.3m.
}
\description{
Largely based on the Fortran IV script appended to the original source.
Does not include certain functions to calculate if age at breast height is unknown,
and has been adapted to R such that it will return one value at a time rather than
print a matrix for height developments.
}
\details{
OBSERVE! Will throw warnings if outside of material, but not stop.

Based on the Chapman-Richards function.
}
\examples{
Hagglund_1972_northern_Sweden_Height_trajectories_Spruce(latitude = 60.1,age = 80, top_height_dm = 120,culture = 0,output = "Equation")
#"y~13+224.224608372892*(1-exp(-age*0.0113938018668758))^1.44412711110086"
Hagglund_1972_northern_Sweden_Height_trajectories_Spruce(latitude = 66.9,age = 80, top_height_dm = 120,culture = 0,output = "Equation")
#"y~13+228.827607753106*(1-exp(-age*0.0116464323317232))^1.52753582497555"

#Remember to divide by 10 to go to metres from dm.
ggplot()+
  geom_function(aes(color="Lat=60.1°"), fun = function(age) (13+224.224608372892*(1-exp(-age*0.0113938018668758))^1.44412711110086)/10)+
  geom_function(aes(color="Lat=66.9°"), fun = function(age) (13+228.827607753106*(1-exp(-age*0.0116464323317232))^1.52753582497555)/10)+
  labs(x="Age at breast height (years)",
       y= "Dominant height (metres)")+
  xlim(c(0,150))+
  ylim(c(0,30))


}
