% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hagglund_1972_time_to_BH.R
\name{Hagglund_1972_Norway_Spruce_time_to_BH_northern_Sweden}
\alias{Hagglund_1972_Norway_Spruce_time_to_BH_northern_Sweden}
\title{Time to breast height for Spruce in Northern Sweden.}
\usage{
Hagglund_1972_Norway_Spruce_time_to_BH_northern_Sweden(
  H100TotalAge,
  culture = 1,
  latitude
)
}
\arguments{
\item{H100TotalAge}{Site Index H100 from Hagglund 1972.}

\item{culture}{1 if culture, otherwise 0.}

\item{latitude}{degrees North.}
}
\description{
Time to breast height for Spruce in Northern Sweden.
}
\examples{
ggplot()+xlim(c(0,200))+geom_function(fun=function(x) Hagglund_1972_northern_Sweden_Height_trajectories_Spruce(16,age=100-Hagglund_1972_Norway_Spruce_time_to_BH_northern_Sweden(16,1,61),age2=x,61,1))+geom_point(aes(x=seq(10,200,10),y=c(27,46,66,86,104,121,137,151,164,175,185,194,203,210,216,222,227,231,235,239)/10))
}
\seealso{
[forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce()]
}
