% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hagglund_1972_time_to_BH.R
\name{Hagglund_1973_Norway_Spruce_time_to_BH_southern_Sweden}
\alias{Hagglund_1973_Norway_Spruce_time_to_BH_southern_Sweden}
\title{Time to breast height for Spruce in Southern Sweden.}
\usage{
Hagglund_1973_Norway_Spruce_time_to_BH_southern_Sweden(H100TotalAge)
}
\arguments{
\item{H100TotalAge}{Site Index H100 from Hagglund 1973.}
}
\description{
Time to breast height for Spruce in Southern Sweden.
}
\examples{
ggplot()+xlim(c(0,100))+geom_function(fun=function(x) Hagglund_1973_southern_Sweden_Height_trajectories_Spruce(28,age=100-Hagglund_1973_Norway_Spruce_time_to_BH_southern_Sweden(28),age2=x))+geom_point(aes(x=seq(10,100,10),y=c(55,100,140,175,204,228,248,264,277,288)/10))+xlab("Age at breast height")+ylab("Dominant Height, m")
}
\seealso{
[forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce()]
}
