% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Hagglund_1973_southern_Sweden_Height_trajectories_Spruce.R
\name{Hagglund_1973_southern_Sweden_Height_trajectories_Spruce}
\alias{Hagglund_1973_southern_Sweden_Height_trajectories_Spruce}
\title{Hägglund 1973 Height growth of Spruce in southern Sweden.}
\source{
Hägglund, Björn (1973) Om övre höjdens utveckling för gran i södra
 Sverige: Site index curves for Norway Spruce in southern Sweden. Diss. Dept.
 of Forest Yield Research. Royal College of Forestry. Report 24. 49 pp. Stockholm.
}
\usage{
Hagglund_1973_southern_Sweden_Height_trajectories_Spruce(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Age of stand or tree at *breast height 1.3 m.*}

\item{age2}{Necessary if output is "Height". The age for which height along
the same curve is to be computed.}

\item{output}{One of "SIH100", "Height", or "Equation".}
}
\value{
If output is "SIH100", the numeric value for the height (m) a stand
will reach at age 100.

 If output is "Height", the numeric value for the height (m) a stand will
 reach at age2.

 If output is "Equation", a named list with 2 elements: I) "Equation": text response with the equation for that height
 curve. II) "T13": The time taken to reach breast height 1.3m.
}
\description{
Largely based on the Fortran IV script appended to the original source.
Does not include certain functions to calculate if age at breast height is unknown,
and has been adapted to R such that it will return one value at a time rather than
print a matrix for height developments.
}
\details{
OBSERVE! Will throw warnings if outside of material, but not stop.

OBSERVE! The SIH100 returned is H100 TOTAL AGE, whereas if you enter age2 as 100 it is age at breast height.

In order to calculate Height at another point in time (age2 at breast height) based on a SIH100 total age (age1 total age) - subtract the time to breast height from age1.
e.g. From Total Age 100, H=16 to age at breast height age2.
Hagglund_1973_southern_Sweden_Height_trajectories_Spruce(16,age=100-forester::Hagglund_1973_Norway_Spruce_time_to_BH_southern_Sweden(16,1,61),age2=140,61,1)

Based on the Chapman-Richards function.
}
\examples{
Hagglund_1973_southern_Sweden_Height_trajectories_Spruce(dominant_height = 22,age = 40,age2 = 40)
}
\seealso{
[forester::Hagglund_1973_Norway_spruce_time_to_BH_southern_Sweden()]
}
