% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hagglund_1974_Sweden_height_trajectories_Pine.R
\name{Hagglund_1974_Sweden_height_trajectories_Pine}
\alias{Hagglund_1974_Sweden_height_trajectories_Pine}
\title{Hägglund 1974 Height growth of Scots Pine in Sweden.}
\source{
Hägglund, Björn (1974) Övre höjdens utveckling i tallbestånd:
Site index curves for Scots Pine in Sweden. Dept.
 of Forest Yield Research. Royal College of Forestry. Report 31. 54 pp. Stockholm.
}
\usage{
Hagglund_1974_Sweden_height_trajectories_Pine(
  dominant_height,
  age,
  age2,
  regeneration = "culture",
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Top height of tree or stand in m.}

\item{age}{Age of stand or tree at breast height 1.3 m.}

\item{age2}{Necessary if output is "Height". The age for which height along
the same curve is to be computed.}

\item{regeneration}{How was the stand established? One of "culture", "natural" or "unknown".}

\item{output}{One of "Height" (default), "SIH100", or "Equation".}
}
\value{
If output is "SIH100", the numeric value for the height (m) a stand
will reach at age 100.

 If output is "Height", the numeric value for the height (m) a stand will
 reach at age2.

 If output is "Equation", a named list with 2 elements: I) "Equation": text response with the equation for that height
 curve. II) "T13": The time taken to reach breast height 1.3m.
}
\description{
Largely based on the Fortran IV script appended to the original source.
Does not include certain functions to calculate if age at breast height is unknown,
and has been adapted to R such that it will return one value at a time rather than
print a matrix for height developments.
}
\details{
OBSERVE! Will throw warnings if outside of material, but not stop.

Based on the Chapman-Richards function. f.4.1

OBSERVE! The SIH100 returned is H100 TOTAL AGE, whereas if you enter age2 as 100 and output == "Height", it is age at breast height.
Example below analogous from Spruce.
In order to calculate Height at another point in time (age2 at breast height) based on a SIH100 total age (age1 total age) - subtract the time to breast height from age1.
e.g. From Total Age 100, H=16 to age at breast height age2.
Hagglund_1972_northern_Sweden_Height_trajectories_Spruce(16,age=100-forester::Hagglund_1972_Norway_Spruce_time_to_BH_northern_Sweden(16,1,61),age2=140,61,1)
}
\examples{
ggplot()+
scale_x_continuous(limits = c(0,120),expand = c(0,0))+
scale_y_continuous(limits=c(0,27),expand = c(0,0))+
#expand_limits(x=c(0,130),y=c(0,30))+
geom_function(aes(linetype="a"), fun = function (age)
 #20
 (13+233.564868164063*(1-exp(-age*0.0172925395397362))^1.14013833378919)/10
)+
geom_function(aes(linetype="b"), fun = function(age)

 #22
 (13+248.657690429688*(1-exp(-age*0.0188389760256776))^1.12794900485356)/10

)+

geom_function(aes(linetype="c"), fun= function(age)

 #24
 (13+264.911499023438*(1-exp(-age*0.0205442213276424))^1.11788897659132)/10

)+

geom_function(aes(linetype="d"), fun= function(age)

  #16
  (13+203.379223632813*(1-exp(-age*0.0143125270333732))^1.17804835993454)/10

)+
theme_classic()+
theme(plot.title = element_text(face="italic",family="serif"),
      plot.margin = unit(c(10,20,10,10), units="points"),
      legend.position = "none")+
labs(
  title="Development of Dominant height of\nScots pine in Sweden",
  caption="After Hägglund (1974) f. 4.1.",
  x="Age at breast height (years)",
  y="Dominant Height (m)"
)+
annotate("text",x=115, y=26,label="24",size=3)+
annotate("text",x=115, y=24, label="22", size=3)+
annotate("text",x=115, y=22, label="20",size=3)+
annotate("text",x=115, y=18, label="16",size=3)
}
\seealso{
[forester::Hagglund_1974_Sweden_Time_to_Breast_Height_Pine]
}
