% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Hagglund_1974_Sweden_Time_to_Breast_Height_Pine.R
\name{Hagglund_1974_Sweden_Time_to_Breast_Height_Pine}
\alias{Hagglund_1974_Sweden_Time_to_Breast_Height_Pine}
\title{Time to breast height for Scots Pine in Sweden.}
\usage{
Hagglund_1974_Sweden_Time_to_Breast_Height_Pine(
  H100TotalAge,
  regeneration = "culture"
)
}
\description{
Time to breast height for Scots Pine in Sweden.
}
\seealso{
[forester::Hagglund_1974_Sweden_height_trajectories_Pine]
}
