% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hagglund_1981_initial_basal_area.R
\name{Hagglund_1981_initial_basal_area}
\alias{Hagglund_1981_initial_basal_area}
\title{Basal area for stand initiation from Hägglund (1981)}
\source{
Hägglund, Björn (1981) Samband mellan ståndortsindex H100 och bonitet för
tall och gran i Sverige. Projekt HUGIN. Rapport 26. Skogsvetenskapliga fakultet.
Sveriges Lantbruksuniversitet. Umeå.
}
\usage{
Hagglund_1981_initial_basal_area(
  species,
  region_code,
  dominant_height_m,
  altitude_m,
  SI,
  vegetation
)
}
\arguments{
\item{species}{"Picea abies" or "Pinus sylvestris".}

\item{region_code}{1-5. 1: Lappmarken, 2: Norr-& Västerbottens kustland & Jämtland, Kopparberg - Sälen-Idre.
3: Västernorrland, Gävleborg, Kopparberg -övrigt , Värmland.
5: Gotland, Skåne, Halland, Blekinge, Göteborg & Bohus.}

\item{dominant_height_m}{Dominant height of stand, meters.}

\item{altitude_m}{Meters above sea level.}

\item{SI}{Site index H100 of species.}

\item{vegetation}{1-18.}
}
\value{
list of 3 elements: 1. Basal Area m^2/ha.
2. Multiple Coefficient of Correlation for used function.
3. Spread about the function, m^2/ha.
}
\description{
Basal area for stand initiation from Hägglund (1981)
}
\examples{
Hagglund_1981_initial_basal_area(species="Picea abies",region_code = 5,dominant_height_m = 10,altitude_m = 0,SI = 36,vegetation = 1)
}
