% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hagglund_1981_initial_stem_count.R
\name{Hagglund_1981_initial_stem_count}
\alias{Hagglund_1981_initial_stem_count}
\title{Stem Count for Stand Initiation from Hägglund (1981)}
\source{
Hägglund, Björn (1981) Samband mellan ståndortsindex H100 och bonitet för
tall och gran i Sverige. Projekt HUGIN. Rapport 26. Skogsvetenskapliga fakultet.
Sveriges Lantbruksuniversitet. Umeå.
}
\usage{
Hagglund_1981_initial_stem_count(
  SI,
  altitude_m,
  vegetation,
  species,
  dominant_height_m,
  region_code
)
}
\arguments{
\item{SI}{Site Index H100 for species.}

\item{altitude_m}{Meters above sea level.}

\item{vegetation}{1-18}

\item{species}{"Picea abies" or "Pinus sylvestris".}

\item{dominant_height_m}{Dominant height of stand, meters.}

\item{region_code}{1-5. 1: Lappmarken, 2: Norr-& Västerbottens kustland & Jämtland, Kopparberg - Sälen-Idre.
3: Västernorrland, Gävleborg, Kopparberg -övrigt , Värmland.
5: Gotland, Skåne, Halland, Blekinge, Göteborg & Bohus.}
}
\value{
list of 3 elements: 1. Stem Count/ha.
2. Multiple Coefficient of Correlation for used function.
3. Spread about the function, m^2/ha.
}
\description{
Stem Count for Stand Initiation from Hägglund (1981)
}
