% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hagglund_age_to_height.R
\name{Hagglund_age_to_height}
\alias{Hagglund_age_to_height}
\title{Age at a certain height according to height trajectories by Hägglund (1972,1973,1974)}
\usage{
Hagglund_age_to_height(
  latitude,
  SI,
  dominant_height_m = 12.5,
  species,
  locality,
  regeneration = "culture"
)
}
\arguments{
\item{latitude}{Degrees N. Only required for "Picea abies" if locality is "northern".}

\item{SI}{Site Index H100}

\item{dominant_height_m}{Given height if interested in height above breast height. Defaults to 12.5 m.}

\item{species}{"Picea abies" or "Pinus sylvestris".}

\item{locality}{For "Picea abies": "northern" or "southern" Sweden.}

\item{regeneration}{How was the stand established? One of "culture", "natural regeneration" or "unknown".}

\item{description}{Wrapper to find age at breast height, total age and age to
breast height at a set SI and/or height for Hagglunds functions.}
}
\value{
List of 3 elements:

"Age at breast height given dominant height", numeric.

"Time to breast height 1.3 m", numeric.

"Total age", numeric.
}
\description{
Age at a certain height according to height trajectories by Hägglund (1972,1973,1974)
}
\seealso{
\code{\link[=Hagglund_1972_northern_Sweden_Height_trajectories_Spruce]{Hagglund_1972_northern_Sweden_Height_trajectories_Spruce()}} ,

\code{\link[=Hagglund_1973_southern_Sweden_Height_trajectories_Spruce]{Hagglund_1973_southern_Sweden_Height_trajectories_Spruce()}},

\code{\link[=Hagglund_1974_Sweden_height_trajectories_Pine]{Hagglund_1974_Sweden_height_trajectories_Pine()}}
}
