% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Hagglund_Lundmark_1979_Site_Index_Stand_Factors.R
\name{Hagglund_Lundmark_1979_Site_Index_Prediction}
\alias{Hagglund_Lundmark_1979_Site_Index_Prediction}
\title{Estimate Site Index H100 for Spruce or Pine in Sweden by stand factors.}
\source{
Corrected functions from Appendix II for estimation of site index
by site factors according to Hägglund and Lundmark (1977), in
Hägglund, B. (1979) Ett system för bonitering av skogsmark - analys,
kontroll och diskussion inför praktisk tillämpning. Rapport 14, Projekt
HUGIN. Skogsvetenskapliga fakulteten, Sveriges Lantbruksuniversitet.
Umeå. Sweden.

Hägglund, B., Lundmark, J-E. (1977) Skattning av höjdboniteten med
ståndortsfaktorer: Tall och gran i Sverige. Research Notes nr. 28. Dept.
Forest Ecology & Forest Soils. Royal College of Forestry. Stockholm. Sweden.
}
\usage{
Hagglund_Lundmark_1979_Site_Index_Prediction(
  species,
  vegetation,
  groundLayer,
  latitude,
  longitude,
  altitude,
  aspectMain = NA,
  inclinePercent = 0,
  soilMoisture,
  soilTexture,
  soilDepth,
  lateralWater,
  ditched,
  climateCode = NA,
  close2Coast = FALSE,
  LimesNorrlandicus = FALSE
)
}
\arguments{
\item{species}{One of "Picea abies" (default) or "Pinus sylvestris".}

\item{vegetation}{integer 1-18 see
[forester::Sweden_vegetation_types(layer="field)]}

\item{groundLayer}{integer 1-6 see
[forester::Sweden_vegetation_types(layer="ground")]}

\item{latitude}{decimal WGS84.}

\item{longitude}{decimal WGS84.}

\item{altitude}{Meters above sea level.}

\item{aspectMain}{Main aspect of the site. One of: NA (default), "North",
"East,"South","West".}

\item{inclinePercent}{Incline of slope, in percent. Default == 0.}

\item{soilMoisture}{integer 1-5 see
[forester::Sweden_soil_types(type="moisture")]}

\item{soilTexture}{integer 1-9 see
[forester::Sweden_soil_types(type="texture")]}

\item{soilDepth}{integer 1-4 see
[forester::Sweden_soil_types(type="depth")]}

\item{lateralWater}{integer 1-3 see
[forester::Sweden_soil_types(type="water")]}

\item{ditched}{boolean TRUE/FALSE (default) if plot affected by ditching.}

\item{climateCode}{One of: NA (default),"M1","M2","K1","K2","K3". see
[forester::Angstrom_1958_local_climate_Sweden()]}

\item{close2Coast}{boolean TRUE/FALSE (default) if plot is closer to coast than 50km. e.g. [forester::coast_distance()]}

\item{LimesNorrlandicus}{boolean TRUE/FALSE (default) if plot is north of Sernanders Limes Norrlandicus (used for Peat soils only.)}
}
\value{
Site Index H100 (meters).
Will return NA & issue a warning message if no method was found.
}
\description{
Estimate Site Index H100 for Spruce or Pine in Sweden by stand factors.
}
\details{
Uses corrected functions from appendix II in Hägglund & Lundmark
1979, except for when used on peat soils, in which case functions f. 8.1
& f. 8.2. pages 138-139 from Hägglund & Lundmark (1977) is used.
}
