% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hart_Becking_relative_spacing_index.R
\name{Hart_Becking_relative_spacing_index}
\alias{Hart_Becking_relative_spacing_index}
\title{Calculate the Hart-Becking Relative Spacing Index.}
\usage{
Hart_Becking_relative_spacing_index(stems_per_ha, dominant_height)
}
\arguments{
\item{stems_per_ha}{Stems per hectare}

\item{dominant_height}{Dominant height of the stand.}
}
\value{
Hart - Becking Spacing Index (percent).
}
\description{
Calculate the Hart-Becking Relative Spacing Index.
}
\examples{
Hart_Becking_relative_spacing_index(stems_per_ha = 1600,dominant_height =8.6 )
}
