% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heidarsson_2012_Iceland_Larix_sibirica.R
\name{Heidarsson_2012}
\alias{Heidarsson_2012}
\alias{Heidarsson_2012_height_trajectory_Iceland_Larix_sibirica}
\alias{Heidarsson_2012_single_tree_height_Iceland_Larix_sibirica}
\alias{Heidarsson_2012_diameter_increment_Iceland_Larix_sibirica}
\alias{Heidarsson_2012_tree_survival_Iceland_Larix_sibirica}
\title{Heidarsson & Pukkala 2012 Individual tree model for Larix sibirica on Iceland.}
\source{
Heidarsson, L. & Pukkala, T. (2012). Models for simulating the
temporal development of Siberian larch (**Larix sibirica**) plantations in
Hallormsstadur Iceland. Icel. Agric. Sci. 25:13-23. Available (2022-04-02):
[skemman.is](https://skemman.is/bitstream/1946/19764/1/Larus%20Heidarsson%20et%20al%202012.pdf)
}
\usage{
Heidarsson_2012_height_trajectory_Iceland_Larix_sibirica(
  dominant_height,
  age = 80,
  age2
)

Heidarsson_2012_single_tree_height_Iceland_Larix_sibirica(
  diameter_cm,
  dominant_height
)

Heidarsson_2012_diameter_increment_Iceland_Larix_sibirica(
  BA_m2_ha,
  BAL,
  diameter_cm,
  stand_age
)

Heidarsson_2012_tree_survival_Iceland_Larix_sibirica(BAL)
}
\arguments{
\item{dominant_height}{Dominant height of stand, in metres.}

\item{age}{Age at breast height, 1.3 m.}

\item{age2}{Age at breast height, 1.3 m at desired output}

\item{diameter_cm}{Diameter of tree at breast height 1.3m, in cm.}

\item{BA_m2_ha}{Basal Area of stand, m^2 / ha.}

\item{BAL}{Basal-area-in-larger-trees, m^2 / ha.}

\item{stand_age}{Age at breast height, 1.3 m.}
}
\value{
Dominant height at age2. (R^2 = 0.980, MSE= 0.173)

Single tree height, m. (R^2 = 0.963, MSE= 0.433)

5 year diameter increment of a tree, cm. (R^2=0.659, MSE= 0.352)

Probability of tree survival for the coming 5 years.
}
\description{
Heidarsson & Pukkala 2012 Individual tree model for Larix sibirica on Iceland.
}
