% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heiskanen_2018_SIS_Finland.R
\name{Heiskanen_2018_SIS_Finland}
\alias{Heiskanen_2018_SIS_Finland}
\title{Correlation between H100 and site factors in Finland}
\source{
Heiskanen J., Hallikainen V., Uusitalo J., Ilvesniemi H. (2018).
Co-variation relations of physical soil properties and site characteristics
of Finnish upland forests. Silva Fennica vol. 52 no. 3 article id 9948.
\url{https://doi.org/10.14214/sf.9948}
}
\usage{
Heiskanen_2018_SIS_Finland(
  site_type = 1,
  temperature_sum = 1425.5,
  fine_fraction_percentage = 32.74
)
}
\arguments{
\item{site_type}{Numeric. 1 (Default): Herb-rich forests (groves), 2: herb-rich heath forests
(OMT). 3: mesic heath forests (MT). 4. sub-xeric heath forests (VT). 5. xeric
heath forests (CT). 6. Barren heath forests. (ClT). \code{\link[=Finland_vegetation_types]{Finland_vegetation_types()}}}

\item{temperature_sum}{Degree days, at 2m, threshold 5 degrees Celsius. Min (807). Max (1746). Mean (1383.91). Median (1425.5).}

\item{fine_fraction_percentage}{Fine fraction (<0.063 mm, \%). Min (2.9). Max (92.57). Mean (32.74). Median (30.98).}
}
\value{
H100 Estimate. (tree-species independent).
}
\description{
OBSERVE: H100 is \emph{tree-species independent}.
Will assume that site type category 1 & 2 are aggregated & that
site type 6 (Barren heath forests) have equivalent H100 to site type 5.
}
\details{
Multiple R^2 = 0.829. adj.R^2 = 0.813. Residual SE= 2.75.
}
