% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hoffmann_1982_Tariff.R
\docType{data}
\name{Hoffmann_1982}
\alias{Hoffmann_1982}
\title{Tariff functions following Hoffmann 1982.}
\format{
A data frame with columns:
\describe{
\item{DBH_cm}{Diameter at breast height 1.3 metres for measured trees, in cm.}
\item{Volumem3}{Volume of the measured trees, in cu.m.}
}
}
\source{

}
\usage{
Hoffmann_1982
}
\description{
This data has been digitalised by hand after figure 2 from the
source. To work this out for yourself, look at the forester vignette:
vignette("Hoffmann_1982_Tariff.Rmd")
}
\keyword{datasets}
