% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUGIN_Thinning_Guide_Sweden.R
\name{HUGIN_Thinning_Guide_Sweden}
\alias{HUGIN_Thinning_Guide_Sweden}
\title{Thinning Recommendations according to the HUGIN Forest model.}
\usage{
HUGIN_Thinning_Guide_Sweden(dominantHeight, H100, species)
}
\arguments{
\item{dominantHeight}{dominant height of the stand (m).}

\item{H100}{Site Index H100 (m) according to Hägglund 1972, 1973. e.g.
[forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce],
[forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce],
[forester::Hagglund_1974_Sweden_height_trajectories_Pine]}

\item{species}{One of "Pinus sylvestris","Picea abies","Pinus contorta",
"Quercus robur,"Fagus sylvatica" or "Deciduous".}
}
\value{
A list of the upper and lower limits for recommended basal area per
hectare before and after thinning.
}
\description{
Thinning Recommendations according to the HUGIN Forest model.
}
\details{
Based on coefficients used in Heureka
 StandWise v. 2.19. e.g. 2022-09-17.

 OBS! For Quercus robur, Fagus sylvatica and Deciduous,
 only two limits are available - unknown if these represent the lower or
 upper limit before and after thinning.
}
\examples{

data.frame(
  dominantHeight=seq(10,36,0.1)
) \%>\% group_by_all() \%>\% mutate(
  data=purrr::map(.x=dominantHeight,.f=HUGIN_Thinning_Guide_Sweden,36,"Picea abies")
) \%>\% unnest_wider(data) \%>\%
  ggplot(aes(x=dominantHeight))+
  geom_ribbon(aes(ymin=BeforeLowerLimit,ymax=BeforeUpperLimit),alpha=0.5)+
  geom_ribbon(aes(ymin=AfterLowerLimit,ymax=AfterUpperLimit),alpha=0.5)
}
