% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_dissection1.R
\name{id_injured}
\alias{id_injured}
\title{ID injured plants}
\usage{
id_injured(
  df,
  plantid,
  year,
  injuries,
  forbidden_values,
  monotonic_increase = FALSE,
  height
)
}
\arguments{
\item{df}{A data frame containing tidy field trial records, i.e. one row per plant and year.}

\item{plantid}{Unique identifier for each plant.}

\item{year}{The year of observation, or other sortable time.}

\item{injuries}{A vector list of variables containing potential injury categories.}

\item{forbidden_values}{A vector list of injury severities which should be marked as injured.}

\item{monotonic_increase}{Logical, uses the is_monotonic function to in a new variable 'monotonic_growth' mark all before a potential first break as TRUE.}

\item{height}{Numeric variable containing the height observations of plants.}
}
\value{
Returns the original data frame with an additional column for:
    the individual plants first observered year of injury, 'firstinjury'
    if the current observation is before or after the first observed injury, 'afterfirstinjury', 1  is equal to or after, 0 for strictly before.
}
\description{
ID injured plants
}
\examples{
#testframe expected
plantid <- rep(c(1,2,3,4,5), times=c(3,3,3,3,3))
year <- rep(1:3, length.out=length(plantid))
set.seed(42)
PrimaryInjury <- sample(c(NA,NA,NA,"Rust","Insect","Snow break"), 15, replace=TRUE)
SecondaryInjury <- rep(NA, length.out=length(plantid))
OtherInjury <- rep(NA, length.out=length(plantid))
height <- c(1,2,3,1,2,3,3,2,1,1,2,3,1,3,2)


expected <- data.frame(plantid,year,PrimaryInjury,SecondaryInjury, OtherInjury,height)

rm(OtherInjury)
rm(plantid)
rm(PrimaryInjury)
rm(SecondaryInjury)
rm(year)
rm(height)


selectlist(df=expected,plantid=plantid, year=year, injuries=c("PrimaryInjury","SecondaryInjury","OtherInjury"),
          forbidden_values=c("Rust","Insect","Snow break"), monotonic_increase = TRUE, height=height)
}
