% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_imputation.R
\name{impute_basic}
\alias{impute_basic}
\title{Imputes the dataset with median of column for numeric data and the class
'other' or most frequent observation for categorical features}
\usage{
impute_basic(
  data,
  na_indicators = c(""),
  categorical_imputation = "other",
  verbose = FALSE
)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{na_indicators}{A list containing the values that will be treated as NA
indicators. By default the list is c(''). WARNING Do not include NA or NaN,
as these are already checked in other criterion.}

\item{categorical_imputation}{A string value describing the imputation method
for categorical features. The user can choose from setting missing values as
'other' or the most frequent value from feature. The respective options are:
'other' or 'frequency'. By default set to 'other'.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
preprocessing process, if FALSE gives none.}
}
\value{
Imputed dataset.
}
\description{
Fast and simple imputation method treating categorical and numerical features
differently. Please, note that imputation is performed on all columns including
the target `y`, as we assume that it has no missing values or these were
handled by removal functions or the user.
}
