% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_imputation.R
\name{impute_knn}
\alias{impute_knn}
\title{Imputes all missing values with the KNN algorithm}
\usage{
impute_knn(data, na_indicators = c(""), k = 10)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{na_indicators}{A list containing the values that will be treated as NA
indicators. By default the list is c(''). WARNING Do not include NA or NaN,
as these are already checked in other criterion.}

\item{k}{An integer describing the number of nearest neighbours to use. By default
set to 10.}
}
\value{
Imputed dataset.
}
\description{
Imputes all missing values with the KNN algorithm
}
