% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_imputation.R
\name{impute_mice}
\alias{impute_mice}
\title{Imputes all missing values with the MICE (Multivariate Imputation by Chained Equations) algorithm}
\usage{
impute_mice(data, na_indicators = c(""), m = 5, verbose = FALSE)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{na_indicators}{A list containing the values that will be treated as NA
indicators. By default the list is c(''). WARNING Do not include NA or NaN,
as these are already checked in other criterion.}

\item{m}{An integer describing the number of multiple imputations to use.
By default set to 5.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
preprocessing process, if FALSE gives none.}
}
\value{
Imputed dataset.
}
\description{
Imputes all missing values with the MICE (Multivariate Imputation by Chained Equations) algorithm
}
