% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_monotonic.R
\name{is_monotonic}
\alias{is_monotonic}
\title{is_monotonic}
\usage{
is_monotonic(df, id, year, height)
}
\arguments{
\item{df}{Tidy dataframe with one row for each plantid and observation.}

\item{id}{Unique identifier variable for each plant.}

\item{year}{Year variable.}

\item{height}{Numeric height variable.}
}
\value{
Returns the original dataframe with an additional column, monotonic_growth: TRUE if not strictly decreasing.
}
\description{
is_monotonic
}
\examples{
data(expected)
is_monotonic(df=expected,id=plantid,year=year,height=height)


}
