% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Johansson_1996_height_trajectory_Sweden_Aspen.R
\name{Johansson_1996_height_trajectory_Sweden_Aspen}
\alias{Johansson_1996_height_trajectory_Sweden_Aspen}
\title{Height trajectory for European Aspen in Sweden from Johansson 1996.}
\source{
Johansson, T. (1996) Site Index Curves for European Aspen (Populus tremula L.)
 Growing on Forest Land of Different Soils in Sweden. Silva Fennica 30(4): 437-458.
 Available: \url{https://doi.org/10.14214/sf.a8503}
}
\usage{
Johansson_1996_height_trajectory_Sweden_Aspen(
  dominant_height,
  age,
  age2,
  model1 = FALSE,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Total age.}

\item{age2}{Total age at output age.}

\item{model1}{Default FALSE. If TRUE, uses model 1 from 1996 instead of recommendation from FAKTA SKOG 13?}

\item{output}{One of "SIH100","Equation" or "Height" (default).}
}
\value{
m.
}
\description{
Abstract:

"Growth data were collected from 40 European aspen (Populus tremula L.) stands
growing on eight localities in Sweden. The stands ranged in latitude from 56 to
66 degrees N. The mean age of the stands was 32 years (range, 12-63), the mean
stand density 1978 stems ha^-1 (range, 300-6000), and the mean diameter at breast height
(on bark) 17 cm (range, 8-34). Site index curves were constructed for total age.
...Curves fitted for H40 totla age have another shape than curves presented by other
Nordic studies. The curves from the present study have slower growth for young aspens
than curves from norwegian and finnish conditions. For 50-70-year-old aspen stands,
curves from the present study indicate taller heights than from Nordic studies."
}
