% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Johansson_2011_height_trajectory_Sweden_Poplar.R
\name{Johansson_2011_height_trajectory_Sweden_Poplar}
\alias{Johansson_2011_height_trajectory_Sweden_Poplar}
\title{Height trajectory for Poplar on former farmland in Sweden from Johansson 2011.}
\source{
Johansson, T. (2011) Site index curves for Poplar growing on former farmland in Sweden. Scandinavian Journal of Forest Research 26(2): 161-170. Available: \url{https://doi.org/10.1080/02827581.2010.543428}
}
\usage{
Johansson_2011_height_trajectory_Sweden_Poplar(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Total age.}

\item{age2}{Total age at output age.}

\item{output}{One of "SIH100","Equation" or "Height".}
}
\value{
m.
}
\description{
Abstract:

"Site index (SI) curves for H 20 (dominant height at 20 years total age) were constructed using growth data from 64 hybrid poplars (Populus spp.) growing in 33 stands planted on farmland in Sweden (Lat. 55–60°N). The mean age of the stands was 24 years (range 14–45), the mean density 993 stems ha−l (155–3493), and the mean diameter at breast height (outside bark) 24 cm (10–45). SI curves fitted for H 20 at total age were well in accordance with the SI curves presented by other studies. A number of dynamic equations for modeling top-height growth from total age were assessed. A generalized algebraic difference approach (ADA) model derived by Cieszewski performed best. The model explained 98% of the observed variation in height development and exhibited no apparent bias across the range of predicted site indices. SI was also examined in relation to soil type. Multiple samples were available for three types of soil: light clay, medium clay, and till. There were no significant differences between the soil types."
}
\examples{

#SIH50 40 m: 40 m at age 50.
Johansson_2011_height_trajectory_Sweden_Poplar(dominant_height = 40,age = 50,output="Equation")

}
