% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Johansson_2013_height_trajectory_Sweden_Hybrid_Aspen.R
\name{Johansson_2013_height_trajectory_Sweden_Hybrid_Aspen}
\alias{Johansson_2013_height_trajectory_Sweden_Hybrid_Aspen}
\title{Height trajectory for Hybrid Aspen, Populus tremula L x Populus tremuloides Michx., in Sweden from Johansson 1999.}
\source{
Johansson, T. (2013) A site dependent top height growth model for hybrid aspen. Journal of Forestry Research. 24. 691-698. Available: \url{https://link.springer.com/article/10.1007\%2Fs11676-013-0365-6}
}
\usage{
Johansson_2013_height_trajectory_Sweden_Hybrid_Aspen(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Total age.}

\item{age2}{Total age at output age.}

\item{output}{One of "SIH100","Equation" or "Height".}
}
\value{

}
\description{
Abstract:

"In this study height growth models for hybrid aspen were developed using
three growth equations. The mean age of the hybrid aspen was 21 years (range
 15−51 years) with a mean stand density of 946 stems ha-1 (87−2374) and a
 mean diameter at breast height (over bark) of 19.6 cm (8.5−40.8 cm). Site
 index was also examined in relation to soil type. Multiple samples were
 collected for three types of soil: light clay, medium clay and till. Site
 index curves were constructed using the col- lected data and compared with
 published reports. A number of dynamic equations were assessed for modeling
 top-height growth from total age. A Generalized Algebraic Difference
 Approach model derived by Cieszewski (2001) performed the best. This model
 explained 99\% of the observed variation in tree height growth and exhibited
 no apparent bias across the range of predicted site indices. There were no
 significant differences between the soil types and site indices."
}
\examples{

#SIH50 40 m: 40 m at age 50.
Johansson_2013_height_trajectory_Sweden_Hybrid_Aspen(dominant_height = 40,age = 50,output="Equation")

}
