% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Johansson_2013_height_trajectory_Sweden_Larch.R
\name{Johansson_2013_height_trajectory_Sweden_Larch}
\alias{Johansson_2013_height_trajectory_Sweden_Larch}
\title{Height trajectory for Larch in Sweden from FAKTA SKOG nr. 14 2013}
\source{
Johansson, U., Ekö, P-M, Elfving, B., Johansson, T., Nilsson, U. (2013). Nya höjdutvecklingskurvor för bonitering. FAKTA SKOG. nr. 14. Available: \url{https://www.slu.se/globalassets/ew/ew-centrala/forskn/popvet-dok/faktaskog/faktaskog13/faktaskog_14_2013.pdf}
}
\usage{
Johansson_2013_height_trajectory_Sweden_Larch(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Total age.}

\item{age2}{Total age at output age.}

\item{output}{One of "SIH100","Equation" or "Height".}
}
\description{
"Larch: Swedish height development curves for Larch have not been constructed earlier. Access to
new data series have made such possible. The curves for larch are based on 21 permanent experimental plots with
European Larch, 14 with Japanese larch, 31 with Siberian Larch, and 11 with Hybrid Larch. The curves
are suitable for use in the entire country and with any of the mentioned species."
}
\details{
77 plots.
130 observations.
}
