% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Johansson_2013_height_trajectory_Sweden_Oak.R
\name{Johansson_2013_height_trajectory_Sweden_Oak}
\alias{Johansson_2013_height_trajectory_Sweden_Oak}
\title{Height trajectory for Oak in Sweden from FAKTA SKOG nr. 14 2013}
\source{
Johansson, U., Ekö, P-M, Elfving, B., Johansson, T., Nilsson, U. (2013). Nya höjdutvecklingskurvor för bonitering. FAKTA SKOG. nr. 14. Available: \url{https://www.slu.se/globalassets/ew/ew-centrala/forskn/popvet-dok/faktaskog/faktaskog13/faktaskog_14_2013.pdf}
}
\usage{
Johansson_2013_height_trajectory_Sweden_Oak(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Total age.}

\item{age2}{Total age at output age.}

\item{output}{One of "SIH100","Equation" or "Height".}
}
\value{
m.
}
\description{
"Beech and Oak: The curves are based on permanent experimental plots in beech and Oak stands,
which are followed with repeat measurements. The same plots were used in the 1970's for develop
height development curves and production models for beech and oak stands, but longer time series
have been available this time."
}
\details{
29 plots.
46 observations.
}
