% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Johnsson_1953_volume_hybrid_aspen.R
\name{Johnsson_1953_volume_hybrid_aspen}
\alias{Johnsson_1953_volume_hybrid_aspen}
\title{Stem Volumes for Hybrid Aspen from Johnsson 1953.}
\source{
Johnsson H. (1953). Hybridaspens ungdomsutveckling och ett försök till framtidsprognos.
Svenska skogsvårdsföreningens tidsskrift. 51:73-96.

Formula copied from Rytter, L., Stener, L-G. (2014). Growth and thinning effects
during a rotation period of hybrid aspen in southern Sweden. Scand. J. For. Res. Vol 29. Issue 8. p. 747-756. \url{https://doi.org/10.1080/02827581.2014.968202}
}
\usage{
Johnsson_1953_volume_hybrid_aspen(diameter.cm, height.m)
}
\arguments{
\item{diameter.cm}{Diameter of tree at breast height in centimeters.}

\item{height.m}{Height of tree in meters.}
}
\value{
Stem Volume over bark in m^3.
}
\description{
Stem Volumes for Hybrid Aspen from Johnsson 1953.
}
