% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jonson_1914.R
\name{Jonson_1914_MAI_to_class}
\alias{Jonson_1914_MAI_to_class}
\title{Jonson 1914 classifier}
\source{
Jonson, Tor (1914). Om bonitering av skogsmark. Svenska skogsvårdsföreningens tidsskrift. Häfte 5. p. 369-392.
}
\usage{
Jonson_1914_MAI_to_class(bonitet)
}
\arguments{
\item{bonitet}{MAI, m3sk.}
}
\value{
Class according to Jonson (1914)
}
\description{
Class according to Jonson (1914). OBS class 9 is given for all sites MAI < 1.2 m3sk.
}
\examples{
Jonson_1914_MAI_to_class(5.3)
}
