% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jonson_1914.R
\name{Jonson_1914_stand_volume}
\alias{Jonson_1914_stand_volume}
\title{Expected volume in fully stocked Pine stands according to Jonson (1914)}
\source{
Jonson, T. (1914) Om boniteringen av skogsmark, eng: On the
estimation of the productivity of forest land. Skogsvårdsföreningens
tidsskrift, nr. 5, p. 379.
}
\usage{
Jonson_1914_stand_volume(
  stand_mean_height,
  species = "Pinus\\n                                      sylvestris"
)
}
\arguments{
\item{stand_mean_height}{metres.}

\item{species}{"Pinus sylvestris" or "Picea abies".}
}
\value{
Stand volume, cubic metres.
}
\description{
Jonson recommends Picea abies only be used as species as
certain support with expressed reservation.
}
