% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jonson_1914_thinned_percentage.R
\name{Jonson_1914_thinned_percentage}
\alias{Jonson_1914_thinned_percentage}
\title{Tor Jonson's smoothing of Maass' experience tables.}
\source{
Jonson, T. (1914) Om boniteringen av skogsmark, eng: On the
estimation of the productivity of forest land. Skogsvårdsföreningens
tidsskrift, nr. 5, p. 382.
}
\usage{
Jonson_1914_thinned_percentage(stand_age)
}
\arguments{
\item{stand_age}{years.}
}
\value{
Percent thinned of total volume.
}
\description{
A smoothing of the experience tables from Maass on the at a
stand age thinned volume as a percentage of the total growth.
}
