% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jonson_1914_to_MAImax.R
\name{Jonson_1914_to_MAImax}
\alias{Jonson_1914_to_MAImax}
\title{Maximum Mean Annual Increment expected from Jonson Site Class (1-9)}
\usage{
Jonson_1914_to_MAImax(jonson_class)
}
\arguments{
\item{jonson_class}{Integer 1 - 9}
}
\value{
Maximum mean annual increment m3sk/yr.
}
\description{
Translates the Jonson site class, an integer 1 - 8. 9 designates impediment.
OBS: The MAI for impediment (class 9) has been set to 0.5.

Can also be roughly delineated according to \url{https://www.heurekaslu.se/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf} p. 5.
MaiMax = 8 * 0.75^(jonson_class-2)
}
