% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jonson_1914.R
\name{Jonson_1914_total_production}
\alias{Jonson_1914_total_production}
\title{Total production including thinned volume according to Jonson (1914)}
\usage{
Jonson_1914_total_production(stand_age, stand_mean_height, species)
}
\arguments{
\item{stand_age}{Years}

\item{stand_mean_height}{Mean Height, meters.}

\item{species}{"Pinus sylvestris" or "Picea abies".}
}
\value{
List, with element 1: Total production in cubic metres; element 2:
Mean Annual Increment.
}
\description{
The author expresses that "Picea abies" only is offered to
provide certain practical support, with expressed reservation.
}
