% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jonsson_1980.R
\name{Jonsson_1980_bark_diameter_over_bark_Birch_5_200}
\alias{Jonsson_1980_bark_diameter_over_bark_Birch_5_200}
\title{Bark function for Birch between 5-200 yr with diameter over bark,
from Jonsson (1980)}
\source{
Jonsson, B. (1980) "Functions for the long-term forecasting of the
size and structure of timber yields". Report #7. Section of Forest Mensuration
and Management. Swedish University of Agricultural Sciences.
ISSN: 0349-2133. ISBN: 91-576-0477-0. Umeå. p. 114.
}
\usage{
Jonsson_1980_bark_diameter_over_bark_Birch_5_200(
  jonson_bonitet,
  diameter_at_breast_height,
  age_at_breast_height,
  latitude,
  altitude
)
}
\arguments{
\item{jonson_bonitet}{3-7+ are available.}

\item{diameter_at_breast_height}{Diameter at breast height over bark in cm}

\item{age_at_breast_height}{Age at breast height.}

\item{latitude}{Latitude, degrees.}

\item{altitude}{Meters above sea level.}
}
\value{
mm, Double bark thickness.
}
\description{
Written to be used with the Swedish National Forest Inventory's 6.64 m radius
plots.
}
\details{
This function originally for ln( increment). Result has been re-transformed.
Log. bias assumed to be included in written coefficient.

Standard Deviation: 0.244
Number of trees: 4322
Coefficient of multiple correlation: 0.85
}
