% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Jonsson_1980_five_year_increment_under_bark_Birch_51_200.R,
%   R/Jonsson_1980_five_year_increment_under_bark_Birch_5_60.R
\name{Jonsson_1980_five_year_increment_under_bark_Birch_5_60}
\alias{Jonsson_1980_five_year_increment_under_bark_Birch_5_60}
\title{Five years diameter increment under bark, mm for Birch between 51-200 yr,
from Jonsson (1980)}
\source{
Jonsson, B. (1980) "Functions for the long-term forecasting of the
size and structure of timber yields". Report #7. Section of Forest Mensuration
and Management. Swedish University of Agricultural Sciences.
ISSN: 0349-2133. ISBN: 91-576-0477-0. Umeå. p. 68.

Jonsson, B. (1980) "Functions for the long-term forecasting of the
size and structure of timber yields". Report #7. Section of Forest Mensuration
and Management. Swedish University of Agricultural Sciences.
ISSN: 0349-2133. ISBN: 91-576-0477-0. Umeå. p. 66.
}
\usage{
Jonsson_1980_five_year_increment_under_bark_Birch_5_60(
  jonson_bonitet,
  diameter_at_breast_height,
  age_at_breast_height,
  diameter_of_largest_tree_on_plot,
  cutting_period,
  plot_basal_area_under_bark
)

Jonsson_1980_five_year_increment_under_bark_Birch_5_60(
  jonson_bonitet,
  diameter_at_breast_height,
  age_at_breast_height,
  diameter_of_largest_tree_on_plot,
  cutting_period,
  plot_basal_area_under_bark
)
}
\arguments{
\item{jonson_bonitet}{1-6 are available.}

\item{diameter_at_breast_height}{Diameter at breast height under bark in cm,}

\item{age_at_breast_height}{Age at breast height.}

\item{cutting_period}{"a", if cutting during first half of five year period;
"b", if cutting during second half of five year period.}

\item{plot_basal_area_under_bark}{Basal area of plot. (under bark).}
}
\value{
mm, diameter increment during five years.

mm, diameter increment during five years.
}
\description{
Written to be used with the Swedish National Forest Inventory's 6.64 m radius
plots.

Written to be used with the Swedish National Forest Inventory's 6.64 m radius
plots.
}
\details{
This function originally for ln( increment). Result has been re-transformed.
Log. bias assumed to be included in written coefficient.

Standard Deviation: 0.7274
Number of trees: 2091
Coefficient of multiple correlation: 0.54

This function originally for ln( increment). Result has been re-transformed.
Log. bias assumed to be included in written coefficient.

Standard Deviation: 0.6284
Number of trees: 1932
Coefficient of multiple correlation: 0.66
}
