% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Jonsson_1980_five_year_increment_under_bark_Pine_5_60.R
\name{Jonsson_1980_five_year_increment_under_bark_Pine_5_60}
\alias{Jonsson_1980_five_year_increment_under_bark_Pine_5_60}
\title{Five years diameter increment under bark, mm for Scots Pine between 5-60 yr,
from Jonsson (1980)}
\source{
Jonsson, B. (1980) "Functions for the long-term forecasting of the
size and structure of timber yields". Report #7. Section of Forest Mensuration
and Management. Swedish University of Agricultural Sciences.
ISSN: 0349-2133. ISBN: 91-576-0477-0. Umeå. p. 54.
}
\usage{
Jonsson_1980_five_year_increment_under_bark_Pine_5_60(
  jonson_bonitet,
  diameter_at_breast_height,
  age_at_breast_height,
  diameter_of_largest_tree_on_plot,
  cutting_period,
  plot_basal_area_under_bark,
  latitude,
  altitude
)
}
\arguments{
\item{jonson_bonitet}{3-7 are available.}

\item{diameter_at_breast_height}{Diameter at breast height under bark in cm,}

\item{age_at_breast_height}{Age at breast height.}

\item{diameter_of_largest_tree_on_plot}{Diameter of thickest tree on plot.}

\item{cutting_period}{"a", if cutting during first half of five year period;
"b", if cutting during second half of five year period.}

\item{plot_basal_area_under_bark}{Basal area of plot. (under bark).}

\item{latitude}{Latitude, degrees.}

\item{altitude}{Meters above sea level.}
}
\value{
mm, diameter increment during five years.
}
\description{
Written to be used with the Swedish National Forest Inventory's 6.64 m radius
plots.
}
\details{
This function originally for ln( increment). Result has been re-transformed.
Log. bias assumed to be included in written coefficient.

Standard Deviation: 0.4719
Number of trees: 4411
Coefficient of multiple correlation: 0.73
}
