% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jonsson_1980_height_Spruce_51_75.R
\name{Jonsson_1980_height_Spruce_51_75}
\alias{Jonsson_1980_height_Spruce_51_75}
\title{Height function for Norway Spruce between 51-75 yr,
from Jonsson (1980)}
\source{
Jonsson, B. (1980) "Functions for the long-term forecasting of the
size and structure of timber yields". Report #7. Section of Forest Mensuration
and Management. Swedish University of Agricultural Sciences.
ISSN: 0349-2133. ISBN: 91-576-0477-0. Umeå. p. 86.
}
\usage{
Jonsson_1980_height_Spruce_51_75(
  jonson_bonitet,
  diameter_at_breast_height,
  diameter_of_largest_tree_on_plot,
  age_at_breast_height,
  latitude,
  altitude
)
}
\arguments{
\item{jonson_bonitet}{1-7 are available.}

\item{diameter_at_breast_height}{Diameter at breast height above bark in cm}

\item{diameter_of_largest_tree_on_plot}{Diameter of thickest tree on plot.}

\item{age_at_breast_height}{Age at breast height.}

\item{latitude}{Latitude, degrees.}

\item{altitude}{Meters above sea level.}
}
\value{
mm, diameter increment during five years.
}
\description{
Written to be used with the Swedish National Forest Inventory's 6.64 m radius
plots.
}
\details{
This function originally for ln( increment). Result has been re-transformed.
Log. bias assumed to be included in written coefficient.

Standard Deviation: 0.1319
Number of trees: 5521
Coefficient of multiple correlation: 0.93
}
