% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jonsson_1980_self_thinning.R
\name{Jonsson_1980_self_thinning}
\alias{Jonsson_1980_self_thinning}
\title{Jonsson (1980) self thinning function}
\source{
Jonsson, B. (1980) "Functions for the long-term forecasting of the
size and structure of timber yields". Report #7. Section of Forest Mensuration
and Management. Swedish University of Agricultural Sciences.
ISSN: 0349-2133. ISBN: 91-576-0477-0. Umeå. p. 120.
}
\usage{
Jonsson_1980_self_thinning(
  stem_number_per_ha,
  jonson_bonitet,
  annual_basal_area_increment_percent
)
}
\arguments{
\item{stem_number_per_ha}{Number of stems per hectare.}

\item{jonson_bonitet}{numeric. Jonson site classes 1-7.}

\item{annual_basal_area_increment_percent}{Annual basal area increment, in percent.}
}
\value{
Self thinned volume in percent of the annual volume increment.
}
\description{
Jonsson (1980) self thinning function
}
\details{
Standard deviation = 30.3
Multiple correlation coefficient = 0.60
Material: Untouched parcels on permanent trial sites, 127 revisions.
}
\examples{
self_thinning_df <- data.frame("Bonitet"=c(1,1,1,1,1,





}
