% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jonsson_to_MAImax.R
\name{Jonsson_to_MAImax}
\alias{Jonsson_to_MAImax}
\title{Maximum Mean Annual Increment expected from Jonsson Site Index (1-9)}
\usage{
Jonsson_to_MAImax(Jonsson_si)
}
\arguments{
\item{Jonsson_si}{}
}
\value{

}
\description{
Translates the Jonsson site index, an integer 1 - 8. 9 designates impediment.

Can also be roughly delineated.
MaiMax = 8 * 0.75^(Jonsson_si-2)
}
