% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Karlberg_1961_douglas_sitka_HL_southern_scandinavia.R
\name{Karlberg_1961_height_trajectory_HL_Scandinavia_Douglas_Fir}
\alias{Karlberg_1961_height_trajectory_HL_Scandinavia_Douglas_Fir}
\title{Lorey's Mean Height (HL) development in stands of Douglas fir in southern
Scandinavia}
\source{
Karlberg, S. (1961) Development and yield of Douglas fir (Pseudotsuga
 taxifolia (Poir.) Britt.) and Sitka spruce (Picea sitchensis (Bong.) Carr.)
 in southern Scandinavia and on the Pacific Coast. Diss. Royal College of
 Forestry. Nr. 34. 141 pages. Emil Kihlströms Tryckeri AB. Stockholm.
 Available Online (2022-04-03):
 \url{https://pub.epsilon.slu.se/9527/1/kungl_skogshogskolans_skrifter_034.pdf}
}
\usage{
Karlberg_1961_height_trajectory_HL_Scandinavia_Douglas_Fir(age, age2, HL)
}
\arguments{
\item{age}{Observed stand age (total)}

\item{age2}{Desired stand age (total)}

\item{HL}{Observed Lorey's Mean Height (HL)}
}
\value{
Predicted HL at age2.
}
\description{
HL was chosen to be able to appropriately compared with Möller
 (1933).

 Site classes were therefore chosen as follows:
 \tabular{lr}{
 Site Class \tab HL50 \cr
 I \tab 32 (30.5-33.5) \cr
 II \tab 29 (27.5-30.5) \cr
 III \tab 26 (24.5-27.5) \cr
 IV \tab 23 (21.5-24.5) \cr
 }
}
