% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Karlsson_2000_height_trajectories_Finland.R
\name{Karlsson_2000_height_trajectory_Finland_coastal_Norway_Spruce}
\alias{Karlsson_2000_height_trajectory_Finland_coastal_Norway_Spruce}
\title{Height trajectories of Coastal Norway Spruce in Finland}
\source{
Karlsson, K. (2000) Height growth patterns of Scots pine and Norway
spruce in the coastal areas of western Finland. For. Ecol. Manage. 135:1-3,
pp. 205-216. DOI: \url{https://doi.org/10.1016/S0378-1127(00)00311-X}
}
\usage{
Karlsson_2000_height_trajectory_Finland_coastal_Norway_Spruce(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of the stand, in metres.}

\item{age}{Age of stand (breast height)}

\item{age2}{Age of stand at output (breast height).}

\item{output}{One of 'Height' (default) or 'Equation'.}
}
\value{
Dominant height at age2, or an equation (character).
}
\description{
Height trajectories of Coastal Norway Spruce in Finland
}
