% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kuehne_2022.R
\name{Kuehne_2022_BA_Norway_Scots_Pine}
\alias{Kuehne_2022_BA_Norway_Scots_Pine}
\title{Basal area in even aged stands of Scots Pine in Norway}
\source{
Kuehne C., McLean J.P., Maleki K., Antón-Fernández C., Astrup R.
(2022). A stand-level growth and yield model for thinned and unthinned
even-aged Scots pine forests in Norway. Silva Fennica vol. 56 no. 1 article
id 10627. \url{https://doi.org/10.14214/sf.10627}
}
\usage{
Kuehne_2022_BA_Norway_Scots_Pine(
  SI40,
  stems_ha,
  stems_ha2,
  BA,
  BA_before_thinning,
  BA_removed,
  age,
  age2,
  age_thin,
  dominant_height,
  dominant_height2
)
}
\arguments{
\item{SI40}{Dominant height of stand at a total age of 40 years.}

\item{stems_ha}{Stem density per hectare.}

\item{stems_ha2}{Stem density per hectare at time point 2.}

\item{BA}{Basal Area at time point 1.}

\item{BA_before_thinning}{Stand Basal Area before thinning m2/ha}

\item{BA_removed}{Stand Basal Area removed by thinning m2/ha}

\item{age}{Total stand age.}

\item{age2}{Total stand age at time point 2.}

\item{age_thin}{Total stand age at time of thinning.}

\item{dominant_height}{Dominant height at time point 1.}

\item{dominant_height2}{Dominant height at time point 2.}
}
\value{
Basal area m2 ha.
}
\description{
Mean error: 0.02192, Mean absolute error: 0.74455, relative MAE: 2.71853
}
