% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Kuehne_2022_height_trajectory_Norway_Scots_Pine.R
\name{Kuehne_2022_height_trajectory_Norway_Scots_Pine}
\alias{Kuehne_2022_height_trajectory_Norway_Scots_Pine}
\title{Height trajectory for Scots Pine in Norway from Kuehne et al 2022.}
\source{
Kuehne C., McLean J.P., Maleki K., Antón-Fernández C.,
 Astrup R. (2022). A stand-level growth and yield model for thinned and
 unthinned even-aged Scots pine forests in Norway. Silva Fennica vol. 56
  no. 1 article id 10627. \url{https://doi.org/10.14214/sf.10627}
}
\usage{
Kuehne_2022_height_trajectory_Norway_Scots_Pine(
  age,
  age2,
  dominant_height,
  output = "Height"
)
}
\arguments{
\item{age}{Total stand age}

\item{age2}{Total stand age at output.}

\item{dominant_height}{Dominant height, metres.}

\item{output}{One of "SIH100","Equation" or "Height" (default).}
}
\value{
One of "SIH100","Equation" or "Height".
}
\description{
Mean error: 0.01458, Mean absolute error: 0.34876, relative MAE: 2.18346
}
\details{
Compared to Tveite (1976) [forester::Tveite_1976_height_trajectory_Norway_Pine()]
and Sharma (2011), estimates a higher Site Index.
}
