% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kuehne_2022.R
\name{Kuehne_2022_stem_density_Norway_Scots_Pine}
\alias{Kuehne_2022_stem_density_Norway_Scots_Pine}
\title{Stem density in even aged stands of Scots Pine in Norway}
\source{
Kuehne C., McLean J.P., Maleki K., Antón-Fernández C., Astrup R.
(2022). A stand-level growth and yield model for thinned and unthinned
even-aged Scots pine forests in Norway. Silva Fennica vol. 56 no. 1 article
id 10627. \url{https://doi.org/10.14214/sf.10627}
}
\usage{
Kuehne_2022_stem_density_Norway_Scots_Pine(
  SI40,
  stems_ha,
  BA_before_thinning,
  BA_after_thinning,
  age,
  age2
)
}
\arguments{
\item{SI40}{Dominant height of stand at a total age of 40 years.}

\item{stems_ha}{Stem density per hectare.}

\item{BA_before_thinning}{Stand Basal Area before thinning m2/ha}

\item{BA_after_thinning}{Stand Basal Area after thinning m2/ha}

\item{age}{Total stand age.}

\item{age2}{Total stand age at time point 2.}
}
\value{
Stem density.
}
\description{
Mean error: -0.11108, Mean absolute error: 46.25402, relative MAE: 2.74474
}
