% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kuehne_2022_volume_Norway_Scots_Pine.R
\name{Kuehne_2022_volume_Norway_Scots_Pine}
\alias{Kuehne_2022_volume_Norway_Scots_Pine}
\title{Volume in even-aged stands of Scots Pine in Norway.}
\source{
Kuehne C., McLean J.P., Maleki K., Antón-Fernández C., Astrup R.
(2022). A stand-level growth and yield model for thinned and unthinned
even-aged Scots pine forests in Norway. Silva Fennica vol. 56 no. 1 article
id 10627. \url{https://doi.org/10.14214/sf.10627}
}
\usage{
Kuehne_2022_volume_Norway_Scots_Pine(
  BA,
  BA_after_thinning,
  BA_before_thinning,
  age_thin,
  age,
  dominant_height
)
}
\arguments{
\item{BA}{Basal Area m2/ha.}

\item{BA_after_thinning}{Basal Area after thinning m2/ha}

\item{BA_before_thinning}{Basal Area before thinning m2 / ha.}

\item{age_thin}{Total stand age at thinning.}

\item{age}{Total stand age}

\item{dominant_height}{Dominant height of stand.}
}
\value{
Volume.
}
\description{
Mean error: 0.13014, Mean absolute error: 3.92983, relative MAE: 1.97324
}
