% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kuehne_2022_quotients_Norway_Scots_Pine.R
\name{Kuehne_quotient}
\alias{Kuehne_quotient}
\alias{Kuehne_2022_stems_quotient_Norway_Scots_Pine}
\alias{Kuehne_2022_BA_quotient_Norway_Scots_Pine}
\title{Reduction in number of trees and basal area in even-aged stands of Scots Pine in Norway}
\source{
Kuehne C., McLean J.P., Maleki K., Antón-Fernández C., Astrup R.
(2022). A stand-level growth and yield model for thinned and unthinned
even-aged Scots pine forests in Norway. Silva Fennica vol. 56 no. 1 article
id 10627. \url{https://doi.org/10.14214/sf.10627}
}
\usage{
Kuehne_2022_stems_quotient_Norway_Scots_Pine(BA_before, BA_after)

Kuehne_2022_BA_quotient_Norway_Scots_Pine(stems_after, stems_before)
}
\arguments{
\item{BA_before}{Basal area before thinning, m2 / ha}

\item{BA_after}{Basal area after thinning, m2 / ha}

\item{stems_after}{Number of stems per ha after thinning.}

\item{stems_before}{Number of stems per ha before thinning}
}
\value{
Quotient
}
\description{
Mean error: -0.00041, Mean absolute error: 0.03386, relative MAE: 5.17959
}
