% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laasasenaho_1975_diameter_stump_height.R
\name{Laasasenaho_1975_diameter_stump_height}
\alias{Laasasenaho_1975_diameter_stump_height}
\title{Laasasenaho 1975 diameter at stump height}
\source{
Laasasenaho, J. 1975. Runkopuun saannon riippuvuus kannon korkeudesta
 ja latvan katkaisuläpimitasta. Summary: Dependence of the amount of
 harvestable timber upon the stump height and the top-logging diameter.
 Folia Forestalia 233. 20 p.
}
\usage{
Laasasenaho_1975_diameter_stump_height(diameter_cm)
}
\arguments{
\item{diameter_cm}{Diameter of the tree (cm) at breast height, 1.3 m.}
}
\value{
Diameter at Stump height (cm)
}
\description{
Laasasenaho 1975 diameter at stump height
}
