% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laasasenaho_1982_root_collar_Finland.R
\name{Laasasenaho_1982_root_collar_Finland}
\alias{Laasasenaho_1982_root_collar_Finland}
\title{Height for uppermost root collar from Laasasenaho 1982.}
\usage{
Laasasenaho_1982_root_collar_Finland(
  diameter_cm,
  height_m,
  species = "Picea abies"
)
}
\arguments{
\item{diameter_cm}{Diameter at breast height 1.3 m in cm.}

\item{height_m}{Height of tree, in metres.}

\item{species}{One of "Picea abies" (default), "Pinus sylvestris",
"Betula pendula"
or "Betula pubescens"}
}
\value{
Height of the uppermost root collar (cm)
}
\description{
"When volume is calculated by means of a taper curve,
the stump height is needed for defining the integration limit.
}
\details{
Pine relative SE = 68 percent.

Spruce relative SE = 38 percent.

Birch relative SE = 44 percent.
}
