% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laasasenaho_1982_tree_volume_2_Finland.R
\name{Laasasenaho_1982_tree_volume_2_Finland}
\alias{Laasasenaho_1982_tree_volume_2_Finland}
\title{Volume curves from diameter only by Laasasenaho 1982.}
\source{
Laasasenaho, J. 1982. Taper curve and volume functions for Pine, Spruce and Birch. Commun. Inst. For. Fenn. 108:1-74. p. 42.
}
\usage{
Laasasenaho_1982_tree_volume_2_Finland(
  diameter_cm,
  height_m,
  bark = TRUE,
  species = "Picea abies"
)
}
\arguments{
\item{diameter_cm}{Diameter at breast height 1.3 m in cm.}

\item{height_m}{Height of the tree in metres.}

\item{bark}{TRUE (default) for volume including bark, FALSE for volume under bark. If FALSE, diameter_cm is assumed to be under bark.}

\item{species}{One of "Picea abies" (default), "Pinus sylvestris",
"Betula pendula"
or "Betula pubescens"}
}
\value{
Volume of the tree (dm^3)
}
\description{
"The accuracy of equations based on diameter at breast height
alone is in any case poor, and the equations are material specific.
In some inventory tasks guideline volume curves, based on diameter only,
are required. The following equations....can be used for this type of task."
}
\details{
Pine relative SE with bark (without bark) = 7.10 (7.5) percent.


Spruce relative SE = 7.47 (8.1) percent.


Birch relative SE = 8.23 (8.4) percent.
}
\seealso{
For diameter only function, see [forester::Laasasenaho_1982_tree_volume_Finland()]
}
