% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laasasenaho_1982_tree_volume_Finland.R
\name{Laasasenaho_1982_tree_volume_Finland}
\alias{Laasasenaho_1982_tree_volume_Finland}
\title{Guideline volume curves from diameter only by Laasasenaho 1982.}
\usage{
Laasasenaho_1982_tree_volume_Finland(diameter_cm, species = "Picea abies")
}
\arguments{
\item{diameter_cm}{Diameter at breast height 1.3 m in cm.}

\item{species}{One of "Picea abies" (default), "Pinus sylvestris",
"Betula pendula"
or "Betula pubescens"}
}
\value{
Volume of the tree (dm^3)
}
\description{
"The accuracy of equations based on diameter at breast height
alone is in any case poor, and the equations are material specific.
In some inventory tasks guideline volume curves, based on diameter only,
are required. The following equations....can be used for this type of task."
}
\details{
Pine relative SE = 17.2 percent.


Spruce relative SE = 18.7 percent.


Birch relative SE = 18.8 percent.
}
\seealso{
For diameter and height function, see [forester::Laasasenaho_1982_tree_volume_2_Finland()]
}
