% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larix_vol_carbonnier_1954.R
\name{larix_sibirica_vol_carbonnier_1954}
\alias{larix_sibirica_vol_carbonnier_1954}
\title{Volume estimation of Larix sibirica}
\usage{
larix_sibirica_vol_carbonnier_1954(
  dbh.cm,
  height.m,
  double.bark.mm,
  crown.base.height.m
)
}
\arguments{
\item{dbh.cm}{Diameter breast height (1.3 m) in centimetres.}

\item{height.m}{Height of tree, in metres.}

\item{double.bark.mm}{Double bark, in mm.}

\item{crown.base.height.m}{Height of crown base, in metres.}
}
\value{
Tree volume
}
\description{
Interface for Carbonnier 1954 function for volume of Larix sibirica
}
\details{
\strong{If double.bark.mm & crown.base.height.m available, will use:}

    0.04588*(dbh.cm^2)*height.m + 0.09006*(dbh.cm^2) - 0.008844*(dbh.cm^3) - 0.06460*dbh.cm*height.m + 0.1488*height.m - 0.001391*dbh.cm*height.m*double.bark.mm + 0.002654*(dbh.cm^2)*dbh.cm*crown.base.height.m

    \strong{otherwise:}

   (0.04801*(dbh.cm^2)*height.m) + (0.08886*(dbh.cm^2)) - (0.01012*(dbh.cm^3)) - (0.08406*dbh.cm*height.m)+(0.1972*height.m)
}
\examples{
larix_sibirica_vol_carbonnier_1954(dbh.cm=10, height=10, double.bark.mm = 9, crown.base.height.m = 5)
larix_sibirica_vol_carbonnier_1954(dbh.cm=10, height=10)
}
