% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leijon_1979_SI_translation_Pine_to_Spruce_1.R
\name{Leijon_1979_SI_translation_Pine_to_Spruce_1}
\alias{Leijon_1979_SI_translation_Pine_to_Spruce_1}
\title{Height trajectories of Norway Spruce and Scots Pine on equal sites.}
\source{
Leijon, B. (1979) Tallen och Granens produktion på lika ståndort. Redovisning till SJFR. SLU, Inst. f. skogsskötsel, intern rapport 100pp.

Funktion 7.2, p. 63.
}
\usage{
Leijon_1979_SI_translation_Pine_to_Spruce_1(SI_Pine)
}
\arguments{
\item{SI_Pine}{SI Pine H100, meters by Hägglund 1974.}
}
\value{
SI Spruce H100, meters by Hägglund 1972,1973.
}
\description{
Height trajectories of Norway Spruce and Scots Pine on equal sites.
}
\details{
Standard deviation, approximately reduced for observational error in H100: 0.1070.


Standard deviation of unlogarithmed residuals (dm) 38.4.
}
\seealso{
[forester::Leijon_1979_SI_translation_Pine_to_Spruce_3]
[forester::Leijon_1979_SI_translation_Pine_to_Spruce_2]
[forester::Leijon_1979_SI_translation_Spruce_to_Pine_1]
[forester::Leijon_1979_SI_translation_Spruce_to_Pine_2]
[forester::Leijon_1979_SI_translation_Spruce_to_Pine_3]
}
