% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leijon_1979_SI_translation_Pine_to_Spruce_3.R
\name{Leijon_1979_SI_translation_Pine_to_Spruce_3}
\alias{Leijon_1979_SI_translation_Pine_to_Spruce_3}
\title{Height trajectories of Norway Spruce and Scots Pine on equal sites.}
\source{
Leijon, B. (1979) Tallen och Granens produktion på lika ståndort. Redovisning till SJFR. SLU, Inst. f. skogsskötsel, intern rapport 100pp.

Function 7.6, p. 68.
}
\usage{
Leijon_1979_SI_translation_Pine_to_Spruce_3(
  SI_Pine,
  latitude,
  altitude,
  continental,
  dry,
  herbs
)
}
\arguments{
\item{SI_Pine}{SI Pine H100, meters by Hägglund 1974.}

\item{latitude}{Degrees N.}

\item{altitude}{Meters above sea level.}

\item{continental}{TRUE/FALSE Is the plot located in a locally continental climate. [forester::local_climate_sweden()]}

\item{dry}{TRUE/FALSE subsoil water depth >2 meters. SWE:"Torr" mark.}

\item{herbs}{TRUE/FALSE field strata of low- or rich-herbs. SWE: "Örttyper".}
}
\value{
SI Spruce H100, meters by Hägglund 1972,1973.
}
\description{
Height trajectories of Norway Spruce and Scots Pine on equal sites.
}
\details{
Standard deviation, approximately reduced for observational error in H100: 0.0782.


Standard deviation of unlogarithmed residuals (dm) 25.8.
}
\seealso{
[forester::Leijon_1979_SI_translation_Pine_to_Spruce_1]
[forester::Leijon_1979_SI_translation_Pine_to_Spruce_2]
[forester::Leijon_1979_SI_translation_Spruce_to_Pine_1]
[forester::Leijon_1979_SI_translation_Spruce_to_Pine_2]
[forester::Leijon_1979_SI_translation_Spruce_to_Pine_3]
}
