% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leijon_1979_SI_translation_Spruce_to_Pine_1.R
\name{Leijon_1979_SI_translation_Spruce_to_Pine_1}
\alias{Leijon_1979_SI_translation_Spruce_to_Pine_1}
\title{Height trajectories of Norway Spruce and Scots Pine on equal sites.}
\source{
Leijon, B. (1979) Tallen och Granens produktion på lika ståndort. Redovisning till SJFR. SLU, Inst. f. skogsskötsel, intern rapport 100pp.

Funktion 7.1, p. 63.
}
\usage{
Leijon_1979_SI_translation_Spruce_to_Pine_1(SI_Spruce)
}
\arguments{
\item{SI_Spruce}{SI Spruce H100, meters by Hägglund 1972,1973.}
}
\value{
SI Pine H100, meters by Hägglund 1974.
}
\description{
Height trajectories of Norway Spruce and Scots Pine on equal sites.
}
\details{
Standard deviation, approximately reduced for observational error in H100: 0.0549


Standard deviation of unlogarithmed residuals (dm) 18.1.
}
\seealso{
[forester::Leijon_1979_SI_translation_Pine_to_Spruce_3]
[forester::Leijon_1979_SI_translation_Pine_to_Spruce_2]
[forester::Leijon_1979_SI_translation_Spruce_to_Pine_1]
[forester::Leijon_1979_SI_translation_Spruce_to_Pine_2]
[forester::Leijon_1979_SI_translation_Spruce_to_Pine_3]
}
