% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexerod_2005_recruitment_prob_Norway.R
\name{Lexerod_2005}
\alias{Lexerod_2005}
\alias{Lexerod_2005_prob_recruitment_5cm_Norway_Norway_Spruce}
\alias{Lexerod_2005_prob_recruitment_5cm_Norway_Scots_Pine}
\alias{Lexerod_2005_prob_recruitment_5cm_Norway_Birch}
\alias{Lexerod_2005_prob_recruitment_5cm_Norway_Broadleaves}
\title{Probability of recruitment (5 cm) for a 5 year period in young norwegian stands.}
\source{
Lexerød, N., Eid, T. (2005) Recruitment models for Norway Spruce,
Scots Pine, Birch and Other Broadleaves in Young Growth Forests in Norway.
Silva Fenn. 39(3): 391-406. Available (2022-02-26):
\url{https://www.silvafennica.fi/pdf/article376.pdf}
}
\usage{
Lexerod_2005_prob_recruitment_5cm_Norway_Norway_Spruce(
  SI40,
  age,
  stems_ha,
  prop_spruce
)

Lexerod_2005_prob_recruitment_5cm_Norway_Scots_Pine(
  SI40,
  age,
  stems_ha,
  prop_pine
)

Lexerod_2005_prob_recruitment_5cm_Norway_Birch(
  age,
  stems_ha,
  prop_birch_broadleaves
)

Lexerod_2005_prob_recruitment_5cm_Norway_Broadleaves(SI40, age, stems_ha)
}
\arguments{
\item{SI40}{Site Index at base-age 40 for species in question. For Broadleaves
, use SI40 Birch (?). e.g.
[forester::Tveite_1976_height_trajectory_Norway_Pine()],
[forester::Tveite_1977_height_trajectory_Norway_Norway_Spruce()], Strand
(1967).}

\item{age}{Total age (?)}

\item{stems_ha}{Stems per hectare.}

\item{prop_spruce}{Proportion (Basal area? Stems?) of Spruce.}

\item{prop_pine}{Proportion (Basal area? Stems?) of Scots Pine.}

\item{prop_birch_broadleaves}{Sum of proportion of of birch and other
broadleaves. (BA% or stems?)}
}
\value{
Probability of recruitment.
}
\description{
Calculates the probability of occurring ingrowth into >5 cm
diameter breast height class) occuring during a 5 years period. These functions
are suitable for development class II (Conifers and broadleaves: max total
stand ages 20, 20, 25, 30, 35, 45, 55 for SI40 23, 20, 17, 14, 11, 8, 6;
Birch: max total stand ages 15, 15, 20, 25, 25, 25, 30 years for SI40 23, 20,
 17, 14, 11, 8, 6).
}
\details{
Norway spruce observations: 897. Misclassified (6.7 percent). Chi^2:
13.9.

Scots Pine observations: 897. Misclassified (7.9 percent). Chi^2:
5.7.

Birch observations: 897. Misclassified (6.9 percent). Chi^2:
9.2.

Broadleaf observations: 897. Misclassified (9 percent). Chi^2:
2.3.
}
