% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexerod_2005_ingrowth.R
\name{Lexerod_2005_count}
\alias{Lexerod_2005_count}
\alias{Lexerod_2005_recruitment_count_5cm_Norway_Norway_Spruce}
\alias{Lexerod_2005_recruitment_count_5cm_Norway_Scots_Pine}
\alias{Lexerod_2005_recruitment_count_5cm_Norway_Birch}
\alias{Lexerod_2005_recruitment_count_5cm_Norway_Broadleaves}
\title{Number of recruits to smallest diameter class (5 cm) for a 5 year period in
young norwegian stands.}
\source{
Lexerød, N., Eid, T. (2005) Recruitment models for Norway Spruce,
Scots Pine, Birch and Other Broadleaves in Young Growth Forests in Norway.
Silva Fenn. 39(3): 391-406. Available (2022-02-26):
\url{https://www.silvafennica.fi/pdf/article376.pdf}
}
\usage{
Lexerod_2005_recruitment_count_5cm_Norway_Norway_Spruce(
  SI40,
  age,
  stems_ha,
  percent_spruce
)

Lexerod_2005_recruitment_count_5cm_Norway_Scots_Pine(
  SI40,
  age,
  stems_ha,
  percent_pine
)

Lexerod_2005_recruitment_count_5cm_Norway_Birch(
  age,
  stems_ha,
  percent_birch_broadleaves
)

Lexerod_2005_recruitment_count_5cm_Norway_Broadleaves(
  altitude,
  latitude,
  SI40,
  age,
  stems_ha,
  percent_birch_broadleaves
)
}
\arguments{
\item{SI40}{Site Index at base-age 40 for species in question. For Broadleaves
, use SI40 Birch (?). e.g.
[forester::Tveite_1976_height_trajectory_Norway_Pine()],
[forester::Tveite_1977_height_trajectory_Norway_Norway_Spruce()], Strand
(1967).}

\item{age}{Stand age, breast height (1.3m)}

\item{stems_ha}{Stems per hectare.}

\item{percent_spruce}{Percent (Basal area? Stems?) of Spruce.}

\item{percent_pine}{Percent (Basal area? Stems?) of Scots Pine.}

\item{percent_birch_broadleaves}{Sum of Percent of of birch and other
broadleaves. (stems)}

\item{altitude}{meters above sea level.}

\item{latitude}{Degrees N.}
}
\value{
Number of recruits
}
\description{
Calculates the number of recruits into >5 cm
diameter breast height class) occuring during a 5 years period. These functions
are suitable for development class II (Conifers and broadleaves: max total
stand ages 20, 20, 25, 30, 35, 45, 55 for SI40 23, 20, 17, 14, 11, 8, 6;
Birch: max total stand ages 15, 15, 20, 25, 25, 25, 30 years for SI40 23, 20,
 17, 14, 11, 8, 6).
}
\details{
Norway spruce observations: 510. Adj.R^2 = 0.396. CV(percent): 13.8.

Scots Pine observations: 228. Adj.R^2 = 0.297. CV(percent): 16.0.

Birch observations: 436. Adj.R^2 = 0.147. CV(percent): 18.2.

Broadleaf observations: 166. Adj.R^2 = 0.159. CV(percent): 18.4.
}
