% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Liziniewicz_2016_height_trajectory_Sweden_Pinus_contorta.R
\name{Liziniewicz_2016_height_trajectory_Sweden_Pinus_contorta}
\alias{Liziniewicz_2016_height_trajectory_Sweden_Pinus_contorta}
\title{Site Index Estimations for P. contorta in Sweden, above 59th parallel.}
\usage{
Liziniewicz_2016_height_trajectory_Sweden_Pinus_contorta(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Total age.}

\item{age2}{Total age at output age.}

\item{output}{One of "SIH100","Equation" or "Height".}
}
\value{
If output is "SIH100", the numeric value for the height (m) a stand
will reach at age 100.

 If output is "Height", the numeric value for the height (m) a stand will
 reach at age_2.

 If output is "Equation", a named list with 2 elements: I) "Equation": text response with the equation for that height
 curve.
}
\description{
Provides interface for the site index function from "\emph{A site index model for lodgepole pine (Pinus contorta Dougl. var. latifolia) in northern Sweden}" (2016),
by Liziniewicz, M., Nilsson, U., Agestam, E., Ekö, P-M & Elfving, B.

"Lodgepole pines have only been cultivated in Sweden
for around 60 years. The experimental data set, therefore, did not contain
much information relating to stands older than 60 years, which may have affected the models'
estimates and means that it is impossible to be certain about the growth patterns
these stands will show as they age further. However, the available growth trajectories for
the oldest Swedish lodgepole pine stands provide no clear evidence of slowing height growth.
Stands aged between 15 and 40 years are well represented in the experimental  data set".

    \strong{Suitable ages for use by species}: P. contorta (20 - 50 yrs)
}
