% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_code_sweden.R
\name{local_climate_sweden}
\alias{local_climate_sweden}
\title{Sweden local climates}
\usage{
local_climate_sweden(latitude, longitude)
}
\arguments{
\item{latitude}{Latitude, decimal}

\item{longitude}{Longitude, decimal}
}
\value{
Local Swedish Climate Code.
}
\description{
Finds the local climate according to Ångström (1958)
\strong{OBSERVE! Latitude and Longitude should be provided in WGS84 reference system.}
}
