% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralHelpers.R
\name{Lorey_mean_height}
\alias{Lorey_mean_height}
\title{Lorey's mean height.}
\usage{
Lorey_mean_height(tree_height_m, basal_area_m2)
}
\arguments{
\item{tree_height_m}{numeric vector. Tree height in meters.}

\item{basal_area_m2}{numeric vector. Basal area in m2.}
}
\value{
Lorey's mean height.
}
\description{
Lorey's mean height is calculated by dividing the sum of the
product of the height and basal area of each tree by the total stand basal area.
}
\details{
This function can handle vectors.
}
\examples{
heights <- c(15,12,14,15,10)
ba <- c(0.4,0.5,0.6,0.5,0.5)

Lorey_mean_height(tree_height_m=heights,basal_area_m2=ba)

}
