% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lundqvist_1957_height_trajectory_Pine_Sweden.R
\name{Lundqvist_1957_height_trajectory_Sweden_Scots_Pine}
\alias{Lundqvist_1957_height_trajectory_Sweden_Scots_Pine}
\title{Height trajectory for Scots Pine in planted cultures in Sweden}
\source{
https://pub.epsilon.slu.se/10009/
}
\usage{
Lundqvist_1957_height_trajectory_Sweden_Scots_Pine(
  yearsToBreastHeight = 7,
  H100,
  age
)
}
\arguments{
\item{yearsToBreastHeight}{Years until 1.3 m.}

\item{H100}{Site Index at 100 years of age (total)}

\item{age}{Total age.}
}
\value{
Height, meters.
}
\description{
Height trajectory for Scots Pine in planted cultures in Sweden
}
