% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ranger.R
\name{make_ranger}
\alias{make_ranger}
\title{Automated function for explaining ranger model}
\usage{
make_ranger(data, target, type)
}
\arguments{
\item{data}{training set for ranger model. It can be data frame, matrix, data.table or dgCMatrix. Note: data should contain target column}

\item{target}{name of a target value. Should be character. Has to be in data frame}

\item{type}{specify weather it is classification task or regression. Should be one of these characters: "classif", "regr"}
}
\value{
An object of the class \code{explainer}.
}
\description{
Function \code{make_ranger} creates ranger model with default parameters using ranger library and
wraps it with the explainer.
}
\examples{
\donttest{
# regression
library(DALEX)
data(apartments, package="DALEX")

exp <- make_ranger(apartments, "m2.price", "regr")
# plot(model_parts(exp))

# binary classification
library(DALEX)
data(titanic_imputed, package="DALEX")
exp <- make_ranger(titanic_imputed, "survived", "classif")
# plot(model_parts(exp))
}
}
\references{
Explanatory Model Analysis. Explore, Explain, and Examine Predictive Models. \url{https://ema.drwhy.ai/}
}
