% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_XGBoost.R
\name{make_xgboost}
\alias{make_xgboost}
\title{Automated Function for Explaining XGBoost model}
\usage{
make_xgboost(data, target, type = "regression")
}
\arguments{
\item{data}{data.frame, matrix, data.table or dgCMatrix - data will be used to run the XGBoost model. NOTE: data has to contain the target column.}

\item{target}{character: name of the target column, placed in quotation marks "...". The input target name is compulsory to be one of column names of input data.}

\item{type}{character: defining the task, placed in quotation marks "...". Two options for type: "regression" and "classification", particularly, binary classification.}
}
\value{
An object of the class \code{explainer} for XGBoost model from given data, target and defined type of problem.
}
\description{
Function \code{make_xgboost} automates the process of applying XGBoost model
for a dataset and simutaneously creats explainer from the use of DALEX package.
The created explainer can be further processed by functions for explanations.
}
\examples{
# simple explainer for regression problem

library(datasets)
library(DALEX)

data("iris")
EXPLAINER <- make_xgboost(iris,"Petal.Length",type="regression")

## Variable importance
variable_importance_xgb <- model_parts(EXPLAINER,type="raw")
plot(variable_importance_xgb)

## Single variable:
sv_xgb_petal_width <- model_profile(EXPLAINER, variable="Petal.Width",type="partial")
plot(sv_xgb_petal_width)

# simple explainer for classification problem

library(DALEX)
## We will predict the survived state: 0 and 1
titanic_explainer <- make_xgboost(titanic_imputed, "survived", "classification")

## Assessment for model performance:
model_performance(titanic_explainer)

## Finding important features:
plot(model_parts(titanic_explainer))

## Plot reverse cumulative distribution of module of residual
plot(model_performance(titanic_explainer))
}
\references{
Explanatory Model Analysis. Explore, Explain and Examine Predictive Models. \url{https://ema.drwhy.ai/}
}
