% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maleki2022BA.R
\name{Maleki2022BA}
\alias{Maleki2022BA}
\alias{Maleki_2022_BA_m2_ha_Norway_Norway_Spruce}
\alias{Maleki_2022_BA_m2_ha_Norway_Scots_Pine}
\alias{Maleki_2022_BA_m2_ha_Norway_Broadleaves}
\title{Basal Area development for main species groups in Norway.}
\source{
Kobra Maleki, Rasmus Astrup, Christian Kuehne, J. Paul McLean &
Clara Antón-Fernández (2022) Stand-level growth models for long-term
projections of the main species groups in Norway, Scandinavian Journal
of Forest Research, DOI: \url{10.1080/02827581.2022.2056632}
}
\usage{
Maleki_2022_BA_m2_ha_Norway_Norway_Spruce(BA, age, age2, SI40, stems, stems2)

Maleki_2022_BA_m2_ha_Norway_Scots_Pine(BA, age, age2, SI40, stems, stems2)

Maleki_2022_BA_m2_ha_Norway_Broadleaves(BA, age, age2, SI40, stems, stems2)
}
\arguments{
\item{BA}{Basal Area, m^2 /ha.}

\item{age}{Total stand age.}

\item{age2}{Desired total stand age.}

\item{SI40}{Site Index H40 (see ?forester::Maleki2022height)}

\item{stems}{Stems at age.}

\item{stems2}{Stems at age2.}
}
\value{
Basal Area at age2.
}
\description{
Basal Area development for main species groups in Norway.
}
\details{
\tabular{lrrrr}{
Species \tab E \tab MAE \tab RMSE \tab R^2 \cr
Norway Spruce \tab -0.067 \tab 1.112 \tab 1.602 \tab 0.984 \cr
Scots Pine \tab 0.032  \tab 0.680 \tab 1.007 \tab 0.991 \cr
Broadleaves \tab -0.138 \tab 0.978 \tab 1.418 \tab 0.978 \cr
}
}
\seealso{
Other {Maleki2022}: 
\code{\link{Maleki2022Volume}},
\code{\link{Maleki2022height}},
\code{\link{Maleki2022ingrowth}},
\code{\link{Maleki2022stems2}},
\code{\link{Maleki2022stems}}
}
\concept{{Maleki2022}}
