% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maleki_2022.R
\name{Maleki2022ingrowth}
\alias{Maleki2022ingrowth}
\alias{Maleki_2022_ingrowth_count_Norway_Norway_Spruce}
\alias{Maleki_2022_ingrowth_count_Norway_Scots_Pine}
\alias{Maleki_2022_ingrowth_probability_Norway_Norway_Spruce}
\alias{Maleki_2022_ingrowth_probability_Norway_Scots_Pine}
\alias{Maleki_2022_ingrowth_probability_Norway_Broadleaves}
\title{Ingrowth for the main tree species groups of Norway.}
\source{
Kobra Maleki, Rasmus Astrup, Christian Kuehne, J. Paul McLean &
Clara Antón-Fernández (2022) Stand-level growth models for long-term
projections of the main species groups in Norway, Scandinavian Journal
of Forest Research, DOI: \url{10.1080/02827581.2022.2056632}
}
\usage{
Maleki_2022_ingrowth_count_Norway_Norway_Spruce(QMD)

Maleki_2022_ingrowth_count_Norway_Scots_Pine(BA)

Maleki_2022_ingrowth_count_Norway_Norway_Spruce(QMD)

Maleki_2022_ingrowth_probability_Norway_Norway_Spruce(QMD, stems)

Maleki_2022_ingrowth_probability_Norway_Scots_Pine(QMD, stems)

Maleki_2022_ingrowth_probability_Norway_Broadleaves(QMD, stems)
}
\arguments{
\item{QMD}{Quadratic Mean Diameter (cm).}

\item{BA}{Basal Area m^2 / ha.}

\item{stems}{Total stems per hectare.}
}
\value{
Probability and count of ingrowth
}
\description{
Ingrowth for the main tree species groups of Norway.
}
\details{
\tabular{lrrrr}{
Total stem density predicted by surviving + ingrowth:\tab\tab\tab\tab\cr
Species \tab E \tab MAE \tab RMSE \tab R^2 \cr
Norway Spruce \tab 2.319 \tab 76.001 \tab 114.614 \tab 0.971 \cr
Scots Pine \tab 1.556  \tab 43.963 \tab 69.482 \tab 0.980 \cr
Broadleaves \tab 2.32 \tab 98.054 \tab 143.261 \tab 0.963 \cr
}
}
\seealso{
Other {Maleki2022}: 
\code{\link{Maleki2022BA}},
\code{\link{Maleki2022Volume}},
\code{\link{Maleki2022height}},
\code{\link{Maleki2022stems2}},
\code{\link{Maleki2022stems}}
}
\concept{{Maleki2022}}
