% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maleki_2022.R
\name{Maleki2022stems}
\alias{Maleki2022stems}
\alias{Maleki_2022_stems_surviving_Norway_Norway_Spruce}
\alias{Maleki_2022_stems_surviving_Norway_Scots_Pine}
\alias{Maleki_2022_stems_surviving_Norway_Broadleaves}
\title{Number of surviving stems at the end of a period.}
\source{
Kobra Maleki, Rasmus Astrup, Christian Kuehne, J. Paul McLean &
Clara Antón-Fernández (2022) Stand-level growth models for long-term
projections of the main species groups in Norway, Scandinavian Journal
of Forest Research, DOI: \url{10.1080/02827581.2022.2056632}
}
\usage{
Maleki_2022_stems_surviving_Norway_Norway_Spruce(age, age2, stems, SI40)

Maleki_2022_stems_surviving_Norway_Scots_Pine(age, age2, stems, SI40)

Maleki_2022_stems_surviving_Norway_Broadleaves(age, age2, stems, SI40)
}
\arguments{
\item{age}{Total stand age.}

\item{age2}{Total stand age at time 2.}

\item{stems}{Stand stem density per hectare.}

\item{SI40}{Site Index H40 (see ?forester::Maleki2022height)}
}
\value{
The number of survived trees at the end of the period
}
\description{
Number of surviving stems at the end of a period.
}
\details{
\tabular{lrrrr}{
Species \tab E \tab MAE \tab RMSE \tab R^2 \cr
Norway Spruce \tab 2.626 \tab 48.467 \tab 71.777 \tab 0.988 \cr
Scots Pine \tab 1.561  \tab 26.192 \tab 37.531 \tab 0.994 \cr
Broadleaves \tab 2.117 \tab 57.370 \tab 82.531 \tab 0.986 \cr
}
}
\seealso{
Other {Maleki2022}: 
\code{\link{Maleki2022BA}},
\code{\link{Maleki2022Volume}},
\code{\link{Maleki2022height}},
\code{\link{Maleki2022ingrowth}},
\code{\link{Maleki2022stems2}}
}
\concept{{Maleki2022}}
