% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maleki_2022.R
\name{Maleki2022stems2}
\alias{Maleki2022stems2}
\alias{Maleki_2022_stem_density_Norway_Norway_Spruce}
\alias{Maleki_2022_stem_density_Norway_Scots_Pine}
\alias{Maleki_2022_stem_density_Norway_Broadleaves}
\title{Stems at age2}
\source{
Kobra Maleki, Rasmus Astrup, Christian Kuehne, J. Paul McLean &
Clara Antón-Fernández (2022) Stand-level growth models for long-term
projections of the main species groups in Norway, Scandinavian Journal
of Forest Research, DOI: \url{10.1080/02827581.2022.2056632}
}
\usage{
Maleki_2022_stem_density_Norway_Norway_Spruce(age, age2, stems, SI40)

Maleki_2022_stem_density_Norway_Scots_Pine(age, age2, stems, SI40)

Maleki_2022_stem_density_Norway_Broadleaves(age, age2, stems, SI40)
}
\arguments{
\item{age}{Total age of stand.}

\item{age2}{Total age of stand desired.}

\item{stems}{Stems at age.}

\item{SI40}{Site Index H40 (see ?forester::Maleki2022height)}
}
\value{
Number of stems at age2.
}
\description{
Stems at age2
}
\details{
\tabular{lrrrr}{
Species \tab E \tab MAE \tab RMSE \tab R^2 \cr
Norway Spruce \tab 10.928 \tab 75.122 \tab 115.755 \tab 0.970 \cr
Scots Pine \tab 4.924  \tab 41.354 \tab 69.141 \tab 0.981 \cr
Broadleaves \tab 12.065 \tab 94.132 \tab 142.359 \tab 0.963 \cr
}
}
\seealso{
Other {Maleki2022}: 
\code{\link{Maleki2022BA}},
\code{\link{Maleki2022Volume}},
\code{\link{Maleki2022height}},
\code{\link{Maleki2022ingrowth}},
\code{\link{Maleki2022stems}}
}
\concept{{Maleki2022}}
