% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maleki2022volume.R
\name{Maleki2022Volume}
\alias{Maleki2022Volume}
\alias{Maleki_2022_volume_m3_ha_Norway_Norway_Spruce}
\alias{Maleki_2022_volume_m3_ha_Norway_Scots_Pine}
\alias{Maleki_2022_volume_m3_ha_Norway_Broadleaves}
\title{Volume Estimation for the main tree species groups in Norway.}
\source{
Kobra Maleki, Rasmus Astrup, Christian Kuehne, J. Paul McLean &
Clara Antón-Fernández (2022) Stand-level growth models for long-term
projections of the main species groups in Norway, Scandinavian Journal
of Forest Research, DOI: \url{10.1080/02827581.2022.2056632}
}
\usage{
Maleki_2022_volume_m3_ha_Norway_Norway_Spruce(dominant_height2, BA2, age2)

Maleki_2022_volume_m3_ha_Norway_Scots_Pine(dominant_height2, BA2, age2)

Maleki_2022_volume_m3_ha_Norway_Broadleaves(dominant_height2, BA2, age2)
}
\arguments{
\item{dominant_height2}{Dominant height at end of a period.}

\item{BA2}{Basal Area m^2 / ha at the end of a period.}

\item{age2}{Stand total age at the end of a period.}
}
\value{
Volume cubic metres / hectare.
}
\description{
Observe, this function is used to calculate the standing volume
at the end of a period.
}
\details{
\tabular{lrrrr}{
Species \tab E \tab MAE \tab RMSE \tab R^2 \cr
Norway Spruce \tab 0.497 \tab 12.653 \tab 18.335 \tab 0.977 \cr
Scots Pine \tab 0.111  \tab 6.915 \tab 9.982 \tab 0.986 \cr
Broadleaves \tab 0.445 \tab 4.62 \tab 7.076 \tab 0.986 \cr
}
}
\seealso{
Other {Maleki2022}: 
\code{\link{Maleki2022BA}},
\code{\link{Maleki2022height}},
\code{\link{Maleki2022ingrowth}},
\code{\link{Maleki2022stems2}},
\code{\link{Maleki2022stems}}
}
\concept{{Maleki2022}}
