% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{manage_missing}
\alias{manage_missing}
\title{Manage missing values}
\usage{
manage_missing(df, y)
}
\arguments{
\item{df}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{y}{A string that indicates a target column name.}
}
\value{
A dataframe with removed and imputed missing values.
}
\description{
Manage missing values
}
