% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managementData.R
\name{managementData}
\alias{managementData}
\title{Create an empty managementData object.}
\usage{
managementData(stand_id, tree_id, species, volume, year, management_event)
}
\arguments{
\item{stand_id}{The Stand ID}

\item{tree_id}{Tree ID}

\item{year}{Year of management event.}

\item{management_event}{Type of management event.}
}
\value{

}
\description{
A data frame detailing the management events applied to every tree.
}
