% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manso_2020_height_trajectory_GB_GADA.R
\name{Manso2020}
\alias{Manso2020}
\alias{Manso_2020_height_trajectory_Great_Britain_Beech}
\alias{Manso_2020_height_trajectory_Great_Britain_Corsican_Pine}
\alias{Manso_2020_height_trajectory_Great_Britain_Douglas_Fir}
\alias{Manso_2020_height_trajectory_Great_Britain_European_Larch}
\alias{Manso_2020_height_trajectory_Great_Britain_Lodgepole_Pine}
\alias{Manso_2020_height_trajectory_Great_Britain_Norway_Spruce}
\alias{Manso_2020_height_trajectory_Great_Britain_Poplar}
\alias{Manso_2020_height_trajectory_Great_Britain_Scots_Pine}
\alias{Manso_2020_height_trajectory_Great_Britain_Sitka_Spruce}
\alias{Manso_2020_height_trajectory_Great_Britain_Western_Hemlock}
\title{New Dynamic top height models for Great Britain}
\source{
Manso, R., McLean J.P., Arcangeli, C., Matthews, R. (2020) Dynamic
top height models for several major forest tree species in Great Britain.
Forestry: An International Journal of Forest Research, Vol 94:2, pp. 181-192.
 Available Online (15/01/2022): \url{https://doi.org/10.1093/foresj/cpaa036}
}
\usage{
Manso_2020_height_trajectory_Great_Britain_Beech(
  dominant_height,
  age,
  age2,
  output = "Height"
)

Manso_2020_height_trajectory_Great_Britain_Corsican_Pine(
  dominant_height,
  age,
  age2,
  output = "Height"
)

Manso_2020_height_trajectory_Great_Britain_Douglas_Fir(
  dominant_height,
  age,
  age2,
  output = "Height"
)

Manso_2020_height_trajectory_Great_Britain_European_Larch(
  dominant_height,
  age,
  age2,
  output = "Height"
)

Manso_2020_height_trajectory_Great_Britain_Lodgepole_Pine(
  dominant_height,
  age,
  age2,
  output = "Height"
)

Manso_2020_height_trajectory_Great_Britain_Norway_Spruce(
  dominant_height,
  age,
  age2,
  output = "Height"
)

Manso_2020_height_trajectory_Great_Britain_Poplar(
  dominant_height,
  age,
  age2,
  output = "Height"
)

Manso_2020_height_trajectory_Great_Britain_Scots_Pine(
  dominant_height,
  age,
  age2,
  output = "Height"
)

Manso_2020_height_trajectory_Great_Britain_Sitka_Spruce(
  dominant_height,
  age,
  age2,
  output = "Height"
)

Manso_2020_height_trajectory_Great_Britain_Western_Hemlock(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Total age.}

\item{age2}{Total age at output.}

\item{output}{One of "SIH100","Equation" or "Height" (default).}
}
\description{
GADA functions to replace Edwards & Christie 1981 used in
'Forest Yield' & 'MOSES_GB'.
}
\examples{
ggplot()+geom_function( fun= function(x) Manso_2020_height_trajectory_Great_Britain_Western_Hemlock(dominant_height=25,age=50,age2=x))+xlim(c(0,200))
}
