% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marklund_1988.R
\name{Marklund_1988_biomass_Sweden}
\alias{Marklund_1988_biomass_Sweden}
\title{Calculate dry-weight biomass for individual trees.}
\source{
Marklund, Lars-Gunnar. 1988. Biomassafunktioner för Tall, Gran och Björk i Sverige: Biomass functions for pine, spruce and birch in Sweden. Report 45.  73 pp. Dept. of Forest Survey. Swedish University of Agricultural Sciences. Umeå. ISSN 0348-0496. ISBN 91-576-3524-2.
}
\usage{
Marklund_1988_biomass_Sweden(
  species = NULL,
  DBH = NULL,
  diameter_3m.cm = NULL,
  diameter_5m.cm = NULL,
  height.m = NULL,
  ageBH = NULL,
  H100 = NULL,
  SI_Species = NULL,
  DBH_i5.mm = NULL,
  double_bark.mm = NULL,
  soil_texture = NULL,
  soil_moisture = NULL,
  soil_water = NULL,
  latitude = NULL,
  longitude = NULL,
  epsg_in = NULL,
  altitude = NULL,
  crown_base_height.m = NULL,
  crown_radius.m = NULL,
  Dmax_10m = NULL
)
}
\arguments{
\item{species}{One of 'Picea abies', 'Pinus sylvestris', 'Betula pendula' or 'Betula pubescens'.}

\item{DBH}{Diameter at breast height (1.3m), in cm.}

\item{diameter_3m.cm}{Diameter at 3 m height, in cm.}

\item{diameter_5m.cm}{Diameter at 5 m height, in cm.}

\item{height.m}{Height of tree, in m.}

\item{ageBH}{Age at breast height (1.3m)}

\item{H100}{Site Index H100 per [forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce()],[forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce()],[forester::Hagglund_1974_Sweden_height_trajectories_Pine()]}

\item{SI_Species}{One of 'Picea abies' or 'Pinus sylvestris'.}

\item{DBH_i5.mm}{Five year diameter increment at breast height, in mm.}

\item{double_bark.mm}{Double bark thickness, mm.}

\item{soil_texture}{Code 1-9. [forester::Sweden_soil_types('texture')]}

\item{soil_moisture}{Code 1-5. [forester::Sweden_soil_types('moisture')]}

\item{soil_water}{Code 1-3. [forester::Sweden_soil_types('water')]}

\item{latitude}{Latitude, decimal.}

\item{longitude}{Longitude, decimal.}

\item{epsg_in}{EPSG of latitude, longitude arguments.}

\item{altitude}{Altitude, meters above sea level.}

\item{crown_base_height.m}{Height from ground to lowest green branch, in meters. See SWE-NFI detailed instruction.}

\item{crown_radius.m}{Crown radius, meters.}

\item{Dmax_10m}{DBH of the thickest tree on a plot of radius 10 m.}
}
\value{
List of dry weights of tree components (kg).
}
\description{
Calculate dry-weight biomass for individual trees.
}
\details{
Unused arguments should be set to NULL.
}
\examples{
Marklund_1988_biomass_Sweden(species='Picea abies',DBH=20)
}
