% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_dead_branches_Birch.R
\name{Marklund_1988_dry_weight_of_dead_branches_Birch}
\alias{Marklund_1988_dry_weight_of_dead_branches_Birch}
\title{Dry weight of dead branches for Birch from Marklund 1988}
\usage{
Marklund_1988_dry_weight_of_dead_branches_Birch(
  diameter.cm,
  height.m,
  altitude,
  latitude
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, in cm.}

\item{height.m}{Height of tree, in metres.}

\item{altitude}{Altitude, in km.}

\item{latitude}{Latitude, RT90}
}
\value{
Dry weight of dead branches, in kilograms.
}
\description{
Dry weight of dead branches for Birch from Marklund 1988
}
