% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_dead_branches_Pine.R
\name{Marklund_1988_dry_weight_of_dead_branches_Pine}
\alias{Marklund_1988_dry_weight_of_dead_branches_Pine}
\title{Dry weight of dead branches from a Pine from Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_dead_branches_Pine(
  diameter.cm,
  height.m,
  latitude,
  altitude,
  age_at_breast_height,
  five_years_diameter_increment.mm,
  Dmax_10_m_radius
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, in cm.}

\item{height.m}{Height of tree, in metres.}

\item{latitude}{Latitude.}

\item{altitude}{Altitude, in km.}

\item{age_at_breast_height}{Age at breast height.}

\item{five_years_diameter_increment.mm}{Increment from the last five years at
breast height, measured in mm.}

\item{Dmax_10_m_radius}{Diameter above bark of the thickest tree on a 10 m radius plot.}
}
\value{
Dry weight of dead branches in kilograms.
}
\description{
Dry weight of dead branches from a Pine from Marklund (1988)
}
