% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_dead_branches_Spruce.R
\name{Marklund_1988_dry_weight_of_dead_branches_Spruce}
\alias{Marklund_1988_dry_weight_of_dead_branches_Spruce}
\title{Dry weight of dead branches for Spruce from Marklund 1988}
\usage{
Marklund_1988_dry_weight_of_dead_branches_Spruce(
  diameter.cm,
  height.m,
  crown_base_height.m,
  Dmax_10_m_radius,
  five_years_diameter_increment.mm
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, cm.}

\item{height.m}{Height of tree, metres.}

\item{crown_base_height.m}{Height to the lowest living green branch not
separated by more than three whorls from the crown.}

\item{Dmax_10_m_radius}{Diameter above bark of the thickest tree on a 10 m radius plot.}

\item{five_years_diameter_increment.mm}{Increment from the last five years at
breast height, measured in mm.}
}
\value{
Dry weight of dead branches for a Spruce, kilograms.
}
\description{
Dry weight of dead branches for Spruce from Marklund 1988
}
