% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_living_branches_and_needles_Spruce.R
\name{Marklund_1988_dry_weight_of_living_branches_and_needles_Spruce}
\alias{Marklund_1988_dry_weight_of_living_branches_and_needles_Spruce}
\title{Dry weight of living branches and needles for Spruce, from Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_living_branches_and_needles_Spruce(
  diameter.cm,
  height.m,
  crown_base_height.m,
  age_at_breast_height,
  Dmax_10_m_radius,
  five_years_diameter_increment.mm,
  dominant_species,
  SI
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, in cm.}

\item{height.m}{Height of tree, in metres.}

\item{crown_base_height.m}{Height to the lowest living green branch not
separated by more than three whorls from the crown.}

\item{age_at_breast_height}{Age at breast height.}

\item{Dmax_10_m_radius}{Diameter of thickest tree on a plot with a 10 metre radius.}

\item{five_years_diameter_increment.mm}{Increment from the last five years at
breast height, measured in mm.}

\item{dominant_species}{Species for which SI was estimated, "Pinus sylvestris"
or "Picea abies".}

\item{SI}{Site index (m.)}
}
\value{
Dry weight of living branches and needles for Spruce, in kilograms.
}
\description{
Dry weight of living branches and needles for Spruce, from Marklund (1988)
}
