% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_living_branches_Birch.R
\name{Marklund_1988_dry_weight_of_living_branches_Birch}
\alias{Marklund_1988_dry_weight_of_living_branches_Birch}
\title{Dry weight of living branches for Birch from Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_living_branches_Birch(
  diameter.cm,
  height.m,
  latitude,
  crown_base_height.m,
  five_years_diameter_increment.mm
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height in cm.}

\item{height.m}{Height of tree in metres.}

\item{latitude}{Latitude, RT90.}

\item{crown_base_height.m}{Height to the lowest living green branch not
separated by more than three whorls from the crown.}

\item{five_years_diameter_increment.mm}{Increment from the last five years at
breast height, measured in mm.}
}
\value{
Dry weight of living branches, in kilograms.
}
\description{
Dry weight of living branches for Birch from Marklund (1988)
}
