% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_living_branches_Pine.R
\name{Marklund_1988_dry_weight_of_living_branches_Pine}
\alias{Marklund_1988_dry_weight_of_living_branches_Pine}
\title{Dry weight of living branches for Pine according to Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_living_branches_Pine(
  diameter.cm,
  height.m,
  age_at_breast_height,
  form_quotient5,
  form_quotient3,
  altitude,
  crown_radius.m,
  crown_base_height.m,
  five_years_diameter_increment.mm,
  latitude
)
}
\arguments{
\item{diameter.cm}{Diameter in cm.}

\item{height.m}{Height of tree in meters.}

\item{age_at_breast_height}{Age at breast height.}

\item{form_quotient5}{Diameter at 5 m / Diameter at breast height.}

\item{form_quotient3}{Diameter at 3 m / Diameter at breast height.}

\item{altitude}{Altitude, in km.}

\item{crown_radius.m}{Maximum crown radius, in metres.}

\item{crown_base_height.m}{Height to the lowest living green branch not
separated by more than three whorls from the crown.}

\item{five_years_diameter_increment.mm}{Increment from the last five years at
breast height, measured in mm.}

\item{latitude}{Latitude.}
}
\value{
Dry weight of living branches for a Pine tree, in kilograms.
}
\description{
Dry weight of living branches for Pine according to Marklund (1988)
}
