% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marklund_1988_dry_weight_of_needles_Pine.R
\name{Marklund_1988_dry_weight_of_needles_Pine}
\alias{Marklund_1988_dry_weight_of_needles_Pine}
\title{Dry weight of needles in a Pine, from Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_needles_Pine(
  diameter.cm,
  height.m,
  age_at_breast_height,
  crown_base_height.m,
  crown_radius.m,
  latitude,
  altitude,
  five_years_diameter_increment.mm
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, in cm.}

\item{height.m}{Height of tree, in m.}

\item{age_at_breast_height}{Age of tree at breast height.}

\item{crown_base_height.m}{Height to the lowest living green branch not
separated by more than three whorls from the crown.}

\item{crown_radius.m}{Maximum crown radius, in metres.}

\item{latitude}{Latitude.}

\item{altitude}{Altitude, in km.}

\item{five_years_diameter_increment.mm}{Increment from the last five years at
breast height, measured in mm.}
}
\value{
Dry weight of needles in kilograms.
}
\description{
Dry weight of needles in a Pine, from Marklund (1988)
}
