% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marklund_1988_dry_weight_of_needles_Spruce.R
\name{Marklund_1988_dry_weight_of_needles_Spruce}
\alias{Marklund_1988_dry_weight_of_needles_Spruce}
\title{Dry weight of needles for Spruce, from Marklund 1988.}
\usage{
Marklund_1988_dry_weight_of_needles_Spruce(
  diameter.cm,
  height.m,
  crown_base_height.m,
  crown_radius.m,
  age_at_breast_height,
  five_years_diameter_increment.mm,
  Dmax_10_m_radius
)
}
\arguments{
\item{diameter.cm}{Diameter of tree at breast height, cm.}

\item{height.m}{Height of tree, metres.}

\item{crown_base_height.m}{Height to the lowest living green branch not
separated by more than three whorls from the crown.}

\item{crown_radius.m}{Maximum crown radius, in metres.}

\item{age_at_breast_height}{Age at breast height.}

\item{five_years_diameter_increment.mm}{Increment from the last five years at
breast height, measured in mm.}

\item{Dmax_10_m_radius}{Diameter above bark of the thickest tree on a 10 m radius plot.}
}
\value{
Dry weight of needles for a Spruce, kilograms.
}
\description{
Dry weight of needles for Spruce, from Marklund 1988.
}
