% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_roots_above_5_Spruce.R
\name{Marklund_1988_dry_weight_of_roots_above_5_Spruce}
\alias{Marklund_1988_dry_weight_of_roots_above_5_Spruce}
\title{Dry weight of roots above 5 cm for Spruce, from Marklund 1988}
\usage{
Marklund_1988_dry_weight_of_roots_above_5_Spruce(
  diameter.cm,
  dominant_species,
  SI,
  age_at_breast_height,
  Moist_soil,
  peat,
  mobile_ground_water_longer_periods
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, cm.}

\item{dominant_species}{"Pinus sylvestris" or "Picea abies";
Species for which SI was estimated.}

\item{SI}{Site index, m.}

\item{age_at_breast_height}{Age of tree at breast height.}

\item{Moist_soil}{If the ground water level is less than 1 m deep, TRUE,
otherwise FALSE.}

\item{peat}{If the humus layer is thicker than 30 cm, TRUE, otherwise
FALSE.}

\item{mobile_ground_water_longer_periods}{If surface/subsurface water flow
occurs during longer periods, TRUE, otherwise FALSE.}
}
\value{
dry weight of roots above 5 cm, kilograms.
}
\description{
Dry weight of roots above 5 cm for Spruce, from Marklund 1988
}
