% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_roots_below_5_Pine.R
\name{Marklund_1988_dry_weight_of_roots_below_5}
\alias{Marklund_1988_dry_weight_of_roots_below_5}
\title{Dry weight of roots below 5 cm diameter for Pine, from Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_roots_below_5(
  diameter.cm,
  dominant_species,
  SI,
  altitude
)
}
\arguments{
\item{diameter.cm}{Diameter of the tree at breast height, cm.}

\item{dominant_species}{Dominant species, for which SI has been estimated.}

\item{SI}{Site index of the site.}

\item{altitude}{Altitude, km.}
}
\value{
Dry weight of roots below 5 cm diameter for Pine, in kilograms.
}
\description{
Dry weight of roots below 5 cm diameter for Pine, from Marklund (1988)
}
