% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_roots_below_5_Spruce.R
\name{Marklund_1988_dry_weight_of_roots_below_5_Spruce}
\alias{Marklund_1988_dry_weight_of_roots_below_5_Spruce}
\title{Dry weight of roots below 5 cm diameter for Spruce from Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_roots_below_5_Spruce(
  diameter.cm,
  dominant_species,
  SI,
  Moist_soil,
  peat
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, cm.}

\item{dominant_species}{"Pinus sylvestris" or "Picea abies";
Species for which SI was estimated.}

\item{SI}{Site index, m.}

\item{Moist_soil}{If the ground water level is less than 1 m deep, TRUE,
otherwise FALSE.}

\item{peat}{If the humus layer is thicker than 30 cm, TRUE, otherwise
FALSE.}
}
\value{
Dry weight of roots below 5 cm, in kilograms.
}
\description{
Dry weight of roots below 5 cm diameter for Spruce from Marklund (1988)
}
