% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_stem_above_bark_Birch.R
\name{Marklund_1988_dry_weight_of_stem_above_bark_Birch}
\alias{Marklund_1988_dry_weight_of_stem_above_bark_Birch}
\title{Dry weight of stem above bark for Birch, from Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_stem_above_bark_Birch(
  diameter.cm,
  height.m,
  age_at_breast_height
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, in cm.}

\item{height.m}{Height of tree, in metres.}

\item{age_at_breast_height}{Age at breast height.}
}
\value{
Dry weight of stem above bark, in kilograms.
}
\description{
Dry weight of stem above bark for Birch, from Marklund (1988)
}
