% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_stem_above_bark_Spruce.R
\name{Marklund_1988_dry_weight_of_stem_above_bark_Spruce}
\alias{Marklund_1988_dry_weight_of_stem_above_bark_Spruce}
\title{Dry weight of stem above bark for Spruce according to Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_stem_above_bark_Spruce(
  diameter.cm,
  height.m,
  age_at_breast_height,
  form_quotient5,
  form_quotient3
)
}
\arguments{
\item{diameter.cm}{Diameter of tree at breast height, cm.}

\item{height.m}{Tree height, m.}

\item{age_at_breast_height}{Age of tree at breast height.}

\item{form_quotient5}{Diameter measured at 5 m / Diameter measured at
breast height.}

\item{form_quotient3}{Diameter measured at 3 m / Diameter measured at
breast height.}
}
\value{
Dry weight of stem above bark, in kilograms.
}
\description{
Dry weight of stem above bark for Spruce according to Marklund (1988)
}
