% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marklund_1988_dry_weight_of_stem_bark_Birch.R
\name{Marklund_1988_dry_weight_of_stem_bark_Birch}
\alias{Marklund_1988_dry_weight_of_stem_bark_Birch}
\title{Dry weight of stem bark for Birch from Marklund 1988}
\usage{
Marklund_1988_dry_weight_of_stem_bark_Birch(
  diameter.cm,
  height.m,
  age_at_breast_height,
  double_bark.mm,
  latitude,
  five_years_diameter_increment,
  Dmax_10_m_radius
)
}
\arguments{
\item{diameter.cm}{Diameter of tree at breast height, in cm-}

\item{height.m}{Height of tree in m.}

\item{age_at_breast_height}{Age of tree at breast height.}

\item{double_bark.mm}{Double bark thickness, mm.}

\item{latitude}{Latitude, RT90.}

\item{five_years_diameter_increment}{Increment from the last five years at
breast height, measured in mm.}

\item{Dmax_10_m_radius}{Diameter of thickest tree on a plot with a 10 metre radius.}
}
\value{
Dry weight of stem bark for Birch, in kilograms.
}
\description{
Dry weight of stem bark for Birch from Marklund 1988
}
