% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marklund_1988_dry_weight_of_stem_bark_Spruce.R
\name{Marklund_1988_dry_weight_of_stem_bark_Spruce}
\alias{Marklund_1988_dry_weight_of_stem_bark_Spruce}
\title{Dry weight of stem bark for Spruce, Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_stem_bark_Spruce(
  diameter.cm,
  height.m,
  age_at_breast_height,
  Dmax_10_m_radius,
  dominant_species,
  SI,
  double_bark.mm
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, in cm.}

\item{height.m}{Height of tree, m.}

\item{age_at_breast_height}{Age of tree at breast height.}

\item{Dmax_10_m_radius}{Diameter above bark of the thickest tree on a 10 m radius plot.}

\item{dominant_species}{"Picea abies" or "Pinus sylvestris". Species for which SI was estimated.}

\item{SI}{Estimated SI.}

\item{double_bark.mm}{Double bark thickness.}
}
\value{

}
\description{
Dry weight of stem bark for Spruce, Marklund (1988)
}
