% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marklund_1988_dry_weight_of_stem_wood_Birch.R
\name{Marklund_1988_dry_weight_of_stem_wood_Birch}
\alias{Marklund_1988_dry_weight_of_stem_wood_Birch}
\title{Dry weight of stem wood for Birch from Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_stem_wood_Birch(
  diameter.cm,
  height.m,
  age_at_breast_height,
  double_bark.mm,
  latitude
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height in cm}

\item{height.m}{Height of tree, in m.}

\item{age_at_breast_height}{Age at breast height.}

\item{double_bark.mm}{Double bark thickness, mm.}

\item{latitude}{Latitude, RT90.}
}
\value{
Dry weight of stem wood, in kilograms.
}
\description{
Dry weight of stem wood for Birch from Marklund (1988)
}
