% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marklund_1988_dry_weight_of_stem_wood_Pine.R
\name{Marklund_1988_dry_weight_of_stem_wood_Pine}
\alias{Marklund_1988_dry_weight_of_stem_wood_Pine}
\title{Dry weight of Stem wood for Pine according to Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_stem_wood_Pine(
  diameter.cm,
  height.m,
  age_at_breast_height,
  double_bark.mm,
  form_quotient5,
  form_quotient3,
  altitude
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, centimetres.}

\item{height.m}{Height of tree, metres.}

\item{age_at_breast_height}{Age of tree at breast height.}

\item{double_bark.mm}{Double thickness of bark, in mm.}

\item{form_quotient5}{Diameter at 5 m / Diameter at breast height.}

\item{form_quotient3}{Diameter at 3 m / Diameter at breast height.}

\item{altitude}{Altitude, in km.}
}
\value{
Dry weight of Pine stem above bark in kilograms.
}
\description{
Dry weight of Stem wood for Pine according to Marklund (1988)
}
