% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marklund_1988_dry_weight_of_stem_wood_Spruce.R
\name{Marklund_1988_dry_weight_of_stem_wood_Spruce}
\alias{Marklund_1988_dry_weight_of_stem_wood_Spruce}
\title{Dry weight of stem wood for Spruce from Marklund (1988)}
\usage{
Marklund_1988_dry_weight_of_stem_wood_Spruce(
  diameter.cm,
  height.m,
  age_at_breast_height,
  form_quotient5,
  form_quotient3
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, in cm.}

\item{height.m}{Height of tree, in m.}

\item{age_at_breast_height}{Age of tree at breast height.}

\item{form_quotient5}{Diameter at 5 m / Diameter at breast height.}

\item{form_quotient3}{Diameter at 3 m / Diameter at breast height.}
}
\value{
Dry weight of stem wood for Spruce, in kilograms.
}
\description{
Dry weight of stem wood for Spruce from Marklund (1988)
}
