% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_stump_and_roots_Spruce.R
\name{Marklund_1988_dry_weight_of_stump_and_roots_Spruce}
\alias{Marklund_1988_dry_weight_of_stump_and_roots_Spruce}
\title{Dry weight of stump and roots for Spruce, from Marklund 1988}
\usage{
Marklund_1988_dry_weight_of_stump_and_roots_Spruce(
  diameter.cm,
  dominant_species,
  SI,
  Moist_soil,
  peat,
  mobile_ground_water_longer_periods
)
}
\arguments{
\item{diameter.cm}{Diameter of tree at breast height, cm.}

\item{dominant_species}{"Pinus sylvestris" or "Picea abies";
Species for which SI was estimated.}

\item{SI}{Site index, m.}

\item{Moist_soil}{If the ground water level is less than 1 m deep, TRUE,
otherwise FALSE.}

\item{peat}{If the humus layer is thicker than 30 cm, TRUE, otherwise
FALSE.}

\item{mobile_ground_water_shorter_periods}{If surface/subsurface water flow
occurs during shorter periods, TRUE, otherwise FALSE.}
}
\value{
Dry weight of stump and roots for Spruce from Marklund 1988
}
\description{
Dry weight of stump and roots for Spruce, from Marklund 1988
}
