% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marklund_1988_dry_weight_of_stump_Spruce.R
\name{Marklund_1988_dry_weight_of_stump_Spruce}
\alias{Marklund_1988_dry_weight_of_stump_Spruce}
\title{Dry weight of stump for Spruce from Marklund 1988}
\usage{
Marklund_1988_dry_weight_of_stump_Spruce(
  diameter.cm,
  dominant_species,
  SI,
  latitude
)
}
\arguments{
\item{diameter.cm}{Diameter of tree at breast height, cm.}

\item{dominant_species}{Species for which SI was estimated. "Pinus sylvestris"
or "Picea abies".}

\item{SI}{Site index, m.}

\item{latitude}{Latitude, RT90.}
}
\value{
Dry weight stump for Spruce, in kilograms.
}
\description{
Dry weight of stump for Spruce from Marklund 1988
}
