% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Marklund_1988_dry_weight_of_stump_and_roots_Pine.R
\name{Marklund_dry_weight_of_stump_and_roots_Pine}
\alias{Marklund_dry_weight_of_stump_and_roots_Pine}
\title{Dry weight of stump and roots of a Pine from Marklund 1988}
\usage{
Marklund_dry_weight_of_stump_and_roots_Pine(
  diameter.cm,
  dominant_species,
  latitude,
  age_at_breast_height,
  Dry_soil,
  Moist_soil,
  peat,
  mobile_ground_water_longer_periods,
  mobile_ground_water_shorter_periods,
  altitude
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, in cm.}

\item{dominant_species}{Dominant species on the plot.}

\item{latitude}{Latitude, RT90.}

\item{age_at_breast_height}{Age at breast height.}

\item{Dry_soil}{If the ground water level is more than 2 m deep, TRUE,
otherwise FALSE.}

\item{Moist_soil}{If the ground water level is less than 1 m deep, TRUE,
otherwise FALSE.}

\item{peat}{If the humus layer is thicker than 30 cm, TRUE, otherwise
FALSE.}

\item{mobile_ground_water_longer_periods}{If surface/subsurface water flow
occurs during longer periods, TRUE, otherwise FALSE.}

\item{mobile_ground_water_shorter_periods}{If surface/subsurface water flow
occurs during shorter periods, TRUE, otherwise FALSE.}

\item{altitude}{Altitude, in km.}
}
\value{
Dry weight of stump and roots, in kilograms.
}
\description{
Dry weight of stump and roots of a Pine from Marklund 1988
}
