% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mensah_2022.R
\name{Mensah_2022_BA}
\alias{Mensah_2022_BA}
\alias{Mensah_2022_BA_Sweden_Scots_Pine}
\alias{Mensah_2022_BA_Sweden_Norway_Spruce}
\alias{Mensah_2022_BA_Sweden_Lodgepole_Pine}
\alias{Mensah_2022_BA_Sweden_Larch}
\title{Basal Area development in conifer stands in Sweden.}
\source{
Mensah, A. A., Holmström, E., Nyström, K., Nilsson, U. 2022.
Modelling potential yield capacity in conifers using Swedish long-term
experiments. For. Ecol. Manage. 512. Available online (2022-04-08):
\url{https://doi.org/10.1016/j.foreco.2022.120162}
}
\usage{
Mensah_2022_BA_Sweden_Scots_Pine(BA, age, age2, stems, H100)

Mensah_2022_BA_Sweden_Norway_Spruce(BA, age, age2, stems, H100)

Mensah_2022_BA_Sweden_Lodgepole_Pine(BA, age, age2, stems, H50)

Mensah_2022_BA_Sweden_Larch(BA, age, age2, stems, H100)
}
\arguments{
\item{BA}{Basal area at breast height, m2 / ha.}

\item{age}{Total stand age at start.}

\item{age2}{Total stand age desired.}

\item{stems}{Number of stems per hectare at start.}

\item{H100}{Site Index H100 e.g. \cr
\tabular{ll}{
Scots Pine\tab [forester::Elfving_Kiviste_1997_height_trajectory_Sweden_Pine()]\cr
Norway Spruce\tab [forester::Elfving_2003_height_trajectory_Sweden_Spruce()]\cr
Lodgepole Pine\tab [forester::Liziniewicz_2016_height_trajectory_Sweden_Pinus_contorta()]\cr
Larch\tab [forester::Johansson_2013_height_trajectory_Sweden_Larch()]\cr
}}

\item{H50}{Site Index H50, see parameter H100}
}
\value{
Expected BA m2 / ha at age2.
}
\description{
Note: BA should be given for age>40 for Norway Spruce or
Scots Pine.For Larch and Lodgepole Pine, age should be 30. The function will
throw a warning.
}
\details{
Table of fit-statistics for basal area functions:
\tabular{lrrr}{
Species\tab RMSE (\%)\tab MD (\%)\tab n\cr
Scots Pine\tab 1.09(3.39)\tab -0.15(-0.47)\tab 3447\cr
Norway Spruce\tab 0.97(2.14)\tab -0.10(-0.22)\tab 984\cr
Lodgepole pine\tab 1.22(3.29)\tab -0.01(-0.03)\tab 237\cr
Larch\tab 2.79(5.78)\tab -0.29(-0.59)\tab 78\cr
}
}
