% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mensah_2022.R
\name{Mensah_2022_MAI}
\alias{Mensah_2022_MAI}
\alias{Mensah_2022_MAI_max_Sweden_Scots_Pine}
\alias{Mensah_2022_MAI_max_Sweden_Norway_Spruce}
\alias{Mensah_2022_MAI_max_Sweden_Lodgepole_Pine}
\alias{Mensah_2022_MAI_max_Sweden_Larch}
\title{Relation between yield capacity MAI max m3/ha/yr and Site Index.}
\source{
Mensah, A. A., Holmström, E., Nyström, K., Nilsson, U. 2022.
Modelling potential yield capacity in conifers using Swedish long-term
experiments. For. Ecol. Manage. 512. Available online (2022-04-08):
\url{https://doi.org/10.1016/j.foreco.2022.120162}
}
\usage{
Mensah_2022_MAI_max_Sweden_Scots_Pine(H100)

Mensah_2022_MAI_max_Sweden_Norway_Spruce(H100)

Mensah_2022_MAI_max_Sweden_Lodgepole_Pine(H50)

Mensah_2022_MAI_max_Sweden_Larch(H100)
}
\arguments{
\item{H100}{Site Index H100 e.g. \cr
\tabular{ll}{
Scots Pine\tab [forester::Elfving_Kiviste_1997_height_trajectory_Sweden_Pine()]\cr
Norway Spruce\tab [forester::Elfving_2003_height_trajectory_Sweden_Spruce()]\cr
Lodgepole Pine\tab [forester::Liziniewicz_2016_height_trajectory_Sweden_Pinus_contorta()]\cr
Larch\tab [forester::Johansson_2013_height_trajectory_Sweden_Larch()]\cr
}}

\item{H50}{Site Index H50, see parameter SI100}
}
\value{
Maximum Annual Increment (m^3 ha^{-1} yr^{-1})
}
\description{
Relation between yield capacity MAI max m3/ha/yr and Site Index.
}
\details{
Fit-statistics \cr
\tabular{lrr}{
Species\tab RMSE(rel,\%)\tab R^2\cr
Scots Pine\tab 0.9757 (16.73)\tab 0.809\cr
Norway Spruce\tab 1.047 (11.14)\tab 0.904\cr
Lodgepole Pine\tab 0.436 (5.52)\tab 0.956\cr
Larch\tab 1.164 (10.82)\tab 0.876\cr
}
}
