% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mensah_2022.R
\name{Mensah_2022_volume}
\alias{Mensah_2022_volume}
\alias{Mensah_2022_volume_yield_m3_ha_Sweden_Scots_Pine}
\alias{Mensah_2022_volume_yield_m3_ha_Sweden_Norway_Spruce}
\alias{Mensah_2022_volume_yield_m3_ha_Sweden_Lodgepole_Pine}
\alias{Mensah_2022_volume_yield_m3_ha_Sweden_Larch}
\title{Species-specific volume (m^3 ha^{-1}) yield functions.}
\source{
Mensah, A. A., Holmström, E., Nyström, K., Nilsson, U. 2022.
Modelling potential yield capacity in conifers using Swedish long-term
experiments. For. Ecol. Manage. 512. Available online (2022-04-08):
\url{https://doi.org/10.1016/j.foreco.2022.120162}
}
\usage{
Mensah_2022_volume_yield_m3_ha_Sweden_Scots_Pine(dominant_height, BA)

Mensah_2022_volume_yield_m3_ha_Sweden_Norway_Spruce(dominant_height, BA)

Mensah_2022_volume_yield_m3_ha_Sweden_Lodgepole_Pine(dominant_height, BA)

Mensah_2022_volume_yield_m3_ha_Sweden_Larch(dominant_height, BA)
}
\arguments{
\item{dominant_height}{\cr
\tabular{ll}{
Scots Pine\tab [forester::Elfving_Kiviste_1997_height_trajectory_Sweden_Pine()]\cr
Norway Spruce\tab [forester::Elfving_2003_height_trajectory_Sweden_Spruce()]\cr
Lodgepole Pine\tab [forester::Liziniewicz_2016_height_trajectory_Sweden_Pinus_contorta()]\cr
Larch\tab [forester::Johansson_2013_height_trajectory_Sweden_Larch()]\cr
}}

\item{BA}{e.g. [forester::Mensah_2022_BA_Sweden_Scots_Pine]}
}
\value{
Volume Yield m^3 ha^-1
}
\description{
Species-specific volume (m^3 ha^{-1}) yield functions.
}
\details{
\cr
\tabular{lrrr}{
Species\tab RMSE(\%)\tab MD(\%)\tab n\cr
Scots Pine\tab 24.27(11.74) \tab -0.1584(-0.0765)\tab 4405 \cr
Norway Spruce\tab 36.98(11.75) \tab -0.1629(-0.0517)\tab 1789 \cr
Lodgepole Pine\tab 22.44(10.96) \tab -0.1238(-0.0626)\tab 725 \cr
Larch\tab 54.18(17.72) \tab 1.1717(0.3834)\tab 453 \cr
}
}
