% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mensah_2022.R
\name{Mensah_initial_BA}
\alias{Mensah_initial_BA}
\alias{Mensah_2022_initial_BA_Sweden_Scots_Pine}
\alias{Mensah_2022_initial_BA_Sweden_Norway_Spruce}
\alias{Mensah_2022_initial_BA_Sweden_Lodgepole_Pine}
\alias{Mensah_2022_initial_BA_Sweden_Larch}
\title{Helper functions for Mensah 2022 initial stand density.}
\source{
Mensah, A. A., Holmström, E., Nyström, K., Nilsson, U. 2022.
Modelling potential yield capacity in conifers using Swedish long-term
experiments. For. Ecol. Manage. 512. Available online (2022-04-08):
\url{https://doi.org/10.1016/j.foreco.2022.120162}
(Unpublished supplement, in print 2022-04-14)
}
\usage{
Mensah_2022_initial_BA_Sweden_Scots_Pine(H100, age)

Mensah_2022_initial_BA_Sweden_Norway_Spruce(H100, age)

Mensah_2022_initial_BA_Sweden_Lodgepole_Pine(H50, age)

Mensah_2022_initial_BA_Sweden_Larch(H100, age)
}
\arguments{
\item{H100}{Site Index H100}

\item{age}{Total stand age}

\item{H50}{Site Index H50}
}
\value{
Initial basal area.
}
\description{
Helper functions for Mensah 2022 initial stand density.
}
\details{
\tabular{lrr}{
Species \tab RMSE m^2/ha (%) \cr
Scots Pine \tab 3.231 (11.14) \cr
Norway Spruce \tab 3.403 (9.72) \cr
Lodgepole Pine \tab 4.534 (15.63) \cr
Larch \tab 4.601 (12.78) \cr
}
}
