% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moren_Perttu_1994_Temperature_Models.R
\name{Moren1994}
\alias{Moren1994}
\alias{Moren_Perttu_1994_Sweden_temperature_sum_5C}
\alias{Moren_Perttu_1994_Sweden_growing_season_length_5C}
\alias{Moren_Perttu_1994_Sweden_start_of_growing_season_5C}
\alias{Moren_Perttu_1994_Sweden_global_radiation_sum}
\title{Regional temperature and radiation indices}
\source{
Morén, Ann-Sofie; Perttu, Kurth L. (1994). Regional temperature and radiation indices and their adjustment to horizontal and inclined forest land. Studia Forestalia Suecica 194. 19 pp. ISSN 0039-3150. ISBN 91-576-4915-4. Url:https://pub.epsilon.slu.se/3910/1/SFS194.pdf
}
\usage{
Moren_Perttu_1994_Sweden_temperature_sum_5C(
  Latitude,
  Altitude,
  Correction = FALSE,
  TmaxMonthly = NULL,
  TminMonthly = NULL
)

Moren_Perttu_1994_Sweden_growing_season_length_5C(Latitude, Altitude)

Moren_Perttu_1994_Sweden_start_of_growing_season_5C(
  Latitude,
  Altitude,
  returnDate = TRUE
)

Moren_Perttu_1994_Sweden_global_radiation_sum(Latitude, Altitude)
}
\arguments{
\item{Latitude}{Degrees North.}

\item{Altitude}{Meters above sea level}

\item{Correction}{Boolean. Apply Continental, Maritime Correction to temperature sum?}

\item{TmaxMonthly}{In case of correction. Maximum monthly mean temperature.}

\item{TminMonthly}{In case of correction. Minimum monthly mean temperature.}

\item{returnDate}{Return as a date?}
}
\value{
Temperature Sum exceeding 5 degrees Celsius.

Length of growing season (> 5 degrees Celsius), days.

Date or numeric.

Global Radiation Sum.
}
\description{
Regional temperature and radiation indices
}
