% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{native}
\alias{native}
\title{Native Species}
\format{
A tibble with 24 rows and 3 variables:
\describe{
\item{strata}{Factor. Fator de estratificação.}
\item{Sample}{Indice identificador da unidade amostral.}
\item{Volume}{Volume, em metros cúbicos}

...
}
}
\source{
Sanquetta
}
\usage{
native
}
\description{
A dataset containing the volumes para dois estratos de floresta
nativa.
}
\keyword{datasets}
