% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nilsen_1992_SI_prediction_vegetation_Norway.R
\name{Nilsen_1992_SI_prediction_vegetation_Norway}
\alias{Nilsen_1992_SI_prediction_vegetation_Norway}
\title{Site index estimation from vegetation type and site properties in Norway.}
\source{
Nilsen, P. & Larsson, J. 1992. Bonitering av skog ved hjelp av
vegetasjonstype og egenskaper ved voksestedet. (Site index estimation from
vegetation type and site properties). Rapp. Skogforsk 22/92:1-43.
Available online (15/03/2022):\url{https://nibio.brage.unit.no/nibio-xmlui/handle/11250/2677459?show=full&locale-attribute=en}
}
\usage{
Nilsen_1992_SI_prediction_vegetation_Norway(
  species = "Picea abies",
  vegetation = 1,
  soil_depth = 3,
  hillside = FALSE,
  latitude,
  altitude
)
}
\arguments{
\item{species}{One of "Picea abies" (default), or "Pinus sylvestris"}

\item{vegetation}{Code from [forester::Nilsen_1992_vegetation_types()]}

\item{soil_depth}{Integer. 1: less than 10 cm, 2: 10-30cm; 3: 30-100cm, 4: 100cm+}

\item{hillside}{TRUE or FALSE (default).}

\item{latitude}{Decimal latitude (only used for Norway Spruce).}

\item{altitude}{Meters above sea level.}
}
\value{
Estimated Site Index H40 [forester::Tveite_1976_height_trajectory_Norway_Pine()][forester::Tveite_1977_height_trajectory_Norway_Norway_Spruce()]
}
\description{
Site index estimation from vegetation type and site properties in Norway.
}
\details{
\itemize{
   \item Spruce: R^2 = 0.62 (sd 16\%)
   \item Pine: R^2 = 0.56 (sd 21\%)
   }
Largest deviation on high site indices for spruce (3 m.)
}
