% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nilsen_1992_SI_prediction_vegetation_Norway.R
\name{Nilsen_1992_vegetation_types}
\alias{Nilsen_1992_vegetation_types}
\title{Vegetation types for predicting site index for Scots Pine and Norway Spruce
in Norway.}
\source{
Nilsen, P. & Larsson, J. 1992. Bonitering av skog ved hjelp av
vegetasjonstype og egenskaper ved voksestedet. (Site index estimation from
vegetation type and site properties). Rapp. Skogforsk 22/92:1-43.
Available online (15/03/2022):\url{https://nibio.brage.unit.no/nibio-xmlui/handle/11250/2677459?show=full&locale-attribute=en}
}
\usage{
Nilsen_1992_vegetation_types(description = FALSE, species = FALSE, code = 1:8)
}
\arguments{
\item{description}{Print vegetation type description? TRUE or FALSE (default)}

\item{species}{Print important species? TRUE or FALSE (default).}

\item{code}{Integer 1 - 8. Default is 1:8}
}
\value{
A tibble with english and norwegian code explanations. If species is TRUE, also a list of the most important species in latin. If description is TRUE, also a translation of the vegetation community.
}
\description{
Vegetation types for predicting site index for Scots Pine and Norway Spruce
in Norway.
}
\details{
N.B. Function authors norwegian is shaky.
}
\seealso{
[forester::Nilsen_1992_SI_prediction_vegetation_Norway()]
}
\author{
Carl Vigren, SLU.
}
