% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NordLarsen_2009_height_trajectories_Denmark.R
\name{nordlarsen2009}
\alias{nordlarsen2009}
\alias{NordLarsen_2009_height_trajectory_Denmark_Beech}
\alias{NordLarsen_2009_height_trajectory_Denmark_Oak}
\alias{NordLarsen_2009_height_trajectory_Denmark_Norway_Spruce}
\alias{NordLarsen_2009_height_trajectory_Denmark_Sitka_Spruce}
\alias{NordLarsen_2009_height_trajectory_Denmark_Silver_Fir}
\alias{NordLarsen_2009_height_trajectory_Denmark_Douglas_Fir}
\title{Site-specific height growth models for six common tree species in Denmark.}
\source{
Thomas Nord-Larsen , Henrik Meilby & Jens Peter Skovsgaard (2009)
Site-specific height growth models for six common tree species in Denmark,
Scandinavian Journal of Forest Research, 24:3, 194-204, DOI:
\url{https://doi.org/10.1080/02827580902795036}
}
\usage{
NordLarsen_2009_height_trajectory_Denmark_Beech(
  dominant_height,
  age,
  age2,
  output = "Height"
)

NordLarsen_2009_height_trajectory_Denmark_Beech(
  dominant_height,
  age,
  age2,
  output = "Height"
)

NordLarsen_2009_height_trajectory_Denmark_Oak(
  dominant_height,
  age,
  age2,
  output = "Height"
)

NordLarsen_2009_height_trajectory_Denmark_Norway_Spruce(
  dominant_height,
  age,
  age2,
  output = "Height"
)

NordLarsen_2009_height_trajectory_Denmark_Sitka_Spruce(
  dominant_height,
  age,
  age2,
  output = "Height"
)

NordLarsen_2009_height_trajectory_Denmark_Silver_Fir(
  dominant_height,
  age,
  age2,
  output = "Height"
)

NordLarsen_2009_height_trajectory_Denmark_Douglas_Fir(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Total age.}

\item{age2}{Total age at output age.}

\item{output}{One of "Equation" or "Height".}
}
\description{
Site-specific height growth models for six common tree species in Denmark.
}
\details{
Beech: R^2=0.991, RMSE cross-validation 1.619

Oak: R^2=0.990, RMSE cross-validation 1.269

Norway Spruce: R^2=0.982, RMSE cross-validation 1.740

Sitka Spruce: R^2=0.984, RMSE cross-validation 1.583

Silver Fir: R^2=0.989, RMSE cross-validation 1.263

Douglas Fir: R^2=0.989, RMSE cross-validation 1.894
}
\examples{
# A ggplot of some common site indices for Beech in Denmark.
ggplot()+
geom_function(fun=function(x) NordLarsen_2009_height_trajectory_Denmark_Beech(dominant_height = 10,age=50,age2 = x))+
geom_function(fun=function(x) NordLarsen_2009_height_trajectory_Denmark_Beech(dominant_height = 15,age=50,age2 = x))+
geom_function(fun=function(x) NordLarsen_2009_height_trajectory_Denmark_Beech(dominant_height = 20,age=50,age2 = x))+
geom_function(fun=function(x) NordLarsen_2009_height_trajectory_Denmark_Beech(dominant_height = 25,age=50,age2 = x))+
geom_function(fun=function(x) NordLarsen_2009_height_trajectory_Denmark_Beech(dominant_height = 30,age=50,age2 = x))+
xlim(c(0,125))+
ylim(c(0,35))+
coord_cartesian(ylim=c(0,40))
}
