% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nyyssonen_1978_stand_volume_increment_Finland.R
\name{Nyssonen_stand_increment}
\alias{Nyssonen_stand_increment}
\alias{Nyyssonen_1978_stand_volume_increment_Finland_Spruce}
\alias{Nyyssonen_1978_stand_volume_increment_Finland_Pine}
\title{Five-year Stand Increment for even-aged stands of Scots Pine and Norway Spruce in Finland.}
\source{
Nyyssönen, A., Mielikäinen, K. 1978. Estimation of stand increment. Act. For. Fenn. 163. ISBN 951-651-038-8.
}
\usage{
Nyyssonen_1978_stand_volume_increment_Finland_Spruce(
  age,
  volume_m3_ha,
  BA_mean_diameter = 0,
  vegetation = "OMT"
)

Nyyssonen_1978_stand_volume_increment_Finland_Pine(
  age,
  volume_m3_ha,
  BA_mean_diameter = 0,
  vegetation = "OMT"
)
}
\arguments{
\item{age}{Age, years.}

\item{volume_m3_ha}{Stem Volume m^3/ha}

\item{BA_mean_diameter}{Optional for improved function. Basal area weighted mean diameter (median of basal area)}

\item{vegetation}{Optional for improved function in Spruce. Necessary for Pine. One of [forester::Finland_vegetation_types()]}
}
\value{
Annual volume-increment percentage for next five year period.
}
\description{
OBS. Controlled, but may differ from tables (errata correction?)
}
\details{
Spruce function no. 18 (diameter)\n n= 146. R^2=0.893. se percent 16.8.\n
Spruce function no. 19 (w/o diameter)\n n= 146. R^2=0.884. se percent 17.3.\n

Pine function no. 16. \n n=352, R^2=0.904, se percent 16.1\n
Pine function no. 17.\n n=352. R^2=0.893, se percent 16.9.\n
}
