% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUGIN_UNGSKOG.R
\name{Nystrom_1987_age_of_individual_young_trees}
\alias{Nystrom_1987_age_of_individual_young_trees}
\title{Determine age at breast height for young trees}
\source{
Nyström, K. & Söderberg, U. 1987. Tillväxtberäkningen för ungskog i Hugin-systemet. SLU, Inst. f. skogsskötsel. Arbetsrapport 18. URL: \url{https://www.heurekaslu.se/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf}
}
\usage{
Nystrom_1987_age_of_individual_young_trees(
  height,
  species,
  mean_height,
  H100_Spruce,
  H100_Pine,
  source_sprout = 0
)
}
\arguments{
\item{height}{Total tree height in meters}

\item{species}{One of 'Pinus sylvestris', 'Picea abies', 'Betula pendula', 'Betula pubescens'}

\item{mean_height}{Arithmetic mean height (m) of future crop trees (1600 stems per ha)}

\item{H100_Spruce}{Meters. for Spruce [forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce()],[forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce()]}

\item{H100_Pine}{Meters. for Pine and Birch [forester::Hagglund_1974_Sweden_height_trajectories_Pine()]}

\item{source_sprout}{1 for sprout, else 0 (seed).}
}
\value{
age at breast height of young trees
}
\description{
Elfving 2010: Functions for determination of age at breast height for young trees were presented by
Nyström & Söderberg (1987). They were based on data from un-damaged sample-trees with
h>1.3 m from the Hugin young stand survey, in total about 10 000 trees. Sampling
probabilities were used as weights in the regressions


\strong{Summary Statistics, Pine}
\itemize{
\item R^2 : 0.96
\item Sres : 2.1
}

\strong{Summary Statistics, Spruce}
\itemize{
\item R^2 : 0.92
\item Sres : 3.0
}

\strong{Summary Statistics, Birch}
\itemize{
\item R^2 : 0.91
\item Sres : 3.1
}
}
