% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUGIN_UNGSKOG.R
\name{Nystrom_1987_DBH_of_individual_young_trees}
\alias{Nystrom_1987_DBH_of_individual_young_trees}
\title{Estimate DBH from height, stand density, etc.}
\source{
Nyström, K. & Söderberg, U. 1987. Tillväxtberäkningen för ungskog i Hugin-systemet. SLU, Inst. f. skogsskötsel. Arbetsrapport 18. URL: \url{https://www.heurekaslu.se/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf}
}
\usage{
Nystrom_1987_DBH_of_individual_young_trees(
  species = "Picea abies",
  height,
  mean_height,
  dominant_height,
  stems,
  proportion_broadleaves,
  sprouts,
  natural_regeneration,
  PCT,
  PCT_years,
  latitude,
  altitude,
  distance_to_coast,
  H100_Pine,
  vegetation
)
}
\arguments{
\item{species}{One of: 'Pinus sylvestris','Picea abies' or 'Betula'}

\item{height}{Tree height, meters}

\item{dominant_height}{Mean height, m, of the three highest trees on the plot.}

\item{stems}{Number of stems per ha.}

\item{proportion_broadleaves}{Proportion (of stems?) which are Broadleaf.}

\item{sprouts}{number of sprouts exceeding 1/2 of dominant_height.}

\item{natural_regeneration}{Boolean.}

\item{PCT}{Boolean. Pre-Commercial thinning performed < 10 years ago?}

\item{PCT_years}{Numeric. Number of years since PCT was performed.}

\item{latitude}{Latitude, degrees N.}

\item{altitude}{Meters above sea level.}

\item{distance_to_coast}{Numeric. (If site is more than 5 (sic?) (TODO: check 50 km) from coast. else 0. E.g. [forester::coast_distance()])}

\item{H100_Pine}{Site index for Scots Pine, e.g. [forester::Hagglund_1974_Sweden_height_trajectories_Pine()]}

\item{vegetation}{Integer. Code from Swedish NFI Field Layer classification: [forester::Sweden_vegetation_types()]}
}
\value{
DBH, cm of individual tree.
}
\description{
Based on plots from the HUGIN young stand survey re-measured 1981-1983, validated with plots from 1984.


Summary Statistics: Pine
\itemize{
\item n = 9524
\item R = 0.91
\item sf = 0.42
}

Summary Statistics: Spruce
\itemize{
\item n = 11706
\item R = 0.95
\item sf = 0.36
}

Summary Statistics Birch
\itemize{
\item n=5476
\item R = 0.93
\item sf = 0.47
}
}
\details{
To calculate sum of squared heights the number of stems has been included.
}
