% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Opdahl_1989_BA_mean_diameter_before_thinning_Norway_Aspen.R
\name{Opdahl_1989_BA_mean_diameter_before_thinning_Norway_Aspen}
\alias{Opdahl_1989_BA_mean_diameter_before_thinning_Norway_Aspen}
\title{Basal area mean diameter before thinning for Aspen stands in Norway from Opdahl 1989.}
\source{
Opdahl, H. 1989. Avsmaling og volum hos osp (Populus tremula L.) i Sør-Norge. (Tapering and volume of Aspen (Populus tremula L.) in South Norway.) Medd. Nor. Inst. Skogforsk. 43.2.:42s. in
Opdahl, H. 1991. Bonitet, vekst og produksjon hos osp (Populus tremula L.) i Sør-Norge. (Site-index, growth and yield in Aspen (Populus tremula L.) stands in South Norway.) Medd. Skogforsk. 44(11):1-44. ISBN 82-7169-527-4. ISSN 0803-2866., page 21. (function 6).
}
\usage{
Opdahl_1989_BA_mean_diameter_before_thinning_Norway_Aspen(
  SIH40,
  dominant_height,
  stems_per_ha_before_thinning,
  correction = TRUE
)
}
\arguments{
\item{SIH40}{Site Index H40, e.g. [forester::Opdahl_1991_height_trajectory_Norway_Aspen()]}

\item{dominant_height}{Dominant height of stand, m. (arithmetic mean of 100 trees with largest diameter per hectare)}

\item{stems_per_ha_before_thinning}{Number of Stems per hectare before thinning.}

\item{correction}{Should the function use the correction used in the original simulation? E.g. return 96.7%. Default =TRUE}
}
\value{
Basal area mean diameter (cm).
}
\description{
Basal area mean diameter before thinning for Aspen stands in Norway from Opdahl 1989.
}
