% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opdahl_1991.R
\name{Opdahl_1989_volume_over_bark_Norway_Aspen}
\alias{Opdahl_1989_volume_over_bark_Norway_Aspen}
\title{Volume over bark for individual Aspen trees in Norway from Opdahl 1991.}
\source{
Opdahl, H. 1991. Bonitet, vekst og produksjon hos osp (Populus tremula L.) i Sør-Norge. (Site-index, growth and yield in Aspen (Populus tremula L.) stands in South Norway.) Medd. Skogforsk. 44(11):1-44. ISBN 82-7169-527-4. ISSN 0803-2866. p. 23
}
\usage{
Opdahl_1989_volume_over_bark_Norway_Aspen(
  diameter_cm,
  height_m,
  age,
  correction = TRUE
)
}
\arguments{
\item{diameter_cm}{Diameter of tree in cm.}

\item{height_m}{Height of tree in m.}

\item{age}{Age (?)}

\item{correction}{Limit as per p.23 if below 8 cm.}
}
\description{
Volume over bark for individual aspen trees in Norway.
}
