% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opdahl_1991_height_trajectory_Norway_Aspen.R
\name{Opdahl_1991_height_trajectory_Norway_Aspen}
\alias{Opdahl_1991_height_trajectory_Norway_Aspen}
\title{Height trajectories for Aspen stands in Norway from Opdahl 1991.}
\source{
Opdahl, H. 1991. Bonitet, vekst og produksjon hos osp (Populus tremula L.) i Sør-Norge. (Site-index, growth and yield in Aspen (Populus tremula L.) stands in South Norway.) Medd. Skogforsk. 44(11):1-44. ISBN 82-7169-527-4. ISSN 0803-2866.
}
\usage{
Opdahl_1991_height_trajectory_Norway_Aspen(dominant_height, age, age2)
}
\arguments{
\item{dominant_height}{Dominant Height, metres.}

\item{age}{Age at breast height.}

\item{age2}{Age at breast height at which to calculate new height.}
}
\value{
Top Height at age at breast height == age2
}
\description{
Functions 1:5, pages 14-15.
}
\examples{
Opdahl_1991_height_trajectory_Norway_Aspen(dominant_height=5,age = 10,age2 = 40)
}
