% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oyen_Tveite_1998_translate_yield_West_Norway.R
\name{Oyen_Tveite_1998_translate_yield_West_Norway}
\alias{Oyen_Tveite_1998_translate_yield_West_Norway}
\title{Stem volume yield between different tree species in West Norway from Øyen & Tveite 1998.}
\source{
Øyen & Tveite 1998. En sammenligning av høydebonitet og produksjonsevne mellom ulika treslag på samme vokstested i Vest-Norge. A comparison of site index class and potential stem volume yield between different tree species growing on equal sites in West Norway. Rapport fra skogforskningen. Nr. 15. 1-32.
}
\usage{
Oyen_Tveite_1998_translate_yield_West_Norway(
  species_from,
  species_to,
  potential_yield_m3_ha_yr
)
}
\arguments{
\item{species_from}{One of : "Picea abies","Pinus sylvestris","Betula pubescens","Picea sitchensis"}

\item{species_to}{One of : "Picea abies","Pinus sylvestris","Betula pubescens","Picea sitchensis"}

\item{potential_yield_m3_ha_yr}{Estimated maximum mean annual yield increment.}
}
\value{
Value.
}
\description{
Stem volume yield between different tree species in West Norway from Øyen & Tveite 1998.
}
\examples{
Oyen_Tveite_1998_translate_yield_West_Norway(species_from = "Picea abies",species_to = "Picea sitchensis",potential_yield_m3_ha_yr = 5)
}
