% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Persson_1992.R
\name{Persson_1992_average_basal_area_Pine}
\alias{Persson_1992_average_basal_area_Pine}
\title{Average Basal area in stands of Scots Pine from Persson, O. (1992)}
\source{
Persson, O. (1992) En produktionsmodell för tallskog i Sverige:
A growth simulator for Scots Pine (Pinus sylvestris, L.) in Sweden.
Report nr. 31. Dept. of Forest Yield Research.
Swedish University of Agricultural Sciences.
ISSN 0348-7636. Garpenberg. p. 59. pp. 206.
}
\usage{
Persson_1992_average_basal_area_Pine(H100, dominant_height)
}
\arguments{
\item{H100}{Site Index e.g. [forester::Hagglund_1974_Sweden_height_trajectories_Pine()]}

\item{dominant_height}{Dominant height m.}
}
\value{
Basal area at the end of the period, m2/ha.
}
\description{
Average Basal area in stands of Scots Pine from Persson, O. (1992)
}
\details{
number of growth periods: 2269

Coefficient of determination: 0.292

Standard deviation about the function (sf): 0.282

(sf)/standard deviation about the mean: 84.0%
}
