% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Persson_1992_BA_increment_Pine.R
\name{Persson_1992_BA_increment_Pine}
\alias{Persson_1992_BA_increment_Pine}
\title{Basal area increment under bark at breast height for Scots Pine from Persson (1992)}
\source{
Persson, O. (1992) En produktionsmodell för tallskog i Sverige:
A growth simulator for Scots Pine (Pinus sylvestris, L.) in Sweden.
Report nr. 31. Dept. of Forest Yield Research.
Swedish University of Agricultural Sciences.
ISSN 0348-7636. Garpenberg. p. 56. pp. 206.
}
\usage{
Persson_1992_BA_increment_Pine(
  basal_area_after_thinning,
  age_at_breast_height,
  SI,
  latitude
)
}
\arguments{
\item{basal_area_after_thinning}{Basal area at the beginning of the period, after thinning, under bark, m2/ha}

\item{age_at_breast_height}{Age at breast height}

\item{SI}{Site Index after Hägglund 1974, m.}

\item{latitude}{Latitude, degrees.}
}
\value{
Annual basal area increment under bark, m2/ha/yr
}
\description{
Basal area increment under bark at breast height for Scots Pine from Persson (1992)
}
\details{
Function no. 2.
More and less detailed functions available.

number of growth periods: 1163

Coefficient of determination: 0.790

Standard deviation about the function (sf): 0.219

(sf)/standard deviation about the mean: 45.9%
}
