% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Persson_1992_bark_Pine.R
\name{Persson_1992_bark_Pine}
\alias{Persson_1992_bark_Pine}
\title{Bark area for Scots Pine from Persson, O. (1992)}
\source{
Persson, O. (1992) En produktionsmodell för tallskog i Sverige:
A growth simulator for Scots Pine (Pinus sylvestris, L.) in Sweden.
Report nr. 31. Dept. of Forest Yield Research.
Swedish University of Agricultural Sciences.
ISSN 0348-7636. Garpenberg. p. 58. pp. 206.
}
\usage{
Persson_1992_bark_Pine(basal_area_above_bark, SI, stems, latitude)
}
\arguments{
\item{basal_area_above_bark}{Basal area over bark, m2/ha}

\item{SI}{Site Index, m. e.g. [forester::Hagglund_1974_Sweden_height_trajectories_Pine()]}

\item{stems}{Number of stems per ha.}

\item{latitude}{Latitude, degrees.}
}
\value{
Bark Area, m2/ha.
}
\description{
Bark area for Scots Pine from Persson, O. (1992)
}
\details{
number of growth periods: 697

Coefficient of determination: 0.896

Standard deviation about the function (sf): 0.103

(sf)/standard deviation about the mean: 32.4%
}
