% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Persson_1992.R
\name{Persson_1992_mortality_diameter_quotient_Pine}
\alias{Persson_1992_mortality_diameter_quotient_Pine}
\title{Diameter ratio of the mortality stems in stands of Scots Pine from Persson (1992)}
\source{
Persson, O. (1992) En produktionsmodell för tallskog i Sverige:
A growth simulator for Scots Pine (Pinus sylvestris, L.) in Sweden.
Report nr. 31. Dept. of Forest Yield Research.
Swedish University of Agricultural Sciences.
ISSN 0348-7636. Garpenberg. p. 61. pp. 206.
}
\usage{
Persson_1992_mortality_diameter_quotient_Pine(dominant_height, stems, thinned)
}
\arguments{
\item{dominant_height}{Dominant height, m.}

\item{stems}{Stems per hectare}

\item{thinned}{TRUE / FALSE if the thinning removal at the start of the period is at least 10\% of the basal area, 1, otherwise 0.}
}
\value{
Ratio between the diameter of self-thinned stems and the diameter of the remaining stems.
}
\description{
Diameter ratio of the mortality stems in stands of Scots Pine from Persson (1992)
}
\details{
number of growth periods: 1134

Coefficient of determination: 0.188

Standard deviation about the function (sf): 0.309

(sf)/standard deviation about the mean: 90.2 \%
}
