% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Persson_1992_natural_mortality_Pine.R
\name{Persson_1992_natural_mortality_Pine}
\alias{Persson_1992_natural_mortality_Pine}
\title{Natural mortality in stands of Scots Pine from Persson, O. (1992)}
\source{
Persson, O. (1992) En produktionsmodell för tallskog i Sverige:
A growth simulator for Scots Pine (Pinus sylvestris, L.) in Sweden.
Report nr. 31. Dept. of Forest Yield Research.
Swedish University of Agricultural Sciences.
ISSN 0348-7636. Garpenberg. p. 60. pp. 206.
}
\usage{
Persson_1992_natural_mortality_Pine(
  basal_area_above_bark,
  dominant_height_m,
  SI,
  stems,
  latitude
)
}
\arguments{
\item{basal_area_above_bark}{Basal area over bark, m2/ha}

\item{dominant_height_m}{Dominant height m}

\item{SI}{Site Index, m. e.g. [forester::Hagglund_1974_Sweden_height_trajectories_Pine()]}

\item{stems}{Number of stems per ha.}

\item{latitude}{Latitude, degrees.}
}
\value{
Annual mortality, basal area on bark + 0.01 m2/ha/yr.
}
\description{
Natural mortality in stands of Scots Pine from Persson, O. (1992)
}
\details{
If basal_area_above_bark denser than average basal area cf. [forester::Persson_1992_average_basal_area_Pine()]

number of growth periods: 301

Coefficient of determination: 0.371

Standard deviation about the function (sf): 1.012

(sf)/standard deviation about the mean: 79.7%

If basal_area_above_bark thinner than average basal area cf. [forester::Persson_1992_average_basal_area_Pine()]

number of growth periods: 1968

Coefficient of determination: 0.173

Standard deviation about the function (sf): 0.981

(sf)/standard deviation about the mean: 91.0%
}
