% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Petersson_1999_biomass.R
\name{Petersson_1999_biomass}
\alias{Petersson_1999_biomass}
\title{Calculate biomass of tree according to Petersson (1999)}
\source{
Petersson, H. (1999). Biomassafunktioner för trädfaktorer av tall,
 gran och björk i Sverige. Eng.: Biomass functions for tree factors of
 Pine, Norway Spruce and Birch in Sweden. Report #59. Skogliga konsekvensanalyser 1999.
 Dept. of Forest Resources and Geomatics. Swedish University of Agricultural Sciences.
 ISSN 1401-1204, ISRN SLU-SRG-AR-59-SE. Umeå, Sweden.

 <https://pub.epsilon.slu.se/8782/1/Petersson_H_120502.pdf>
}
\usage{
Petersson_1999_biomass(
  species,
  diameter.cm,
  height.m,
  age_at_breast_height,
  SI,
  dominant_species,
  five_years_radial_increment.mm,
  peat,
  latitude,
  longitude,
  epsg_in = 4326,
  altitude
)
}
\arguments{
\item{species}{One of "Picea abies","Pinus sylvestris","Betula pendula" or
"Betula pubescens".}

\item{diameter.cm}{Diameter at breast height, in cm.}

\item{height.m}{Height of tree, in metres.}

\item{age_at_breast_height}{Age at breast height.}

\item{SI}{Site index of dominating tree species according to Hägglund &
Lundmark (1977).}

\item{dominant_species}{For which the SI was calculated. "Picea abies" or
"Pinus sylvestris".}

\item{five_years_radial_increment.mm}{five years radial increment, in mm.}

\item{peat}{TRUE if soil type is peat, otherwise FALSE. Soil type is peat
if the tree plot is more than 30 cm deep peat.}

\item{latitude}{latitude}

\item{longitude}{longitude}

\item{altitude}{Metres above sea level.}

\item{espg_in}{EPSG of Latitude and Longitude, default=4326.}
}
\value{
List of weights, in g.
}
\description{
Forests at high altitudes and close to coast are not represented.
}
