% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Petersson_Stahl_2006.R
\name{Petersson_Stahl_2006_below_ground_biomass_Birch}
\alias{Petersson_Stahl_2006_below_ground_biomass_Birch}
\title{Below-ground biomass for Birch from Petersson, H. & Stahl. G (2006)}
\source{
Petersson, H., Ståhl, G. (2006) "Functions for below-ground biomass of
Pinus sylvestris, Picea abies, Betula pendula and Betula pubescens in Sweden.
Scandinavian Journal of Forest Research. Vol. 21. pp. 84-93. DOI:
<https://doi.org/10.1080/14004080500486864>
}
\usage{
Petersson_Stahl_2006_below_ground_biomass_Birch(diameter.cm, Root_detail)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, in cm.}

\item{Root_detail}{2 or 5 mm resolution of smallest included roots.}
}
\value{
Below-ground biomass (g)
}
\description{
n=13
}
\details{
Betula pendula & Betula pubescens.
}
