% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Petersson_Stahl_2006.R
\name{Petersson_Stahl_2006_below_ground_biomass_Spruce}
\alias{Petersson_Stahl_2006_below_ground_biomass_Spruce}
\title{Below-ground biomass for Spruce from Petersson, H. & Stahl. G (2006)}
\source{
Petersson, H., Ståhl, G. (2006) "Functions for below-ground biomass of
Pinus sylvestris, Picea abies, Betula pendula and Betula pubescens in Sweden.
Scandinavian Journal of Forest Research. Vol. 21. pp. 84-93. DOI:
<https://doi.org/10.1080/14004080500486864>
}
\usage{
Petersson_Stahl_2006_below_ground_biomass_Spruce(
  diameter.cm,
  age_at_breast_height,
  height.m,
  crown_base_height.m,
  height.m.basal_area_weighted,
  age_basal_area_weighted,
  basal_area,
  altitude,
  Dry_soil,
  Root_detail
)
}
\arguments{
\item{diameter.cm}{Diameter at breast height, in cm.}

\item{age_at_breast_height}{Age at breast height.}

\item{height.m}{Tree height, m.}

\item{crown_base_height.m}{Height to the lowest living green branch not
separated by more than three whorls from the crown.}

\item{height.m.basal_area_weighted}{Basal area weighted height of stand.}

\item{age_basal_area_weighted}{Basal area weighted age of stand.}

\item{basal_area}{Basal area of the stand, m2 per hectare.}

\item{altitude}{Meters above sea level.}

\item{Dry_soil}{If Dry or Mesic : 1, otherwise 0.}

\item{Root_detail}{2 or 5 mm resolution of smallest included roots.}
}
\description{
n=339
}
