% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pettersson_1955_funcs.R
\name{Petterson_1955_DiameterClasses}
\alias{Petterson_1955_DiameterClasses}
\title{Petterson 1955 Diameter Classes for truncated normal distributions.}
\usage{
Petterson_1955_DiameterClasses(Phi0 = 3, Sigman0, alpha0, nClasses)
}
\arguments{
\item{Phi0}{Range of the truncated distribution from upper limit to
truncation expressed in standard deviations of the normal distribution. e.g.
untruncated normal distribution Phi0 is 6.}

\item{Sigman0}{Standard deviation of the normal distribution.}

\item{alpha0}{The truncation point.}

\item{nClasses}{Number of desired diameter classes.}
}
\description{
Petterson 1955 Diameter Classes for truncated normal distributions.
}
\details{
The parameters suffix '0' indicates that this is the initial state.
}
