% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pettersson_1955_funcs.R
\name{Petterson_1955_time_to_BH}
\alias{Petterson_1955_time_to_BH}
\title{Pettersson time to breast height.}
\source{
Pettersson, H. 1955. Barrskogens volymproduktion (Die
Massenproduktion des Nadelwaldes). Reports of the forest research institute
of Sweden. Vol 45:1. Centraltryckeriet Esselte AB. Stockholm. Available
Online (07/08/2022): \href{PUB EPSILON SLU}{https://pub.epsilon.slu.se/9982/1/medd_statens_skogsforskningsinst_045_01_a.pdf}
}
\usage{
Petterson_1955_time_to_BH(
  species,
  H100 = 20,
  planted = 0,
  northern = 0,
  q_val = 0.7
)
}
\arguments{
\item{species}{One of "Picea abies" or "Pinus sylvestris}

\item{H100}{Dominant height at total age 100.}

\item{planted}{1 if planted, 0 if natural regeneration.}

\item{northern}{1 if in northern sweden, else 0 if in southern sweden.}
}
\value{
time to breast height (years).
}
\description{
Pettersson time to breast height.
}
\details{
This is a smoothed function over calculated years to breast height
1.3 m in Petterson 1955. On the form: \eqn{timeToBH = Asymptote + (R0 - Asymptote) * exp(-exp(lrc) * H100)}
Which is included in the height_trajectory functions.
}
