% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pettersson_1955_funcs.R
\name{Pettersson_1955_height}
\alias{Pettersson_1955_height}
\alias{Pettersson_1955_height_trajectory_Northern_Sweden_Pine}
\alias{Pettersson_1955_height_trajectory_southern_Sweden_Pine}
\alias{Pettersson_1955_height_trajectory_southern_Sweden_Spruce}
\alias{Pettersson_1955_height_trajectory_northern_Sweden_Spruce}
\title{Dominant height development of Norway Spruce and Scots Pine in Sweden.}
\source{
Pettersson, H. 1955. Barrskogens volymproduktion (Die
Massenproduktion des Nadelwaldes). Reports of the forest research institute
of Sweden. Vol 45:1. Centraltryckeriet Esselte AB. Stockholm. Available
Online (07/08/2022): \href{PUB EPSILON SLU}{https://pub.epsilon.slu.se/9982/1/medd_statens_skogsforskningsinst_045_01_a.pdf}
}
\usage{
Pettersson_1955_height_trajectory_Northern_Sweden_Pine(
  dominant_height,
  age,
  age2,
  ageType = 1,
  planted = 0,
  q_val = 0.7
)

Pettersson_1955_height_trajectory_southern_Sweden_Pine(
  dominant_height,
  age,
  age2,
  ageType = 1,
  planted = 0,
  q_val = 0.7
)

Pettersson_1955_height_trajectory_southern_Sweden_Spruce(
  dominant_height,
  age,
  age2,
  ageType = 1,
  planted = 0,
  q_val = 0.7
)

Pettersson_1955_height_trajectory_northern_Sweden_Spruce(
  dominant_height,
  age,
  age2,
  ageType = 1,
  planted = 0,
  q_val = 0.7
)
}
\arguments{
\item{dominant_height}{Dominant height of the stand (Height corresponding to \eqn{3\sigma} of the diameter distribution)}

\item{age}{Age of stand.}

\item{age2}{Desired age of stand.}

\item{ageType}{If 1: Total age. If 2: Age at breast height (1.3m).}

\item{planted}{If 0 (default), no. If 1, planted.}

\item{q_val}{Relation between the time to breast height for planted / non-planted stands.}
}
\value{
Dominant height of the stand in metres at desired age.
}
\description{
Dominant height development of Norway Spruce and Scots Pine in Sweden.
}
