% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Pettersson_1992_BA_weighted_mean_diameter_after_PCT_Pine.R
\name{Pettersson_1992_BA_weighted_mean_diameter_after_PCT_Pine}
\alias{Pettersson_1992_BA_weighted_mean_diameter_after_PCT_Pine}
\title{Basal area weighted mean diameter after PCT in young stands of Scots Pine, from Pettersson 1992.}
\source{
Pettersson Nils (1992) The effect on stand development of different
spacing after planting and precommercial thinning in Norway Spruce (Picea abies
 (L.) Karst.) and Scots Pine (Pinus sylvestris L.) stands: The effect of density
 after precommercial thinning on volume and structure in Pinus sylvestris and
 Picea abies stands. Diss. Report no. 34, Dept. of Forest Yield Research.
 ISSN 0348-7636. p. 9.

 @details

 R^2 = 0.95
 F=1563
}
\usage{
Pettersson_1992_BA_weighted_mean_diameter_after_PCT_Pine(
  dominant_height,
  diameter_mean_basal_area_stem
)
}
\arguments{
\item{dominant_height}{Dominant height, m.}

\item{diameter_mean_basal_area_stem}{Diameter corresponding to the mean basal area stem, cm.}
}
\value{
Basal area weighted mean diameter, cm after thinning
}
\description{
Basal area weighted mean diameter after PCT in young stands of Scots Pine, from Pettersson 1992.
}
