% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pettersson_1992.R
\name{Pettersson_1992_form_quotient_after_PCT_Spruce}
\alias{Pettersson_1992_form_quotient_after_PCT_Spruce}
\title{Form quotient after PCT in stands of Norway Spruce, from Pettersson 1992.}
\source{
Pettersson Nils (1992) The effect on stand development of different
spacing after planting and precommercial thinning in Norway Spruce (Picea abies
 (L.) Karst.) and Scots Pine (Pinus sylvestris L.) stands: The effect of density
 after precommercial thinning on volume and structure in Pinus sylvestris and
 Picea abies stands. Diss. Report no. 34, Dept. of Forest Yield Research.
 ISSN 0348-7636. p. 9.
}
\usage{
Pettersson_1992_form_quotient_after_PCT_Spruce(dominant_height, stems)
}
\arguments{
\item{dominant_height}{Dominant height, metres.}

\item{stems}{Stems per hectare}
}
\value{
Form quotient Spruce after PCT
}
\description{
Form quotient after PCT in stands of Norway Spruce, from Pettersson 1992.
}
\details{
F=362
R^2 = 0.81
}
