% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pettersson_1992.R
\name{Pettersson_1992_total_volume_yield_after_PCT_Pine}
\alias{Pettersson_1992_total_volume_yield_after_PCT_Pine}
\title{Total volume yield after PCT in young stands of Scots Pine, from Pettersson 1992.}
\source{
Pettersson Nils (1992) The effect on stand development of different
spacing after planting and precommercial thinning in Norway Spruce (Picea abies
 (L.) Karst.) and Scots Pine (Pinus sylvestris L.) stands: The effect of density
 after precommercial thinning on volume and structure in Pinus sylvestris and
 Picea abies stands. Diss. Report no. 34, Dept. of Forest Yield Research.
 ISSN 0348-7636. p. 7.

 @details R^2 = 0.97, for both classes.
}
\usage{
Pettersson_1992_total_volume_yield_after_PCT_Pine(stems, dominant_height, SI)
}
\arguments{
\item{stems}{Number of stems per hectare}

\item{dominant_height}{Dominant height, metres}

\item{SI}{Site Index, according to Hägglund 1974 [forester::Hagglund_1974_Sweden_height_trajectories_Pine]}
}
\value{
Total volume yield m3/ha.
}
\description{
Total volume yield after PCT in young stands of Scots Pine, from Pettersson 1992.
}
