% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Pettersson_1992_young_stands_BA_weighted_diameter_Spruce.R
\name{Pettersson_1992_young_stands_BA_weighted_diameter_Spruce}
\alias{Pettersson_1992_young_stands_BA_weighted_diameter_Spruce}
\title{Basal area weighted mean diameter in young stands of Norway Spruce, from Pettersson 1992}
\source{
Pettersson Nils (1992) The effect on stand development of different
spacing after planting and precommercial thinning in Norway Spruce (Picea abies
 (L.) Karst.) and Scots Pine (Pinus sylvestris L.) stands. Diss. Report no. 34, Dept. of Forest Yield Research.
 ISSN 0348-7636. p. 25.
}
\usage{
Pettersson_1992_young_stands_BA_weighted_diameter_Spruce(
  dominant_height,
  diameter_mean_basal_area_stem
)
}
\arguments{
\item{dominant_height}{Dominant height of the stand, in meters.}

\item{diameter_mean_basal_area_stem}{Diameter corresponding to the mean stem basal area.}
}
\value{
Basal area weighted diameter, in cm.
}
\description{
Basal area weighted mean diameter in young stands of Norway Spruce, from Pettersson 1992
}
\details{
function 18.

No. of observations = 100
R^2 = 0.95
}
