% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pettersson_1992.R
\name{Pettersson_1992_young_stands_form_quotient_Spruce}
\alias{Pettersson_1992_young_stands_form_quotient_Spruce}
\title{Form quotient in young stands of Norway Spruce, from Pettersson 1992}
\source{
Pettersson Nils (1992) The effect on stand development of different
spacing after planting and precommercial thinning in Norway Spruce (Picea abies
 (L.) Karst.) and Scots Pine (Pinus sylvestris L.) stands. Diss. Report no. 34, Dept. of Forest Yield Research.
 ISSN 0348-7636. p. 25.
}
\usage{
Pettersson_1992_young_stands_form_quotient_Spruce(
  dominant_height,
  initial_stems,
  SI
)
}
\arguments{
\item{dominant_height}{Dominant height of the stand, in meters.}

\item{initial_stems}{Initial number of stems.}

\item{SI}{SI from Hägglund 1972,1973.}
}
\value{
Form quotient
}
\description{
Form quotient in young stands of Norway Spruce, from Pettersson 1992
}
\details{
function 14.

No. of observations = 198
R^2 = 0.97
}
