% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pettersson_1992.R
\name{Pettersson_1992_young_stands_living_basal_area_Pine}
\alias{Pettersson_1992_young_stands_living_basal_area_Pine}
\title{Living basal area in young stands of Scots Pine, from Pettersson 1992}
\source{
Pettersson Nils (1992) The effect on stand development of different
spacing after planting and precommercial thinning in Norway Spruce (Picea abies
 (L.) Karst.) and Scots Pine (Pinus sylvestris L.) stands. Diss. Report no. 34, Dept. of Forest Yield Research.
 ISSN 0348-7636. p. 25.
}
\usage{
Pettersson_1992_young_stands_living_basal_area_Pine(
  dominant_height,
  initial_stems,
  basal_area_m2_ha
)
}
\arguments{
\item{dominant_height}{Dominant height of the stand, in meters.}

\item{initial_stems}{Initial number of stems.}

\item{basal_area_m2_ha}{Basal Area m2/ha.}
}
\value{
Living basal area
}
\description{
Living basal area in young stands of Scots Pine, from Pettersson 1992
}
\details{
function 15.

No. of observations = 86
R^2 = 0.99
}
