% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pettersson_1992.R
\name{Pettersson_1992_young_stands_mean_diameter_x_trees_Pine}
\alias{Pettersson_1992_young_stands_mean_diameter_x_trees_Pine}
\title{Mean diameter of the thickest trees per hectare in young stands of Scots Pine, from Pettersson 1992}
\source{
Pettersson Nils (1992) The effect on stand development of different
spacing after planting and precommercial thinning in Norway Spruce (Picea abies
 (L.) Karst.) and Scots Pine (Pinus sylvestris L.) stands. Diss. Report no. 34, Dept. of Forest Yield Research.
 ISSN 0348-7636. p. 27.
}
\usage{
Pettersson_1992_young_stands_mean_diameter_x_trees_Pine(
  dominant_height,
  diameter_mean_basal_area_stem,
  thickest_x_trees
)
}
\arguments{
\item{dominant_height}{Dominant height of the stand, in meters.}

\item{diameter_mean_basal_area_stem}{Diameter corresponding to the mean stem basal area.}

\item{thickest_x_trees}{One of '800', '400' or '100'.}
}
\value{
Basal area weighted diameter, in cm.
}
\description{
Mean diameter of the thickest trees per hectare in young stands of Scots Pine, from Pettersson 1992
}
\details{
function 19. Mean diameter of thickest 800 trees per hectare.

No. of observations = 86
R^2 = 0.91


function 20. Mean diameter of thickest 400 trees per hectare.

No. of observations = 86
R^2 =  0.92


function 21. Mean diameter of thickest 100 trees per hectare.

No. of observations = 86
R^2 = 0.87
}
