% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pettersson_1992.R
\name{Pettersson_1992_young_stands_mean_height_Lorey_Pine}
\alias{Pettersson_1992_young_stands_mean_height_Lorey_Pine}
\title{Relationship between dominant height and Lorey's mean height in young stands of Scots Pine}
\source{
Pettersson Nils (1992) The effect on stand development of different
spacing after planting and precommercial thinning in Norway Spruce (Picea abies
 (L.) Karst.) and Scots Pine (Pinus sylvestris L.) stands. Diss. Report no. 34, Dept. of Forest Yield Research.
 ISSN 0348-7636. p. 20.
}
\usage{
Pettersson_1992_young_stands_mean_height_Lorey_Pine(
  dominant_height,
  initial_stems
)
}
\arguments{
\item{dominant_height}{Dominant height of the stand, in meters.}

\item{initial_stems}{Initial number of stems.}
}
\value{
Lorey's Mean height of the stand.
}
\description{
Relationship between dominant height and Lorey's mean height in young stands of Scots Pine
}
\details{
Number of observations: 86.

R^2 = 0.98.
}
