% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pettersson_1992.R
\name{Pettersson_1992_young_stands_mean_height_Lorey_Spruce}
\alias{Pettersson_1992_young_stands_mean_height_Lorey_Spruce}
\title{Relationship between dominant height and Lorey's mean height in young stands of Norway Spruce}
\source{
Pettersson Nils (1992) The effect on stand development of different
spacing after planting and precommercial thinning in Norway Spruce (Picea abies
 (L.) Karst.) and Scots Pine (Pinus sylvestris L.) stands. Diss. Report no. 34, Dept. of Forest Yield Research.
 ISSN 0348-7636. p. 21.
}
\usage{
Pettersson_1992_young_stands_mean_height_Lorey_Spruce(
  dominant_height,
  initial_stems,
  SI
)
}
\arguments{
\item{dominant_height}{Dominant height of the stand, in meters.}

\item{initial_stems}{Initial number of stems.}

\item{SI}{Site Index for Spruce, according to Hägglund 1972,1973, e.g.
[forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce]
[forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce]}
}
\value{
Lorey's Mean height of the stand.
}
\description{
Relationship between dominant height and Lorey's mean height in young stands of Norway Spruce
}
\details{
Number of observations: 100.

R^2 = 0.97.
}
