% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Pettersson_1992_young_stands_total_volume_Pine.R
\name{Pettersson_1992_young_stands_total_volume_Pine}
\alias{Pettersson_1992_young_stands_total_volume_Pine}
\title{Total volume production in young stands of Scots Pine, from Pettersson 1992}
\source{
Pettersson Nils (1992) The effect on stand development of different
spacing after planting and precommercial thinning in Norway Spruce (Picea abies
 (L.) Karst.) and Scots Pine (Pinus sylvestris L.) stands. Diss. Report no. 34, Dept. of Forest Yield Research.
 ISSN 0348-7636. p. 21.
}
\usage{
Pettersson_1992_young_stands_total_volume_Pine(dominant_height, initial_stems)
}
\arguments{
\item{dominant_height}{Dominant height of the stand, in meters.}

\item{initial_stems}{Initial number of stems.}
}
\value{
Total volume production, m3sk / ha.
}
\description{
Total volume production in young stands of Scots Pine, from Pettersson 1992
}
\details{
function 3.
}
