% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Pettersson_1992_young_stands_volume_distribution_Spruce.R
\name{Pettersson_1992_young_stands_volume_distribution_Spruce}
\alias{Pettersson_1992_young_stands_volume_distribution_Spruce}
\title{Volume distribution in young stands of Scots Pine, from Pettersson 1992}
\source{
Pettersson Nils (1992) The effect on stand development of different
spacing after planting and precommercial thinning in Norway Spruce (Picea abies
 (L.) Karst.) and Scots Pine (Pinus sylvestris L.) stands. Diss. Report no. 34, Dept. of Forest Yield Research.
 ISSN 0348-7636. p. 24.
}
\usage{
Pettersson_1992_young_stands_volume_distribution_Spruce(
  dominant_height,
  diameter_corresponding_to_mean_basal_area_cm,
  initial_stems
)
}
\arguments{
\item{dominant_height}{Dominant height of the stand, in meters.}

\item{diameter_corresponding_to_mean_basal_area_cm}{Diameter corresponding to the mean basal area stem, in cm.}

\item{initial_stems}{Initial number of stems.}
}
\value{
List of distribution metrics.
}
\description{
Volume distribution in young stands of Scots Pine, from Pettersson 1992
}
\details{
function 9-12.

f. 9. mean Volume:

No. of observations = 100
R^2= 0.94



f. 10 Standard Deviation of Volume:

No. of observations = 100
R^2 = 0.55



f. 11 Volume skew

No. of observations = 100
R^2 = 0.32



f. 12 Volume kurtosis

No. of observations = 100
R^2 = 0.42
}
