% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pettersson_1955_distribution.R
\name{PhiTable}
\alias{PhiTable}
\title{Pettersson Structural Values related to Phi.}
\source{
Pettersson, H. 1955. Barrskogens volymproduktion (Die
Massenproduktion des Nadelwaldes). Reports of the forest research institute
of Sweden. Vol 45:1. Centraltryckeriet Esselte AB. Stockholm. Available
Online (07/08/2022): \href{PUB EPSILON SLU}{https://pub.epsilon.slu.se/9982/1/medd_statens_skogsforskningsinst_045_01_a.pdf}
}
\usage{
PhiTable(Phi)
}
\arguments{
\item{Phi}{Distance from the normal distributions upper boundary (cut-off),
at mean + 3 sigma.}
}
\value{
A list of values.
}
\description{
#Compare to table on p. 325.
}
\details{
This is used in Henrik Pettersson growth and yield model to
describe stem distribution.
}
