% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picea_spp_vol.R
\name{picea_spp_vol}
\alias{picea_spp_vol}
\title{Picea volume}
\usage{
picea_spp_vol(height.m, dbh.cm, latitude, crown.base.height.m)
}
\arguments{
\item{height.m}{Tree height in metres.}

\item{dbh.cm}{Diameter breast height \strong{(1.3 m)}, in cm.}

\item{latitude}{The latitudinal placement of the tree, in degrees N.}

\item{crown.base.height.m}{The lowest height of the living tree crown, in metres.}

\item{species.latin}{One of "Picea ...", "Pinus sylvestris", "Larix sibirica", "Abies ...".}

\item{double.bark.mm}{The sum of bark thickness on polar ends of the diameter as measured at 1.3m height, in millimetres.}
}
\value{
Volume of the tree, in dm3
}
\description{
Background function for tree.vol.sweden.
}
\examples{
picea_spp_vol(height.m=3, dbh.cm=4, latitude=61)
}
