% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pinus}
\alias{pinus}
\title{\emph{Pinus Taeda} Data}
\source{
Mente, S.R. and Lombardo, F. (2005). A recursive-partitioning model
for blood-brain barrier permeation, \emph{Journal of Computer-Aided Molecular Design},
Vol. 19, pg. 465-481.
}
\value{
um data.frame com dados de pinus.
}
\description{
Mente and Lombardo (2005) develop models to predict the log
of the ratio of the concentration of a compound in the brain
and the concentration in blood. For each compound, they computed
three sets of molecular descriptors: MOE 2D, rule-of-five and
Charge Polar Surface Area (CPSA). In all, 134 descriptors were
calculated. Included in this package are 208 non-proprietary
literature compounds. The vector \code{desc} contains the concentration
}
\keyword{datasets}
