% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinus_spp_vol.R
\name{pinus_spp_vol}
\alias{pinus_spp_vol}
\title{Pinus spp volume}
\usage{
pinus_spp_vol(
  species.latin,
  dbh.cm,
  height.m,
  crown.base.height.m,
  double.bark.mm,
  latitude
)
}
\arguments{
\item{species.latin}{Species name in latin, e.g. "Pinus sylvestris"}

\item{dbh.cm}{Diameter breast height in centimetres.}

\item{height.m}{Height in metres.}

\item{crown.base.height.m}{Height of crown base, in metres.}

\item{double.bark.mm}{Double bark, in mm.}

\item{latitude}{Latitude.}
}
\value{
Tree volume
}
\description{
Volume of Pinus spp.

Uses Brandel 1990 100-4 for all Pinus above 4.5 cm, for P. contorta 100-1.
For tree under 4.5 cm, Andersson 1954.
}
\examples{
pinus_spp_volume(species.latin="Pinus sylvestris", dbh.cm=10, height.m=10,latitude=63)
}
