% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedling_spacing.R
\name{plant_spacing}
\alias{plant_spacing}
\title{Plant Spacing}
\usage{
plant_spacing(
  distance,
  distance_a,
  distance_b,
  system = "square",
  stems_per_ha,
  area_m2 = 10000,
  output = "spacing"
)
}
\arguments{
\item{distance}{The \strong{square} system distance between plants}

\item{distance_a}{The \strong{in-row} distance between plants}

\item{distance_b}{The \strong{between-row} distance.}

\item{system}{Planting system, one of \emph{square} or \emph{rectangular}. Defaults to \strong{square}}

\item{stems_per_ha}{The number of steams planted per hectare}

\item{output}{One of \emph{stems per ha} or \emph{spacing}}

\item{area}{area in square metres. Defaults to 1 hectare, 10'000 sq. metres.}
}
\value{
Prints the output value.
}
\description{
Plant Spacing
}
\examples{

#Plant spacing for 2500 seedlings per hectare.
plant_spacing(stems_per_ha=2500, output="spacing")

}
