% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification.R
\name{plot.classification}
\alias{plot.classification}
\title{Plot of Metrics for Binary Classification Models}
\usage{
\method{plot}{classification}(train_output, models = NULL, type = "comparison", metric = "rmse")
}
\arguments{
\item{train_output}{An object, returned from \code{train()} function.}

\item{models}{A character or numeric that indicates which models
will be presented. If `NULL` (the default option) then the three best models
will be presented.}

\item{type}{A character one of `TO DO`
indicates the type of chart.}

\item{metric}{A character one of `accuracy`/`auc`/`f1` indicates the metric on the plots.}
}
\value{
a ggplot2 object
}
\description{
Function plot.classification plots measure of quality of regression models.
}
