% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiclass.R
\name{plot.multiclass}
\alias{plot.multiclass}
\title{Plot of Metrics for Multiclass Classification Models}
\usage{
\method{plot}{multiclass}(x, models = NULL, type = "comparison", metric = "accuracy", ...)
}
\arguments{
\item{x}{An object, returned from \code{train()} function.}

\item{models}{A character or numeric that indicates which models
will be presented. If `NULL` (the default option) then the three best models
will be presented.}

\item{type}{A character one of `comparison`, `confusion-matrix`, `train-test`
indicates the type of chart.}

\item{metric}{A character such as `accuracy` indicates the metric on the plots.}

\item{...}{Other parameters that are necessary for consistency with generic plot function.}
}
\value{
a ggplot2 object
}
\description{
Function plot.multiclass plots measure of quality of regression models.
}
\examples{
\dontrun{
library('forester')
data('iris')

x <- train(iris, 'Species', bayes_iter = 0, random_evals = 0)
plot(x)
}

}
