% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eko_PM_1985_basal_area_5_year_increment.R
\name{PM_BA}
\alias{PM_BA}
\alias{Eko_PM_1985_basal_area_5_year_increment_northern_Sweden_Pine}
\alias{Eko_PM_1985_basal_area_5_year_increment_central_Sweden_Pine}
\alias{Eko_PM_1985_basal_area_5_year_increment_southern_Sweden_Pine}
\alias{Eko_PM_1985_basal_area_5_year_increment_northern_Sweden_Spruce}
\alias{Eko_PM_1985_basal_area_5_year_increment_central_Sweden_Spruce}
\alias{Eko_PM_1985_basal_area_5_year_increment_southern_Sweden_Spruce}
\alias{Eko_PM_1985_basal_area_5_year_increment_northern_central_Sweden_Birch}
\alias{Eko_PM_1985_basal_area_5_year_increment_southern_Sweden_Birch}
\alias{Eko_PM_1985_basal_area_5_year_increment_northern_central_Sweden_Broadleaves}
\alias{Eko_PM_1985_basal_area_5_year_increment_southern_Sweden_Broadleaves}
\alias{Eko_PM_1985_basal_area_5_year_increment_Sweden_Beech}
\alias{Eko_PM_1985_basal_area_5_year_increment_Sweden_Oak}
\title{Basal area growth functions for even-aged monocultures or mixed stands in Sweden.}
\source{
Ekö, Per Magnus (1985) "En produktionsmodell för skog i Sverige, baserad på bestånd från
riksskogstaxeringens provytor: A growth simulator for Swedish forests, based on
data from the national forest survey. Rapporter nr. 16. Swedish University of
Agricultural Sciences, dept. of Silviculture. Umeå. ISBN 91-576-2386-4. ISSN
 0348-8969.  p. 63.
}
\usage{
Eko_PM_1985_basal_area_5_year_increment_northern_Sweden_Pine(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  ba_quotient_chronic_mortality,
  ba_quotient_acute_mortality
)

Eko_PM_1985_basal_area_5_year_increment_central_Sweden_Pine(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  fertilised,
  ba_quotient_chronic_mortality,
  ba_quotient_acute_mortality
)

Eko_PM_1985_basal_area_5_year_increment_southern_Sweden_Pine(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  fertilised,
  latitude,
  ba_quotient_chronic_mortality,
  ba_quotient_acute_mortality
)

Eko_PM_1985_basal_area_5_year_increment_northern_Sweden_Spruce(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  fertilised,
  latitude,
  ba_quotient_chronic_mortality = 0,
  ba_quotient_acute_mortality = 0
)

Eko_PM_1985_basal_area_5_year_increment_central_Sweden_Spruce(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  fertilised,
  latitude,
  ba_quotient_chronic_mortality,
  ba_quotient_acute_mortality
)

Eko_PM_1985_basal_area_5_year_increment_southern_Sweden_Spruce(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  fertilised,
  latitude,
  ba_quotient_chronic_mortality,
  ba_quotient_acute_mortality
)

Eko_PM_1985_basal_area_5_year_increment_northern_central_Sweden_Birch(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  fertilised,
  latitude,
  ba_quotient_chronic_mortality,
  ba_quotient_acute_mortality
)

Eko_PM_1985_basal_area_5_year_increment_southern_Sweden_Birch(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  fertilised,
  latitude,
  ba_quotient_chronic_mortality,
  ba_quotient_acute_mortality
)

Eko_PM_1985_basal_area_5_year_increment_northern_central_Sweden_Broadleaves(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  fertilised,
  latitude,
  ba_quotient_chronic_mortality
)

Eko_PM_1985_basal_area_5_year_increment_southern_Sweden_Broadleaves(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  fertilised,
  latitude,
  ba_quotient_chronic_mortality,
  ba_quotient_acute_mortality
)

Eko_PM_1985_basal_area_5_year_increment_Sweden_Beech(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  fertilised,
  latitude,
  ba_quotient_acute_mortality
)

Eko_PM_1985_basal_area_5_year_increment_Sweden_Oak(
  basal_area.m2,
  stem_number_ha,
  age_at_breast_height,
  basal_area_other_species,
  SI,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species,
  thinned,
  thinned_previous_5_years,
  vegetation,
  south_sweden,
  soil_moisture,
  altitude,
  TAX77 = 0,
  fertilised,
  latitude,
  ba_quotient_acute_mortality
)
}
\arguments{
\item{basal_area.m2}{Basal area over bark (m2/ha)}

\item{stem_number_ha}{Number of stems per hectare.}

\item{age_at_breast_height}{Age at breast height, calculated as the mean age of the two thickest trees (years).}

\item{basal_area_other_species}{Basal area of other species on the plot. (m2/ha)}

\item{SI}{Site index H100 Pine , m.}

\item{basal_area_weighted_mean_diameter}{Diameter corresponding to mean basal area (m).}

\item{basal_area_weighted_mean_diameter_other_species}{Diameter coressponding to mean basal area (m) of other species on plot.}

\item{thinned}{1 if the stand been thinned, otherwise 0.}

\item{thinned_previous_5_years}{1 if the stand has been thinned within 5 years, otherwise 0.}

\item{vegetation}{Variable indicating vegetation type scaled from -5 to +4 as follows:

\tabular{llr}{
Field Layer Code (NFI) \tab Description \tab Index \cr
1\tab Tall herbs w/o dwarf shrubs \tab 4 \cr
2\tab Tall herbs with bilberry \tab 2.5 \cr
3\tab Tall herbs with cowberry \tab 2 \cr
4\tab Low herbs w/o dwarf shrubs \tab 3 \cr
5\tab Low herbs with bilberry \tab 2.5 \cr
6\tab Low herbs with cowberry \tab 2 \cr
7\tab No field layer \tab 3 \cr
8\tab broad-leafed grasses \tab 2.5 \cr
9\tab narrow-leaved grasses \tab 1.5 \cr
10\tab Sedge, tall \tab -3 \cr
11\tab Sedge, low \tab -3 \cr
12\tab Horsetail \tab 1 \cr
13\tab Bilberry \tab 0 \cr
14\tab Cowberry \tab -0.5 \cr
15\tab Crowberry \tab -3 \cr
16\tab Poor shrub \tab -5 \cr
17\tab Lichen-rich \tab -0.5 \cr
18\tab Lichen-dominated \tab -1 \cr
}}

\item{south_sweden}{1 if the plot is situated in southern Sweden.}

\item{soil_moisture}{Type 1="Dry/torr",2="Mesic/frisk",3="Mesic-moist/frisk-fuktig",4="Moist/fuktig",5="Wet/Blöt"}

\item{altitude}{Metres above sea level}

\item{TAX77}{Always 0 for simulation. Otherwise see original text.}

\item{ba_quotient_chronic_mortality}{Quotient of basal area (total?) expected to die during the next five-year period, from e.g. crowding, drought (factors that have affected the growth of the stand)}

\item{ba_quotient_acute_mortality}{Quotient of basal area (total?) expected to die during the next five-year period, from e.g. wind or snowbreak, which do not affect the growth.}

\item{fertilised}{1 if the stand has been fertilised, otherwise 0.}

\item{latitude}{Latitude, degrees.}
}
\value{
Basal area growth over bark during the five year growth period.
}
\description{
\strong{Pine, northern Sweden}
Number of observations: 3461

\tabular{lrrr}{
Thinning \tab SI<160 \tab 160≤SI<200 \tab SI≥200 \cr
Unthinned \tab 920 \tab 1014 \tab 523 \cr
Thinned \tab 319 \tab 441 \tab 244 \cr
}

Multiple correlation coefficient: 0.91

Standard deviation about the function, Sf : 0.411

Sf/ Standard deviation about the mean, (\%): 41

\strong{Pine, central Sweden}
Number of observations: 1631

\tabular{lrrrr}{
Thinning \tab SI<180 \tab 180≤SI<220 \tab SI≥220  \cr
Unthinned \tab 309 \tab 296 \tab 391 \cr
Thinned \tab 150 \tab 186 \tab 299 \cr
}

Multiple correlation coefficient: 0.92


Standard deviation about the function, Sf : 0.376


Sf/ Standard deviation about the mean, (\%): 39

\strong{Pine, southern Sweden}
Number of observations: 1631

\tabular{lrrrr}{
Thinning \tab SI<160 \tab 160≤SI<200 \tab 200≤SI<240 \tab SI≥240  \cr
Unthinned \tab 223 \tab 461 \tab 835 \tab 1114 \cr
Thinned \tab 84 \tab 244 \tab 523 \tab 656  \cr
}

Multiple correlation coefficient: 0.92


Standard deviation about the function, Sf : 0.409


Sf/ Standard deviation about the mean, (\%): 40


\strong{Norway Spruce, northern Sweden}

Number of observations: 3959

\tabular{lrrrr}{
Thinning \tab SI<160 \tab 160≤SI<200 \tab SI<200   \cr
Unthinned \tab 1405 \tab 1012 \tab 431  \cr
Thinned \tab 476 \tab 455 \tab 180 \cr
}

Multiple correlation coefficient: 0.94


Standard deviation about the function, Sf : 0.391


Sf/ Standard deviation about the mean, (\%): 36

\strong{Norway Spruce, central Sweden}

Number of observations: 1773

\tabular{lrrrr}{
Thinning \tab SI<180 \tab 180≤SI<220 \tab 220≤SI<260 \tab SI≥260  \cr
Unthinned \tab 212 \tab 291 \tab 322 \tab 270 \cr
Thinned \tab 102 \tab 169 \tab 235 \tab 172  \cr
}

Multiple correlation coefficient: 0.93


Standard deviation about the function, Sf : 0.427


Sf/ Standard deviation about the mean, (\%): 39


\strong{Norway Spruce, southern Sweden}

Number of observations: 4864

\tabular{lrrrr}{
Thinning \tab SI<220 \tab 220≤SI<260 \tab 260≤SI<300 \tab SI≥300  \cr
Unthinned \tab 333 \tab 643 \tab 1121 \tab 1031 \cr
Thinned \tab 163 \tab 404 \tab 685 \tab 484  \cr
}

Multiple correlation coefficient: 0.93


Standard deviation about the function, Sf : 0.433


Sf/ Standard deviation about the mean, (\%): 48

\strong{Birch, northern and central Sweden}

Number of observations: 4429

\tabular{lrrrr}{
Thinning \tab SI<140 \tab 140≤SI<180 \tab 180≤SI<220 \tab SI>220   \cr
Unthinned \tab 859 \tab 1157 \tab 792 \tab 482  \cr
Thinned \tab 243 \tab 378 \tab 269 \tab 249 \cr
}

Multiple correlation coefficient: 0.87


Standard deviation about the function, Sf : 0.607


Sf/ Standard deviation about the mean, (\%): 50

\strong{Birch, southern Sweden}

Number of observations: 3072

\tabular{lrrrr}{
Thinning \tab SI<220 \tab 220≤SI<260 \tab 260≤SI<300 \tab SI>300   \cr
Unthinned \tab 249 \tab 406 \tab 772 \tab 756  \cr
Thinned \tab 89 \tab 181 \tab 334 \tab 285 \cr
}

Multiple correlation coefficient: 0.88


Standard deviation about the function, Sf : 0.599


Sf/ Standard deviation about the mean, (\%): 48

\strong{Broadleaves, northern and central Sweden}
Number of observations: 1367

\tabular{lrrrr}{
Thinning \tab SI<160 \tab 160≤SI<200 \tab 200≤SI<240 \tab SI>240   \cr
Unthinned \tab 319 \tab 335 \tab 225 \tab 160  \cr
Thinned \tab 85 \tab 100 \tab 75 \tab 68 \cr
}

Multiple correlation coefficient: 0.89


Standard deviation about the function, Sf : 0.608


Sf/ Standard deviation about the mean, (\%): 47

\strong{Broadleaves, southern Sweden}

Number of observations: 1404

\tabular{lrrrr}{
Thinning \tab SI<240 \tab 240≤SI<280 \tab 280≤SI<320 \tab SI>320   \cr
Unthinned \tab 131 \tab 209 \tab 424 \tab 266 \cr
Thinned \tab 60 \tab 75 \tab 151 \tab 88 \cr
}

Multiple correlation coefficient: 0.90


Standard deviation about the function, Sf : 0.622


Sf/ Standard deviation about the mean, (\%): 45

\strong{Beech}
Number of observations: 298

\tabular{lrr}{
Thinning \tab SI<310 \tab SI≥310 \cr
Unthinned \tab 67 \tab 103 \cr
Thinned \tab 61 \tab 67 \cr
}

Multiple correlation coefficient: 0.89


Standard deviation about the function, Sf : 0.603


Sf/ Standard deviation about the mean, (\%): 46

\strong{Oak}
Number of observations: 678

\tabular{lrrrr}{
Thinning \tab SI<280 \tab 280≤SI<320 \tab SI>320   \cr
Unthinned \tab 117 \tab 185 \tab 165  \cr
Thinned \tab 54 \tab 93 \tab 64 \cr
}

Multiple correlation coefficient: 0.89


Standard deviation about the function, Sf : 0.603


Sf/ Standard deviation about the mean, (\%): 46
}
