% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eko_PM_1985_volume.R
\name{PM_Vol}
\alias{PM_Vol}
\alias{Eko_PM_1985_volume_central_Sweden_Pine}
\alias{Eko_PM_1985_volume_northern_Sweden_Pine}
\alias{Eko_PM_1985_volume_southern_Sweden_Pine}
\alias{Eko_PM_1985_volume_northern_Sweden_Spruce}
\alias{Eko_PM_1985_volume_central_Sweden_Spruce}
\alias{Eko_PM_1985_volume_southern_Sweden_Spruce}
\alias{Eko_PM_1985_volume_northern_central_Sweden_Birch}
\alias{Eko_PM_1985_volume_southern_Sweden_Birch}
\alias{Eko_PM_1985_volume_northern_central_Sweden_Broadleaves}
\alias{Eko_PM_1985_volume_southern_Sweden_Broadleaves}
\alias{Eko_PM_1985_volume_Sweden_Beech}
\alias{Eko_PM_1985_volume_Sweden_Oak}
\title{Volume function for for even-aged monocultures or mixed stands in Sweden.}
\source{
Ekö, Per Magnus (1985) "En produktionsmodell för skog i Sverige, baserad på bestånd från
riksskogstaxeringens provytor: A growth simulator for Swedish forests, based on
data from the national forest survey. Rapporter nr. 16. Swedish University of
Agricultural Sciences, dept. of Silviculture. Umeå. ISBN 91-576-2386-4. ISSN
 0348-8969.  p. 105

 @description
 \strong{Pine, northern Sweden}
 Multiple correlation coefficient: 0.98

 Standard deviation about the function: 0.23

 Sf/Standard deviation about the mean: 0.17
 \strong{Pine, central Sweden}
 Multiple correlation coefficient: 0.98

 Standard deviation about the function: 0.23

 Sf/standard deviation about the mean: 0.17

 \strong{Pine, southern Sweden}
 Multiple correlation coefficient: 0.98

 Standard deviation about the function: 0.23

 Sf/standard deviation about the mean: 0.18

 \strong{Norway Spruce, northern Sweden}
 Multiple correlation coefficient: 0.98

 Standard deviation about the function: 0.29

 Sf/standard deviation about the mean: 0.20
 \strong{Norway Spruce, central Sweden}
 Multiple correlation coefficient: 0.98

 Standard deviation about the function: 0.29

 Sf/standard deviation about the mean: 0.20

 \strong{Norway Spruce, southern Sweden}
 Multiple correlation coefficient: 0.99

 Standard deviation about the function: 0.25

 Sf/standard deviation about the mean: 0.17

 \strong{Birch, northern and central Sweden}
 Multiple correlation coefficient: 0.95

 Standard deviation about the function: 0.39

 Sf/Standard deviation about the mean: 0.32

 \strong{Birch, southern Sweden}
 Multiple correlation coefficient: 0.96

 Standard deviation about the function: 0.35

 Sf/Standard deviation about the mean: 0.26

 \strong{Broadleaves, northern and central Sweden}
 Multiple correlation coefficient: 0.94

 Standard deviation about the function: 0.41

 Sf/Standard deviation about the mean: 0.33

 \strong{Broadleaves, southern Sweden}
 Multiple correlation coefficient: 0.97

 Standard deviation about the function: 0.37

 Sf/Standard deviation about the mean: 0.26

 \strong{Beech}
 Multiple correlation coefficient: 0.99

 Standard deviation about the function: 0.28

 Sf/Standard deviation about the mean: 0.17


 \strong{Oak}
 Multiple correlation coefficient: 0.97

 Standard deviation about the function: 0.39

 Sf/Standard deviation about the mean: 0.26
}
\usage{
Eko_PM_1985_volume_central_Sweden_Pine(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  thinned,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)

Eko_PM_1985_volume_northern_Sweden_Pine(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  thinned_previous_five_years,
  thinned_before_previous_five_years,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)

Eko_PM_1985_volume_southern_Sweden_Pine(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  thinned,
  vegetation,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)

Eko_PM_1985_volume_northern_Sweden_Spruce(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  thinned,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)

Eko_PM_1985_volume_central_Sweden_Spruce(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  vegetation,
  thinned,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)

Eko_PM_1985_volume_southern_Sweden_Spruce(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  thinned,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)

Eko_PM_1985_volume_northern_central_Sweden_Birch(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  latitude,
  altitude,
  fertilised,
  thinned,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)

Eko_PM_1985_volume_southern_Sweden_Birch(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  latitude,
  fertilised,
  thinned,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)

Eko_PM_1985_volume_northern_central_Sweden_Broadleaves(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  latitude,
  altitude,
  thinned,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)

Eko_PM_1985_volume_southern_Sweden_Broadleaves(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  latitude,
  thinned,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)

Eko_PM_1985_volume_Sweden_Beech(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  thinned,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)

Eko_PM_1985_volume_Sweden_Oak(
  basal_area.m2,
  basal_area_other_species,
  age_at_breast_height,
  stem_number_ha,
  SI,
  thinned,
  basal_area_weighted_mean_diameter,
  basal_area_weighted_mean_diameter_other_species
)
}
\arguments{
\item{basal_area.m2}{Basal area over bark of Pine (m2/ha)}

\item{basal_area_other_species}{Basal area of other species on the plot. (m2/ha)}

\item{age_at_breast_height}{Age at breast height, calculated as the mean age of the two thickest trees (years).}

\item{stem_number_ha}{Number of stems per hectare.}

\item{SI}{Site index H100 Pine , m.}

\item{thinned}{1 if the stand has been thinned before, otherwise 0.}

\item{basal_area_weighted_mean_diameter}{Diameter corresponding to mean basal area (m).}

\item{basal_area_weighted_mean_diameter_other_species}{Diameter corresponding to mean basal area (m) of other species on plot.}

\item{thinned_previous_five_years}{1 if the stand has been thinned during the last five years, otherwise 0.}

\item{thinned_before_previous_five_years}{1 if the stand has been thinned before the last five years, otherwise 0.}

\item{vegetation}{Variable indicating vegetation type scaled from -5 to +4 as follows:

\tabular{llr}{
Field Layer Code (NFI) \tab Description \tab Index \cr
1\tab Tall herbs w/o dwarf shrubs \tab 4 \cr
2\tab Tall herbs with bilberry \tab 2.5 \cr
3\tab Tall herbs with cowberry \tab 2 \cr
4\tab Low herbs w/o dwarf shrubs \tab 3 \cr
5\tab Low herbs with bilberry \tab 2.5 \cr
6\tab Low herbs with cowberry \tab 2 \cr
7\tab No field layer \tab 3 \cr
8\tab broad-leafed grasses \tab 2.5 \cr
9\tab narrow-leaved grasses \tab 1.5 \cr
10\tab Sedge, tall \tab -3 \cr
11\tab Sedge, low \tab -3 \cr
12\tab Horsetail \tab 1 \cr
13\tab Bilberry \tab 0 \cr
14\tab Cowberry \tab -0.5 \cr
15\tab Crowberry \tab -3 \cr
16\tab Poor shrub \tab -5 \cr
17\tab Lichen-rich \tab -0.5 \cr
18\tab Lichen-dominated \tab -1 \cr
}}

\item{latitude}{Degrees.}

\item{altitude}{Meters above sea level.}

\item{fertilised}{1 if the stand has been fertilised, otherwise 0.}
}
\value{
m3sk.
}
\description{
Volume function for for even-aged monocultures or mixed stands in Sweden.
}
