% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Automatic Function For Preparing Data}
\usage{
prepare_data(data_train, target, type, fill_na = TRUE, num_features = NULL)
}
\arguments{
\item{data_train}{dataset, used for training models. Class of data_train is one of those classes: data.frame, matrix, data.table or dgCMatrix. NOTE: data_train includes target column.}

\item{target}{character, indicating name of the target column in data_train.}

\item{type}{character, defining the task. Option is "regression" or "classification", namely binary classification.}

\item{fill_na}{logical, default is FALSE. If TRUE, missing values in target column are removed, missing values in categorical columns are replaced by mode and
missing values in numeric columns are substituted by median of corresponding columns.}

\item{num_features}{numeric, default is NULL. Parameter indicates number of most important features, which are chosen from the train dataset. Automatically, those important
features will be kept in the train and test datasets.}
}
\value{
\code{forester_prepared_data} object which stores data as a data frame and recipe with performed modifications
}
\description{
Function \code{prepare_data} covers fundamental techniques in ML Feature Engineering process.
}
\details{
With adjustable arguments, users can independently decide how function deals with NAs value
and important features selection. Furthermore, options for sampling methods will pop up in
case of classification problem.
}
\references{
forester library \url{https://modeloriented.github.io/forester/}
\donttest{
library(DALEX)
data(apartments, package="DALEX")

# Preparing data
prepared_data <- prepare_data(apartments,
                             target = "m2.price",
                             type = "regression",
                             fill_na = TRUE,
                              num_features = 3)
}
}
