% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Conducts preprocessing processes}
\usage{
preprocessing(data, y)
}
\arguments{
\item{data}{A data source, that is one of major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string which indicates a target column name.}
}
\value{
A preprocessed dataset.
}
\description{
Conducts preprocessing processes
}
\examples{
data(compas)
prep_data <- preprocessing(compas,'Two_yr_Recidivism')
}
