% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_feature_selection.R
\name{preprocessing_feature_selection}
\alias{preprocessing_feature_selection}
\title{Conducts a feature selection process with one out of four proposed methods}
\usage{
preprocessing_feature_selection(
  data,
  y,
  feature_selection_method = "VI",
  max_features = "default",
  nperm = 1,
  cutoffPermutations = 20,
  threadsNumber = NULL,
  method = "estevez",
  verbose = FALSE
)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string that indicates a target column name.}

\item{feature_selection_method}{A string value indication the feature selection method.
The imputation method must be one of 'VI', 'MCFS', 'MI', or 'BORUTA'.}

\item{max_features}{A positive integer value describing the desired number of
selected features. Initial value set as 'default' which is min(10, ncol(data) - 1)
for `VI` and `MI`, and NULL (number of relevant features chosen by the method)
for `MCFS`. Only `MCFS` can use the NULL value. `BORUTA` doesn't use this parameter.}

\item{nperm}{An integer describing the number of permutations performed, relevant
for the `VI` method. By default set to 1.}

\item{cutoffPermutations}{An non-negative integer value that determines the number of permutation
runs. It needs at least 20 permutations for a statistically significant result.
Minimum value of this parameter is 3, however if it is 0 then permutations
method is turned off. Relevant for the `MCFS` method.}

\item{threadsNumber}{A positive integer value describing the number of threads to use
in computation. More threads needs more CPU cores as well as memory usage is
a bit higher. It is recommended to set this value equal to or less than CPU
available cores. By default set to NULL, which will use maximal number of cores
minus 1. Relevant for the `MCFS` method.}

\item{method}{A string that indicates which algorithm will be used for `MI` method. Available
options are the default 'estevez' which works well for smaller datasets, but can
raise errors for bigger ones, and simpler 'peng'. More details present in the
documentation of ?varrank method.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
preprocessing process, if FALSE gives none.}
}
\value{
A list containing two objects:
\itemize{
\item \code{`data`} A dataset with selected columns,
\item \code{`idx`} The indexes of removed columns.
}
}
\description{
\itemize{
\item \code{`VI`} The variable importance method based on random forest - short time,
\item \code{`MCFS`} The Monte Carlo Feature Selection - long time,
\item \code{`MI`} The Varrank method based on mutual information scores - moderate time,
if we set too big `max_features` it can work really long,
\item \code{`BORUTA`} The BORUTA algorithm - short time.
}
}
