% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_imputation.R
\name{preprocessing_imputation}
\alias{preprocessing_imputation}
\title{Imputes missing values according to one of four prepared methods}
\usage{
preprocessing_imputation(
  data,
  na_indicators = c(""),
  imputation_method = "median-other",
  k = 10,
  m = 5
)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{na_indicators}{A list containing the values that will be treated as NA
indicators. By default the list is c(''). WARNING Do not include NA or NaN,
as these are already checked in other criterion.}

\item{imputation_method}{A string value indication the imputation method. The
imputation method must be one of 'median-other', 'median-frequency', 'knn', or 'mice'.}

\item{k}{An integer describing the number of nearest neighbours to use. By default
set to 10. The parameter applicable only if selection `imputation_method` is 'knn'.}

\item{m}{An integer describing the number of multiple imputations to use.
By default set to 5. The parameter applicable only if selection
`imputation_method` is 'mice'.}
}
\value{
Imputed dataset.
}
\description{
\itemize{
\item \code{`median-other`} The numeric features are imputed with median value,
whereas the categorical ones with the 'other' string. It is a fast method,
\item \code{`median-frequency`} The numeric features are imputed with median value,
whereas the categorical ones with the most frequent value. It is a fast method,
\item \code{`knn`} All features are imputed with KNN algorithm. It is a moderately fast method,
\item \code{`mice`} All features are imputed with MICE algorithm. It is a slow method.
}
}
