% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pukkala_2009.R
\name{Pukkala_2009_diameter_increment_Finland_Pine}
\alias{Pukkala_2009_diameter_increment_Finland_Pine}
\alias{Pukkala_2009_diameter_increment_Finland_Spruce}
\alias{Pukkala_2009_diameter_increment_Finland_Betula}
\title{Diameter Increment for Uneven-sized Spruce forests in Finland from Pukkala 2009}
\source{
Pukkala, T., Lähde, E., Laiho, O. (2009). Growth and Yield Models of uneven-sized forest stands in Finland. For. Ecol. Manage. 258: 207-216. Available (25/12/2021) \url{https://doi.org/10.1016/j.foreco.2009.03.052}
}
\usage{
Pukkala_2009_diameter_increment_Finland_Pine(
  BAL_Spruce,
  BAL_other,
  BA,
  diameter_cm,
  vegetation,
  temperature_sum
)

Pukkala_2009_diameter_increment_Finland_Spruce(
  BAL_Spruce,
  BAL_other,
  BA,
  diameter_cm,
  vegetation,
  temperature_sum
)

Pukkala_2009_diameter_increment_Finland_Betula(
  BAL_Spruce,
  BAL_other,
  BA,
  diameter_cm,
  vegetation,
  temperature_sum
)
}
\arguments{
\item{BAL_Spruce}{Basal area of Spruce trees larger than the subject tree (m2 / ha)}

\item{BAL_other}{Basal area of trees other than Spruce larger than the subject tree (m2/ha)}

\item{BA}{Basal Area of all trees larger than 5 cm dbh. m2/ha.}

\item{diameter_cm}{Diameter of subject tree at breast height 1.3m, cm.}

\item{vegetation}{One of \code{\link[=Finland_vegetation_types]{Finland_vegetation_types()}}}

\item{temperature_sum}{Degree days}
}
\value{
5 year over-bark diameter increment.
}
\description{
Diameter Increment for Uneven-sized Spruce forests in Finland from Pukkala 2009
}
\details{
Pine model is based on 15282 trees. R^2 = 0.4, SD of resid.= 0.527. Snowdon correction = 1.110

Spruce model is based on 24014 trees. R^2 = 0.568, SD of resid.= 0.604. Snowdon correction = 1.124

@details Betula model is based on 8219 trees. R^2 = 0.541, SD of resid.= 0.599. Snowdon correction = 1.127
}
