% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pukkala_2009_diameter_ingrowth_Finland.R
\name{Pukkala_2009_diameter_ingrowth_Finland_Spruce}
\alias{Pukkala_2009_diameter_ingrowth_Finland_Spruce}
\alias{Pukkala_2009_diameter_ingrowth_Finland_Betula_Pine}
\title{Diameter of Ingrowth in Uneven-sized stands in Finland}
\source{
Pukkala, T., Lähde, E., Laiho, O. (2009). Growth and Yield Models of uneven-sized forest stands in Finland. For. Ecol. Manage. 258: 207-216. Available (25/12/2021) \url{https://doi.org/10.1016/j.foreco.2009.03.052}
}
\usage{
Pukkala_2009_diameter_ingrowth_Finland_Spruce(
  BA_m2_ha,
  stems_ha_spruce,
  stems_ha_other,
  vegetation
)

Pukkala_2009_diameter_ingrowth_Finland_Betula_Pine(
  BA_m2_ha,
  stems_ha_spruce,
  stems_ha_other,
  vegetation
)
}
\arguments{
\item{BA_m2_ha}{Total Basal Area (m2/ha) of trees >5 cm dbh.}

\item{stems_ha_spruce}{Spruce stems per hectare >5 cm dbh.}

\item{stems_ha_other}{Stems other than Spruce per hectare > 5 cm dbh.}

\item{vegetation}{One of [forester::Finland_vegetation_types()]}
}
\value{
Diameter cm dbh.
}
\description{
Diameter of Ingrowth in Uneven-sized stands in Finland
}
\details{
Spruce model is based on 170 plot observations, R^2= 0.586, SD of resid. = 0.0473.

@details Pine & Betula model is based on 141 plot observations, R^2= 0.469, SD of resid. = 0.0730.
}
