% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pukkala_2009_height_model_Finland.R
\name{Pukkala_2009_height_model_Finland_Pine}
\alias{Pukkala_2009_height_model_Finland_Pine}
\alias{Pukkala_2009_height_model_Finland_Betula}
\alias{Pukkala_2009_height_model_Finland_Norway_Spruce}
\title{Height model for Individual Trees from Pukkala 2009.}
\source{
Pukkala, T., Lähde, E., Laiho, O. (2009). Growth and Yield Models of uneven-sized forest stands in Finland. For. Ecol. Manage. 258: 207-216. Available (25/12/2021) \url{https://doi.org/10.1016/j.foreco.2009.03.052}
}
\usage{
Pukkala_2009_height_model_Finland_Pine(vegetation, diameter_cm)

Pukkala_2009_height_model_Finland_Betula(vegetation, diameter_cm)

Pukkala_2009_height_model_Finland_Norway_Spruce(vegetation, diameter_cm)
}
\arguments{
\item{vegetation}{Code from \code{\link[=Finland_vegetation_types]{Finland_vegetation_types()}}.}

\item{diameter_cm}{Diameter of the tree at breast height (1.3 m)}
}
\value{
Tree height.
}
\description{
Height model for Individual Trees from Pukkala 2009.
}
\details{
Pine model based on 8622 observations. R^2=0.779. SD resid. = 2.360

Betula model based on 1200 observations. R^2=0.802, SD resid.=2.001

Norway Spruce model based on 12144 observations. R^2 = 0.856, SD resid. = 2.104.
}
