% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pukkala_2009_ingrowth_Finland.R
\name{Pukkala_2009_ingrowth_Finland_Spruce}
\alias{Pukkala_2009_ingrowth_Finland_Spruce}
\alias{Pukkala_2009_ingrowth_Finland_Betula_Pine}
\title{Ingrowth model for Uneven-sized stands in Finland}
\source{
Pukkala, T., Lähde, E., Laiho, O. (2009). Growth and Yield Models of uneven-sized forest stands in Finland. For. Ecol. Manage. 258: 207-216. Available (25/12/2021) \url{https://doi.org/10.1016/j.foreco.2009.03.052}
}
\usage{
Pukkala_2009_ingrowth_Finland_Spruce(
  stems_ha_spruce,
  stems_ha_other,
  vegetation,
  BA_m2_ha
)

Pukkala_2009_ingrowth_Finland_Betula_Pine(
  stems_ha_spruce,
  stems_ha_other,
  vegetation,
  BA_m2_ha
)
}
\arguments{
\item{stems_ha_spruce}{Spruce stems per hectare >5 cm dbh.}

\item{stems_ha_other}{Stems other than Spruce per hectare > 5 cm dbh.}

\item{vegetation}{One of [forester::Finland_vegetation_types()]}

\item{BA_m2_ha}{Total Basal Area (m2/ha) of trees >5 cm dbh.}
}
\value{
Number of ingrowth per hectare.
}
\description{
Baskerville Correction has been added to the constant.
}
\details{
The Spruce model is based on 10545 living trees, 1127 dead trees. Best threshold probability is c. 0.5-0.6. Percentage of correct predictions 90.7.

Betula and Pine model is based on 3230 living trees, 289 dead trees. Best threshold probability = 0.6. Percentage of correct predictions = 92.6
}
