% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Pukkala_2019_form_quotient_Norway_Norway_Spruce.R
\name{Pukkala_2019_form_quotient_Norway_Norway_Spruce}
\alias{Pukkala_2019_form_quotient_Norway_Norway_Spruce}
\title{Diameter at a height or height from diameter for Norway Spruce in Norway.}
\source{
Pukkala, T., Holt Hanssen, K., Andreassen, K. (2019) Stem taper and bark functions for Norway spruce in Norway. Silva Fennica vol. 53 no. 3 article id 10187. \url{https://doi.org/10.14214/sf.10187}
}
\usage{
Pukkala_2019_form_quotient_Norway_Norway_Spruce(
  height_m,
  dbh_cm,
  get_d_at_height_m = NA,
  get_h_at_diameter_cm = NA,
  age = NA,
  SIH40 = NA,
  output = "diameter"
)
}
\arguments{
\item{height_m}{Height of tree in metres.}

\item{dbh_cm}{Diameter of tree at breast height.}

\item{get_d_at_height_m}{If output is 'diameter', the height at which diameter is desired. Otherwise NA.}

\item{get_h_at_diameter_cm}{If output is 'height', the diameter at which height is desired. Otherwise NA.}

\item{age}{Default (NA). Otherwise (total?) age of tree.}

\item{SIH40}{Default (NA). Otherwise Site Index H40 for Norway Spruce, e.g. [forester::Tveite_1977_height_trajectory_Norway_Norway_Spruce()]}

\item{output}{One of: "diameter" (default) or "height".}
}
\value{
If output is 'diameter', the diameter at height get_d_at_height_m; if output is 'height', the height when the tree will have the diameter 'get_h_at_diameter_cm'.
}
\description{
Uses Kozak (1997) stem taper function (function 1995 in original publication) to describe the diameter of Norway Spruce in Norway.
}
\details{
This function has two functions. i) If output is 'diameter', to find the diameter of a tree at the height 'get_d_at_height_m'. ii) If output is 'height', to find the height at which the tree is 'get_h_at_diameter_cm'.
Will return NA if the desired height is below 0 m or above the height of the tree.
Will return NA if the desired diameter is larger than the expected diameter at ground level (height==0) or the desired diameter is below 0.

The function will, if no age or SIH40 is supplied, use the 'baseline' parameters from the source.
If age or SIH40 is supplied, it will prefer the parameters from each of these (table 3, 4 in publication).
If both age and SIH40 are supplied, function will prefer the site-index dependent fixed effects (table 3).
}
\examples{
#Get the diameter at ground level for a Spruce with dbh==18 cm and height==14 m.
Pukkala_2019_form_quotient_Norway_Norway_Spruce(height_m = 14,dbh_cm = 18,output = "diameter",get_d_at_height_m = 0)
}
