% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralHelpers.R
\name{quadratic_mean_diameter}
\alias{quadratic_mean_diameter}
\title{Calculate the Quadratic Mean Diameter of a stand, cm.}
\source{
Curtis, R.O., Marshall, D.D. 2000. Technical Note: Why Quadratic Mean Diameter? West. J. Appl. For. 15(3):137-139. Available: \url{https://www.fs.fed.us/pnw/olympia/silv/publications/opt/436_CurtisMarshall2000.pdf}
}
\usage{
quadratic_mean_diameter(Basal_area_m2_ha, stems_per_ha)
}
\arguments{
\item{Basal_area_m2_ha}{Basal area m2 / ha.}

\item{stems_per_ha}{Stems per ha.}
}
\value{
QMD, cm.
}
\description{
Calculate the Quadratic Mean Diameter of a stand, cm.
}
