% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_search.R
\name{random_search}
\alias{random_search}
\title{Random optimization of hyperparameters}
\usage{
random_search(train_data, y, models = NULL, engine, type, max_evals = 10)
}
\arguments{
\item{train_data}{A training data for models created by `prepare_data()` function.}

\item{y}{A string that indicates a target column name.}

\item{models}{A list of models trained by `train_models()` function.
They will be compered with models trained with different hyperparameters}

\item{engine}{A vector of tree-based models that shall be created. Possible
values are: `ranger`, `xgboost`, `desicion tree`, `lightgbm`, `catboost`.}

\item{type}{A string that determines if Machine Learning task is the
`binary_clf` or `regression` task.}

\item{max_evals}{The number of trained models for each model type in `engine`.}
}
\value{
A list consisting of models created via random search and ranked list
of models scores.
}
\description{
Random optimization of hyperparameters
}
