% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_regeneration_stocking_level.R
\name{regeneration_stocking}
\alias{regeneration_stocking}
\title{Stocking in Natural regenerations}
\usage{
regeneration_stocking(
  altitude,
  latitude,
  age,
  proportion_cultivated,
  N_full,
  ground_water,
  county,
  scarification,
  burnt,
  no_treatment,
  uncleaned,
  number_seed_trees,
  SIH,
  SIS,
  area
)
}
\arguments{
\item{altitude}{altitude masl}

\item{latitude}{latitude decimal}

\item{age}{age, years}

\item{proportion_cultivated}{percent}

\item{area}{}
}
\value{

}
\description{
Stocking in Natural regenerations
}
