% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reineke_1933.R
\docType{data}
\name{Reineke_1933}
\alias{Reineke_1933}
\title{Data from Reineke article on SDI, 1933.}
\format{
A dataframe with the following columns:
\describe{
\item{AverageDiameterInches}{QMD}
\item{NperAcre}{Stems per acre}
\item{SPP}{Species}
}
}
\source{
Reineke, L. H. 1933. Perfecting a stand-density index for even-aged forests. Journal of Agricultural Research. Vol. 46. Number 7. Washington, D.C.
Available Online: \url{https://www.fs.usda.gov/psw/publications/cfres/cfres_1933_reineke001.pdf}
}
\usage{
Reineke_1933
}
\description{
Data digitised from graphs from original source.
}
\keyword{datasets}
