% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_removal.R
\name{rm_correlated_columns}
\alias{rm_correlated_columns}
\title{Remove strongly correlated columns (Spearman's rank for numerical, Crammer's V for
categorical)}
\usage{
rm_correlated_columns(data, y, threshold = 0.7)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{y}{A string that indicates a target column name.}

\item{threshold}{A numeric value from [0,1] range, which indicates when we consider
the correlation to be high. If feature surpasses this threshold it is going to
be removed. By default set to 0.7.}
}
\value{
A list containing four objects
\itemize{
\item \code{`data`} A dataset with deleted columns.
\item \code{`idx`} The indexes of removed columns.
\item \code{`rm_num`} The indexes of removed numerical columns.
\item \code{`rm_cat`} The indexes of removed categorical columns.
}
}
\description{
Remove strongly correlated columns (Spearman's rank for numerical, Crammer's V for
categorical)
}
