% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_removal.R
\name{rm_corrupted_observations}
\alias{rm_corrupted_observations}
\title{Remove observations with too much empty values or empty target}
\usage{
rm_corrupted_observations(data, y, threshold = 0.7, na_indicators = c(""))
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{y}{A string that indicates a target column name.}

\item{threshold}{A numeric value from [0,1] range, which indicates the maximum
threshold of missing values for observation. If observation has more missing fields
it is going to be removed. By default set to 0.7.}

\item{na_indicators}{A list containing the values that will be treated as NA
indicators. By default the list is c(''). WARNING Do not include NA or NaN,
as these are already checked in other criterion.}
}
\value{
A list containing two objects
\itemize{
\item \code{`data`} A dataset with deleted observations.
\item \code{`idx`} The indexes of removed observations.
}
}
\description{
Remove observations with too much empty values or empty target
}
