% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_removal.R
\name{rm_duplicate_columns}
\alias{rm_duplicate_columns}
\title{Remove duplicate columns iteratively from data set.}
\usage{
rm_duplicate_columns(data, y)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{y}{A string that indicates a target column name.}
}
\value{
A list containing two objects
\itemize{
\item \code{`data`} A dataset with deleted columns.
\item \code{`idx`} The indexes of removed columns.
}
}
\description{
Remove duplicate columns iteratively from data set.
}
