% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_removal.R
\name{rm_id_like_columns}
\alias{rm_id_like_columns}
\title{Remove ID-like columns.}
\usage{
rm_id_like_columns(
  data,
  id_names = c("id", "nr", "number", "idx", "identification", "index")
)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{id_names}{A vector of strings indicating which column names are perceived
as ID-like. By default the list is: ['id', 'nr', 'number', 'idx', 'identification', 'index'].}
}
\value{
A list containing two objects
\itemize{
\item \code{`data`} A dataset with deleted columns.
\item \code{`idx`} The indexes of removed columns.
}
}
\description{
The columns are perceived as ID-like, when they are present in the id_names
vector fo strings or the column values are unique integers.
}
