% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_removal.R
\name{rm_sparse_columns}
\alias{rm_sparse_columns}
\title{Remove columns with too much empty values}
\usage{
rm_sparse_columns(
  data,
  y,
  threshold = 0.3,
  na_indicators = c(""),
  verbose = FALSE
)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{y}{A string that indicates a target column name.}

\item{threshold}{A numeric value from [0,1] range, which indicates the maximum
threshold of missing values for columns If column has more missing fields
it is going to be removed. By default set to 0.3.}

\item{na_indicators}{A list containing the values that will be treated as NA
indicators. By default the list is c(''). WARNING Do not include NA or NaN,
as these are already checked in other criterion.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
preprocessing process, if FALSE gives none.}
}
\value{
A list containing two objects
\itemize{
\item \code{`data`} A dataset with deleted columns.
\item \code{`idx`} The indexes of removed columns.
}
}
\description{
Remove columns with too much empty values
}
