% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_removal.R
\name{rm_static_cols}
\alias{rm_static_cols}
\title{Remove columns which are dominated by one value}
\usage{
rm_static_cols(data, y, threshold = 1, verbose = FALSE)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string that indicates a target column name.}

\item{threshold}{A numeric value from [0,1] range, which indicates the maximum
threshold of dominating values for column If feature has more dominating
values it is going to be removed. By default set to 1, which indicates that
all values are equal.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
preprocessing process, if FALSE gives none.}
}
\value{
A list containing two objects:
\itemize{
\item \code{`data`} A dataset with deleted columns
\item \code{`idx`} The indexes of removed columns
}
}
\description{
Remove columns which are dominated by one value
}
