% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_forest}
\alias{save_forest}
\title{Save elements from forester}
\usage{
save_forest(train, file = NULL, verbose = TRUE, return_name = FALSE)
}
\arguments{
\item{train}{The return from `train` function.}

\item{file}{A string describing the name and path of the file. By default set to
NULL which creates a `forester_{timestamp}` in current working directory.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
training process, if FALSE gives none.}

\item{return_name}{A logical value, if set to TRUE, function returns \
full path and name of the saved file.}
}
\description{
Save elements from forester
}
\examples{
\dontrun{
# Training the model.
train <- train(data        = iris,
               y           = 'Species',
               bayes_iter  = 0,
               random_eval = 3)

# Saving the outcomes.
save_forest(train, 'saved_train')

# Reading saved file.
train2 <- readRDS('saved_train.RData')
}
}
