% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDI_plot.R
\name{SDIplot}
\alias{SDIplot}
\title{SDIplot}
\usage{
SDIplot(
  slope,
  start = 100,
  stop = 1000,
  interval = 100,
  maxD = 40,
  xlabel = "Average Diameter in Inches",
  ylabel = "Trees per Acre"
)
}
\arguments{
\item{slope}{Slope of reference curve for the species when on a log-log scale.}

\item{start}{Lowest SDI to plot}

\item{stop}{Highest SDI to plot}

\item{interval}{Curve interval}

\item{maxD}{maximum diameter to plot to}

\item{xlabel}{Change the label of the x-axis}

\item{ylabel}{Change the label of the y-axis}
}
\value{
A plot of the functions SDI=start to SDI=stop on a log log scale after Reineke (1933)
}
\description{
SDIplot
}
\examples{

SDI_plot(slope=-1.605)

}
