% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_feature_selection.R
\name{select_boruta}
\alias{select_boruta}
\title{Perform Boruta algorithm for selecting most important features}
\usage{
select_boruta(data, y)
}
\arguments{
\item{data}{A data source before preprocessing, that is one of the major R formats:
data.table, data.frame, matrix, and so on.}

\item{y}{A string that indicates a target column name.}
}
\value{
A list containing two objects:
\itemize{
\item \code{`data`} A dataset with selected columns,
\item \code{`idx`} The indexes of removed columns.
}
}
\description{
Perform Boruta algorithm for selecting most important features
}
