% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_feature_selection.R
\name{select_mcfs}
\alias{select_mcfs}
\title{Feature Selection performed by MCFS (Monte Carlo Feature Selection) algorithm}
\usage{
select_mcfs(
  data,
  y,
  cutoffPermutations = 20,
  threadsNumber = NULL,
  max_features = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string that indicates a target column name.}

\item{cutoffPermutations}{An non-negative integer value that determines the number of permutation
runs. It needs at least 20 permutations for a statistically significant result.
Minimum value of this parameter is 3, however if it is 0 then permutations
method is turned off.}

\item{threadsNumber}{A positive integer value describing the number of threads to use
in computation. More threads needs more CPU cores as well as memory usage is
a bit higher. It is recommended to set this value equal to or less than CPU
available cores. By default set to NULL, which will use maximal number of cores
minus 1.}

\item{max_features}{A positive integer value describing the desired number of
selected features. By default set to NULL, then the MCFS algorithm uses all
features that seems important to it.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
preprocessing process, if FALSE gives none.}
}
\value{
A list containing two objects:
\itemize{
\item \code{`data`} A dataset with selected columns,
\item \code{`idx`} The indexes of removed columns.
}
}
\description{
Unfortunately it is impossible to turn off the prints, as they come from used
rmcfs package's back-end written in Java.
}
