% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_feature_selection.R
\name{select_mi_varrank}
\alias{select_mi_varrank}
\title{Feature selection performed by varrank package}
\usage{
select_mi_varrank(data, y, max_features = 10)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string that indicates a target column name.}

\item{max_features}{A positive integer value describing the desired number of
selected features. By default set to 10.}
}
\value{
A list containing two objects:
\itemize{
\item \code{`data`} A dataset with selected columns,
\item \code{`idx`} The indexes of removed columns.
}
}
\description{
It heuristically estimates the variables ranks based on mutual information
with multiple model and search schemes.
}
