% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_models.R
\name{select_models}
\alias{select_models}
\title{Select models from train() output.}
\usage{
select_models(train_output, models)
}
\arguments{
\item{train_output}{The output created with train() function.}

\item{models}{The vector including either the names of models or their indexes,
that we want to select.}
}
\value{
The train_output with selected models.
}
\description{
The function provides a necessary interface for minimizing the size of output
created with train(). It not only selects the models, but also the data,
predictions, and all other necessary information.
}
\examples{
\dontrun{
out      <- train(iris, 'Species', random_evals = 30, bayes_iters = 0)
new_out  <- select_models(out, c('random_forest_RS_1', 'xgboost_RS_14'))
new_out2 <- select_models(out, c(1, 15, 30, 45))
}
}
