% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_feature_selection.R
\name{select_vi}
\alias{select_vi}
\title{Feature Selection performed by Variable Importance algorithm based on random forest}
\usage{
select_vi(data, y, nperm = 1, max_features = 10)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string that indicates a target column name.}

\item{nperm}{An integer describing the number of permutations performed.
By default set to 1.}

\item{max_features}{A positive integer value describing the desired number of
selected features. By default set to 10.}
}
\value{
A list containing two objects:
\itemize{
\item \code{`data`} A dataset with selected columns,
\item \code{`idx`} The indexes of removed columns.
}
}
\description{
Feature Selection performed by Variable Importance algorithm based on random forest
}
