% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sharma_2011_height_trajectories_Norway.R
\name{Sharma2011}
\alias{Sharma2011}
\alias{Sharma_2011_height_trajectory_Norway_Norway_Spruce}
\alias{Sharma_2011_height_trajectory_Norway_Scots_Pine}
\title{Height trajectories for Norway Spruce and Scots Pine in Norway}
\source{
Sharma, R.P., Brunner, A., Eid, T., Øyen B-H. (2011) Modelling dominant
height growth from national forest inventory individual tree data with short
time series and large age errors. For. Ecol. Manage. 262:12. 2162-2175. DOI:
\url{https://doi.org/10.1016/j.foreco.2011.07.037}
}
\usage{
Sharma_2011_height_trajectory_Norway_Norway_Spruce(
  age,
  age2,
  dominant_height,
  output = "Height"
)

Sharma_2011_height_trajectory_Norway_Scots_Pine(
  age,
  age2,
  dominant_height,
  output = "Height"
)
}
\arguments{
\item{age}{Age at breast height (1.3m)}

\item{age2}{Age at breast height (1.3m) at output.}

\item{dominant_height}{Dominant height, metres.}

\item{output}{One of "Equation" or "Height" (default).}
}
\value{
One of "Equation" or "Height".
}
\description{
Height trajectories for Norway Spruce and Scots Pine in Norway
}
\details{
Prediction statistics for all trees (MPE=-0.0269 m, R^2adj. = 0.9931,
N=36 for Spruce; MPE = 0.0007m, R^2adj.=0.9909, N = 37 for Pine)
}
