% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_to_breast_height.R
\name{shift_to_breast_height}
\alias{shift_to_breast_height}
\title{Shift a height trajectory function to age at breast height 1.3m}
\usage{
shift_to_breast_height(
  height_trajectory,
  dominant_height,
  age,
  age2,
  interval_low = 0,
  interval_high = 10,
  ...
)
}
\arguments{
\item{height_trajectory}{A height-trajectory function containing the arguments 'dominant_height', 'age','age2','output'.}

\item{dominant_height}{Dominant height}

\item{age2}{Age at which to return height.}

\item{interval_low}{Default=0,Lowest possible time to breast height. (If a function already has an intercept at breast height)}

\item{interval_high}{Default=10, highest possible time to breast height.}

\item{...}{A list of additional arguments passed to height_trajectory.}

\item{reference_age}{Age}
}
\value{
Height at breast height age2.
}
\description{
Shift a height trajectory function to age at breast height 1.3m
}
\examples{
ggplot()+
xlim(c(0,100))+
 geom_function(
   aes(linetype="Shift"),
   fun=function(x) shift_to_breast_height(height_trajectory = Eriksson_1997_height_trajectory_Sweden_Birch,dominant_height = 17,reference_age = 40,age2 = x)
 )+
 geom_function(
   aes(linetype="Normal"),
   fun=function(x) Eriksson_1997_height_trajectory_Sweden_Birch(dominant_height = 17,age = 40,age2 = x,output = "Height")
 )+
 geom_point(
  aes(x=40,y=17)
)

}
