% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_100_estimate.R
\name{si_100_estimate}
\alias{si_100_estimate}
\title{Site Index Estimations for Sweden}
\usage{
si_100_estimate(
  species = c("NorwaySpruce", "ScotsPine", "Larch", "Beech", "Oak", "Birch",
    "AlnusIncana", "AlnusGlutinosa", "Aspen"),
  age,
  height
)
}
\arguments{
\item{species}{One of "NorwaySpruce", "ScotsPine", "Larch", "Beech","Oak","Birch","AlnusIncana", "AlnusGlutinosa", "Aspen".}

\item{age}{Average total age in years. For birch age at breast height.}

\item{height}{Average Height in metres of dominant trees.}

\item{data}{Tidy data frame.}

\item{...}{}
}
\value{
data frame with an additional column with the estimate of the height of the tree at age 100.
}
\description{
Provides interface for the site index functions from "\emph{Nya höjdutvecklingskurvor för bonitering}" (2013),
by Johansson, U., Ekö, P-M, Elfving, B., Johansson, T. and Nilsson, U.

    \strong{Reference ages by species}: Picea abies (100);Pinus sylvestris (100); Larix (100); Fagus sylvatica (100); Quercus (100); Betula (50); Alnus (50); Populus (50).

    \strong{Suitable ages for use by species}: Picea abies (10-80); Pinus sylvestris (10-80); Larix (10-100); Fagus sylvatica (20-150); Quercus (20-150);  Betula (10-90); Alnus glutinosa (<100); Alnus incana (<70); Populus tremula x tremuloides (<50); Populus tremula (<50-60)
}
\examples{
si_100_estimate(species="NorwaySpruce",age=30, height=6.2)
}
