% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silver_fir_1921.R
\docType{data}
\name{Silver_fir_1921}
\alias{Silver_fir_1921}
\title{Silver fir growth}
\format{
A data frame with columns:
\describe{
\item{Site_class}{A value between I-V}
\item{Height}{Mean height of the main stand, metres}
\item{Stem number}{Stem number}
\item{Total stem wood m3}{Total stem wood m3}
}
}
\source{
\strong{Gehrhardt, E. (1923) Ertragstafeln für Eiche, Buche, Tanne, Fichte und Riefer. Berlag von Julius Springer. Berlin. 46 pages. pp. 35-40.}
}
\usage{
Silver_fir_1921
}
\description{
Gehrhardt extended Eichhorn's rule so that not only was the mean height directly related to the standing volume regardless of site class, but the mean height was directly related to the total volume production.

To work this out for yourself, look at the forester vignette: vignette("Eichhorn")
}
\keyword{datasets}
