% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_tree_age_estimation.R
\name{singleTreeAgeEstimation}
\alias{singleTreeAgeEstimation}
\title{Functions for estimation of single tree ages at breast height}
\usage{
singleTreeAgeEstimation(
  diameter,
  diameter_ba,
  latitude,
  altitude,
  lnd,
  total_stand_age,
  SIS,
  total_ba,
  species,
  field_vegetation,
  even_aged,
  county,
  peat,
  ditch
)
}
\arguments{
\item{diameter}{diameter at breast height, cm}

\item{diameter_ba}{diameter of basal area weighted mean tree, cm.}

\item{latitude}{latitude}

\item{altitude}{altitude, metres above sea level.}

\item{lnd}{log mean diameter of co-dominant trees. Output of mean_diam_codominant.}

\item{total_stand_age}{total stand age}

\item{SIS}{site index according site factors, given as top height at age 100, m.}

\item{total_ba}{As estimated with relascope in field. m^2 / ha}

\item{species}{Tree species}

\item{field_vegetation}{type, FsKod.}

\item{even_aged}{TRUE / FALSE,  above 80 percent of volume within 20 yr span.}

\item{county}{Swedish county}

\item{ditch}{TRUE / FALSE}
}
\value{
age at breast height
}
\description{
Presented in Elfving, B. (2003) Ålderstilldelning till enskilda träd i skogliga tillväxtprognoser. SLU, inst. f. skogsskötsel. Arbetsrapport 182.
}
