% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS_estimation.R
\name{SIS_estimate}
\alias{SIS_estimate}
\title{Site index estimation by site factors}
\usage{
SIS_estimate(
  species,
  vegetation,
  latitude,
  longitude,
  altitude,
  aspect.main,
  incline.percent,
  soil_moisture,
  soil_texture,
  climate_code,
  county
)
}
\arguments{
\item{species}{Spruce or Pine}

\item{vegetation}{Type}

\item{latitude}{Decimal degrees.}

\item{longitude}{Decimal degrees.}

\item{altitude}{Metres above sea level}

\item{soil_moisture}{Type}

\item{soil_texture}{Type}

\item{climate_code}{Optional, else from lat, lon.}

\item{county}{Optional, else from lat/lon}
}
\value{
Site index, in metres.
}
\description{
Corrected functions from Appendix II for estimation of site index by site factors according to Hägglund and Lundmark (1977), in
Hägglund, B. (1979) Ett system för bonitering av skogsmark - analys, kontroll och diskussion inför praktisk tillämpning. Rapport 14, Projekt HUGIN. Skogsvetenskapliga fakulteten, Sveriges Lantbruksuniversitet. Umeå. Sweden.
}
