% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKS_1984_Thinning_Guide_Sweden.R
\name{SKS_1984_Thinning_Guide_Sweden}
\alias{SKS_1984_Thinning_Guide_Sweden}
\title{Thinning Recommendations according to the Swedish Forest Agency 1984.}
\usage{
SKS_1984_Thinning_Guide_Sweden(
  H100,
  dominantHeight,
  northernSweden = TRUE,
  species
)
}
\arguments{
\item{H100}{Site Index H100 according to Hägglund 1972, 1973. e.g.
[forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce],
[forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce],
[forester::Hagglund_1974_Sweden_height_trajectories_Pine]}

\item{dominantHeight}{dominant height of the stand.}

\item{northernSweden}{boolean. TRUE if stand is located in northern Sweden
(default).}

\item{species}{One of "Pinus sylvestris","Picea abies","Pinus contorta",
"Quercus robur,"Fagus sylvatica" or "Deciduous".}
}
\value{
A list of the maximum reccomended basal area per hectare before and
the lowest reccomended basal area after thinning.
}
\description{
Thinning Recommendations according to the Swedish Forest Agency 1984.
}
\details{
Based on a regression of the original diagrams used in Heureka
 StandWise v. 2.19. e.g. 2022-09-17.
}
\examples{


data.frame(
  "dominantHeight"=seq(10,32,0.1)
) \%>\% group_by_all() \%>\%
  mutate(data=purrr::map(.x=dominantHeight,
                        .f = SKS_1984_Thinning_Guide_Sweden,H100=36,
                        northernSweden = FALSE,
                        species="Picea abies")) \%>\%
  unnest_wider(data) \%>\%
  ggplot() +
  xlim(c(10, 32)) +
  ylim(c(18, 44)) +
  geom_line(aes(x=dominantHeight,y=Before))+
  geom_line(aes(x=dominantHeight,y=After))+
  ylab("Basal Area m^2/ha")
}
