% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soderberg_1986_Sweden_BAI5.R
\name{Soderberg1986BAI5}
\alias{Soderberg1986BAI5}
\alias{Soderberg_1986_BAI5_Sweden_Oak}
\alias{Soderberg_1986_BAI5_Sweden_Beech}
\alias{Soderberg_1986_BAI5_southern_Sweden_under_65_Spruce}
\alias{Soderberg_1986_BAI5_southern_Sweden_under_65_Pine}
\alias{Soderberg_1986_BAI5_southern_Sweden_under_65_Birch}
\alias{Soderberg_1986_BAI5_southern_Sweden_over_45_Spruce}
\alias{Soderberg_1986_BAI5_southern_Sweden_over_45_Pine}
\alias{Soderberg_1986_BAI5_southern_Sweden_over_45_Birch}
\alias{Soderberg_1986_BAI5_southern_Sweden_Broadleaves}
\alias{Soderberg_1986_BAI5_northern_Sweden_under_65_Spruce}
\alias{Soderberg_1986_BAI5_northern_Sweden_under_65_Pine}
\alias{Soderberg_1986_BAI5_northern_Sweden_over_45_Spruce}
\alias{Soderberg_1986_BAI5_northern_Sweden_over_45_Pine}
\alias{Soderberg_1986_BAI5_northern_central_Sweden_under_65_Birch}
\alias{Soderberg_1986_BAI5_northern_central_Sweden_over_45_Birch}
\alias{Soderberg_1986_BAI5_northern_central_Sweden_Broadleaves}
\alias{Soderberg_1986_BAI5_central_Sweden_under_65_Spruce}
\alias{Soderberg_1986_BAI5_central_Sweden_under_65_Pine}
\alias{Soderberg_1986_BAI5_central_Sweden_over_45_Spruce}
\alias{Soderberg_1986_BAI5_central_Sweden_over_45_Pine}
\title{Basal Area Increment of 5 years for Individual trees in Sweden from Söderberg 1986.}
\source{
Söderberg, U. (1986) Funktioner för skogliga produktionsprognoser -
Tillväxt och formhöjd för enskilda träd av inhemska trädslag i Sverige. /
Functions for forecasting of timber yields - Increment and form height for
individual trees of native species in Sweden. Report 14. Section of Forest
Mensuration and Management. Swedish University of Agricultural Sciences.
Umeå. ISBN 91-576-2634-0. ISSN 0349-2133. pp.251.
}
\usage{
Soderberg_1986_BAI5_Sweden_Oak(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  aspect,
  maritime
)

Soderberg_1986_BAI5_Sweden_Beech(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Birch.m2_ha,
  BA.m2_ha,
  SI_species,
  SI100,
  age,
  thinned,
  last_thinned,
  latitude,
  altitude,
  divided_plot = 0
)

Soderberg_1986_BAI5_southern_Sweden_under_65_Spruce(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA_Birch.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  aspect,
  soil_moisture,
  county,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_southern_Sweden_under_65_Pine(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  soil_moisture,
  county,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_southern_Sweden_under_65_Birch(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  soil_moisture,
  county,
  divided_plot = 0,
  fertilised_plot = 0
)

Soderberg_1986_BAI5_southern_Sweden_over_45_Spruce(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA_Birch.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  aspect,
  soil_moisture,
  county,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0
)

Soderberg_1986_BAI5_southern_Sweden_over_45_Pine(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  soil_moisture,
  latitude,
  altitude,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_southern_Sweden_over_45_Birch(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA_Birch.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  soil_moisture,
  county,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_southern_Sweden_Broadleaves(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA_Birch.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  aspect,
  latitude,
  divided_plot = 0
)

Soderberg_1986_BAI5_northern_Sweden_under_65_Spruce(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  aspect,
  latitude,
  altitude,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_northern_Sweden_under_65_Pine(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  latitude,
  aspect,
  soil_moisture,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_northern_Sweden_over_45_Spruce(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  aspect,
  soil_moisture,
  latitude,
  altitude,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_northern_Sweden_over_45_Pine(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Spruce.m2_ha,
  BA_Birch.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  latitude,
  altitude,
  aspect,
  soil_moisture,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_northern_central_Sweden_under_65_Birch(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA_Birch.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  vegetation,
  aspect,
  altitude,
  soil_moisture,
  county,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_northern_central_Sweden_over_45_Birch(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA_Birch.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  vegetation,
  aspect,
  latitude,
  soil_moisture,
  county,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_northern_central_Sweden_Broadleaves(
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  aspect,
  latitude,
  altitude,
  divided_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_central_Sweden_under_65_Spruce(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA_Birch.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  soil_moisture,
  latitude,
  altitude,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_central_Sweden_under_65_Pine(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  latitude,
  altitude,
  soil_moisture,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)

Soderberg_1986_BAI5_central_Sweden_over_45_Spruce(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  soil_moisture,
  latitude,
  altitude,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0
)

Soderberg_1986_BAI5_central_Sweden_over_45_Pine(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_tree.m2,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  age,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  latitude,
  altitude,
  soil_moisture,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0,
  plot_inventoried_76_77 = 0
)
}
\arguments{
\item{DBH.cm}{Diameter at breast height}

\item{DBH_largest_tree_on_plot.cm}{Diameter at breast height of the largest tree on the plot.}

\item{BA_tree.m2}{Basal area of the tree, m^2.}

\item{BA.m2_ha}{Basal area of all tree species the plot, m^2 / ha.}

\item{age}{Age at breast height of the tree.}

\item{thinned}{TRUE if the stand has been thinned, otherwise FALSE.}

\item{last_thinned}{Number of growing seasons since last thinning.}

\item{aspect}{If more than 2:20 / 5 percent, one of the following. Otherwise 0.

\tabular{cl}{
1 \tab North  \cr
2 \tab North-East \cr
3 \tab East \cr
4 \tab South-East \cr
5 \tab South \cr
6 \tab South-West \cr
7 \tab West \cr
8 \tab North-West \cr
}}

\item{maritime}{TRUE, if the plot is situated in a maritime climatic region.
cf. Ångstrom 1958. e.g. \code{\link[=Angstrom_1958_local_climate_Sweden]{Angstrom_1958_local_climate_Sweden()}} .
Otherwise FALSE.}

\item{BA_Pine.m2_ha}{Basal area Pine on the plot, m^2 / ha.}

\item{BA_Birch.m2_ha}{Basal area Birch on the plot, m^2 / ha.}

\item{SI_species}{Species for which SIH100 was estimated.
One of : 'Picea abies' or 'Pinus sylvestris'.}

\item{SI100}{Site Index H100, m.}

\item{latitude}{Latitude, degrees.}

\item{altitude}{Altitude, meters above sea level.}

\item{divided_plot}{1 for plots described in different parts,
which appears when the original plot consists of different land classes,
density classes or cutting classes or belongs to different owners. 0 for
full plots (default).}

\item{BA_Spruce.m2_ha}{Basal area Spruce on the plot, m^2 / ha.}

\item{soil_moisture}{1-5. e.g. \link{Sweden_soil_types('moisture')}}

\item{county}{Character string. One of:
\tabular{cc}{
Northern Sweden \tab \cr
\tab Norrbottens lappmark \cr
\tab Norrbottens kustland \cr
\tab Västerbottens lappmark \cr
\tab Västerbottens kustland \cr
\tab Västernorrland - Ångermanlands landskap \cr
\tab Västernorrland - Medelpads landskap \cr
\tab Jämtland - Jämtlands landskap \cr
\tab Jämtland - Härjedalens landskap \cr
\tab Kopparberg - Sälen-Idre. \cr

Central Sweden \tab \cr
\tab Kopparberg - övriga \cr
\tab Gävleborg - Hälsinglands landskap \cr
\tab Gävleborg - övriga \cr
\tab Värmland \cr

Southern Sweden \tab \cr
\tab Stockholm \cr
\tab Uppsala \cr
\tab Västmanland \cr
\tab Södermanland \cr
\tab Örebro \cr
\tab Östergötland \cr
\tab Skaraborg \cr
\tab Älvsborg - Västergötlands landskap \cr
\tab Älvsborg - Dalslands landskap \cr
\tab Jönköping \cr
\tab Kronoberg \cr
\tab Kalmar \cr
\tab Halland \cr
\tab Kristianstad \cr
\tab Malmöhus \cr
\tab Blekinge \cr
\tab Gotland \cr
}}

\item{peatland}{1 if plot is Peatland, 0 for others (default).}

\item{fertilised_plot}{1 for fertilised plots, 0 for others (default).}

\item{plot_inventoried_76_77}{1 for plots measured in the years 1976-77,
0 for others (default).}

\item{vegetation}{1-18. See \code{\link[=Sweden_vegetation_types]{Sweden_vegetation_types()}} for
description.}
}
\value{
Basal area increment during 5 years, m2.
}
\description{
\strong{Applicable counties:}
\tabular{cc}{
Northern Sweden \tab \cr
\tab Norrbottens lappmark \cr
\tab Norrbottens kustland \cr
\tab Västerbottens lappmark \cr
\tab Västerbottens kustland \cr
\tab Västernorrland - Ångermanlands landskap \cr
\tab Västernorrland - Medelpads landskap \cr
\tab Jämtland - Jämtlands landskap \cr
\tab Jämtland - Härjedalens landskap \cr
\tab Kopparberg - Sälen-Idre. \cr

Central Sweden \tab \cr
\tab Kopparberg - övriga \cr
\tab Gävleborg - Hälsinglands landskap \cr
\tab Gävleborg - övriga \cr
\tab Värmland \cr

Southern Sweden \tab \cr
\tab Stockholm \cr
\tab Uppsala \cr
\tab Västmanland \cr
\tab Södermanland \cr
\tab Örebro \cr
\tab Östergötland \cr
\tab Skaraborg \cr
\tab Älvsborg - Västergötlands landskap \cr
\tab Älvsborg - Dalslands landskap \cr
\tab Jönköping \cr
\tab Kronoberg \cr
\tab Kalmar \cr
\tab Halland \cr
\tab Kristianstad \cr
\tab Malmöhus \cr
\tab Blekinge \cr
\tab Gotland \cr
}

\strong{Regarding the regional division of the models:}

Freely translated from Söderberg 1986 p. 29:
'The regional division is motivated in the following.
For Pine the country has been divided into 3 regions (see figure 4.1),
which are based on the different forms of Scots Pine which have been
distinguished by Sylvén 1917. The distribution of the different forms of
Scots Pine is assumed to depend on the expansion-history of the Pine,
according to which the Pine immigrated both from the north and the south
with a transitional zone in central Sweden, where both forms are present.
For other species, the change in genotype is more continuous over the
country (Kiellander 1974). For Norway Spruce, it is assumed that it has had
time for 30-50 generations in northern Sweden, contrasted against only 10-20
in southern Sweden (Kiellander 1966). Therefore the same regional division
is used for Norway Spruce as for Scots Pine. For the other species, the
regional division is limited to two regions, wherein the northern and central
regions have been merged.

\strong{NB in Söderberg (1986), no correction for logarithmic bias was introduced}
, as, (freely translated) p. 114: "At the presentation of the functions we
lightly touched on the effects of the measurement errors in the variables
which had been included in the regression. It was then concluded
that several factors contrived to that the spread about the functions is
overestimated. Therefore no correction for logarithmic bias was carried out,
primarily because the effect of the errors in the indipendent variables on
the spread about the function are difficult to establish. If one were to
assume that the spread about the function was overestimated by 10 percent,
the predicted growth from the functions would be increased by roughly 2
percent. For 20 percents overestimation, an increase of predicted growth by
4.5 percent."
}
\details{
\strong{Beech Southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.87 \cr
Spread about the function sf \tab 0.586 \cr
sf/Spread about the mean \tab 0.53 \cr
Number of observations \tab 178 \cr
Sum of weights \tab 161.5 \cr
}

\strong{Norway Spruce Southern Sweden under 65}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.85 \cr
Spread about the function sf \tab 0.526 \cr
sf/Spread about the mean \tab 0.52 \cr
Number of observations \tab 3564 \cr
Sum of weights \tab 2997.7 \cr
}

\strong{Norway Spruce Southern Sweden over 45}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.84 \cr
Spread about the function sf \tab 0.559 \cr
sf/Spread about the mean \tab 0.55 \cr
Number of observations \tab 2651 \cr
Sum of weights \tab 2446.5 \cr
}

\strong{Birch Southern Sweden under 65}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.78 \cr
Spread about the function sf \tab 0.658 \cr
sf/Spread about the mean \tab 0.63 \cr
Number of observations \tab 1034 \cr
Sum of weights \tab 892.9 \cr
}

\strong{Birch Southern Sweden over 45}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.74 \cr
Spread about the function sf \tab 0.697 \cr
sf/Spread about the mean \tab 0.68 \cr
Number of observations \tab 704 \cr
Sum of weights \tab 648.6 \cr
}

\strong{Broadleaves Southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.79 \cr
Spread about the function sf \tab 0.630 \cr
sf/Spread about the mean \tab 0.62 \cr
Number of observations \tab 643 \cr
Sum of weights \tab 541.1 \cr
}

\strong{Oak Southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.90 \cr
Spread about the function sf \tab 0.417 \cr
sf/Spread about the mean \tab 0.45 \cr
Number of observations \tab 222 \cr
Sum of weights \tab 187.9 \cr
}

\strong{Scots Pine Southern Sweden under 65}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.80 \cr
Spread about the function sf \tab 0.506 \cr
sf/Spread about the mean \tab 0.60 \cr
Number of observations \tab 1662 \cr
Sum of weights \tab 1417.6 \cr
}

\strong{Scots Pine Southern Sweden over 45}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.72 \cr
Spread about the function sf \tab 0.514 \cr
sf/Spread about the mean \tab 0.70 \cr
Number of observations \tab 2373 \cr
Sum of weights \tab 2221.5 \cr
}

\strong{Scots Pine Central Sweden under 65}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.84 \cr
Spread about the function sf \tab 0.476 \cr
sf/Spread about the mean \tab 0.55 \cr
Number of observations \tab 815 \cr
Sum of weights \tab 692.9 \cr
}

\strong{Norway Spruce Northern Sweden under 65}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.87 \cr
Spread about the function sf \tab 0.470 \cr
sf/Spread about the mean \tab 0.49 \cr
Number of observations \tab 1790 \cr
Sum of weights \tab 1610.4 \cr
}

\strong{Norway Spruce Northern Sweden over 45}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.84 \cr
Spread about the function sf \tab 0.526 \cr
sf/Spread about the mean \tab 0.55 \cr
Number of observations \tab 3261 \cr
Sum of weights \tab 3035.4 \cr
}

\strong{Norway Spruce Central Sweden under 65}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.88 \cr
Spread about the function sf \tab 0.505 \cr
sf/Spread about the mean \tab 0.47 \cr
Number of observations \tab 1565 \cr
Sum of weights \tab 1370.1 \cr
}

\strong{Norway Spruce Central Sweden over 45}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.86 \cr
Spread about the function sf \tab 0.546 \cr
sf/Spread about the mean \tab 0.52 \cr
Number of observations \tab 1733 \cr
Sum of weights \tab 1606.5 \cr
}

\strong{Broadleaves Northern or Central Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.83 \cr
Spread about the function sf \tab 0.595 \cr
sf/Spread about the mean \tab 0.57 \cr
Number of observations \tab 246  \cr
Sum of weights \tab 216.0 \cr
}

\strong{Birch Northern or Central Sweden under 65}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.76 \cr
Spread about the function sf \tab 0.637 \cr
sf/Spread about the mean \tab 0.65 \cr
Number of observations \tab 755 \cr
Sum of weights \tab 706.8 \cr
}

TO DO: Same as under 65?
\strong{Birch Northern or Central Sweden over 45}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.76 \cr
Spread about the function sf \tab 0.637 \cr
sf/Spread about the mean \tab 0.65 \cr
Number of observations \tab 755 \cr
Sum of weights \tab 706.8 \cr
}

\strong{Scots Pine Northern Sweden under 65}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.88 \cr
Spread about the function sf \tab 0.419 \cr
sf/Spread about the mean \tab 0.48 \cr
Number of observations \tab 1249 \cr
Sum of weights \tab 1084.2 \cr
}

\strong{Scots Pine Northern Sweden over 45}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.80 \cr
Spread about the function sf \tab 0.502 \cr
sf/Spread about the mean \tab 0.60 \cr
Number of observations \tab 2495 \cr
Sum of weights \tab 2351.3 \cr
}

#' \strong{Scots Pine Central Sweden over 45}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.81 \cr
Spread about the function sf \tab 0.476 \cr
sf/Spread about the mean \tab 0.58 \cr
Number of observations \tab 1440 \cr
Sum of weights \tab 1350.9 \cr
}
}
