% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soderberg_1986_1992.R
\name{Soderberg1986doublebark}
\alias{Soderberg1986doublebark}
\alias{Soderberg_1986_double_bark_central_Sweden_Pine}
\alias{Soderberg_1986_double_bark_central_Sweden_Spruce}
\alias{Soderberg_1986_double_bark_northern_central_Sweden_Birch}
\alias{Soderberg_1986_double_bark_northern_central_Sweden_Broadleaves}
\alias{Soderberg_1986_double_bark_northern_Sweden_Pine}
\alias{Soderberg_1986_double_bark_northern_Sweden_Spruce}
\alias{Soderberg_1986_double_bark_southern_Sweden_Birch}
\alias{Soderberg_1986_double_bark_southern_Sweden_Broadleaves}
\alias{Soderberg_1986_double_bark_southern_Sweden_Spruce}
\alias{Soderberg_1986_double_bark_Sweden_Beech}
\alias{Soderberg_1986_double_bark_Sweden_Betula_pendula}
\alias{Soderberg_1986_double_bark_Sweden_Betula_pubescens}
\alias{Soderberg_1986_double_bark_Sweden_Oak}
\title{Double bark for Individual trees in Sweden from Söderberg 1986.}
\source{
Söderberg, U. (1986) Funktioner för skogliga produktionsprognoser -
Tillväxt och formhöjd för enskilda träd av inhemska trädslag i Sverige. /
Functions for forecasting of timber yields - Increment and form height for
individual trees of native species in Sweden. Report 14. Section of Forest
Mensuration and Management. Swedish University of Agricultural Sciences.
Umeå. ISBN 91-576-2634-0. ISSN 0349-2133. pp.251.
}
\usage{
Soderberg_1986_double_bark_central_Sweden_Pine(
  DBH.u.b.cm,
  age,
  latitude,
  altitude,
  soil_moisture,
  vegetation,
  continental,
  SI100
)

Soderberg_1986_double_bark_central_Sweden_Spruce(
  DBH.u.b.cm,
  age,
  soil_moisture,
  distance_to_coast_km,
  SI100
)

Soderberg_1986_double_bark_northern_central_Sweden_Birch(
  DBH.u.b.cm,
  age,
  latitude,
  vegetation,
  soil_moisture,
  SI100,
  distance_to_coast_km
)

Soderberg_1986_double_bark_northern_central_Sweden_Broadleaves(
  DBH.u.b.cm,
  age,
  latitude,
  altitude,
  distance_to_coast_km,
  SI100
)

Soderberg_1986_double_bark_northern_Sweden_Pine(
  DBH.u.b.cm,
  age,
  latitude,
  altitude,
  soil_moisture,
  distance_to_coast_km,
  vegetation,
  continental,
  SI100
)

Soderberg_1986_double_bark_northern_Sweden_Spruce(
  DBH.u.b.cm,
  age,
  latitude,
  soil_moisture,
  SI100
)

Soderberg_1986_double_bark_southern_Sweden_Birch(
  DBH.u.b.cm,
  age,
  latitude,
  altitude,
  vegetation,
  soil_moisture,
  SI100,
  distance_to_coast_km
)

Soderberg_1986_double_bark_southern_Sweden_Broadleaves(
  DBH.u.b.cm,
  age,
  latitude,
  altitude,
  soil_moisture,
  vegetation
)

Soderberg_1986_double_bark_central_Sweden_Pine(
  DBH.u.b.cm,
  age,
  latitude,
  altitude,
  soil_moisture,
  vegetation,
  continental,
  SI100
)

Soderberg_1986_double_bark_southern_Sweden_Spruce(
  DBH.u.b.cm,
  age,
  latitude,
  soil_moisture,
  SI100
)

Soderberg_1986_double_bark_Sweden_Beech(DBH.u.b.cm, age, latitude)

Soderberg_1986_double_bark_Sweden_Betula_pendula(
  DBH.u.b.cm,
  age,
  latitude,
  vegetation,
  soil_moisture,
  SI100,
  distance_to_coast_km
)

Soderberg_1986_double_bark_Sweden_Betula_pubescens(
  DBH.u.b.cm,
  age,
  latitude,
  vegetation,
  soil_moisture,
  SI100,
  distance_to_coast_km
)

Soderberg_1986_double_bark_Sweden_Oak(DBH.u.b.cm, age, latitude)
}
\arguments{
\item{DBH.u.b.cm}{Diameter under bark of tree, in cm.}

\item{age}{Age at breast height of the tree.}

\item{latitude}{Latitude, degrees.}

\item{altitude}{Altitude, metres.}

\item{soil_moisture}{1-5. e.g. \link{Sweden_soil_types('moisture')}}

\item{vegetation}{Vegetation type according to follows Swedish National
forest inventory FALTSKIKT. e.g. \code{\link[=Sweden_vegetation_types]{Sweden_vegetation_types()}}}

\item{continental}{TRUE, if the plot is situated in a continental climatic
region. cf. Ångström 1958. e.g.
\code{\link[=Angstrom_1958_local_climate_Sweden]{Angstrom_1958_local_climate_Sweden()}}. Otherwise FALSE.}

\item{SI100}{Site Index H100, m. TO DO: Ask P-M Ekö for clarification.}

\item{distance_to_coast_km}{Closest distance to coast, in km, e.g.
\code{\link[=coast_distance]{coast_distance()}}}
}
\value{
Double bark thickness, mm.
}
\description{
\strong{Applicable counties:}
\tabular{cc}{
Northern Sweden \tab \cr
\tab Norrbottens lappmark \cr
\tab Norrbottens kustland \cr
\tab Västerbottens lappmark \cr
\tab Västerbottens kustland \cr
\tab Västernorrland - Ångermanlands landskap \cr
\tab Västernorrland - Medelpads landskap \cr
\tab Jämtland - Jämtlands landskap \cr
\tab Jämtland - Härjedalens landskap \cr
\tab Kopparberg - Sälen-Idre. \cr

Central Sweden \tab \cr
\tab Kopparberg - övriga \cr
\tab Gävleborg - Hälsinglands landskap \cr
\tab Gävleborg - övriga \cr
\tab Värmland \cr

Southern Sweden \tab \cr
\tab Stockholm \cr
\tab Uppsala \cr
\tab Västmanland \cr
\tab Södermanland \cr
\tab Örebro \cr
\tab Östergötland \cr
\tab Skaraborg \cr
\tab Älvsborg - Västergötlands landskap \cr
\tab Älvsborg - Dalslands landskap \cr
\tab Jönköping \cr
\tab Kronoberg \cr
\tab Kalmar \cr
\tab Halland \cr
\tab Kristianstad \cr
\tab Malmöhus \cr
\tab Blekinge \cr
\tab Gotland \cr
}

\strong{Regarding the regional division of the models:}

Freely translated from Söderberg 1986 p. 29:
'The regional division is motivated in the following.
For Pine the country has been divided into 3 regions (see figure 4.1),
which are based on the different forms of Scots Pine which have been
distinguished by Sylvén 1917. The distribution of the different forms of
Scots Pine is assumed to depend on the expansion-history of the Pine,
according to which the Pine immigrated both from the north and the south
with a transitional zone in central Sweden, where both forms are present.
For other species, the change in genotype is more continuous over the
country (Kiellander 1974). For Norway Spruce, it is assumed that it has had
time for 30-50 generations in northern Sweden, contrasted against only 10-20
in southern Sweden (Kiellander 1966). Therefore the same regional division
is used for Norway Spruce as for Scots Pine. For the other species, the
regional division is limited to two regions, wherein the northern and central
regions have been merged.

\strong{NB in Söderberg (1986), no correction for logarithmic bias was introduced}
, as, (freely translated) p. 114: "At the presentation of the functions we
lightly touched on the effects of the measurement errors in the variables
which had been included in the regression. It was then concluded
that several factors contrived to that the spread about the functions is
overestimated. Therefore no correction for logarithmic bias was carried out,
primarily because the effect of the errors in the indipendent variables on
the spread about the function are difficult to establish. If one were to
assume that the spread about the function was overestimated by 10 percent,
the predicted growth from the functions would be increased by roughly 2
percent. For 20 percents overestimation, an increase of predicted growth by
4.5 percent."
}
\details{
\strong{Oak}

\tabular{ll}{
Multiple correlation coefficient R \tab 0.89 \cr
Spread about the function sf \tab 0.20 \cr
sf/Spread about the mean \tab 0.38 \cr
Number of observations \tab 315 \cr
}

\strong{Beech}

\tabular{ll}{
Multiple correlation coefficient R \tab 0.88 \cr
Spread about the function sf \tab 0.21 \cr
sf/Spread about the mean \tab 0.41 \cr
Number of observations \tab 291 \cr
}

\strong{Norway Spruce}
\strong{For Norway Spruce, single function is reported area-wise due to use
of dummy variables in only some locations. Regression statistics for entire
country}

\strong{Norway Spruce in Sweden}

\tabular{ll}{
Multiple correlation coefficient R \tab 0.84 \cr
Spread about the function sf \tab 0.22 \cr
sf/Spread about the mean \tab 0.48 \cr
Number of observations \tab 12279 \cr
}

\strong{Scots Pine in Northern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.87 \cr
Spread about the function sf \tab 0.22 \cr
sf/Spread about the mean \tab 0.43 \cr
Number of observations \tab 2316 \cr
}

\strong{Scots Pine in Central Sweden} ##

\tabular{ll}{
Multiple correlation coefficient R \tab 0.85 \cr
Spread about the function sf \tab 0.23 \cr
sf/Spread about the mean \tab 0.46 \cr
Number of observations \tab 2014 \cr
}

\strong{Scots Pine in southern Sweden}

\tabular{ll}{
Multiple correlation coefficient R \tab 0.83 \cr
Spread about the function sf \tab 0.24 \cr
sf/Spread about the mean \tab 0.49 \cr
Number of observations \tab 4833 \cr
}

\strong{Broadleaves in southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.82 \cr
Spread about the function sf \tab 0.30 \cr
sf/Spread about the mean \tab 0.53 \cr
Number of observations \tab 403 \cr
}

\strong{Broadleaves in Northern and Central Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.84 \cr
Spread about the function sf \tab 0.29 \cr
sf/Spread about the mean \tab 0.52 \cr
Number of observations \tab 244 \cr
}

*\strong{Birch species: Betula pendula}

NB: \emph{Betula pendula} is in the original source referred as
\emph{Betula verrucosa}

\tabular{ll}{
Multiple correlation coefficient R \tab 0.83 \cr
Spread about the function sf \tab 0.32 \cr
sf/Spread about the mean \tab 0.47 \cr
Number of observations \tab 846 \cr
}

*\strong{Birch species: Betula pubescens}

\tabular{ll}{
Multiple correlation coefficient R \tab 0.83 \cr
Spread about the function sf \tab 0.23 \cr
sf/Spread about the mean \tab 0.46 \cr
Number of observations \tab 1852 \cr
}

\strong{Birch in Northern or Central Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.85 \cr
Spread about the function sf \tab 0.22 \cr
sf/Spread about the mean \tab 0.43 \cr
Number of observations \tab 1093 \cr
}

\strong{Birch in southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.81 \cr
Spread about the function sf \tab 0.27 \cr
sf/Spread about the mean \tab 0.47 \cr
Number of observations \tab 1605 \cr
}
}
