% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soderberg_1986_1992.R
\name{Soderberg1986formfactor}
\alias{Soderberg1986formfactor}
\alias{Soderberg_1986_form_factor_Sweden_Oak}
\alias{Soderberg_1986_form_factor_Sweden_Beech}
\alias{Soderberg_1986_form_factor_southern_Sweden_Spruce}
\alias{Soderberg_1986_form_factor_southern_Sweden_Pine}
\alias{Soderberg_1986_form_factor_southern_Sweden_Broadleaves}
\alias{Soderberg_1986_form_factor_southern_Sweden_Birch}
\alias{Soderberg_1986_form_factor_northern_Sweden_Pine}
\alias{Soderberg_1986_form_factor_northern_central_Sweden_Spruce}
\alias{Soderberg_1986_form_factor_northern_central_Sweden_Broadleaves}
\alias{Soderberg_1986_form_factor_northern_central_Sweden_Birch}
\alias{Soderberg_1986_form_factor_Gotland_Sweden_Pine}
\alias{Soderberg_1986_form_factor_central_Sweden_Pine}
\title{Form factor for Individual trees in Sweden from Söderberg 1986.}
\source{
Söderberg, U. (1986) Funktioner för skogliga produktionsprognoser -
Tillväxt och formhöjd för enskilda träd av inhemska trädslag i Sverige. /
Functions for forecasting of timber yields - Increment and form height for
individual trees of native species in Sweden. Report 14. Section of Forest
Mensuration and Management. Swedish University of Agricultural Sciences.
Umeå. ISBN 91-576-2634-0. ISSN 0349-2133. pp.251.
}
\usage{
Soderberg_1986_form_factor_Sweden_Oak(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_Beech.m2_ha,
  BA_Birch.m2_ha,
  BA.m2_ha,
  SI_species,
  SI100,
  age,
  vegetation,
  soil_moisture,
  lateral_water,
  altitude,
  divided_plot = 0,
  county
)

Soderberg_1986_form_factor_Sweden_Beech(
  DBH.cm,
  BA_Beech.m2_ha,
  BA_Oak.m2_ha,
  BA.m2_ha,
  SI100_Spruce,
  age,
  aspect,
  altitude,
  divided_plot = 0
)

Soderberg_1986_form_factor_southern_Sweden_Spruce(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  continental,
  maritime,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  SI_species,
  SI100,
  age,
  aspect,
  latitude,
  altitude,
  lateral_water,
  divided_plot = 0,
  county
)

Soderberg_1986_form_factor_southern_Sweden_Pine(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  distance_to_coast_km,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  SI_species,
  SI100,
  vegetation,
  age,
  latitude,
  altitude,
  county,
  maritime,
  soil_moisture,
  divided_plot = 0,
  fertilised_plot
)

Soderberg_1986_form_factor_southern_Sweden_Broadleaves(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  SI_species,
  SI100,
  age,
  latitude,
  vegetation,
  soil_moisture,
  aspect,
  divided_plot = 0
)

Soderberg_1986_form_factor_southern_Sweden_Birch(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  maritime,
  BA_Birch.m2_ha,
  BA_Pine.m2_ha,
  BA.m2_ha,
  SI_species,
  SI100,
  vegetation,
  age,
  latitude,
  altitude,
  divided_plot = 0,
  fertilised_plot = 0,
  county
)

Soderberg_1986_form_factor_northern_Sweden_Pine(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  distance_to_coast_km,
  BA_Pine.m2_ha,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  SI_species,
  SI100,
  vegetation,
  age,
  latitude,
  altitude,
  aspect,
  soil_moisture,
  divided_plot = 0
)

Soderberg_1986_form_factor_northern_central_Sweden_Spruce(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  distance_to_coast_km,
  continental,
  BA_Spruce.m2_ha,
  BA_Pine.m2_ha,
  BA.m2_ha,
  SI_species,
  SI100,
  vegetation,
  age,
  aspect,
  latitude,
  altitude,
  soil_moisture,
  lateral_water,
  divided_plot = 0
)

Soderberg_1986_form_factor_northern_central_Sweden_Broadleaves(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  maritime,
  BA_Spruce.m2_ha,
  BA.m2_ha,
  SI_species,
  SI100,
  age,
  latitude,
  altitude,
  lateral_water,
  divided_plot = 0
)

Soderberg_1986_form_factor_northern_central_Sweden_Birch(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_Spruce.m2_ha,
  BA_Pine.m2_ha,
  BA.m2_ha,
  SI_species,
  SI100,
  vegetation,
  aspect,
  age,
  latitude,
  altitude,
  soil_moisture,
  lateral_water,
  fertilised_plot,
  divided_plot = 0
)

Soderberg_1986_form_factor_Gotland_Sweden_Pine(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  BA_Pine.m2_ha,
  BA.m2_ha,
  SI100_Spruce,
  age,
  altitude,
  divided_plot = 0
)

Soderberg_1986_form_factor_central_Sweden_Pine(
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  distance_to_coast_km,
  BA_Pine.m2_ha,
  BA.m2_ha,
  SI_species,
  SI100,
  age,
  latitude,
  altitude,
  soil_moisture,
  fertilised_plot,
  divided_plot = 0
)
}
\arguments{
\item{DBH.cm}{Diameter at breast height}

\item{DBH_largest_tree_on_plot.cm}{Diameter at breast height of the
largest tree on the plot.}

\item{BA_Beech.m2_ha}{Basal area Beech on the plot, m^2 / ha.}

\item{BA_Birch.m2_ha}{Basal area Birch on the plot, m^2 / ha.}

\item{BA.m2_ha}{Basal area of all tree species the plot, m^2 / ha.}

\item{SI_species}{Species for which SIH100 was estimated. One of :
'Picea abies' or 'Pinus sylvestris'.}

\item{SI100}{Site Index H100, m.}

\item{age}{Age at breast height of the tree.}

\item{vegetation}{1-18. See \code{\link[=Sweden_vegetation_types]{Sweden_vegetation_types()}} for
description.}

\item{soil_moisture}{1-5. e.g. \code{\link[=Sweden_soil_types]{Sweden_soil_types()}}}

\item{lateral_water}{\tabular{cl}{
Code \tab Description \cr
1 \tab Missing \cr
2 \tab Seldom \cr
3 \tab Shorter periods \cr
4 \tab Longer periods \cr
5 \tab Slope \cr
}}

\item{altitude}{Altitude, meters above sea level.}

\item{divided_plot}{1 for plots described in different parts,
which appears when the original plot consists of different land classes,
density classes or cutting classes or belongs to different owners. 0 for
full plots (default).}

\item{county}{Character string. One of:
\tabular{cc}{
Northern Sweden \tab \cr
\tab Norrbottens lappmark \cr
\tab Norrbottens kustland \cr
\tab Västerbottens lappmark \cr
\tab Västerbottens kustland \cr
\tab Västernorrland - Ångermanlands landskap \cr
\tab Västernorrland - Medelpads landskap \cr
\tab Jämtland - Jämtlands landskap \cr
\tab Jämtland - Härjedalens landskap \cr
\tab Kopparberg - Sälen-Idre. \cr

Central Sweden \tab \cr
\tab Kopparberg - övriga \cr
\tab Gävleborg - Hälsinglands landskap \cr
\tab Gävleborg - övriga \cr
\tab Värmland \cr

Southern Sweden \tab \cr
\tab Stockholm \cr
\tab Uppsala \cr
\tab Västmanland \cr
\tab Södermanland \cr
\tab Örebro \cr
\tab Östergötland \cr
\tab Skaraborg \cr
\tab Älvsborg - Västergötlands landskap \cr
\tab Älvsborg - Dalslands landskap \cr
\tab Jönköping \cr
\tab Kronoberg \cr
\tab Kalmar \cr
\tab Halland \cr
\tab Kristianstad \cr
\tab Malmöhus \cr
\tab Blekinge \cr
\tab Gotland \cr
}}

\item{BA_Oak.m2_ha}{Basal area Oak on the plot, m^2 / ha.}

\item{SI100_Spruce}{Site Index H100, m for Norway Spruce. See description.#'}

\item{aspect}{If more than 2:20 / 5\\%, one of the following. Otherwise 0.
\tabular{cl}{
1 \tab North  \cr
2 \tab North-East \cr
3 \tab East \cr
4 \tab South-East \cr
5 \tab South \cr
6 \tab South-West \cr
7 \tab West \cr
8 \tab North-West \cr
}}

\item{continental}{TRUE, if the plot is situated in a continental climatic region.
cf. Ångstrom 1958. e.g. \code{\link[=Angstrom_1958_local_climate_Sweden]{Angstrom_1958_local_climate_Sweden()}} .
Otherwise FALSE.}

\item{maritime}{TRUE, if the plot is situated in a maritime climatic region.
cf. Ångstrom 1958. e.g. \code{\link[=Angstrom_1958_local_climate_Sweden]{Angstrom_1958_local_climate_Sweden()}} .
Otherwise FALSE.}

\item{BA_Spruce.m2_ha}{Basal area Norway Spruce on the plot, m^2 / ha.}

\item{latitude}{Latitude, degrees.}

\item{distance_to_coast_km}{Closest distance to coast, in km, e.g. \link{coast_distance}}

\item{fertilised_plot}{1 for fertilised plots, 0 for others (default).}

\item{BA_Pine.m2_ha}{Basal area Scots Pine on the plot, m^2 / ha.}
}
\value{
Form quotient of tree
}
\description{
\strong{Applicable counties:}
\tabular{lr}{
Northern Sweden \tab \cr
\tab Norrbottens lappmark \cr
\tab Norrbottens kustland \cr
\tab Västerbottens lappmark \cr
\tab Västerbottens kustland \cr
\tab Västernorrland - Ångermanlands landskap \cr
\tab Västernorrland - Medelpads landskap \cr
\tab Jämtland - Jämtlands landskap \cr
\tab Jämtland - Härjedalens landskap \cr
\tab Kopparberg - Sälen-Idre. \cr

Central Sweden \tab \cr
\tab Kopparberg - övriga \cr
\tab Gävleborg - Hälsinglands landskap \cr
\tab Gävleborg - övriga \cr
\tab Värmland \cr

Southern Sweden \tab \cr
\tab Stockholm \cr
\tab Uppsala \cr
\tab Västmanland \cr
\tab Södermanland \cr
\tab Örebro \cr
\tab Östergötland \cr
\tab Skaraborg \cr
\tab Älvsborg - Västergötlands landskap \cr
\tab Älvsborg - Dalslands landskap \cr
\tab Jönköping \cr
\tab Kronoberg \cr
\tab Kalmar \cr
\tab Halland \cr
\tab Kristianstad \cr
\tab Malmöhus \cr
\tab Blekinge \cr
\tab Gotland \cr
}

\strong{Regarding the regional division of the models:}

Freely translated from Söderberg 1986 p. 29:
'The regional division is motivated in the following.
For Pine the country has been divided into 3 regions (see figure 4.1),
which are based on the different forms of Scots Pine which have been
distinguished by Sylvén 1917. The distribution of the different forms of
Scots Pine is assumed to depend on the expansion-history of the Pine,
according to which the Pine immigrated both from the north and the south
with a transitional zone in central Sweden, where both forms are present.
For other species, the change in genotype is more continuous over the
country (Kiellander 1974). For Norway Spruce, it is assumed that it has had
time for 30-50 generations in northern Sweden, contrasted against only 10-20
in southern Sweden (Kiellander 1966). Therefore the same regional division
is used for Norway Spruce as for Scots Pine. For the other species, the
regional division is limited to two regions, wherein the northern and central
regions have been merged.

\strong{NB in Söderberg (1986), no correction for logarithmic bias was introduced}
, as, (freely translated) p. 114: "At the presentation of the functions we
lightly touched on the effects of the measurement errors in the variables
which had been included in the regression. It was then concluded
that several factors contrived to that the spread about the functions is
overestimated. Therefore no correction for logarithmic bias was carried out,
primarily because the effect of the errors in the indipendent variables on
the spread about the function are difficult to establish. If one were to
assume that the spread about the function was overestimated by 10 percent,
the predicted growth from the functions would be increased by roughly 2
percent. For 20 percents overestimation, an increase of predicted growth by
4.5 percent."
}
\details{
\strong{Oak}

\strong{OBSERVE:} The function has a typo in the book. Although the
variable description states that it is Altitude, the symbol shows LAT. I have
interpreted this as that it should be Altitude, since the size of the
coefficient is most alike coefficients for ALT in other related functions.

\tabular{ll}{
Multiple correlation coefficient R \tab 0.86 \cr
Spread about the function sf \tab 0.180 \cr
sf/Spread about the mean \tab 0.516 \cr
Number of observations \tab 292 \cr
}

\strong{Beech}

\strong{OBSERVE:} The function has a typo in the book. Although the
variable description states that it is Altitude^2, the symbol shows LAT^2.
I have interpreted this as that it should be Altitude^2, since the size of
the coefficient is most alike coefficients for ALT^2 in other related
functions.

\strong{OBSERVE:} The following interpretation was made of the function -
since only a factor was given for the Site Index in the case that the
indicating species was Picea abies, this was interpreted as mandatory.
Therefore, SI_Species and SI100 have been replaced by SI100_Spruce.

\tabular{ll}{
Multiple correlation coefficient R \tab 0.90 \cr
Spread about the function sf \tab 0.165 \cr
sf/Spread about the mean \tab 0.442 \cr
Number of observations \tab 219 \cr
}

\strong{Norway Spruce in southern Sweden}

\strong{OBSERVE:} As per p. 64 , if county is equal to "Gotland": "Residual
studies for these species indicate the functions for the mainland (Spruce
south, Birch south) can be used with a reduction of the constant term
amounting to -0.089 and -0.067, respectively".

\tabular{ll}{
Multiple correlation coefficient R \tab 0.94 \cr
Spread about the function sf \tab 0.128 \cr
sf/Spread about the mean \tab 0.346 \cr
Number of observations \tab 7311 \cr
}

\strong{Norway Spruce in Central or Northern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.93 \cr
Spread about the function sf \tab 0.132 \cr
sf/Spread about the mean \tab 0.365 \cr
Number of observations \tab 7410 \cr
}

\strong{Scots Pine in Northern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.94 \cr
Spread about the function sf \tab 0.107 \cr
sf/Spread about the mean \tab 0.352 \cr
Number of observations \tab 3785 \cr
}

\strong{Scots Pine in Central Sweden}

\tabular{ll}{
Multiple correlation coefficient R \tab 0.94 \cr
Spread about the function sf \tab 0.105 \cr
sf/Spread about the mean \tab 0.328 \cr
Number of observations \tab 2248 \cr
}

\strong{Scots Pine in southern Sweden}

\strong{NB for the island of Gotland, use
\code{\link[=Soderberg_1986_form_factor_Gotland_Sweden_Pine]{Soderberg_1986_form_factor_Gotland_Sweden_Pine()}}}

\tabular{ll}{
Multiple correlation coefficient R \tab 0.92 \cr
Spread about the function sf \tab 0.120 \cr
sf/Spread about the mean \tab 0.395 \cr
Number of observations \tab 4231 \cr
}

\strong{Scots Pine on the island Gotland}

\tabular{ll}{
Multiple correlation coefficient R \tab 0.84 \cr
Spread about the function sf \tab 0.150 \cr
sf/Spread about the mean \tab 0.545 \cr
Number of observations \tab 351 \cr
}

\strong{Broadleaves in southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.80 \cr
Spread about the function sf \tab 0.187 \cr
sf/Spread about the mean \tab 0.609 \cr
Number of observations \tab 831 \cr
}

\strong{Broadleaves in Central or Northern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.87 \cr
Spread about the function sf \tab 0.174 \cr
sf/Spread about the mean \tab 0.499 \cr
Number of observations \tab 317 \cr
}

\strong{Birch in southern Sweden}

\strong{OBSERVE:} As per p. 64 , if county is equal to "Gotland": "Residual
studies for these species indicate the functions for the mainland (Spruce
south, Birch south) can be used with a reduction of the constant term
amounting to -0.089 and -0.067, respectively".

\tabular{ll}{
Multiple correlation coefficient R \tab 0.85 \cr
Spread about the function sf \tab 0.161 \cr
sf/Spread about the mean \tab 0.533 \cr
Number of observations \tab 1725 \cr
}

\strong{Birch in Central or Northern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.91 \cr
Spread about the function sf \tab 0.138 \cr
sf/Spread about the mean \tab 0.416 \cr
Number of observations \tab 1735 \cr
}
}
