% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soderberg_1992_height_Sweden.R
\name{Soderberg1992height}
\alias{Soderberg1992height}
\alias{Soderberg_1992_height_southern_Sweden_Beech}
\alias{Soderberg_1992_height_southern_Sweden_Pine}
\alias{Soderberg_1992_height_southern_Sweden_Birch}
\alias{Soderberg_1992_height_southern_Sweden_Broadleaves}
\alias{Soderberg_1992_height_southern_Sweden_Oak}
\alias{Soderberg_1992_height_southern_Sweden_Spruce}
\alias{Soderberg_1992_height_central_Sweden_Pine}
\alias{Soderberg_1992_height_northern_central_Sweden_Birch}
\alias{Soderberg_1992_height_northern_central_Sweden_Broadleaves}
\alias{Soderberg_1992_height_northern_central_Sweden_Spruce}
\alias{Soderberg_1992_height_northern_Sweden_Pine}
\title{Height function for Individual trees in Sweden from Söderberg 1992.}
\source{
Söderberg, U. (1992) Funktioner för skogsindelning: Höjd,
formhöjd och barktjocklek för enskilda träd. / Functions for forest
management: Height, form height and bark thickness of individual trees.
Report 52. Dept. of Forest Survey. Swedish University of Agricultural
Sciences. ISSN 0348-0496.
}
\usage{
Soderberg_1992_height_southern_Sweden_Beech(
  DBH.cm,
  total_age,
  BA_Spruce.m2,
  BA.m2,
  latitude,
  altitude,
  divided_plot = 0,
  county
)

Soderberg_1992_height_southern_Sweden_Pine(
  SI100_Pine,
  distance_to_coast_km,
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  total_age,
  BA_Pine.m2,
  BA_Spruce.m2,
  BA_Birch.m2,
  BA.m2,
  latitude,
  altitude,
  divided_plot = 0,
  county
)

Soderberg_1992_height_southern_Sweden_Birch(
  SI100_Pine,
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  total_age,
  BA_Pine.m2,
  BA_Spruce.m2,
  BA.m2,
  latitude,
  altitude,
  divided_plot = 0,
  county
)

Soderberg_1992_height_southern_Sweden_Broadleaves(
  SI100_Pine,
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  total_age,
  BA_Pine.m2,
  BA_Spruce.m2,
  BA.m2,
  latitude,
  altitude,
  divided_plot = 0
)

Soderberg_1992_height_southern_Sweden_Oak(
  SI100_Pine,
  DBH.cm,
  total_age,
  BA_Spruce.m2,
  BA.m2,
  latitude,
  altitude,
  divided_plot = 0,
  county
)

Soderberg_1992_height_southern_Sweden_Spruce(
  SI100_Pine,
  distance_to_coast_km,
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  total_age,
  BA_Pine.m2,
  BA_Spruce.m2,
  BA.m2,
  latitude,
  altitude,
  divided_plot = 0
)

Soderberg_1992_height_central_Sweden_Pine(
  SI100_Pine,
  distance_to_coast_km,
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  total_age,
  BA_Pine.m2,
  BA_Spruce.m2,
  BA_Birch.m2,
  BA.m2,
  latitude,
  altitude,
  divided_plot = 0
)

Soderberg_1992_height_northern_central_Sweden_Birch(
  SI100_Pine,
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  total_age,
  BA_Pine.m2,
  BA_Spruce.m2,
  BA.m2,
  latitude,
  altitude,
  divided_plot = 0
)

Soderberg_1992_height_northern_central_Sweden_Broadleaves(
  SI100_Pine,
  distance_to_coast_km,
  DBH.cm,
  total_age,
  BA_Spruce.m2,
  BA.m2,
  latitude,
  divided_plot = 0
)

Soderberg_1992_height_northern_central_Sweden_Spruce(
  SI100_Pine,
  distance_to_coast_km,
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  total_age,
  BA_Pine.m2,
  BA_Spruce.m2,
  BA.m2,
  latitude,
  altitude,
  divided_plot = 0
)

Soderberg_1992_height_northern_Sweden_Pine(
  SI100_Pine,
  distance_to_coast_km,
  DBH.cm,
  DBH_largest_tree_on_plot.cm,
  total_age,
  BA_Pine.m2,
  BA.m2,
  latitude,
  altitude,
  divided_plot = 0
)
}
\arguments{
\item{DBH.cm}{Diameter over bark of tree, in cm.}

\item{total_age}{Total age of the stand.}

\item{BA_Spruce.m2}{Basal Area m2 of Norway Spruce}

\item{BA.m2}{Basal area m2 on plot.}

\item{latitude}{Latitude, degrees.}

\item{altitude}{Altitude, metres.}

\item{divided_plot}{0 for full plots, 1 for divided plots.}

\item{county}{County name.}

\item{SI100_Pine}{SIH 100 Pine according to Hägglund 1974.
e.g. \link{Hagglund_1974_Sweden_height_trajectories_Pine}}

\item{distance_to_coast_km}{Closest distance to coast, in km,
e.g. \link{coast_distance}}

\item{DBH_largest_tree_on_plot.cm}{Diameter over bark of the tree
with the greatest diameter on the plot, cm.}

\item{BA_Pine.m2}{Basal Area m2 of Scots Pine}

\item{BA_Birch.m2}{Basal Area m2 of Birch}
}
\value{
Height of tree, m.
}
\description{
\strong{Applicable counties:}
\tabular{lr}{
Northern Sweden \tab \cr
\tab Norrbottens lappmark \cr
\tab Norrbottens kustland \cr
\tab Västerbottens lappmark \cr
\tab Västerbottens kustland \cr
\tab Västernorrland - Ångermanlands landskap \cr
\tab Västernorrland - Medelpads landskap \cr
\tab Jämtland - Jämtlands landskap \cr
\tab Jämtland - Härjedalens landskap \cr
\tab Kopparberg - Sälen-Idre. \cr

Central Sweden \tab \cr
\tab Kopparberg - övriga \cr
\tab Gävleborg - Hälsinglands landskap \cr
\tab Gävleborg - övriga \cr
\tab Värmland \cr

Southern Sweden \tab \cr
\tab Stockholm \cr
\tab Uppsala \cr
\tab Västmanland \cr
\tab Södermanland \cr
\tab Örebro \cr
\tab Östergötland \cr
\tab Skaraborg \cr
\tab Älvsborg - Västergötlands landskap \cr
\tab Älvsborg - Dalslands landskap \cr
\tab Jönköping \cr
\tab Kronoberg \cr
\tab Kalmar \cr
\tab Halland \cr
\tab Kristianstad \cr
\tab Malmöhus \cr
\tab Blekinge \cr
\tab Gotland \cr
}

\strong{Regarding the regional division of the models:}

Freely translated from Söderberg 1986 p. 29:
'The regional division is motivated in the following.
For Pine the country has been divided into 3 regions (see figure 4.1),
which are based on the different forms of Scots Pine which have been distinguished
by Sylvén 1917. The distribution of the different forms of Scots Pine is assumed
to depend on the expansion-history of the Pine, according to which the Pine
immigrated both from the north and the south with a transitional zone in central
Sweden, where both forms are present. For other species, the change in genotype
is more continuous over the country (Kiellander 1974). For Norway Spruce, it is
assumed that it has had time for 30-50 generations in northern Sweden, contrasted
against only 10-20 in southern Sweden (Kiellander 1966). Therefore the same regional
division is used for Norway Spruce as for Scots Pine. For the other species, the regional
division is limited to two regions, wherein the northern and central regions have been merged.
}
\details{
\strong{Beech Southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.92 \cr
Spread about the function sf \tab 0.161 \cr
sf/Spread about the mean \tab 0.41 \cr
Number of observations \tab 220 \cr
}

\strong{Norway Spruce Northern or Central Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.94 \cr
Spread about the function sf \tab 0.148 \cr
sf/Spread about the mean \tab 0.33 \cr
Number of observations \tab 7470 \cr
}

\strong{Norway Spruce Southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.95 \cr
Spread about the function sf \tab 0.145 \cr
sf/Spread about the mean \tab 0.32 \cr
Number of observations \tab 7467 \cr
}

\strong{Birch Northern or Central Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.90 \cr
Spread about the function sf \tab 0.158 \cr
sf/Spread about the mean \tab 0.43 \cr
Number of observations \tab 1747 \cr
}

\strong{Birch Southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.89 \cr
Spread about the function sf \tab 0.164 \cr
sf/Spread about the mean \tab 0.46 \cr
Number of observations \tab 1766 \cr
}

\strong{Broadleaves Northern or Central Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.88 \cr
Spread about the function sf \tab 0.191 \cr
sf/Spread about the mean \tab 0.48 \cr
Number of observations \tab 318 \cr
}
\strong{Broadleaves Southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.84 \cr
Spread about the function sf \tab 0.195 \cr
sf/Spread about the mean \tab 0.54 \cr
Number of observations \tab 837 \cr
}

\strong{Oak Southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.88 \cr
Spread about the function sf \tab 0.178 \cr
sf/Spread about the mean \tab 0.47 \cr
Number of observations \tab 292 \cr
}

\strong{Scots Pine Northern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.92 \cr
Spread about the function sf \tab 0.152 \cr
sf/Spread about the mean \tab 0.39 \cr
Number of observations \tab 3832 \cr
}

\strong{Scots Pine Central Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.94 \cr
Spread about the function sf \tab 0.137 \cr
sf/Spread about the mean \tab 0.34 \cr
Number of observations \tab 2275 \cr
}

\strong{Scots Pine Southern Sweden}
\tabular{ll}{
Multiple correlation coefficient R \tab 0.92 \cr
Spread about the function sf \tab 0.159 \cr
sf/Spread about the mean \tab 0.40 \cr
Number of observations \tab 4628 \cr
}
}
