% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Soderberg_1986_BA_increment_central_Sweden_over_45_Spruce.R
\name{Soderberg_1986_BA_increment_central_Sweden_over_45_Spruce}
\alias{Soderberg_1986_BA_increment_central_Sweden_over_45_Spruce}
\title{5 year Basal area increment over bark, Spruce, central Sweden, over 45 years at breast height from Söderberg (1986)}
\source{
Söderberg, U. (1986) Funktioner för skogliga produktionsprognoser - Tillväxt och formhöjd för enskilda träd av inhemska trädslag i Sverige. / Functions for forecasting of timber yields - Increment and form height for individual trees of native species in Sweden. Report 14. Section of Forest Mensuration and Management. Swedish University of Agricultural Sciences. Umeå. ISBN 91-576-2634-0. ISSN 0349-2133. pp.251. p. 184.
}
\usage{
Soderberg_1986_BA_increment_central_Sweden_over_45_Spruce(
  diameter_cm,
  diameter_largest_tree_on_plot_cm,
  Basal_area_of_tree_m2,
  Basal_area_Pine_m2_ha,
  Basal_area_Spruce_m2_ha,
  Basal_area_plot_m2_ha,
  age_at_breast_height,
  thinned,
  last_thinned,
  SI100,
  SI_species,
  soil_moisture,
  latitude,
  altitude,
  peatland = 0,
  divided_plot = 0,
  fertilised_plot = 0
)
}
\arguments{
\item{diameter_cm}{Diameter at breast height}

\item{diameter_largest_tree_on_plot_cm}{Diameter at breast height of the largest tree on the plot.}

\item{Basal_area_of_tree_m2}{Basal area of the tree, m^2.}

\item{Basal_area_Pine_m2_ha}{Basal area Pine on the plot, m^2 / ha.}

\item{Basal_area_Spruce_m2_ha}{Basal area Spruce on the plot, m^2 / ha.}

\item{Basal_area_plot_m2_ha}{Basal area of all tree species the plot, m^2 / ha.}

\item{age_at_breast_height}{Age at breast height of the tree.}

\item{thinned}{TRUE if the stand has been thinned, otherwise FALSE.}

\item{last_thinned}{Number of growing seasons since last thinning.}

\item{SI100}{Site Index H100, m.}

\item{SI_species}{Species for which SIH100 was estimated. One of : 'Picea abies' or 'Pinus sylvestris'.}

\item{soil_moisture}{Type 1="Dry/torr",2="Mesic/frisk",3="Mesic-moist/frisk-fuktig",4="Moist/fuktig",5="Wet/Blöt"}

\item{latitude}{Latitude, degrees.}

\item{altitude}{Altitude, meters.}

\item{peatland}{1 if plot is Peatland, 0 for others (default).}

\item{divided_plot}{1 for plots described in different parts, which appears when the original plot consists of different land classes, density classes or cutting classes or belongs to different owners. 0 for full plots (default).}

\item{fertilised_plot}{1 for fertilised plots, 0 for others (default).}
}
\value{
Basal area increment during 5 years, m2.
}
\description{
\strong{Applicable counties:}

*Kopparberg - övriga
*Gävleborg - Hälsinglands landskap
*Gävleborg - övriga
*Värmland

\strong{Regarding the regional division of the models:}

Freely translated from Söderberg 1986 p. 29:
'The regional division is motivated in the following.
For Pine the country has been divided into 3 regions (see figure 4.1),
which are based on the different forms of Scots Pine which have been distinguished
by Sylvén 1917. The distribution of the different forms of Scots Pine is assumed
to depend on the expansion-history of the Pine, according to which the Pine
immigrated both from the north and the south with a transitional zone in central
Sweden, where both forms are present. For other species, the change in genotype
is more continuous over the country (Kiellander 1974). For Norway Spruce, it is
assumed that it has had time for 30-50 generations in northern Sweden, contrasted
against only 10-20 in southern Sweden (Kiellander 1966). Therefore the same regional
division is used for Norway Spruce as for Scots Pine. For the other species, the regional
division is limited to two regions, wherein the northern and central regions have been merged.
}
\details{
Multiple correlation coefficient R = 0.86

Spread about the function sf = 0.546

sf/Spread about the mean = 0.52

Number of observations = 1733

Sum of weights: 1606.5

NB in Söderberg (1986), no correction for logarithmic bias was introduced, as, (freely translated) p. 114:
"At the presentation of the functions we lightly touched on the effects of the measurement
errors in the variables which had been included in the regression. It was then concluded
that several factors contrived to that the spread about the functions is overestimated.
Therefore no correction for logarithmic bias was carried out, primarily because the
effect of the errors in the indipendent variables on the spread about the function
are difficult to establish. If one were to assume that the spread about the function
was overestimated by 10 percent, the predicted growth from the functions would be increased by
roughly 2 percent. For 20 percents overestimation, an increase of predicted growth by 4.5 perpcent."
}
