% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soderberg_1986_BA_increment_Sweden_Oak.R
\name{Soderberg_1986_BA_increment_Sweden_Oak}
\alias{Soderberg_1986_BA_increment_Sweden_Oak}
\title{5 year Basal area increment over bark, Oak, Sweden from Söderberg (1986)}
\source{
Söderberg, U. (1986) Funktioner för skogliga produktionsprognoser - Tillväxt och formhöjd för enskilda träd av inhemska trädslag i Sverige. / Functions for forecasting of timber yields - Increment and form height for individual trees of native species in Sweden. Report 14. Section of Forest Mensuration and Management. Swedish University of Agricultural Sciences. Umeå. ISBN 91-576-2634-0. ISSN 0349-2133. pp.251. p. 204.
}
\usage{
Soderberg_1986_BA_increment_Sweden_Oak(
  diameter_cm,
  diameter_largest_tree_on_plot_cm,
  Basal_area_of_tree_m2,
  Basal_area_plot_m2_ha,
  age_at_breast_height,
  thinned,
  last_thinned,
  aspect,
  maritime
)
}
\arguments{
\item{diameter_cm}{Diameter at breast height}

\item{diameter_largest_tree_on_plot_cm}{Diameter at breast height of the largest tree on the plot.}

\item{Basal_area_of_tree_m2}{Basal area of the tree, m^2.}

\item{Basal_area_plot_m2_ha}{Basal area of all tree species the plot, m^2 / ha.}

\item{age_at_breast_height}{Age at breast height of the tree.}

\item{thinned}{TRUE if the stand has been thinned, otherwise FALSE.}

\item{last_thinned}{Number of growing seasons since last thinning.}

\item{aspect}{If more than 2:20 / 5\%, one of the following. Otherwise 0.

\tabular{cl}{
1 \tab North  \cr
2 \tab North-East \cr
3 \tab East \cr
4 \tab South-East \cr
5 \tab South \cr
6 \tab South-West \cr
7 \tab West \cr
8 \tab North-West \cr
}}

\item{maritime}{TRUE, if the plot is situated in a maritime climatic region. cf. Ångstrom 1958. e.g. [forester::local_climate_sweden()] . Otherwise FALSE.}
}
\value{
Basal area increment during 5 years, m2.
}
\description{
5 year Basal area increment over bark, Oak, Sweden from Söderberg (1986)
}
\details{
Multiple correlation coefficient R = 0.90

Spread about the function sf = 0.417

sf/Spread about the mean = 0.45

Number of observations = 222

Sum of weights: 187.9

NB in Söderberg (1986), no correction for logarithmic bias was introduced, as, (freely translated) p. 114:
"At the presentation of the functions we lightly touched on the effects of the measurement
errors in the variables which had been included in the regression. It was then concluded
that several factors contrived to that the spread about the functions is overestimated.
Therefore no correction for logarithmic bias was carried out, primarily because the
effect of the errors in the indipendent variables on the spread about the function
are difficult to establish. If one were to assume that the spread about the function
was overestimated by 10 percent, the predicted growth from the functions would be increased by
roughly 2 percent. For 20 percents overestimation, an increase of predicted growth by 4.5 perpcent."
}
