% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soderberg_1986_1992.R
\name{Soderberg_1986_BA_percentual_mortality_per_annum}
\alias{Soderberg_1986_BA_percentual_mortality_per_annum}
\title{Mortality in self-thinned stands}
\source{
From Table 7.2, Söderberg, U. (1986) Funktioner för skogliga produktionsprognoser - Tillväxt och formhöjd för enskilda träd av inhemska trädslag i Sverige. / Functions for forecasting of timber yields - Increment and form height for individual trees of native species in Sweden. Report 14. Section of Forest Mensuration and Management. Swedish University of Agricultural Sciences. Umeå. ISBN 91-576-2634-0. ISSN 0349-2133. pp.251. p. 86.
}
\usage{
Soderberg_1986_BA_percentual_mortality_per_annum(
  SI_species,
  SIH100,
  total_age_stand,
  Basal_area_plot_m2_ha,
  Basal_area_Spruce_m2_ha,
  Basal_area_Deciduous_m2_ha
)
}
\arguments{
\item{SI_species}{Species for which SIH100 was estimated. One of 'Picea abies' or 'Pinus sylvestris'.}

\item{SIH100}{Site Index H100, metres.}

\item{total_age_stand}{Total age of the stand.}

\item{Basal_area_plot_m2_ha}{Basal area of all tree species on the plot, m^2 / ha.}

\item{Basal_area_Spruce_m2_ha}{Basal area Spruce on the plot, m^2 / ha.}

\item{Basal_area_Deciduous_m2_ha}{Basal area Deciduous trees on the plot, m^2 / ha.}
}
\value{
Mortality as annual proportion of basal area.
}
\description{
Material from Unthinned permanent plots.

\strong{N.B. As unthinned plots} present a relatively extreme case, this function is not suggested to be used other than in plots with very high basal area. (p. 86 & p. 90)
}
\details{
Multiple correlation coefficient R = 0.40

Spread about the function sf = 0.0059

sf/Spread about the mean = 92.7

Number of observations = 532.
}
