% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soderberg_1986_BA_self_thinned_stands.R
\name{Soderberg_1986_BA_self_thinned_stands}
\alias{Soderberg_1986_BA_self_thinned_stands}
\title{Basal area in self-thinned stands}
\source{
From Table 7.1, Söderberg, U. (1986) Funktioner för skogliga produktionsprognoser - Tillväxt och formhöjd för enskilda träd av inhemska trädslag i Sverige. / Functions for forecasting of timber yields - Increment and form height for individual trees of native species in Sweden. Report 14. Section of Forest Mensuration and Management. Swedish University of Agricultural Sciences. Umeå. ISBN 91-576-2634-0. ISSN 0349-2133. pp.251. p. 83.
}
\usage{
Soderberg_1986_BA_self_thinned_stands(
  SI_species,
  SIH100,
  total_age_stand,
  BA_quotient_Spruce,
  BA_quotient_Deciduous,
  stems_per_ha
)
}
\arguments{
\item{SI_species}{Species for which SIH100 was estimated. One of 'Picea abies' or 'Pinus sylvestris'.}

\item{SIH100}{Site Index H100, metres.}

\item{total_age_stand}{Total age of the stand.}

\item{BA_quotient_Spruce}{Basal area Spruce / Basal area}

\item{BA_quotient_Deciduous}{Basal area Deciduous trees / Basal area}

\item{stems_per_ha}{Number of stems per hectare.}
}
\value{
Basal area m2/ha.
}
\description{
Material from Unthinned permanent plots.
}
\details{
Multiple correlation coefficient R = 0.87

Spread about the function sf = 0.15

sf/Spread about the mean = 50.2

Number of observations = 532.
}
