% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soderberg_1986_double_bark_Sweden_Beech.R
\name{Soderberg_1986_double_bark_Sweden_Beech}
\alias{Soderberg_1986_double_bark_Sweden_Beech}
\title{Double bark thickness for Beech in Sweden from Söderberg 1986.}
\source{
Söderberg, U. (1986) Funktioner för skogliga produktionsprognoser -
  Tillväxt och formhöjd för enskilda träd av inhemska trädslag i Sverige. /
  Functions for forecasting of timber yields - Increment and form height for
  individual trees of native species in Sweden. Report 14. Section of Forest
  Mensuration and Management. Swedish University of Agricultural Sciences.
  Umeå. ISBN 91-576-2634-0. ISSN 0349-2133. pp.251. p. 251.
}
\usage{
Soderberg_1986_double_bark_Sweden_Beech(
  diameter_cm_under_bark,
  age_at_breast_height,
  latitude
)
}
\arguments{
\item{diameter_cm_under_bark}{Diameter under bark of tree, in cm.}

\item{age_at_breast_height}{Age at breast height of the tree.}

\item{latitude}{Latitude, degrees.}
}
\value{
Double bark thickness, mm.
}
\description{
Double bark thickness for Beech in Sweden from Söderberg 1986.
}
\details{
Multiple correlation coefficient R = 0.88

Spread about the function sf = 0.21

sf/Spread about the mean = 0.41

Number of observations = 291
}
