% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Soderberg_1986_double_bark_Sweden_Betula_pendula.R
\name{Soderberg_1986_double_bark_Sweden_Betula_pendula}
\alias{Soderberg_1986_double_bark_Sweden_Betula_pendula}
\title{Double bark thickness in Sweden for Betula pendula from Söderberg 1986.}
\source{
Söderberg, U. (1986) Funktioner för skogliga produktionsprognoser -
  Tillväxt och formhöjd för enskilda träd av inhemska trädslag i Sverige. /
  Functions for forecasting of timber yields - Increment and form height for
  individual trees of native species in Sweden. Report 14. Section of Forest
  Mensuration and Management. Swedish University of Agricultural Sciences.
  Umeå. ISBN 91-576-2634-0. ISSN 0349-2133. pp.251. p. 248.
}
\usage{
Soderberg_1986_double_bark_Sweden_Betula_pendula(
  diameter_cm_under_bark,
  age_at_breast_height,
  latitude,
  vegetation,
  soil_moisture,
  SI100,
  distance_to_coast_km
)
}
\arguments{
\item{diameter_cm_under_bark}{Diameter under bark of tree, in cm.}

\item{age_at_breast_height}{Age at breast height of the tree.}

\item{latitude}{Latitude, degrees.}

\item{vegetation}{Vegetation type according to follows Swedish National forest inventory FALTSKIKT:
\tabular{ll}{
Code \tab Vegetation \cr
1 \tab  Rich-herb without shrubs \cr
2 \tab Rich-herb with shrubs/bilberry \cr
3 \tab Rich-herb with shrubs/lingonberry \cr
4 \tab Low-herb without shrubs \cr
5 \tab Low-herb with shrubs/bilberry \cr
6 \tab Low-herb with shrubs/lingonberry \cr
7 \tab No field layer \cr
8 \tab Broadleaved grass \cr
9 \tab Thinleaved grass \cr
10 \tab Sedge, high \cr
11 \tab Sedge, low \cr
12 \tab Horsetail, Equisetum ssp. \cr
13 \tab Bilberry \cr
14 \tab Lingonberry \cr
15 \tab Crowberry \cr
16 \tab Poor shrub \cr
17 \tab Lichen, frequent occurrence \cr
18 \tab Lichen, dominating \cr
}}

\item{soil_moisture}{Type 1="Dry/torr",2="Mesic/frisk",3="Mesic-moist/frisk-fuktig",4="Moist/fuktig",5="Wet/Blöt"}

\item{SI100}{Site Index H100, m.}

\item{distance_to_coast_km}{Closest distance to coast, in km, e.g. [forester::coast_distance()]}
}
\value{
Double bark thickness, mm.
}
\description{
Double bark thickness in Sweden for Betula pendula from Söderberg 1986.
}
\details{
Multiple correlation coefficient R = 0.83

Spread about the function sf = 0.32

sf/Spread about the mean = 0.47

Number of observations = 846

NB: \emph{Betula pendula} is in the original source referred as \emph{Betula verrucosa}
}
