% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soderberg_1986_form_quotient_Sweden_Beech.R
\name{Soderberg_1986_form_quotient_Sweden_Beech}
\alias{Soderberg_1986_form_quotient_Sweden_Beech}
\title{Form quotient for Beech, Sweden from Söderberg (1986)}
\source{
Söderberg, U. (1986) Funktioner för skogliga produktionsprognoser -
  Tillväxt och formhöjd för enskilda träd av inhemska trädslag i Sverige. /
  Functions for forecasting of timber yields - Increment and form height for
  individual trees of native species in Sweden. Report 14. Section of Forest
  Mensuration and Management. Swedish University of Agricultural Sciences.
  Umeå. ISBN 91-576-2634-0. ISSN 0349-2133. pp.251. p. 228.
}
\usage{
Soderberg_1986_form_quotient_Sweden_Beech(
  diameter_cm,
  Basal_area_Beech_m2_ha,
  Basal_area_Oak_m2_ha,
  Basal_area_plot_m2_ha,
  SI100_Spruce,
  age_at_breast_height,
  aspect,
  altitude,
  divided_plot = 0
)
}
\arguments{
\item{diameter_cm}{Diameter at breast height}

\item{Basal_area_Beech_m2_ha}{Basal area Beech on the plot, m^2 / ha.}

\item{Basal_area_Oak_m2_ha}{Basal area Oak on the plot, m^2 / ha.}

\item{Basal_area_plot_m2_ha}{Basal area of all tree species the plot, m^2 /
ha.}

\item{SI100_Spruce}{Site Index H100 for Spruce, m.}

\item{age_at_breast_height}{Age at breast height of the tree.}

\item{aspect}{If more than 2:20 / 5\%, one of the following. Otherwise 0.

\tabular{cl}{
1 \tab North  \cr
2 \tab North-East \cr
3 \tab East \cr
4 \tab South-East \cr
5 \tab South \cr
6 \tab South-West \cr
7 \tab West \cr
8 \tab North-West \cr
}}

\item{altitude}{Altitude, meters above sea level.}

\item{divided_plot}{1 for plots described in different parts, which appears
when the original plot consists of different land classes, density classes
or cutting classes or belongs to different owners. 0 for full plots
(default).}
}
\value{
Form quotient, metres.
}
\description{
\strong{OBSERVE:} The function has a typo in the book. Although the variable
description states that it is Altitude^2, the symbol shows LAT^2. I have
interpreted this as that it should be Altitude^2, since the size of the
coefficient is most alike coefficients for ALT^2 in other related functions.

\strong{OBSERVE:} The following interpretation was made of the function - since only
a factor was given for the Site Index in the case that the indicating species was
Picea abies, this was interpreted as mandatory. Therefore, SI_Species and SI100 have
been replaced by SI100_Spruce.
}
\details{
Multiple correlation coefficient R = 0.90

Spread about the function sf = 0.165

sf/Spread about the mean = 0.442

Number of observations = 219


NB in Söderberg (1986), no correction for logarithmic bias was introduced,
as, (freely translated) p. 114: "At the presentation of the functions we
lightly touched on the effects of the measurement errors in the variables
which had been included in the regression. It was then concluded that several
factors contrived to that the spread about the functions is overestimated.
Therefore no correction for logarithmic bias was carried out, primarily
because the effect of the errors in the indipendent variables on the spread
about the function are difficult to establish. If one were to assume that the
spread about the function was overestimated by 10 percent, the predicted
growth from the functions would be increased by roughly 2 percent. For 20
percents overestimation, an increase of predicted growth by 4.5 perpcent."
}
