% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soderberg_1986_form_quotient_Sweden_Oak.R
\name{Soderberg_1986_form_quotient_Sweden_Oak}
\alias{Soderberg_1986_form_quotient_Sweden_Oak}
\title{Form quotient for Oak, Sweden from Söderberg (1986)}
\source{
Söderberg, U. (1986) Funktioner för skogliga produktionsprognoser -
  Tillväxt och formhöjd för enskilda träd av inhemska trädslag i Sverige. /
  Functions for forecasting of timber yields - Increment and form height for
  individual trees of native species in Sweden. Report 14. Section of Forest
  Mensuration and Management. Swedish University of Agricultural Sciences.
  Umeå. ISBN 91-576-2634-0. ISSN 0349-2133. pp.251. p. 230.
}
\usage{
Soderberg_1986_form_quotient_Sweden_Oak(
  diameter_cm,
  diameter_largest_tree_on_plot_cm,
  Basal_area_Beech_m2_ha,
  Basal_area_Birch_m2_ha,
  Basal_area_plot_m2_ha,
  SI_species,
  SI100,
  age_at_breast_height,
  vegetation,
  soil_moisture,
  lateral_water,
  altitude,
  divided_plot = 0,
  county
)
}
\arguments{
\item{diameter_cm}{Diameter at breast height}

\item{Basal_area_Beech_m2_ha}{Basal area Beech on the plot, m^2 / ha.}

\item{Basal_area_Birch_m2_ha}{Basal area Birch on the plot, m^2 / ha.}

\item{Basal_area_plot_m2_ha}{Basal area of all tree species the plot, m^2 /
ha.}

\item{SI_species}{Species for which SIH100 was estimated. One of : 'Picea abies' or 'Pinus sylvestris'.}

\item{SI100}{Site Index H100, m.}

\item{age_at_breast_height}{Age at breast height of the tree.}

\item{vegetation}{Vegetation type according to follows Swedish National forest inventory FALTSKIKT:
\tabular{ll}{
Code \tab Vegetation \cr
1 \tab  Rich-herb without shrubs \cr
2 \tab Rich-herb with shrubs/bilberry \cr
3 \tab Rich-herb with shrubs/lingonberry \cr
4 \tab Low-herb without shrubs \cr
5 \tab Low-herb with shrubs/bilberry \cr
6 \tab Low-herb with shrubs/lingonberry \cr
7 \tab No field layer \cr
8 \tab Broadleaved grass \cr
9 \tab Thinleaved grass \cr
10 \tab Sedge, high \cr
11 \tab Sedge, low \cr
12 \tab Horsetail, Equisetum ssp. \cr
13 \tab Bilberry \cr
14 \tab Lingonberry \cr
15 \tab Crowberry \cr
16 \tab Poor shrub \cr
17 \tab Lichen, frequent occurrence \cr
18 \tab Lichen, dominating \cr
}}

\item{soil_moisture}{Type 1="Dry/torr",2="Mesic/frisk",3="Mesic-moist/frisk-fuktig",4="Moist/fuktig",5="Wet/Blöt"}

\item{lateral_water}{Type 1="Missing", 2="Seldom",3="Shorter periods", 4="Longer periods",5="Slope".}

\item{altitude}{Altitude, meters above sea level.}

\item{divided_plot}{1 for plots described in different parts, which appears
when the original plot consists of different land classes, density classes
or cutting classes or belongs to different owners. 0 for full plots
(default).}

\item{diameter_largest_tree_on_plot}{Diameter at breast height of the largest tree on the plot.}
}
\value{
Form quotient, metres.
}
\description{
\strong{OBSERVE:} The function has a typo in the book. Although the variable
description states that it is Altitude, the symbol shows LAT. I have
interpreted this as that it should be Altitude, since the size of the
coefficient is most alike coefficients for ALT in other related functions.
}
\details{
Multiple correlation coefficient R = 0.86

Spread about the function sf = 0.180

sf/Spread about the mean = 0.516

Number of observations = 292


NB in Söderberg (1986), no correction for logarithmic bias was introduced,
as, (freely translated) p. 114: "At the presentation of the functions we
lightly touched on the effects of the measurement errors in the variables
which had been included in the regression. It was then concluded that several
factors contrived to that the spread about the functions is overestimated.
Therefore no correction for logarithmic bias was carried out, primarily
because the effect of the errors in the indipendent variables on the spread
about the function are difficult to establish. If one were to assume that the
spread about the function was overestimated by 10 percent, the predicted
growth from the functions would be increased by roughly 2 percent. For 20
percents overestimation, an increase of predicted growth by 4.5 percent."
}
