% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Soderberg_1992_double_bark_southern_Sweden_Birch.R
\name{Soderberg_1992_double_bark_southern_Sweden_Birch}
\alias{Soderberg_1992_double_bark_southern_Sweden_Birch}
\title{Double bark thickness in southern Sweden for Birch from Söderberg 1992.}
\source{
Söderberg, U. (1992) Funktioner för skogsindelning: Höjd, formhöjd och barktjocklek för enskilda träd. / Functions for forest management: Height, form height and bark thickness of individual trees. Report 52. Dept. of Forest Survey. Swedish University of Agricultural Sciences. ISSN 0348-0496. p. 67.
}
\usage{
Soderberg_1992_double_bark_southern_Sweden_Birch(
  SI100_Pine,
  diameter_cm,
  diameter_largest_tree_on_plot_cm,
  total_age_stand,
  Basal_area_Pine,
  Basal_area_Birch,
  Basal_area_plot,
  latitude,
  altitude,
  county
)
}
\arguments{
\item{SI100_Pine}{SIH 100 Pine according to Hägglund 1974. e.g. [forester::Hagglund_1974_Sweden_height_trajectories_Pine]}

\item{diameter_cm}{Diameter over bark of tree, in cm.}

\item{diameter_largest_tree_on_plot_cm}{Diameter over bark of the tree with the greatest diameter on the plot, cm.}

\item{total_age_stand}{Total age of the stand.}

\item{Basal_area_Pine}{Basal Area m2 of Scots Pine}

\item{Basal_area_Birch}{Basal Area m2 of Birch}

\item{Basal_area_plot}{Basal area m2 on plot.}

\item{latitude}{Latitude, degrees.}

\item{altitude}{Altitude, metres.}

\item{county}{County name.}
}
\value{
Double bark thickness, mm.
}
\description{
\strong{Applicable counties:}

*Stockholm
*Uppsala
*Västmanland
*Södermanland
*Örebro
*Östergötland
*Skaraborg
*Älvsborg - Västergötlands landskap
*Älvsborg - Dalslands landskap
*Jönköping
*Kronoberg
*Kalmar
*Halland
*Kristianstad
*Malmöhus
*Blekinge
*Gotland

\strong{Regarding the regional division of the models:}

Freely translated from Söderberg 1986 p. 29:
'The regional division is motivated in the following.
For Pine the country has been divided into 3 regions (see figure 4.1),
which are based on the different forms of Scots Pine which have been distinguished
by Sylvén 1917. The distribution of the different forms of Scots Pine is assumed
to depend on the expansion-history of the Pine, according to which the Pine
immigrated both from the north and the south with a transitional zone in central
Sweden, where both forms are present. For other species, the change in genotype
is more continuous over the country (Kiellander 1974). For Norway Spruce, it is
assumed that it has had time for 30-50 generations in northern Sweden, contrasted
against only 10-20 in southern Sweden (Kiellander 1966). Therefore the same regional
division is used for Norway Spruce as for Scots Pine. For the other species, the regional
division is limited to two regions, wherein the northern and central regions have been merged.
}
\details{
Multiple correlation coefficient R = 0.82

Spread about the function sf = 0.354

sf/Spread about the mean = 0.58

Number of observations = 1766
}
