% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soderberg_1992_height_southern_Sweden_Beech.R
\name{Soderberg_1992_height_southern_Sweden_Beech}
\alias{Soderberg_1992_height_southern_Sweden_Beech}
\title{Height function for Beech in southern Sweden from Söderberg 1992.}
\source{
Söderberg, U. (1992) Funktioner för skogsindelning: Höjd, formhöjd och barktjocklek för enskilda träd. / Functions for forest management: Height, form height and bark thickness of individual trees. Report 52. Dept. of Forest Survey. Swedish University of Agricultural Sciences. ISSN 0348-0496. p. 42.
}
\usage{
Soderberg_1992_height_southern_Sweden_Beech(
  diameter_cm,
  total_age_stand,
  Basal_area_Spruce,
  Basal_area_plot,
  latitude,
  altitude,
  divided_plot = 0,
  county
)
}
\arguments{
\item{diameter_cm}{Diameter over bark of tree, in cm.}

\item{total_age_stand}{Total age of the stand.}

\item{Basal_area_Spruce}{Basal Area m2 of Norway Spruce}

\item{Basal_area_plot}{Basal area m2 on plot.}

\item{latitude}{Latitude, degrees.}

\item{altitude}{Altitude, metres.}

\item{divided_plot}{0 for full plots, 1 for divided plots.}

\item{county}{County name.}
}
\value{
Height of tree, m.
}
\description{
\strong{Applicable counties:}

*Stockholm
*Uppsala
*Västmanland
*Södermanland
*Örebro
*Östergötland
*Skaraborg
*Älvsborg - Västergötlands landskap
*Älvsborg - Dalslands landskap
*Jönköping
*Kronoberg
*Kalmar
*Halland
*Kristianstad
*Malmöhus
*Blekinge
*Gotland

\strong{Regarding the regional division of the models:}

Freely translated from Söderberg 1986 p. 29:
'The regional division is motivated in the following.
For Pine the country has been divided into 3 regions (see figure 4.1),
which are based on the different forms of Scots Pine which have been distinguished
by Sylvén 1917. The distribution of the different forms of Scots Pine is assumed
to depend on the expansion-history of the Pine, according to which the Pine
immigrated both from the north and the south with a transitional zone in central
Sweden, where both forms are present. For other species, the change in genotype
is more continuous over the country (Kiellander 1974). For Norway Spruce, it is
assumed that it has had time for 30-50 generations in northern Sweden, contrasted
against only 10-20 in southern Sweden (Kiellander 1966). Therefore the same regional
division is used for Norway Spruce as for Scots Pine. For the other species, the regional
division is limited to two regions, wherein the northern and central regions have been merged.
}
\details{
Multiple correlation coefficient R = 0.92

Spread about the function sf = 0.161

sf/Spread about the mean = 0.41

Number of observations = 220
}
