% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Simple random sampling}
\format{
Um data.frame (tibble) com:
\describe{
\item{Specie}{Nome da espécie}
\item{Sample}{Numero de unidades de amostras.}
\item{Volume}{Volume total em cada unidade amostral.}

...
}
}
\usage{
species
}
\description{
O conjunto de dados contendo amostragem de duas espécies.
}
\keyword{datasets}
