% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species2}
\alias{species2}
\title{Multiple Species}
\format{
A data frame with 54 rows and 3 variables:
\describe{
\item{Specie}{Factor. Wood species.}
\item{Sample}{sample unit, em metros quadrados}
\item{Volume}{Volume, em metros cúbicos}

...
}
}
\source{
Simulated data
}
\usage{
species2
}
\description{
A dataset containing the volumes for multiple species.
}
\keyword{datasets}
