% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Splitting Data Into Train And Test Set}
\usage{
split_data(data, target, type, ratio = 0.8)
}
\arguments{
\item{data}{data.frame, matrix, data.table or dgCMatrix - training data set to create model, if data_test = NULL, then data will be
automatically divided into training and testing dataset. NOTE: data has to contain the target column.}

\item{target}{character: name of the target column, should be character and has to be column name in data.}

\item{type}{character: defining the task. Two options are: "regression" and "classification", particularly, binary classification.}

\item{ratio}{numeric, ranged from between 0 and 1, indicating the proportion of data train over original data set.}
}
\value{
A list of data train and data test.
}
\description{
\code{split_data} divide the data frame with given target and type into 
data train and data test set with adjustable ratio.
}
