% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spruce_Growth.R
\docType{data}
\name{Spruce_Growth}
\alias{Spruce_Growth}
\title{Growth laws of normal spruce stands, 1904}
\format{
A data frame with columns:
\describe{
\item{Age}{A value between 10-130 with 10 yr increments}
\item{Height}{Height of the crop, metres}
\item{Height increment}{Increment since former yr}
\item{Bonitet}{Site productivity class}
}
}
\source{
*Wuchsgesetze normaler Fichtenbestände* Mitth. aus d. forstl.
 Vers.-wesen Österreichs 1904. XXIX h. sid. 1 1 o. f.
 in Wahlgren (1922) *Skogsskötsel: handledning vid uppdragande, vård och fo̊ryngring av skog*,p. 494, 2nd Ed.
 Kungl. Hofboktr. Iduns tryckeri-A.-B. Stockholm, Sweden. Swedish.
}
\usage{
Spruce_Growth
}
\description{
Translated from Wahlgren (1922): A. Schiffel summarised the German experience tables from **Swappach**, Lorey & Baur** such that the height development on differing site classes are as follows.
As can be seen from the table, the height growth of the spruce and other tree species culminates earlier on the richer sites than the poorer.

Note: What is referred to as height here is likely mean height of the stand.
}
\keyword{datasets}
