% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS.R
\name{SRS}
\alias{SRS}
\title{Stratified Random Sampling}
\usage{
SRS(data = NULL, x, strata, A, a, SA = "PA", FP = FALSE, DT = FALSE, ...)
}
\arguments{
\item{data}{Conjunto de dados que contém as informações de inventário florestal.}

\item{x}{Um vetor numérico com a variável de interesse.}

\item{strata}{Fator de estratificação da característica de interesse.}

\item{A}{Área total da população, em hectare.}

\item{a}{Área da unidade de amostra, em hectare.}

\item{SA}{Método de alocação das unidades de amostras nos estratos.}

\item{FP}{Lógico. Fator de proporcionalidade para transformar para hectare. (default = FALSE)}

\item{DT}{Lógico. Imprime a saída no painel de visualização. (default = FALSE)}

\item{...}{Parâmetros adicionais}
}
\value{
Uma lista com estatística descritiva, anova, estimativas de amostragem aleatória estratificada e informações de base para cálculo das estimativas.
}
\description{
Stratified Random Sampling
}
\examples{
\dontrun{
data(native)
SRS(x=native$Volume, strata=native$Strata, A = c(650, 350), a = 1, SA = "PA")
}

}
