% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralHelpers.R
\name{stand_basal_area_QMD}
\alias{stand_basal_area_QMD}
\title{Stand Basal Area from Quadratic mean diameter and stems per hectare.}
\source{
Curtis, R.O., Marshall, D.D. 2000. Technical Note: Why Quadratic Mean Diameter? West. J. Appl. For. 15(3):137-139. Available: \url{https://www.fs.fed.us/pnw/olympia/silv/publications/opt/436_CurtisMarshall2000.pdf}
}
\usage{
stand_basal_area_QMD(stems_per_ha, QMD)
}
\arguments{
\item{stems_per_ha}{Stems per hectare.}

\item{QMD}{Quadratic mean diameter, cm.}
}
\value{
Basal area, m2/ha.
}
\description{
Stand Basal Area from Quadratic mean diameter and stems per hectare.
}
