% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standData.R
\name{standData}
\alias{standData}
\title{Create a standData object.}
\usage{
standData(siteData, treelistData, weatherData)
}
\arguments{
\item{siteData}{a siteData object.}

\item{treelistData}{a treelist object.}

\item{weatherData}{a weatherData object.}
}
\value{
standData object.
}
\description{
Create a standData object.
}
