% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Hagglund_Lundmark_1979_Site_Index_Stand_Factors.R
\name{SWE_NFI_FieldLayer_from_Site}
\alias{SWE_NFI_FieldLayer_from_Site}
\title{Determine swedish field layer from site index and soil moisture}
\source{
https://www.heurekaslu.se/wiki/VegetationType_from_site_features
}
\usage{
SWE_NFI_FieldLayer_from_Site(H100_Spruce, soilMoisture)
}
\arguments{
\item{H100_Spruce}{Site Index for Norway Spruce}

\item{soilMoisture}{integer 1-5 see
[forester::Sweden_soil_types(type="moisture")]}
}
\value{
Swedish NFI Field Layer Code
}
\description{
A helper function in the case that Swedish NFI Field Layer is unknown.
TODO: Check if in fact H100_Spruce, or H100_Pine.
}
\examples{
SWE_NFI_FieldLayer_from_Site(15,1)
}
