% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweden_local_climates.R
\docType{data}
\name{sweden_local_climates}
\alias{sweden_local_climates}
\title{Swedish Local Climate Shape Files}
\format{A large SpatialPolygonsDataFrame}
\source{
\url{https://www.heurekaslu.se/wiki/Definition:ClimateCode}
}
\usage{
sweden_local_climates
}
\description{
Contains polygons corresponding to Swedish local climates by Ångström (1958).
}
\keyword{datasets}
