% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swedish_NFI_coding.R
\name{Sweden_soil_types}
\alias{Sweden_soil_types}
\title{Coding and description of swedish categorical soil parameters.}
\usage{
Sweden_soil_types(type = "texture")
}
\arguments{
\item{type}{One of "texture","moisture" or "depth"}
}
\value{
A tibble.
}
\description{
Coding and description of swedish categorical soil parameters.
}
\examples{
Sweden_soil_types()
}
