% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swedish_NFI_coding.R
\name{Sweden_vegetation_types}
\alias{Sweden_vegetation_types}
\title{Swedish vegetation types (NFI).}
\usage{
Sweden_vegetation_types(layer = "field")
}
\arguments{
\item{layer}{One of "field" or "ground".}
}
\value{
A tibble.
}
\description{
Swedish vegetation types (NFI).
}
\examples{
Sweden_vegetation_types()
}
