% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tamminen_1993_SI_prediction_Finland.R
\name{tamminen_1993}
\alias{tamminen_1993}
\alias{Tamminen_1993_SI_prediction_southern_Finland_Norway_Spruce}
\alias{Tamminen_1993_SI_prediction_southern_Finland_Scots_Pine}
\title{SI prediction for Norway Spruce and Scots Pine in southern Finland.}
\source{
Tamminen, P. 1993. Pituusboniteetin ennustaminen kasvupaikan ominaisuuksien avulla
Etelä-Suomen kangasmetsissä. Summary: Estimation of site index for Scots Pine and Norway
spruce stands in South Finland using site properties. Folia Forestalia 819. 26 pp. Available
(16/03/2022): \url{https://jukuri.luke.fi/handle/10024/522402}
}
\usage{
Tamminen_1993_SI_prediction_southern_Finland_Norway_Spruce(
  vegetation,
  temperature_sum,
  fine_fraction_lower_than_15_percent = FALSE,
  clay_content_higher_than_10_percent = FALSE
)

Tamminen_1993_SI_prediction_southern_Finland_Scots_Pine(
  vegetation,
  temperature_sum,
  humus_thickness_cm,
  rod_penetration_cm,
  shallow_soil = FALSE
)
}
\arguments{
\item{vegetation}{One of "OMT"<"MT"<"VT"<"CT". For Spruce one of: ("LH" equal growth as "OMT")<"MT"<"VT"}

\item{temperature_sum}{Effective temperature sum (>5 C), e.g. Ojansuu & Henttonen 1983.}

\item{fine_fraction_lower_than_15_percent}{Boolean. TRUE if fine fraction of soil is lower than than 15 percent, otherwise FALSE (default)}

\item{clay_content_higher_than_10_percent}{Boolean. TRUE if clay content of soil is higher  than 10 percent, otherwise FALSE (default)}

\item{humus_thickness_cm}{Thickness of humus layer, in cm.}

\item{rod_penetration_cm}{Penetration of stone rod, e.g. Viro 1952.}

\item{shallow_soil}{Boolean. TRUE if shallow soil (<30 cm to bedrock) else FALSE (Default).}
}
\description{
SI prediction for Norway Spruce and Scots Pine in southern Finland.
}
