% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Tegnhammar_1992_adjusted_SI_by_stand_variables.R
\name{Tegnhammar_1992_adjusted_SI_by_stand_variables}
\alias{Tegnhammar_1992_adjusted_SI_by_stand_variables}
\title{Calculate Tegnhammar's adjusted Site Index from stand variables}
\source{
Tegnhammar, L. (1992) "Om skattningen av ståndortsindex för gran",
eng:"On the estimation of site index for Norway spruce". Report 53. Dept. of
Forest Survey. Swedish University of Agricultural Sciences. Umeå. p.80.
ISSN 0348-0496
}
\usage{
Tegnhammar_1992_adjusted_SI_by_stand_variables(
  species,
  latitude,
  longitude,
  altitude,
  vegetation,
  ground_layer,
  aspect_main,
  incline_percent,
  soil_moisture,
  soil_depth,
  soil_texture,
  humidity,
  ditched,
  lateral_water,
  peat_humification = "Medium",
  epsg = "EPSG:4326"
)
}
\arguments{
\item{latitude}{decimal degrees. By default WGS84, see parameter 'epsg'.}

\item{longitude}{decimal degrees. By default WGS84, see parameter 'epsg'.}

\item{altitude}{metres above sea level}

\item{vegetation}{Vegetation type according to follows Swedish National forest inventory FALTSKIKT:
\tabular{ll}{
Code \tab Vegetation \cr
1 \tab  Rich-herb without shrubs \cr
2 \tab Rich-herb with shrubs/bilberry \cr
3 \tab Rich-herb with shrubs/lingonberry \cr
4 \tab Low-herb without shrubs \cr
5 \tab Low-herb with shrubs/bilberry \cr
6 \tab Low-herb with shrubs/lingonberry \cr
7 \tab No field layer \cr
8 \tab Broadleaved grass \cr
9 \tab Thinleaved grass \cr
10 \tab Sedge, high \cr
11 \tab Sedge, low \cr
12 \tab Horsetail, Equisetum ssp. \cr
13 \tab Bilberry \cr
14 \tab Lingonberry \cr
15 \tab Crowberry \cr
16 \tab Poor shrub \cr
17 \tab Lichen, frequent occurrence \cr
18 \tab Lichen, dominating \cr
}}

\item{ground_layer}{According to Swedish NFI, roughly:

  \tabular{ll}{
Code \tab Ground layer \cr
1 \tab Lichen type (>50\% of existing ground layer)\cr
2 \tab Lichen-rich bogmoss type (>25\% lichen + >50\% Sphagnum of existing ground layer)\cr
3 \tab Lichen rich (>25\% lichen, & not >50\% Sphagnum of existing ground layer) \cr
4 \tab Bogmoss type (\emph{Sphagnum} > 50\% of existing ground layer) \cr
5 \tab Swamp moss type (\emph{Polytrichum commune, P. gracile, P. strictum, Sphagnum, Depranocladus, Scorpidium, Paludella, Calliergon, Tomentypnum, Campylium.}) \cr
6 \tab Fresh moss type (\emph{Hylocomium splendens, Ptilium crista-castrensis, Dicranum ssp.}) \cr
}}

\item{soil_moisture}{Type 1="Dry/torr",2="Mesic/frisk",3="Mesic-moist/frisk-fuktig",4="Moist/fuktig",5="Wet/Blöt"}

\item{soil_depth}{Type 1="Deep, >70cm",2="Rather shallow, 20-70cm",3="Shallow <20cm", 4= "Varying".}

\item{soil_texture}{\tabular{ll}{
Code \tab Short \cr
1 \tab Boulder \cr
2 \tab Gravel \cr
3 \tab Coarse sand \cr
4 \tab Medium sand \cr
5 \tab Fine sand \cr
6 \tab Coarse silt \cr
7 \tab Fine silt \cr
8 \tab Clay \cr
9 \tab Peat \cr
}}

\item{humidity}{Humidity during the \emph{vegetation period}. mm. If not provided, will be calculated from latitude, longitude with [forester::Eriksson_1986_humidity()] from Eriksson, B. (1986) fig.34. Available online (26/10/2021) \url{https://www.smhi.se/polopoly_fs/1.166663!/RMK_46\%20Nederbörds-\%20och\%20humiditetsklimat\%20i\%20Sverige\%20under\%20vegetationsperioden..pdf}}

\item{ditched}{TRUE/FALSE if affected by ditching.}

\item{lateral_water}{Type 1="Missing",Type 2="Seldom",3="shorter periods",4="longer periods", 5="slope".}

\item{peat_humification}{"Low", if organic remnants clearly visible: squeezing sample gives somewhat cloudy water; "Medium", if with some difficulty organic remnants can be distinguished, if sample is squeezed the water is cloudy; or "High", if no organic remnants can be distinguished. Water and humus cannot be separated by squeezing. Humus is porridge-like.}

\item{epsg}{Default is WGS84 'EPSG:4326'. For SWEREF-99TM - 'EPSG:3006'.}

\item{aspect}{If more than 1:20 (5\%), one of the following, otherwise 0.

 \tabular{cl}{
1 \tab North \cr
2 \tab North-East \cr
3 \tab East \cr
4 \tab South-East \cr
5 \tab South \cr
6 \tab South-West \cr
7 \tab West \cr
8 \tab North-West \cr
}}
}
\value{
Tegnhammars adjusted SI, meters.
}
\description{
Stand variables were regressed to estimate Tegnhammar's adjusted site index SIHjust.

 \strong{SIHjust} corrects height development functions by Hägglund
 (1972, 1973) for the 'apparent' trend, an age bias, wherein the functions
 would indicate higher SIH in young forests as compared to older forests,
 even if the site index is the same.

 Tegnhammar's dissertation investigated a trend in the site index
 curves for spruce in Sweden, possibly caused by stem sectioning of elderly
 trees that were suppressed in their youth; young planted forests are
 composed of better plant material; young forests have recieved a better
 start as a result of methods introduced during the 1920-1950's; older stands
 which were not suitable for SI estimation by height were measured in such
 a manner nonetheless; The risk of including hidden top breakages is larger
 for older trees; & the productivity has increased with time due to nitrogen
 deposition - younger trees have been affected for a longer period of their
 lifetime.
}
\details{
N.B. Not for use on Öland / Gotland, or soil moisture types 'Wet'/'Blöt'.
No underlying material for these sites. The material was too small to determine the exact size of the error, but indicate that estimates are much too high for these sites.

s = 30.71662
R^2  = 0.7045
}
