% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testing_data.R
\docType{data}
\name{testing_data}
\alias{testing_data}
\title{Testing dataset}
\format{
A data frame with 1000 rows and 12 variables
}
\usage{
data(testing_data)
}
\description{
\code{testing_data} is an artificially generated dataset, based on
Friedman regression problem,that was created for testing preprocessing process on
ugly data with missing values, outliers, and so on.
}
