% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_theme_forester.R
\name{theme_forester}
\alias{theme_forester}
\title{Forester Theme for ggplot2 objects}
\usage{
theme_forester()
}
\value{
theme for ggplot2 objects
}
\description{
Forester Theme for ggplot2 objects
}
