% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_response.R
\name{thinning_response_Hynynen_2019}
\alias{thinning_response_Hynynen_2019}
\title{Thinning response function}
\usage{
thinning_response_Hynynen_2019(
  BasalArea,
  height_weighted_w_ba,
  REM_percent = 33,
  management = 1,
  isOMT,
  degreedays,
  CUT0.5 = F,
  CUT5.10 = F,
  CUT10.15 = F
)
}
\arguments{
\item{BasalArea}{stand basal area, m^2ha^-1}

\item{height_weighted_w_ba}{mean height weighted with basal area, m.}

\item{REM_percent}{removed percentage in cutting, percent of stand basal area.}

\item{management}{1 if evenaged, 2 if unevenaged}

\item{isOMT}{Binary, if site type is Oxalis-Myrtillus (mesic) )== 1}

\item{degreedays}{Effective temperature sum, threshold +5°C}

\item{CUT0.5}{True or False, if between 0-5 yr since thinning}

\item{CUT5.10}{True or False, if between 5-10 yr since thinning}

\item{CUT10.15}{True or False, if between 10-15 yr since thinning}
}
\value{
Mean annual stand basal area growth during 5-year growth period, m^2ha^{-1}year^{-1}
}
\description{
After table 3 in Hynynen, J., Eerikäinen, K., Mäkinen, H., Valkonen, S. (2019) Growth response to cuttings in Norway Spruce stands under even-aged and uneven-aged management.

Available at \url{https://www.sciencedirect.com/science/article/pii/S0378112718318838}


\strong{N.B. : Returns response in period, i.e. 0-5, 5-10, 10-15}
}
\examples{
thinning_response_Hynonen_2019(BasalArea=30, height_weighted_w_ba=15, isOMT=F, degreedays=1577, CUT10.15=T)
}
