% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thor_2005_Heterobasidion_rate.R
\name{Thor_2005_Heterobasidion_Decay_Probability_Individual_trees}
\alias{Thor_2005_Heterobasidion_Decay_Probability_Individual_trees}
\title{Probability of incidence of Heterobasidion Root Rot in individual trees of Norway Spruce.}
\source{
Thor, M. 2005. Heterobasidion Root Rot in Norway Spruce: Modelling Incidence, Control Efficacy and Economic Consequences in Swedish Forestry. Diss. Swedish University of Agricultural Sciences. Uppsala. p. 22. URL: https://pub.epsilon.slu.se/729/1/Avh.nr.5_.pdf
}
\usage{
Thor_2005_Heterobasidion_Decay_Probability_Individual_trees(
  Latitude,
  Longitude,
  Altitude,
  TotalAge,
  DBHcm,
  SIS,
  soilMoisture,
  soilTexture,
  BA_percent_NorwaySpruce,
  TsumCorrection = FALSE,
  TmaxMonthly = NULL,
  TminMonthly = NULL,
  Region = "auto",
  AtStumpHeight = FALSE
)
}
\arguments{
\item{Latitude}{Decimal Degrees WGS84.}

\item{Longitude}{Decimal Degrees WGS84.}

\item{Altitude}{Meters above sea level}

\item{TotalAge}{Total age, years.}

\item{DBHcm}{Diameter at Breast height, 1.3 meters.}

\item{SIS}{Site Index Prediction according to [forester::Hagglund_Lundmark_1979_Site_Index_Stand_Factors()]}

\item{soilMoisture}{Integer 1-5 [forester::Sweden_soil_types('moisture')]}

\item{soilTexture}{Integer 1-9 [forester::Sweden_soil_types('texture')]}

\item{BA_percent_NorwaySpruce}{Percent of basal area which is Norway Spruce in stand.}

\item{TsumCorrection}{Apply continental/maritime correction to Tsum? See [forester::Moren_Perttu_1994_Sweden_temperature_sum_5C()]}

\item{TmaxMonthly}{Required for TsumCorrection. Else NULL.}

\item{TminMonthly}{Required for TsumCorrection. Else NULL.}

\item{Region}{One of 'auto' for automatic detection from latitude, or 'national'}

\item{AtStumpHeight}{Boolean. If True, gives probability at stump. If FALSE (Default), provides probability of incidence at breast height 1.3 m.}
}
\value{
Numeric. Probability of incidence.
}
\description{
Probability of incidence of Heterobasidion Root Rot in individual trees of Norway Spruce.
}
