% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_reach_breast_height.R
\name{time_to_reach_breast_height}
\alias{time_to_reach_breast_height}
\title{Time to reach breast height}
\usage{
time_to_reach_breast_height(main_species, SIS, latitude)
}
\arguments{
\item{main_species}{Main species, Spruce or Pine.}

\item{SIS}{Site Index according to Site factors.}

\item{latitude}{Decimal degrees.}
}
\value{
Time to reach breast height, years.
}
\description{
Number of years for basal area mean trees to reach breast height, 1.3metres, from a smoothed function to estimate a table by NFI.
For Oak (Quercus robur) and Beech (Fagus sylvatica), spruce will be used. For other broadleaves, e.g. (birch, aspen), time to breast height is computed 150/SIS.
}
\examples{
time_to_reach_breast_height(main_species="Spruce",SIS=14,latitude=57)
}
