% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train models with forester}
\usage{
train(
  data,
  y,
  type = "guess",
  engine = c("ranger", "xgboost", "decision_tree", "lightgbm", "catboost"),
  loss = "default",
  validation = "default",
  tuning = "default",
  keep = FALSE,
  verbose = TRUE,
  bayes_iter = 10,
  random_iter = 10,
  advanced_preprocessing = FALSE
)
}
\arguments{
\item{data}{A `data.frame` or `matrix` - data which will be
used to build models. By default model will be trained
on all columns in the `data`.}

\item{y}{A target variable. It can be either
(1) a vector of the same number of observations as `data` or
(2) a character name of variable in the `data` that contains
the target variable.}

\item{type}{A character, one of `classification`/`regression`/`guess` that
sets the type of the task. If `guess` (the default option) then
forester will figure out `type` based on the number of unique values
in the `y` variable.}

\item{engine}{A vector of tree-based models that shall be tested.
Possible values are: `ranger`, `xgboost`, `decision_tree`, `lightgbm`, `catboost`.
All models from this vector will be trained and the best one will be returned.}

\item{loss}{A character with the name of the loss function.
If `default` then the default loss function for a given task will be used.}

\item{validation}{A validation/model selection a character of
data.frame. If `data.frame` then this data will be used for the model
assessment and selection. If `character`, then it will be a name of
the hyperparameter.}

\item{tuning}{A character with the name of the tuning procedure.
one out of `default`/`portfolio`/`bayesian`/`random search`.
If `default` then no tuning, just default hyperparameteres are used.
If `portfolio` then static portfolio of hyperparametes are tried.
If `bayesian` then Bayesian optimization is performed.
If `random search` then random search is performed.}

\item{keep}{Shall all models be returned (`keep = FALSE`) or only the best one
(`keep = FALSE`, default option)?}

\item{verbose}{A logical value, if set to TRUE, provides all information about
training process, if FALSE gives none.}

\item{bayes_iter}{An integer value describing number of optimization rounds
used by Bayesian optimization.}

\item{random_iter}{An integer value describing number of optimization rounds
used by random search.}

\item{advanced_preprocessing}{A logical value describing, whether the user wants to use
advanced preprocessing methods (ex. deleting correlated values).}
}
\value{
A list of all necessary objects for other functions.
}
\description{
The `train()` function is the core function of this package.
The only obligatory arguments are `data` and `target`.
Setting and changing other arguments will affect model
validation strategy, tested model families, and so on.
}
\examples{
library(forester)
data('lisbon')
train_output <- train(lisbon, 'Price')
train_output$ranked_list
}
