% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{train_models}
\alias{train_models}
\title{Train models from given engines}
\usage{
train_models(data, y, engine, type)
}
\arguments{
\item{data}{A training data for models created by `prepare_data()` function.}

\item{y}{A string that indicates a target column name.}

\item{engine}{A vector of tree-based models that shall be created. Possible
values are: `ranger`, `xgboost`,`decision_tree`, `lightgbm`, `catboost`.}

\item{type}{A string that determines if Machine Learning task is the
`binary_clf` or `regression`.}
}
\value{
A list of models for every engine.
}
\description{
Train models from given engines
}
