% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_balance.R
\name{train_test_balance}
\alias{train_test_balance}
\title{Balancess and splits dataset}
\usage{
train_test_balance(data, y, type, balance = TRUE, fractions = c(0.6, 0.2, 0.2))
}
\arguments{
\item{data}{A data source, that is one of major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string which indicates a target column name.}

\item{type}{A string which determines if machine learning task is the
`classification` or `regression`.}

\item{balance}{A logical value, determines if we want to balance the dataset.}

\item{fractions}{A vector with 3 numeric values that sums to 1 which
determine sizes of train, test, validation datasets. DEFAULT: c(0.6, 0.2, 0.2).}
}
\value{
A list of train, test and validation datasets.
}
\description{
Balancess and splits dataset
}
\examples{
data(lisbon)
b_lisbon <- train_test_balance(lisbon,'Price',balance = FALSE,'regression',
fractions = c(train = 0.6, valid = 0.2, test = 0.2))
}
