% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_swe_function.R
\name{tree.volume}
\alias{tree.volume}
\title{Volume estimation of trees}
\usage{
tree.volume(
  dbh.cm,
  height.m,
  crown.base.height.m,
  double.bark.mm,
  species.latin,
  latitude
)
}
\arguments{
\item{dbh.cm}{Diameter breast height \strong{(1.3 m)}, in cm.}

\item{height.m}{Tree height in metres.}

\item{crown.base.height.m}{The lowest height of the living tree crown, in metres.}

\item{double.bark.mm}{The sum of bark thickness on polar ends of the diameter as measured at 1.3m height, in millimetres.}

\item{species.latin}{One of "Picea ...", "Pinus sylvestris", "Larix sibirica", "Abies ...".}

\item{latitude}{The latitudinal placement of the tree, in degrees N.}
}
\value{
Volume of the tree, in \strong{dm^3}
}
\description{
Provides interface for several functions from the collection of individual tree volume functions in  "\emph{Fältdatasystem för skogliga fältförsökö}" (2012),
by Karlsson, K., Ekö, Mossberg, M., Ulvcrona, T.

    \strong{If used on array, must use dplyr::rowwise()}

    \strong{Supported species}: Picea spp. ; Abies spp. ;  Pinus sylvestris ; Larix sibirica; Pinus contorta; Picea abies

    \strong{If diameter double bark is available}
    Will use a function following:

    If Pinus sylvestris, larger than 4.5 cm dbh: \strong{Brandel, 1990: function 100-04}

    If Pinus sylvestris, below 4.5 cm dbh :\strong{Andersson, S-O, 1954}

    If Picea abies (Picea spp., Abies spp.), larger than 4.5 cm dbh : \strong{Brandel. 1990: function 100-02}

    If Picea abies (Picea spp., Abies spp.), smaller than 4.5 cm dbh: \strong{Andersson, S-O, 1954}

    If European/Siberian larch... >= 5 cm dbh: \strong{Carbonnier, 1954}

    If Pinus contorta: \strong{Eriksson, 1973}


    \strong{If diameter double bark is Not available}
    Will use a function following:

    If Pinus...  larger than 4.5 cm dbh, \strong{Brandel 1990: 100-01}

    If Pinus... below 4.5 cm dbh, \strong{Andersson, S-O, 1954}

    If Picea/Abies... larger than 4.5 cm dhb, \strong{Brandel 1990: 100-01}

    If Picea/Abies... below 4.5 cm dbh, \strong{Andersson, S-O, 1954}

    If Larix sibirica... >=5 cm dbh: \strong{Carbonnier, 1954}.B
}
\examples{
tree.volume(species.latin=SPP, height.m=height, dbh.cm=dbh, crown.base.height.m=cbhm, double.bark.mm=dbmm, latitude=latitude)
}
