% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_type.R
\name{tree_type}
\alias{tree_type}
\title{Is a tree species deciduous or coniferous?}
\usage{
tree_type(species)
}
\arguments{
\item{species}{Latin name}
}
\value{
character: "Deciduous" or "Coniferous"
}
\description{
This will try to grab the genus name of a given tree species (in latin). The first letter of family name MUST be capital.
It will try to match this name against a group of well known Genus and respond with if they are coniferous, or if not, deciduous. If none is true it will return NA and throw a warning.
}
\examples{
tree_type("Picea abies") # Inserting the latin name for Norway spruce will return "Coniferous".
tree_type("Pisum sativum") # Inserting the latin name for 'pea' will throw an error and return NA.
}
