% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeInfo.R
\name{treeInfo}
\alias{treeInfo}
\title{Construct a tree.}
\usage{
treeInfo(
  tree_ID,
  year,
  species,
  age,
  height.m,
  diameter.cm,
  bark.mm,
  double_bark.mm,
  crown_base_height.m,
  coord_x,
  coord_y
)
}
\arguments{
\item{tree_ID}{Unique tree identifier}

\item{year}{Year of measurements}

\item{species}{Species, latin.}

\item{age}{Age, years.}

\item{height.m}{Height, metres.}

\item{diameter.cm}{Diameter at breast height, cm.}

\item{bark.mm}{Bark thickness, mm.}

\item{double_bark.mm}{Double bark thickness, mm. If not provided, 2x bark.mm}

\item{coord_x}{X-coordinate}

\item{coord_y}{Y-coordinate}

\item{crown.base.height.m}{Height from ground to base of tree crown.}
}
\value{

}
\description{
Initiates a treeInfo object, containing information about a single tree.
}
