% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treelistData.R
\name{treelistData}
\alias{treelistData}
\title{Create a treelistData object.}
\usage{
treelistData(
  treeID,
  origin_year,
  observation_year,
  species,
  age,
  height.m,
  diameter.cm,
  bark.mm,
  double_bark.mm,
  crown.base.height.m,
  coordx,
  coordy
)
}
\arguments{
\item{treeID}{The unique ID of a tree.}

\item{origin_year}{Year of planting.}

\item{observation_year}{Year of observation.}

\item{species}{Tree species}

\item{age}{Tree total age}

\item{height.m}{Tree height in metres}

\item{diameter.cm}{Tree diameter in cm.}

\item{bark.mm}{Tree bark thickness, in mm.}

\item{double_bark.mm}{Tree double bark thickness, in mm.}

\item{crown.base.height.m}{Height to lowest living green branch on a tree.}

\item{coordx}{X coordinate}

\item{coordy}{Y coordinate}

\item{generation}{Does this tree belong to the current generation of management? 1== current, 2== previous, 3..}

\item{basal_area}{Basal area of tree in m2.}

\item{volume}{Tree volume, m3sk.}
}
\value{

}
\description{
A dataframe containing the treelist for a stand.
}
