% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuhus_1995_natural_mortality_Norway.R
\name{Tuhus_1995_natural_mortality_Norway_Norway_Spruce}
\alias{Tuhus_1995_natural_mortality_Norway_Norway_Spruce}
\alias{Tuhus_1995_natural_mortality_Norway_Pine}
\alias{Tuhus_1995_natural_mortality_Norway_Birch_Aspen}
\alias{Tuhus_1995_natural_mortality_Norway_Broadleaf}
\title{Natural Mortality Rate for Individual Trees in Norway.}
\source{
Tuhus, E. 1995. Naturlig avgang av trær. (Natural mortality of trees) Rapp. Skogforsk. 6/97: 1-28. ISBN 82-7169-836-2. ISSN 0803-2858.
}
\usage{
Tuhus_1995_natural_mortality_Norway_Norway_Spruce(
  diameter_cm,
  BA_m2_ha,
  BA_m2_ha_Spruce,
  period_length
)

Tuhus_1995_natural_mortality_Norway_Pine(
  diameter_cm,
  BA_m2_ha,
  dominant_height_m,
  period_length
)

Tuhus_1995_natural_mortality_Norway_Birch_Aspen(
  diameter_cm,
  BA_m2_ha,
  stand_age,
  period_length
)

Tuhus_1995_natural_mortality_Norway_Broadleaf(
  diameter_cm,
  BA_m2_ha,
  stand_age,
  period_length
)
}
\arguments{
\item{diameter_cm}{Diameter of the tree at breast height (1.3 m)}

\item{BA_m2_ha}{Basal Area m2/ha.}

\item{BA_m2_ha_Spruce}{Basal Area m2/ha which is Spruce.}

\item{period_length}{Length of the period.}

\item{dominant_height_m}{Dominant height of the stand in metres.}

\item{stand_age}{Age of the stand (at breast height?)}
}
\value{
Mortality Rate.
}
\description{
Natural Mortality Rate for Individual Trees in Norway.
}
\details{
Calculates the natural mortality rate for Norway Spruce,
Scots Pine, Betula spp., European Aspen and other Broadleaves in Norway.
Based on 1771 permanent plots of size 100 m^2.
}
