% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Tveite_1967_Loreys_mean_height_Norway_Norway_Spruce.R
\name{Tveite_1967_Loreys_mean_height_Norway_Norway_Spruce}
\alias{Tveite_1967_Loreys_mean_height_Norway_Norway_Spruce}
\title{Lorey's mean height for Picea abies from Tveite 1967.}
\source{
Tveite, B. 1967. Sambandet mellom grunnflateveid middelhøyde (HL) og noen andre bastandshøyder i gran- og furuskog. Meddr. norske SkogforsVes. 22:483-538. p. 512, table 8. Function 8.22.
}
\usage{
Tveite_1967_Loreys_mean_height_Norway_Norway_Spruce(
  dominant_height,
  stems_per_ha,
  Basal_area_m2_ha,
  quadratic_mean_diameter
)
}
\arguments{
\item{dominant_height}{dominant height}

\item{stems_per_ha}{Stems per hectare}

\item{Basal_area_m2_ha}{Basal Area m2/ha.}

\item{quadratic_mean_diameter}{Quadratic Mean Diameter at start. e.g. [forester::quadratic_mean_diameter()]}
}
\value{
Lorey's Mean height for Spruce stands in Norway.
}
\description{
Lorey's mean height for Picea abies from Tveite 1967.
}
