% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tveite_1967_Loreys_mean_height_Norway_Pine.R
\name{Tveite_1967_Loreys_mean_height_Norway_Pine}
\alias{Tveite_1967_Loreys_mean_height_Norway_Pine}
\title{Lorey's mean height for stands of Scots Pine in Norway from Tveite 1967.}
\source{
Tveite, B. 1967. Sambandet mellom grunnflateveid middelhøyde (HL) og noen andra bestandshøyder i gran- og furuskog. Meddr norske SkogforsVes. 22:483-538. p. 513. table 9. function 9.18. in Braastad, Helge 1977. Tilvekstmodellprogram for furu. Growth model computer program for Pinus sylvestris. Meddr. Norsk inst. skogforsk 35:265-359. p 273 function (2)
}
\usage{
Tveite_1967_Loreys_mean_height_Norway_Pine(
  dominant_height,
  stems_per_ha,
  basal_area_m2_ha,
  diameter_mean_basal_area_stem
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{stems_per_ha}{Number of stems per hectare}

\item{basal_area_m2_ha}{Basal area (m^2/ha)}

\item{diameter_mean_basal_area_stem}{Diameter corresponding to the mean basal area stem.}
}
\value{
Lorey's mean height for stands of Scots Pine in Norway.
}
\description{
Lorey's mean height for stands of Scots Pine in Norway from Tveite 1967.
}
