% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tveite_1976_height_trajectory_Norway_Pine.R
\name{Tveite_1976_height_trajectory_Norway_Pine}
\alias{Tveite_1976_height_trajectory_Norway_Pine}
\title{Height trajectories for Scots Pine in Norway from Tveite 1976}
\source{
Tveite, B. 1976. Bonitetskurver for furu. Unpublished. in: Braastad, Helge 1977. Tilvekstmodellprogram for furu. Growth model computer program for Pinus sylvestris. Meddr. Norsk inst. skogforsk 35:265-359. p 273 function (1)
}
\usage{
Tveite_1976_height_trajectory_Norway_Pine(dominant_height, age, age2)
}
\arguments{
\item{dominant_height}{Dominant height, metres.}

\item{age}{Age at breast height.}

\item{age2}{Age at breast height at which to calculate new height.}
}
\value{
Dominant height at age at breast height == age2
}
\description{
Height trajectories for Scots Pine in Norway from Tveite 1976
}
\examples{
#' #What was the height of a Pine currently 17m and 40 years old when it was 20 years at breast height?.
Tveite_1976_height_trajectory_Norway_Pine(dominant_height = 17,age = 40,age2 = 20)
}
