% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Tveite_1977_height_trajectory_Norway_Norway_Spruce.R
\name{Tveite_1977_height_trajectory_Norway_Norway_Spruce}
\alias{Tveite_1977_height_trajectory_Norway_Norway_Spruce}
\title{Height trajectories for Norway Spruce in Norway from Tveite 1977}
\source{
Tveite, Bjørn (1977) Bonitetskurver for gran. (Site-index curves for Norway Spruce (Picea abies (L.) Karst).) Meddr norsk inst. skogforsk. 33:1-84.
}
\usage{
Tveite_1977_height_trajectory_Norway_Norway_Spruce(dominant_height, age, age2)
}
\arguments{
\item{dominant_height}{Dominant Height, metres.}

\item{age}{Age at breast height.}

\item{age2}{Age at breast height at which to calculate new height.}
}
\value{
Dominant Height at age at breast height == age2
}
\description{
Based on permanent plot records from the eastern and middle regions of Norway - intended for use in whole country apart from Western regions. Data consists of 128 height/age series of different length. Construction method makes the reference age arbitrary. 40 years typical in Norway. At ages above 100, the ratio between the anamorphic curves is held constant.
Not recommended for use on stands younger than 15-20 years of age at breast height.
This function returns the regular height trajectories (Equations VII.8 and VII.9)
}
\examples{
#What was the height of a Spruce currently 17m and 40 years old when it was 20 years at breast height?.
Tveite_1977_height_trajectory_Norway_Norway_Spruce(Height = 17,age = 40,age2 = 20)
}
