% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R, R/verbose_cat.R
\name{verbose_cat}
\alias{verbose_cat}
\title{Print the provided cat-like input if verbose is TRUE}
\usage{
verbose_cat(..., sep = " ", verbose = TRUE)

verbose_cat(..., sep = " ", verbose = TRUE)
}
\arguments{
\item{...}{R objects - strings, (see `cat` documentation).}

\item{sep}{A character vector of strings to append after each element.}

\item{verbose}{A logical value indicating whether we want to print the string
or not.}
}
\description{
Print the provided cat-like input if verbose is TRUE

Print the provided cat-like input if verbose is TRUE
}
