% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norway_double_bark_functions.R
\name{Vestjordet_1967_double_bark_thickness_Norway_Norway_Spruce}
\alias{Vestjordet_1967_double_bark_thickness_Norway_Norway_Spruce}
\title{Double bark thickness for Norway Spruce in Norway}
\source{
Vestjordet, E. 1967. Functions and tables for volume of standing
trees. Norway spruce. Reports of the Norwegian Forest Research Institute, 22,
 539-574.
}
\usage{
Vestjordet_1967_double_bark_thickness_Norway_Norway_Spruce(
  diameter_cm,
  height_m
)
}
\arguments{
\item{diameter_cm}{Diameter of tree at breast height, 1.3m.}

\item{height_m}{Total height of tree, m.}
}
\value{
Double bark, in mm.
}
\description{
Double bark thickness for Norway Spruce in Norway
}
\author{
Modified from hansoleorka/skogR on the 2022-04-02
}
