% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Vestjordet_1967_volume_function_Norway_Norway_Spruce.R
\name{Vestjordet_1967_volume_Norway_Norway_Spruce}
\alias{Vestjordet_1967_volume_Norway_Norway_Spruce}
\title{Simple Volume function for Norway Spruce in Norway}
\source{
Vestjordet, E. 1967. Functions and tables for volume of standing
trees. Norway spruce. Reports of the Norwegian Forest Research Institute,
22, 539-574.
}
\usage{
Vestjordet_1967_volume_Norway_Norway_Spruce(
  diameter_cm,
  height_m,
  above_bark = TRUE
)
}
\arguments{
\item{diameter_cm}{Diameter of tree in cm at breast height, 1.3 m.}

\item{height_m}{Total tree height, m.}

\item{above_bark}{TRUE (default) or FALSE.}
}
\value{
Volume of tree in dm^3.
}
\description{
Simple Volume function for Norway Spruce in Norway
}
