% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vuokila_1983_Larix_sibirica_Finland.R
\name{Vuokila_1983_BAI5_Larix_sibirica_Finland}
\alias{Vuokila_1983_BAI5_Larix_sibirica_Finland}
\title{Basal Area Increment for 5 year period, Larix sibirica in Finland.}
\usage{
Vuokila_1983_BAI5_Larix_sibirica_Finland(SI100, DominantHeightM, BasalAreaM2)
}
\arguments{
\item{SI100}{Site Index H100, e.g. [forester::Vuokila_1983_height_trajectory_Finland_Larix_sibirica_GADA]}

\item{DominantHeightM}{Dominant height in meters.}

\item{BasalAreaM2}{Basal Area, m^2 / ha.}
}
\value{
Mean Annual BAI during the future 5 year period, in percent (!)
}
\description{
Basal Area Increment for 5 year period, Larix sibirica in Finland.
}
