% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vuokila_1983_Larix_sibirica_Finland.R
\name{Vuokila_1983_volume_increment_Finland_Larix_sibirica}
\alias{Vuokila_1983_volume_increment_Finland_Larix_sibirica}
\title{Volume Increment in percent, Larix sibirica in Finland.}
\usage{
Vuokila_1983_volume_increment_Finland_Larix_sibirica(
  volumeM3,
  dominantHeightM,
  SI100,
  period
)
}
\arguments{
\item{volumeM3}{Volume per ha, m^3/ha including bark.}

\item{dominantHeightM}{Dominant height of stand, meters.}

\item{SI100}{Site index H100, e.g. [forester::Vuokila_1983_height_trajectory_Finland_Larix_sibirica_GADA]}

\item{period}{Length of calculation period, years.}
}
\value{
Mean annual volume increment percentage of the future calculation period.
}
\description{
Volume Increment in percent, Larix sibirica in Finland.
}
