% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NYSKOG_young_stand_height_development.R
\name{young_stand_height_development_Elfving_NYSKOG}
\alias{young_stand_height_development_Elfving_NYSKOG}
\title{Height development in young stands NYSKOG}
\usage{
young_stand_height_development_Elfving_NYSKOG(species, si, total_age)
}
\arguments{
\item{species}{One of : Pinus sylvestris, Picea abies, Betula pendula, Betula pubescens, Populus tremula, Fagus sylvatica, Quercus robur, Pinus contorta.}

\item{si}{Site index H100.
\tabular{lr}{
\strong{Species} \tab \strong{SI species} \cr
Pine    \tab Pine \cr
Spruce \tab Spruce \cr
Birch \tab Pine \cr
Aspen \tab Pine \cr
Beech \tab Beech \cr
Oak \tab Oak

}}

\item{total_age}{}
}
\value{
Estimated height, m.
}
\description{
A group of functions for mean height development in young stands of scots pine, norway spruce, birch (B. pubescens), birch (B. pendula), aspen, beech, oak or pinus contorta.
}
\examples{
young_stand_height_development_Elfving_NYSKOG(species=species, si=si, total_age=total_age)
}
