% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{basic_info}
\alias{basic_info}
\title{Provides basic dataset information}
\usage{
basic_info(df, y)
}
\arguments{
\item{df}{A data source, that is one of major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string which indicates a target column name.}
}
\value{
A list with every line of the sub-report.
}
\description{
Provides basic dataset information
}
