% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_cor}
\alias{check_cor}
\title{Searches for strongly correlated values (Spearman for numerical, Crammer V for
categorical)}
\usage{
check_cor(df, y)
}
\arguments{
\item{df}{A data source, that is one of major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string which indicates a target column name.}
}
\value{
A list with every line of the sub-report.
}
\description{
Searches for strongly correlated values (Spearman for numerical, Crammer V for
categorical)
}
