% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Run data check pipeline to seek for potential problems with the data}
\usage{
check_data(
  data,
  y = NULL,
  time = NULL,
  status = NULL,
  type = "auto",
  verbose = TRUE
)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{y}{A string that indicates a target column name for regression or classification.
Either y, or pair: time, status can be used. By default NULL.}

\item{time}{A string that indicates a time column name for survival analysis task.
Either y, or pair: time, status can be used. By default NULL.}

\item{status}{A string that indicates a status column name for survival analysis task.
Either y, or pair: time, status can be used. By default NULL.}

\item{type}{A character, one of `binary_clf`/`regression`/`survival`/`auto`/`multiclass` that
sets the type of the task. If `auto` (the default option) then
the function will figure out `type` based on the number of unique values
in the `y` variable, or the presence of time/status columns.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
the process, if FALSE gives none.}
}
\value{
A list with two vectors: lines of the report (str) and the outliers (outliers).
}
\description{
Run data check pipeline to seek for potential problems with the data
}
\examples{
check_data(lisbon, 'Price')
}
