% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Run data check pipeline to seek for potential problems with the data}
\usage{
check_data(data, y, verbose = TRUE)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix, and so on.}

\item{y}{A string that indicates a target column name.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
the process, if FALSE gives none.}
}
\value{
A list with two vectors: lines of the report (str) and the outliers (outliers).
}
\description{
Run data check pipeline to seek for potential problems with the data
}
\examples{
check_data(iris[1:100, ], 'Species')
check_data(lisbon, 'Price')
check_data(compas, 'Two_yr_Recidivism')
check_data(iris, 'Species')
check_data(lymph, 'class')
}
