% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_dim}
\alias{check_dim}
\title{Searches for dimensionality problems in the dataset}
\usage{
check_dim(df)
}
\arguments{
\item{df}{A data source, that is one of major R formats: data.table, data.frame,
matrix and so on.}
}
\value{
A list with every line of the sub-report.
}
\description{
Searches for dimensionality problems in the dataset
}
