% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_static}
\alias{check_static}
\title{Searches for columns dominated by a single value}
\usage{
check_static(df)
}
\arguments{
\item{df}{A data source, that is one of major R formats: data.table, data.frame,
matrix and so on.}
}
\value{
A list with every line of the sub-report.
}
\description{
Searches for columns dominated by a single value
}
