% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_y_balance}
\alias{check_y_balance}
\title{Checks wheter target column is unbalanced (for regression it bins values
via quantiles)}
\usage{
check_y_balance(df, y)
}
\arguments{
\item{df}{A data source, that is one of major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string which indicates a target column name.}
}
\value{
A list with every line of the sub-report.
}
\description{
Checks wheter target column is unbalanced (for regression it bins values
via quantiles)
}
