% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_models.R
\name{choose_best_models}
\alias{choose_best_models}
\title{Choose the bests models, according to the score data frame}
\usage{
choose_best_models(models, engine, score, number)
}
\arguments{
\item{models}{A list of models trained by `train_models()` function.}

\item{engine}{A vector of tree-based models that shall be tested.
Possible values are: `ranger`, `xgboost`, `decision_tree`, `lightgbm`, `catboost`.}

\item{score}{A data frame with metrics values for given models, by `score_models`.}

\item{number}{The number of models to return.}
}
\value{
The list of chosen best models.
}
\description{
Choose the bests models, according to the score data frame
}
