% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_sweden.R
\name{county_sweden}
\alias{county_sweden}
\title{Coordinate to county Sweden}
\usage{
county_sweden(latitude, longitude)
}
\arguments{
\item{latitude}{Latitude, decimal}

\item{longitude}{Longitude, decimal}
}
\value{
Text, Swedish County
}
\description{
Based on the shape file \link[forester]{county_shape_file} returns the Swedish county which the coordinate is placed in.
 If there is no match, county will be 'Other', and a warning will be shown.
}
