% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ranked_list.R
\name{create_ranked_list}
\alias{create_ranked_list}
\title{Create final ranked_list}
\usage{
create_ranked_list(basic, random, Bayes, type)
}
\arguments{
\item{basic}{A data.frame with ranked list of basic models.}

\item{random}{A data.frame with ranked list of random search models.}

\item{Bayes}{A data.frame with ranked list of Bayes Optimization models.}

\item{type}{A string which determines if Machine Learning task is the
`classification` or `regression`.}
}
\value{
A data.frame ranked list sorted by default metric.
}
\description{
Create final ranked_list
}
