% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{draw_scatterplot}
\alias{draw_scatterplot}
\title{Draw scatterplot of true vs predicted values of target for training and test data for one model}
\usage{
draw_scatterplot(
  train_observed,
  train_predictions,
  test_observed,
  test_predictions
)
}
\arguments{
\item{train_observed}{A vector of true values for training data.}

\item{train_predictions}{A vector of predicted values for training data.}

\item{test_observed}{A vector of true values for test data.}

\item{test_predictions}{A vector of predicted values for test data.}
}
\description{
Draw scatterplot of true vs predicted values of target for training and test data for one model
}
