% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elfving_2003_mean_diameter_codominant_trees.R
\name{Elfving_2003_mean_diameter_codominant_trees}
\alias{Elfving_2003_mean_diameter_codominant_trees}
\title{Mean diameter of co-dominant trees}
\usage{
Elfving_2003_mean_diameter_codominant_trees(total_stand_age, SIS, total_ba)
}
\arguments{
\item{total_stand_age}{total stand age}

\item{SIS}{site index according site factors, given as top height at age 100, m.}

\item{total_ba}{As estimated with relascope in field. m^2 / ha}
}
\value{

}
\description{
Presented in Elfving, B. (2003) Ålderstilldelning till enskilda träd i skogliga tillväxtprognoser. SLU, inst. f. skogsskötsel. Arbetsrapport 182.
}
