% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eriksson_1997_height_trajectory_Sweden_BIrch.R
\name{Eriksson_1997_height_trajectory_Sweden_Birch}
\alias{Eriksson_1997_height_trajectory_Sweden_Birch}
\title{Height trajectory for Birch in Sweden from Eriksson et al 1997.}
\source{
Eriksson, H., Johansson, U., Kiviste, A. (1997) A site-index model for
pure and mixed stands of Betula pendula and Betula pubescens in Sweden.
Scandinavian Journal of Forest Research. 12:2, 149-156.
Available: \url{https://doi.org/10.1080/02827589709355396}
}
\usage{
Eriksson_1997_height_trajectory_Sweden_Birch(
  dominant_height,
  age,
  age2,
  output = "Height"
)
}
\arguments{
\item{dominant_height}{Dominant height of stand, m.}

\item{age}{Age at breast height.}

\item{age2}{Output age at breast height.}

\item{output}{One of "SIH100","Equation" or "Height".}
}
\value{
One of output.
}
\description{
Abstract

A site‐index model was constructed based on stem analysis data for 266 top‐height
trees of Betula pendula Roth and Betula pubescens Ehrh. growing on 155 temporary
sample plots and 12 remeasured, permanent sample plots in pure and mixed stands
scattered throughout Sweden. Different growth functions and techniques for
modelling the top‐height growth over breast‐height age were assessed. A difference model
based on the Hossfeld IV growth equation performed best, and its practical
application was therefore recommended. For birch stands over 40 yrs of age,
top‐height growth predicted by the new model was significantly slower than
that predicted by a site‐index model used previously.
}
\examples{

ErikssonVect <- Vectorize(Eriksson_1997_height_trajectory_Sweden_Birch)

plot(NULL,
    xlim=c(0,100),
    ylim=c(0,35))

for(i in 1:11){
 lines(seq(0,100,0.1),ErikssonVect(seq(10,30,2)[i],age = 50,age2 = seq(0,100,0.1)))
}

points(rep(50,11),seq(10,30,2))
}
