% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forester.R
\name{forester}
\alias{forester}
\title{Function for training models}
\usage{
forester(data, target, type, metric = NULL, data_test = NULL)
}
\arguments{
\item{metric}{string containing name of mertic based on which the model will be selected.
Metric shoud be written in small letters}
}
\value{
An object of the class \code{forester_model}.
}
\description{
Function \code{train} automates the whole ML pipelines ....
}
\examples{
\donttest{
library(DALEX)
data(apartments, package="DALEX")

best_model <- forester(apartments, "m2.price", "regression")
}
}
\references{
Explanatory Model Analysis. Explore, Explain, and Examine Predictive Models. \url{https://ema.drwhy.ai/}
}
