% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forester.R
\name{forester}
\alias{forester}
\title{Create a forest plot column in a table}
\usage{
forester(
  left_side_data,
  estimate,
  ci_low,
  ci_high,
  right_side_data = NULL,
  estimate_precision = 1,
  ggplot_width = 30,
  null_line_at = 0,
  file_path = here::here("forester_plot.png"),
  dpi = 600,
  display = TRUE,
  blank_na = TRUE,
  font_family = "mono",
  estimate_col_name = "Estimate",
  stripe_colour = "#eff3f2",
  x_scale_linear = TRUE,
  xlim = NULL,
  xbreaks = NULL,
  nudge_y = NULL,
  arrows = FALSE,
  arrow_labels = c("Lower", "Higher"),
  add_plot = NULL,
  add_plot_width = 3
)
}
\arguments{
\item{left_side_data}{Data frame (required). The information to be displayed to the left of the forest plot.}

\item{estimate}{Vector. The point estimates to be displayed in the forest plot.}

\item{ci_low}{Vector. The lower confidence bounds.}

\item{ci_high}{Vector. The upper confidence bounds.}

\item{right_side_data}{Data frame (optional). Information to be displayed on the right side of the table. If not supplied, an Estimate column is generated automatically.}

\item{estimate_precision}{Integer. The number of decimal places on the estimate (default 1)}

\item{ggplot_width}{Integer. The width of forest plot in characters (default 30)}

\item{null_line_at}{Numeric. Default 0. Change to 1 if using relative measures such as OR, RR.}

\item{file_path}{String. Where to save the image, default "forestable_plot.png" in the current working directory.}

\item{dpi}{Numeric. The image resolution in dpi, default 600}

\item{display}{Logical. Show the table in RStudio viewer? Default TRUE}

\item{blank_na}{Logical. Should missing values in the left side table be displayed as blank? Default TRUE, if FALSE, NA values will be shown}

\item{font_family}{String. The font to use for the ggplot and table. Default "mono".}

\item{estimate_col_name}{String. The name for the generated estimate column. Default "Estimate"}

\item{stripe_colour}{Hex String. Colour to use for the table stripes, default "#eff3f2"}

\item{x_scale_linear}{Logical. Default TRUE, change to FALSE for log scale}

\item{xlim}{Vector. Manually specify limits for the x axis as a vector length 2, i.e. c(low, high)}

\item{xbreaks}{Vector. X axis breaks to label. Specify limits in xlim if using this option.}

\item{nudge_y}{Numeric. Allows small changes to the vertical alignment of the forest plot points. 1 unit is approximately the height of 1 row.}

\item{arrows}{Logical. Should there be arrows displayed below the ggplot? Default FALSE. Specify xlim if using arrows.}

\item{arrow_labels}{String Vector, length 2. Labels for the arrows. Set arrows to TRUE or this will have no effect.}

\item{add_plot}{A ggplot object to add to the right side of the table. To align correctly with rows, 1 unit is the height of a row and y = 0 for the center of the bottom row.}

\item{add_plot_width}{Integer (2 - 5). Width to display add_plot. Relative to the width of the whole figure, where 10 would be the whole width. Default 3.}
}
\value{
image
}
\description{
Creates an png image with the forest plot column inserted in the supplied data frame.
}
