% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_models_details.R
\name{format_models_details}
\alias{format_models_details}
\title{Format info about models}
\usage{
format_models_details(models)
}
\arguments{
\item{models}{The result of `choose_best_models()` function or just a list of models
from engine.}
}
\value{
Prints formatted text.
}
\description{
Format info about models
}
