% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Fries_1964_height_trajectory_southern_central_Sweden_Birch.R
\name{Fries_1964_height_trajectory_southern_central_Sweden_Birch}
\alias{Fries_1964_height_trajectory_southern_central_Sweden_Birch}
\title{Height for diameter classes in Birch stands according to Fries 1964}
\usage{
Fries_1964_height_trajectory_southern_central_Sweden_Birch(
  dominant_height,
  diameter_cm,
  D10
)
}
\arguments{
\item{dominant_height}{Dominant height, mean height of the 10% thickest trees.}

\item{diameter_cm}{Diameter above bark at breast height 1.3 m.}

\item{D10_cm}{Diameter of the 10% thickest trees above bark.}
}
\value{

}
\description{
Height for diameter classes in Birch stands according to Fries 1964
}
