% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_type.R
\name{guess_type}
\alias{guess_type}
\title{Guess task type by the target value from dataset}
\usage{
guess_type(data, y, max_unique_numeric = 5, max_unique_not_numeric = 15)
}
\arguments{
\item{data}{A data source, that is one of major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string which indicates a target column name.}

\item{max_unique_numeric}{An integer describing maximal number of unique
values in `y` if `y` is numeric.}

\item{max_unique_not_numeric}{An integer describing maximal number of unique
values in `y` if `y` is not numeric.}
}
\value{
A string describing type of ml task: `binary_clf`, `multi_clf` or
`regression`.
}
\description{
For now, supported types are binary classification and regression.
Multilabel classification is planned to be added later on.
}
\examples{
data(compas)
guess_type(compas,'Two_yr_Recidivism')
}
