% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hagglund_1981_si_to_bonitet.R
\name{Hagglund_1981_si_to_bonitet}
\alias{Hagglund_1981_si_to_bonitet}
\title{Smoothed Productivity estimates in m^3sk (cu.m.) from Hägglund 1981.}
\usage{
Hagglund_1981_si_to_bonitet(H100, main_species, vegetation, altitude, county)
}
\arguments{
\item{H100}{Estimated stand top height at 100 years age. e.g. [forester::Hagglund_1972_northern_Sweden_Height_trajectories_Spruce()],[forester::Hagglund_1973_southern_Sweden_Height_trajectories_Spruce()],[forester::Hagglund_1974_Sweden_height_trajectories_Pine()]}

\item{main_species}{Main species, one of: 'Picea abies', or 'Pinus sylvestris'}

\item{vegetation}{Vegetation class [forester::Sweden_vegetation_types()]}

\item{altitude}{Metres above sea level.}

\item{county}{Swedish county [forester::county_sweden()]}
}
\value{
Estimates the mean volume growth in m3sk / ha yr-1 at the time of culmination (or if late culmination, at 150yrs of age)
}
\description{
Smoothed productivity estimates in m3sk from Hägglund 1981.
}
\details{
This function conveniently chooses the appropriate function.

 Productive potential as a function of site index

 a - Spruce, southern Sweden.

 b - Spruce, middle Sweden, herb types, grass types and ground without field layer.

 c - Spruce, middle Sweden, bilberry type and worse.

 d - Spruce, northern Sweden, herb types, grass types and ground without field layer.

 e - Spruce, northern Sweden, bilberry type and worse.

 f - Pine, southern and middle Sweden as well as lowlands of northern Sweden.

 g - Pine, northern Sweden, more than 200 meters above sea level.
}
