% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heidarsson_2012_Iceland_Larix_sibirica.R
\name{Heidarsson_2011_stem_taper_Iceland_Larix_sibirica}
\alias{Heidarsson_2011_stem_taper_Iceland_Larix_sibirica}
\alias{Heidarsson_2011}
\alias{Heidarsson_2011_stem_taper_Iceland_Pinus_contorta}
\title{Stem taper of Lodgepole Pine and Siberian Larch on Iceland}
\source{

}
\usage{
Heidarsson_2011_stem_taper_Iceland_Larix_sibirica(
  diameter_cm,
  height_m,
  output_height
)

Heidarsson_2011_stem_taper_Iceland_Pinus_contorta(
  diameter_cm,
  height_m,
  output_height
)
}
\arguments{
\item{diameter_cm}{Diameter of tree at breast height 1.3m.}

\item{height_m}{Total height of tree.}

\item{output_height}{Output height.}
}
\value{
Diameter of tree in cm at output height.
}
\description{
Stem taper of Lodgepole Pine and Siberian Larch on Iceland
}
\details{
\tabular{lrr}{
Species \tab R^2 \tab RMSE \cr
Larix sibirica \tab 0.986 \tab 1.024 \cr
Pinus contorta \tab 0.979 \tab 0.888 \cr
}
}
