% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_catboost.R
\name{make_catboost}
\alias{make_catboost}
\title{Automated function for explaining catboost models}
\usage{
make_catboost(
  data,
  target,
  type,
  fill_na = TRUE,
  num_features = NULL,
  tune = FALSE,
  tune_metric = NULL,
  tune_iter = 20,
  label = "Catboost"
)
}
\arguments{
\item{data}{data.frame, matrix, data.table or dgCMatrix - data will be used to run the XGBoost model. NOTE: data has to contain the target column.}

\item{target}{character: name of the target column, should be character and has to be a column name in data.}

\item{type}{character: defining the task. Two options are: "regression" and "classification", particularly, binary classification.}

\item{fill_na}{logical, default is FALSE. If TRUE, missing values in target column are removed, missing values in categorical columns are replaced by mode and
missing values in numeric columns are substituted by median of corresponding columns.}

\item{num_features}{numeric, default is NULL. Parameter indicates number of most important features, which are chosen from the train dataset. Automatically, those important
features will be kept in the train and test datasets.}

\item{tune}{logical. If TRUE, function will perform the hyperparameter tuning steps for each model inside.}

\item{tune_metric}{character, name of metric used for evaluating best model. For regression, options are: "mse", "rmse", "mad" and "r2".
For classification, options are: "auc", "recall", "precision", "f1" and "accuracy".}

\item{tune_iter}{number (default: 20) - total number of times the optimization step is to repeated. This argument is used when tune = TRUE.}

\item{label}{string indicating the name of the model. Might be usefull while comparing different models of the same type.}
}
\value{
An object of the class \code{explainer}.
}
\description{
Function \code{make_catboost} creates catboost model with default parameters using ranger library and
wraps it with the explainer.
}
\examples{
\donttest{
# regression
library(DALEX)
data(apartments, package="DALEX")

exp <- make_catboost(apartments, "m2.price", "regression")
# plot(model_parts(exp))

# binary classification
library(DALEX)
data(titanic_imputed, package="DALEX")
exp <- make_catboost(titanic_imputed, "survived", "classification")
# plot(model_parts(exp))
}
}
\references{
Explanatory Model Analysis. Explore, Explain, and Examine Predictive Models. \url{https://ema.drwhy.ai/}
}
