% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern_Hagberg_1975_volume_Oak_beech.R
\name{MaternVolume}
\alias{MaternVolume}
\alias{Matern_1975_volume_Sweden_Oak}
\alias{Matern_1975_volume_Sweden_Beech}
\title{Volumes of Beech or Oak trees, from Matérn 1975.}
\source{
Matérn, B. 1975. Volymfunktioner för stående träd av ek och bok. Skogshögskolan, institutionen för skoglig matematisk statistik. Rapp. o. Upps. nr 15.

PM for Heureka 2004-01-20 Björn Elfving. Available: \url{https://www.heurekaslu.org/w/images/9/93/Heureka_prognossystem_\%28Elfving_rapportutkast\%29.pdf}
}
\usage{
Matern_1975_volume_Sweden_Oak(diameter_cm, height_m)

Matern_1975_volume_Sweden_Beech(diameter_cm, height_m)
}
\arguments{
\item{diameter_cm}{Diameter at breast height, cm.}

\item{height_m}{Height of tree, metres.}
}
\value{
Volume, dm3.
}
\description{
Volumes of Beech or Oak trees, from Matérn 1975.
}
