% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naslund_kubering.R
\name{NaslundKubering}
\alias{NaslundKubering}
\alias{Naslund_1947_volume_southern_Sweden_Pine}
\alias{Naslund_1947_form_factor_southern_Sweden_Pine}
\alias{Naslund_1947_volume_southern_Sweden_Spruce}
\alias{Naslund_1947_form_factor_southern_Sweden_Spruce}
\alias{Naslund_1947_volume_southern_Sweden_Birch}
\alias{Naslund_1947_form_factor_southern_Sweden_Birch}
\alias{Naslund_1947_volume_Sweden_Spruce}
\alias{Naslund_1947_form_factor_Sweden_Pine}
\alias{Naslund_1947_form_factor_Sweden_Spruce}
\alias{Naslund_1947_form_factor_Sweden_Birch}
\alias{Naslund_1947_volume_Sweden_Birch}
\alias{Naslund_1940_volume_northern_Sweden_Pine}
\alias{Naslund_1940_form_factor_northern_Sweden_Pine}
\alias{Naslund_1940_volume_northern_Sweden_Spruce}
\alias{Naslund_1940_form_factor_northern_Sweden_Spruce}
\alias{Naslund_1940_volume_northern_Sweden_Birch}
\alias{Naslund_1940_form_factor_northern_Sweden_Birch}
\title{Funktioner och Tabeller för kubering av stående träd.}
\source{
Northern Sweden: Näslund, Manfred. 1941. Funktioner och tabeller för
kubering av stående träd. Tall, gran och björk i norra Sverige. Meddelanden
från statens skogsförsöksanstalt. Häfte 32. Nr. 4. Centraltryckeriet, Esselte,
 Stockholm. p. 87-142 pp. \url{https://pub.epsilon.slu.se/10185/1/medd_statens_skogsforskningsanst_032_04.pdf}

Southern Sweden and entire country: Näslund, Manfred. 1947. Funktioner
och tabeller för kubering av stående träd. Tall, gran och björk i södra Sverige samt i hela landet.
Meddelanden från statens skogsforskningsinstitut. Band 36. Nr. 3. Centraltryckeriet,
Esselte, Stockholm. p. 1-80. \url{https://pub.epsilon.slu.se/9900/1/medd_statens_skogsforskningsinst_036_03.pdf}
}
\usage{
Naslund_1947_volume_southern_Sweden_Pine(
  height,
  DBH,
  double_bark_mm,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1947_form_factor_southern_Sweden_Pine(
  height,
  DBH,
  double_bark_mm,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1947_volume_southern_Sweden_Spruce(
  height,
  DBH,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1947_form_factor_southern_Sweden_Spruce(
  height,
  DBH,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1947_volume_southern_Sweden_Birch(
  height,
  DBH,
  double_bark_mm = NULL,
  over_bark = TRUE
)

Naslund_1947_form_factor_southern_Sweden_Birch(
  height,
  DBH,
  double_bark_mm = NULL,
  over_bark = TRUE
)

Naslund_1947_volume_Sweden_Spruce(
  height,
  DBH,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1947_form_factor_Sweden_Pine(
  height,
  DBH,
  double_bark_mm = NULL,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1947_form_factor_Sweden_Spruce(
  height,
  DBH,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1947_volume_Sweden_Spruce(
  height,
  DBH,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1947_form_factor_Sweden_Birch(
  height,
  DBH,
  double_bark_mm = NULL,
  over_bark = TRUE
)

Naslund_1947_volume_Sweden_Birch(
  height,
  DBH,
  double_bark_mm = NULL,
  over_bark = TRUE
)

Naslund_1940_volume_northern_Sweden_Pine(
  height,
  DBH,
  double_bark_mm = NULL,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1940_form_factor_northern_Sweden_Pine(
  height,
  DBH,
  double_bark_mm = NULL,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1940_volume_northern_Sweden_Spruce(
  height,
  DBH,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1940_form_factor_northern_Sweden_Spruce(
  height,
  DBH,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1940_volume_northern_Sweden_Birch(
  height,
  DBH,
  double_bark_mm = NULL,
  crownBaseheight = NULL,
  over_bark = TRUE
)

Naslund_1940_form_factor_northern_Sweden_Birch(
  height,
  DBH,
  double_bark_mm = NULL,
  crownBaseheight = NULL,
  over_bark = TRUE
)
}
\arguments{
\item{height}{Tree height, meters.}

\item{DBH}{Diameter at breast height, over or respectively under bark.}

\item{double_bark_mm}{Double bark thickness, mm.}

\item{crownBaseheight}{Height until crown base.}

\item{over_bark}{Boolean. If input diameter is over/under bark, and if returned value should be over or under bark.}
}
\value{
Form factor (unspecified.) or Volume, dm^3.
}
\description{
Funktioner och Tabeller för kubering av stående träd.
}
\details{
If only required for more detailed function, arguments will be default=NULL.
}
