% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pinus}
\alias{pinus}
\title{Pinus Taeda Data}
\format{
A data frame with 12 rows and 2 variables:
\describe{
\item{Parcela}{sample unit, em metros quadrados}
\item{Volume}{Volume, em metros cúbicos}

...
}
}
\source{
Sanquetta
}
\usage{
pinus
}
\description{
A dataset containing the volumes per sample.
}
\keyword{datasets}
