% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification.R
\name{plot.binary_clf}
\alias{plot.binary_clf}
\title{Plot of Metrics for Binary Classification Models}
\usage{
\method{plot}{binary_clf}(x, models = NULL, type = "comparison", metric = "accuracy", ...)
}
\arguments{
\item{x}{An object, returned from \code{train()} function.}

\item{models}{A character or numeric that indicates which models
will be presented. If `NULL` (the default option) then the three best models
will be presented.}

\item{type}{A character one of `comparison`, `roc`, `confusion-matrix`, `train-test`
indicates the type of chart.}

\item{metric}{A character one of `accuracy`/`auc`/`f1` indicates the metric on the plots.}

\item{...}{Other parameters that are necessary for consistency with generic plot function.}
}
\value{
a ggplot2 object
}
\description{
Function plot.classification plots measure of quality of regression models.
}
\examples{
\dontrun{
library('forester')
data('adult')

x <- train(adult, 'salary', bayes_iter = 0, random_evals = 0)
plot(x)
}

}
