% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression.R
\name{plot.regression}
\alias{plot.regression}
\title{Plot of Metrics for Regression Models}
\usage{
\method{plot}{regression}(x, models = NULL, type = "residuals", metric = "rmse", ...)
}
\arguments{
\item{x}{An object, returned from \code{train()} function.}

\item{models}{A character or numeric that indicates which models
will be presented. If `NULL` (the default option) then the three best models
will be presented.}

\item{type}{A character one of `residuals`/`train-test-observed-predicted`/`train-test-rmse`
indicates the type of chart.}

\item{metric}{A character one of `rmse`/`mse`/`r2`/`mae` indicates the metric on the plots.}

\item{...}{Other parameters that are necessary for consistency with generic plot function.}
}
\value{
a ggplot2 object
}
\description{
Function plot.regression plots measure of quality of regression models.
}
\examples{
\dontrun{
library('forester')
data('lisbon')

x <- train(lisbon, 'Price', bayes_iter = 0, random_evals = 0)
plot(x)
}

}
