% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_models_all.R
\name{predict_models_all}
\alias{predict_models_all}
\title{Predictions for a list of models with multiple occurrences of the same types of models}
\usage{
predict_models_all(models, data, y, type)
}
\arguments{
\item{models}{A list of models trained by `train_models()` function.}

\item{data}{A test data for models created by `prepare_data()` function.}

\item{y}{A string that indicates a target column name.}

\item{type}{A string that determines if Machine Learning task is the
`binary_clf`, `regression`, `survival`, or `multiclass` task.}
}
\value{
A list of predictions for every engine without names.
}
\description{
Predictions for a list of models with multiple occurrences of the same types of models
}
