% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_models_all.R
\name{predict_models_all}
\alias{predict_models_all}
\title{Predictions for a list of models with multiple occurrences of the same types of models}
\usage{
predict_models_all(models, data, y, type)
}
\arguments{
\item{models}{A list of models trained by `train_models()` function.}

\item{data}{A test data for models created by `prepare_data()` function.}

\item{y}{A string that indicates a target column name.}

\item{type}{A string that determines if Machine Learning task is the
`binary_clf` or `regression` task.}
}
\value{
A list of predictions for every engine without names.
}
\description{
Predictions for a list of models with multiple occurrences of the same types of models
}
\examples{
data(iris)
iris_bin          <- iris[1:100, ]
type              <- guess_type(iris_bin, 'Species')
preprocessed_data <- preprocessing(iris_bin, 'Species', type)
preprocessed_data <- preprocessed_data$data
split_data <-
  train_test_balance(preprocessed_data,
                     'Species',
                     balance = FALSE)
train_data <-
  prepare_data(split_data$train,
               'Species',
               engine = c('ranger', 'xgboost', 'decision_tree', 'lightgbm', 'catboost'))
test_data <-
  prepare_data(split_data$test,
               'Species',
               engine = c('ranger', 'xgboost', 'decision_tree', 'lightgbm', 'catboost'),
               predict = TRUE,
               train = split_data$train)

model <-
  train_models(train_data,
               'Species',
               engine = c('ranger', 'xgboost', 'decision_tree', 'lightgbm', 'catboost'),
               type = type)
predictions <-
  predict_models_all(model,
                 test_data,
                 'Species',
                 type = type)
}
