% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data into format correct for the selected model engine}
\usage{
prepare_data(
  data,
  type,
  y = NULL,
  time = NULL,
  status = NULL,
  engine = c("ranger", "xgboost", "decision_tree", "lightgbm", "catboost"),
  predict = FALSE,
  train = NULL
)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix and so on.}

\item{type}{A character, one of `binary_clf`/`regression`/`survival`/`auto`/`multiclass` that
sets the type of the task.}

\item{y}{A string that indicates a target column name for regression or classification.
Either y, or pair: time, status can be used. By default NULL.}

\item{time}{A string that indicates a time column name for survival analysis task.
Either y, or pair: time, status can be used. By default NULL.}

\item{status}{A string that indicates a status column name for survival analysis task.
Either y, or pair: time, status can be used. By default NULL.}

\item{engine}{A vector of tree-based models that shall be created. Possible
values are: `ranger`, `xgboost`, `lightgbm`, `catboost`, `decision_tree`.
Determines which models will be later learnt. It doesn't matter for survival analysis.}

\item{predict}{A logical value, determines whether the data set will be used
for prediction or training. It is necessary, because lightgbm model can't predict
on training dataset.}

\item{train}{A train data, if predict is TRUE you have to provide training
dataset from split data here.}
}
\value{
A dataset in format proper for the selected engines.
}
\description{
Prepare data into format correct for the selected model engine
}
