% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Conduct preprocessing processes}
\usage{
preprocessing(data, y, type, advanced = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{A data source, that is one the of major R formats: data.table, data.frame,
matrix, and so on.}

\item{y}{A string that indicates a target column name.}

\item{type}{A string that determines if Machine Learning task is the
`binary_clf` or `regression`.}

\item{advanced}{A logical value describing, whether the user wants to use
advanced preprocessing methods (deleting correlated values).}

\item{verbose}{A logical value, if set to TRUE, provides all information about
the process, if FALSE gives none.}
}
\value{
A preprocessed dataset.
}
\description{
Conduct preprocessing processes
}
