% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Conduct preprocessing processes}
\usage{
preprocessing(
  data,
  y = NULL,
  time = NULL,
  status = NULL,
  type,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data source, that is one the of major R formats: data.table, data.frame,
matrix, and so on.}

\item{y}{A string that indicates a target column name for regression or classification.
Either y, or pair: time, status can be used. By default NULL.}

\item{time}{A string that indicates a time column name for survival analysis task.
Either y, or pair: time, status can be used. By default NULL.}

\item{status}{A string that indicates a status column name for survival analysis task.
Either y, or pair: time, status can be used. By default NULL.}

\item{type}{A string that determines if Machine Learning task is the
`binary_clf`, `regression`, or `survival`.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
the process, if FALSE gives none.}
}
\value{
A list containing three objects:
\itemize{
\item \code{`data`} A preprocessed dataset,
\item \code{`rm_colnames`} The names of removed columns,
\item \code{`bin_labels`} The text labels before target binarization.
}
}
\description{
Conduct preprocessing processes
}
