% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_search.R
\name{random_search}
\alias{random_search}
\title{Random optimization of hyperparameters}
\usage{
random_search(
  train_data,
  y = NULL,
  time = NULL,
  status = NULL,
  engine,
  type,
  max_evals = 10,
  parallel = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{train_data}{A training data for models created by `prepare_data()` function.}

\item{y}{A string that indicates a target column name for regression or classification.
Either y, or pair: time, status can be used.}

\item{time}{A string that indicates a time column name for survival analysis task.
Either y, or pair: time, status can be used.}

\item{status}{A string that indicates a status column name for survival analysis task.
Either y, or pair: time, status can be used.}

\item{engine}{A vector of tree-based models that shall be created. Possible
values are: `ranger`, `xgboost`,`decision_tree`, `lightgbm`, `catboost`. Doesn't
matter for survival analysis.}

\item{type}{A string that determines if Machine Learning task is the
`binary_clf`, `regression`, `survival`, or `multiclass` task.}

\item{max_evals}{The number of trained models for each model type in `engine`.}

\item{parallel}{A logical value, if set to TRUE, the function will use parallel computing.
By default set to FALSE.}

\item{verbose}{A logical value, if set to TRUE, provides all information about
the process, if FALSE gives none. Set to FALSE by default.}
}
\value{
A list consisting of models created via random search and ranked list
of models scores.
}
\description{
Random optimization of hyperparameters
}
