% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regeneration_stocking_level.R
\name{regeneration_stocking}
\alias{regeneration_stocking}
\title{Stocking in regenerations}
\usage{
regeneration_stocking(
  altitude,
  latitude,
  age,
  regeneration_method,
  proportion_cultivated,
  N_full,
  ground_water,
  county = "county",
  scarification,
  burnt,
  no_treatment,
  uncleaned,
  number_seed_trees,
  SIH,
  SIS,
  area
)
}
\arguments{
\item{altitude}{altitude masl}

\item{latitude}{latitude decimal}

\item{age}{age, number of vegetation periods between cut and inventory.}

\item{regeneration_method}{"natural" or "cultivated-sown" or "cultivated-planted"}

\item{proportion_cultivated}{percent cultivated, for partly planted sites.}

\item{N_full}{number seedlings demanded for full stocking. In thousands?}

\item{county}{Swedish county}

\item{scarification}{1 if scarified, 0 if no treatment.}

\item{burnt}{1 if prescribed burning, 0 if no treatment.}

\item{no_treatment}{1 if no measures were taken to promote regeneration.}

\item{uncleaned}{1 if bushes and small trees remain after clear-cut.}

\item{number_seed_trees}{number of seed trees per hectare}

\item{area}{size of regeneration area in hectares.}

\item{main_species}{Main species "Picea abies" or "Pinus sylvestris".}

\item{soil_moisture}{Soil moisture code (SWE-NFI 'FUKTIGHET')}

\item{square_spacing}{Average square spacing used for sowing/planting}
}
\value{

}
\description{
Stocking in regenerations
}
