% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generate report after training}
\usage{
report(
  train_output,
  output_file = NULL,
  output_format = "pdf_document",
  output_dir = getwd(),
  check_data = TRUE,
  metric = NULL
)
}
\arguments{
\item{train_output}{The output of `train()` function.}

\item{output_file}{The output file name.}

\item{output_format}{Format of the output file ('pdf_document', 'html_document' or both).}

\item{output_dir}{The path where the report will be saved, by default - the working
directory.}

\item{check_data}{If TRUE, prints results of `check_data()` function.}

\item{metric}{A character string describing the main metric for sorting the
ranked list of models.}
}
\value{
Report generated to the local file. It contains table with
metrics for every model, scatter plot for chosen metric
and result of `check_data()` function - about used data.
}
\description{
Generate report after training
}
\examples{
\dontrun{
library(forester)
data('lisbon')
train_output <- train(lisbon, 'Price', random_evals = 10, bayes_iter = 1)
report(train_output)
}
}
