% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{RS}
\alias{RS}
\title{Simple Random Sampling}
\usage{
RS(data = NULL, x, by = NULL, A, a, FP = FALSE, DT = FALSE, ...)
}
\arguments{
\item{data}{Conjunto de dados que contém as informações de inventário florestal.}

\item{x}{Um vetor numérico com a variável de interesse.}

\item{by}{Um fator para cálculo.}

\item{A}{Área total da população, em hectare.}

\item{a}{Área da unidade de amostra, em hectare.}

\item{FP}{Lógico. Fator de proporcionalidade para transformar para hectare. (default = FALSE)}

\item{DT}{Lógico. Imprime a saída no painel de visualização. (default = FALSE)}

\item{...}{Parâmetros adicionais}
}
\value{
Uma lista com estatística descritiva, estimativas de amostragem aleatória simples e informações de base para cálculo das estimativas.
}
\description{
Calcula as estimativas de amostragem aleatória simples para dados de inventário florestal.
}
\details{
Um detalhe importante.
}
\examples{
\dontrun{
data(pinus)
SRS(x = pinus$Volume, A = 40, a = 0.06, FP = T, DT = F)
}

}
