% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS_estimation.R
\name{SIS_estimate}
\alias{SIS_estimate}
\title{Site index estimation by site factors}
\usage{
SIS_estimate(
  plotid,
  species,
  vegetation,
  ground_layer,
  latitude,
  longitude,
  altitude,
  aspect_main,
  incline_percent,
  soil_moisture,
  soil_texture,
  soil_depth,
  lateral_water,
  ditched,
  local_climate,
  county
)
}
\arguments{
\item{plotid}{PlotID}

\item{species}{"Picea abies" or "Pinus sylvestris"}

\item{vegetation}{Type 1-18.}

\item{ground_layer}{Type 1="Lichen type", 2="Lichen-rich bogmoss type", 3="Lichen-rich",4="Bogmoss type" Sphagnum,5="Swamp moss type",6="Fresh moss type"}

\item{latitude}{Decimal degrees.}

\item{longitude}{Decimal degrees.}

\item{altitude}{Metres above sea level}

\item{aspect_main}{Main aspect of site, one of :"North","South","East","West" or 0.}

\item{incline_percent}{Incline of slope, in percent.}

\item{soil_moisture}{Type 1="Dry/torr",2="Mesic/frisk",3="Mesic-moist/frisk-fuktig",4="Moist/fuktig",5="Wet/Blöt"}

\item{soil_texture}{Type 1-9}

\item{soil_depth}{Type 1="Deep, >70cm", 2 = "Rather shallow, 20-70 cm", 3 = "Shallow <20", 4 = "Varying"}

\item{lateral_water}{Type 1="Missing", 2="Seldom",3="shorter periods",4="longer periods",5="slope".}

\item{ditched}{TRUE/FALSE if affected by ditching.}

\item{county}{Optional, else from lat/lon}

\item{climate_code}{Optional, else from lat, lon.}
}
\value{
Site index, in metres.
}
\description{
Corrected functions from Appendix II for estimation of site index by site factors according to Hägglund and Lundmark (1977), in
Hägglund, B. (1979) Ett system för bonitering av skogsmark - analys, kontroll och diskussion inför praktisk tillämpning. Rapport 14, Projekt HUGIN. Skogsvetenskapliga fakulteten, Sveriges Lantbruksuniversitet. Umeå. Sweden.
}
