% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteData.R
\name{siteData}
\alias{siteData}
\title{siteData}
\usage{
siteData(
  altitude,
  latitude,
  longitude,
  polygon,
  county,
  local_climate,
  vegetation,
  soil_texture,
  soil_moisture,
  temperature_sum.c,
  aspect,
  incline
)
}
\arguments{
\item{altitude}{Metres above sea level}

\item{latitude}{Latitude decimal WGS84}

\item{longitude}{Longitude decimal WGS84}

\item{polygon}{Optional List of boundary coordinates.}

\item{county}{County in Sweden. If not supplied, will calculate through \link[forester]{county_sweden}}

\item{local_climate}{Swedish Local Climate Code. If not supplied, will calculate through \link[forester]{local_climate_sweden}}

\item{vegetation}{NFI Vegetation code (FsKod) 1-18}

\item{soil_texture}{1-9}

\item{soil_moisture}{1-5}

\item{temperature_sum.c}{Degrees celsius. If not supplied, will calculate from \link[forester]{odin_tsum_1983}}
}
\value{

}
\description{
Initiates a siteData object, containing information about the plot site.
}
