% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS.R
\name{SRS}
\alias{SRS}
\title{Simple Random Sampling}
\usage{
SRS(x, by = NULL, A, a, FP = TRUE, DT = TRUE)
}
\arguments{
\item{x}{Um vetor numérico com a variável de interesse.}

\item{by}{Um fator para cálculo.}

\item{A}{Área total da população, em metros quadrados.}

\item{a}{Área da amostra, em metros quadrados.}

\item{FP}{Lógico. Fator de proporcionalidade para transformar para hectare. (default=TRUE)}

\item{DT}{Lógico. imprime a saída no painel de visualização. (default=TRUE)}
}
\value{
Uma tibble com as estimativas de amostragem aleatória simples.
}
\description{
Calcula as estimativas de amostragem aleatória simples para dados de inventário florestal.
}
\details{
Um detalhe importante.
}
\examples{
\dontrun{
data(pinus)
SRS(x = pinus$Volume, A = 400000, a = 600, DT=F)
}

}
