% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS.R
\name{SRS}
\alias{SRS}
\title{Simple random sampling}
\usage{
SRS(x, by = NULL, A, a, DT = TRUE)
}
\arguments{
\item{x}{Um vetor numérico com a variável de interesse.}

\item{by}{Um fator para cálculo.}

\item{A}{Área total da população, em metros quadrados.}

\item{a}{Área da amostra, em metros quadrados.}

\item{DT}{Lógico. imprime a saída no painel de visualização. (default=TRUE)}
}
\value{
Uma tibble com as estimativas de amostragem aleatória simples.
}
\description{
Esta função aplica os estimadores da amostragem aleatória simples.
}
\details{
Um detalhe importante.
}
