% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_balance.R
\name{train_test_balance}
\alias{train_test_balance}
\title{Balance and split the dataset}
\usage{
train_test_balance(
  data,
  y,
  type,
  balance = TRUE,
  fractions = c(0.6, 0.2, 0.2),
  seed = NULL
)
}
\arguments{
\item{data}{A data source, that is one of the major R formats: data.table, data.frame,
matrix and so on.}

\item{y}{A string that indicates a target column name.}

\item{type}{A string that determines if Machine Learning task is the
`classification` or `regression`.}

\item{balance}{A logical value, determines if we want to balance the dataset.}

\item{fractions}{A vector with 3 numeric values that sum to 1 which
determine sizes of train, test and validation datasets. DEFAULT: c(0.6, 0.2, 0.2).}

\item{seed}{An integer random seed. It allows for comparable results. If it
is NULL, the split is random.}
}
\value{
A list of train, test and validation datasets.
}
\description{
Balance and split the dataset
}
\examples{
data(lisbon)
b_lisbon <- train_test_balance(lisbon, 'Price', balance = FALSE, 'regression',
                               fractions = c(train = 0.6, valid = 0.2, test = 0.2))
}
